% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_apply_archive_rule}
\alias{accessanalyzer_apply_archive_rule}
\title{Retroactively applies the archive rule to existing findings that meet
the archive rule criteria}
\usage{
accessanalyzer_apply_archive_rule(analyzerArn, ruleName, clientToken = NULL)
}
\arguments{
\item{analyzerArn}{[required] The Amazon resource name (ARN) of the analyzer.}

\item{ruleName}{[required] The name of the rule to apply.}

\item{clientToken}{A client token.}
}
\description{
Retroactively applies the archive rule to existing findings that meet the archive rule criteria.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_apply_archive_rule/} for full documentation.
}
\keyword{internal}
