% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_cancel_policy_generation}
\alias{accessanalyzer_cancel_policy_generation}
\title{Cancels the requested policy generation}
\usage{
accessanalyzer_cancel_policy_generation(jobId)
}
\arguments{
\item{jobId}{[required] The \code{JobId} that is returned by the
\code{\link[=accessanalyzer_start_policy_generation]{start_policy_generation}}
operation. The \code{JobId} can be used with
\code{\link[=accessanalyzer_get_generated_policy]{get_generated_policy}} to
retrieve the generated policies or used with
\code{\link[=accessanalyzer_cancel_policy_generation]{cancel_policy_generation}} to
cancel the policy generation request.}
}
\description{
Cancels the requested policy generation.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_cancel_policy_generation/} for full documentation.
}
\keyword{internal}
