% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_create_archive_rule}
\alias{accessanalyzer_create_archive_rule}
\title{Creates an archive rule for the specified analyzer}
\usage{
accessanalyzer_create_archive_rule(
  analyzerName,
  ruleName,
  filter,
  clientToken = NULL
)
}
\arguments{
\item{analyzerName}{[required] The name of the created analyzer.}

\item{ruleName}{[required] The name of the rule to create.}

\item{filter}{[required] The criteria for the rule.}

\item{clientToken}{A client token.}
}
\description{
Creates an archive rule for the specified analyzer. Archive rules automatically archive new findings that meet the criteria you define when you create the rule.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_create_archive_rule/} for full documentation.
}
\keyword{internal}
