% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_create_service_linked_analyzer}
\alias{accessanalyzer_create_service_linked_analyzer}
\title{Creates a service-linked analyzer managed by an Amazon Web Services
service}
\usage{
accessanalyzer_create_service_linked_analyzer(
  type,
  archiveRules = NULL,
  clientToken = NULL,
  configuration = NULL
)
}
\arguments{
\item{type}{[required] The type of analyzer to create. Valid values are \code{ACCOUNT_UNUSED_ACCESS} and \code{ORGANIZATION_UNUSED_ACCESS}.}

\item{archiveRules}{Specifies the archive rules to add for the analyzer. Archive rules automatically archive findings that meet the criteria you define for the rule.}

\item{clientToken}{A client token.}

\item{configuration}{Specifies the configuration of the analyzer. The specified scope of unused access is used for the configuration.}
}
\description{
Creates a service-linked analyzer managed by an Amazon Web Services service. This operation can only be invoked by authorized Amazon Web Services services. Direct customer invocation returns \code{AccessDeniedException}.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_create_service_linked_analyzer/} for full documentation.
}
\keyword{internal}
