% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_delete_service_linked_analyzer}
\alias{accessanalyzer_delete_service_linked_analyzer}
\title{Deletes a service-linked analyzer}
\usage{
accessanalyzer_delete_service_linked_analyzer(analyzerName, clientToken = NULL)
}
\arguments{
\item{analyzerName}{[required] The name of the service-linked analyzer to delete. Service-linked analyzer names follow the format \verb{_AccessAnalyzerFor\{ServiceName\}-\{Id\}}.}

\item{clientToken}{A client token.}
}
\description{
Deletes a service-linked analyzer. This operation can be invoked by both authorized Amazon Web Services services and customers.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_delete_service_linked_analyzer/} for full documentation.
}
\keyword{internal}
