% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_generate_finding_recommendation}
\alias{accessanalyzer_generate_finding_recommendation}
\title{Creates a recommendation for an unused permissions finding}
\usage{
accessanalyzer_generate_finding_recommendation(analyzerArn, id)
}
\arguments{
\item{analyzerArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access-analyzer-getting-started.html#permission-resources}{ARN of the analyzer}
used to generate the finding recommendation.}

\item{id}{[required] The unique ID for the finding recommendation.}
}
\description{
Creates a recommendation for an unused permissions finding.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_generate_finding_recommendation/} for full documentation.
}
\keyword{internal}
