% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_get_analyzed_resource}
\alias{accessanalyzer_get_analyzed_resource}
\title{Retrieves information about a resource that was analyzed}
\usage{
accessanalyzer_get_analyzed_resource(analyzerArn, resourceArn)
}
\arguments{
\item{analyzerArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access-analyzer-getting-started.html#permission-resources}{ARN of the analyzer}
to retrieve information from.}

\item{resourceArn}{[required] The ARN of the resource to retrieve information about.}
}
\description{
Retrieves information about a resource that was analyzed.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_get_analyzed_resource/} for full documentation.
}
\keyword{internal}
