% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_get_archive_rule}
\alias{accessanalyzer_get_archive_rule}
\title{Retrieves information about an archive rule}
\usage{
accessanalyzer_get_archive_rule(analyzerName, ruleName)
}
\arguments{
\item{analyzerName}{[required] The name of the analyzer to retrieve rules from.}

\item{ruleName}{[required] The name of the rule to retrieve.}
}
\description{
Retrieves information about an archive rule.

See \url{https://paws-r.github.io/docs/accessanalyzer/get_archive_rule.html} for full documentation.
}
\keyword{internal}
