% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_get_finding_recommendation}
\alias{accessanalyzer_get_finding_recommendation}
\title{Retrieves information about a finding recommendation for the specified
analyzer}
\usage{
accessanalyzer_get_finding_recommendation(
  analyzerArn,
  id,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{analyzerArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access-analyzer-getting-started.html#permission-resources}{ARN of the analyzer}
used to generate the finding recommendation.}

\item{id}{[required] The unique ID for the finding recommendation.}

\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{A token used for pagination of results returned.}
}
\description{
Retrieves information about a finding recommendation for the specified analyzer.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_get_finding_recommendation/} for full documentation.
}
\keyword{internal}
