% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_get_findings_statistics}
\alias{accessanalyzer_get_findings_statistics}
\title{Retrieves a list of aggregated finding statistics for an external access
or unused access analyzer}
\usage{
accessanalyzer_get_findings_statistics(analyzerArn)
}
\arguments{
\item{analyzerArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access-analyzer-getting-started.html#permission-resources}{ARN of the analyzer}
used to generate the statistics.}
}
\description{
Retrieves a list of aggregated finding statistics for an external access or unused access analyzer.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_get_findings_statistics/} for full documentation.
}
\keyword{internal}
