% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_list_analyzed_resources}
\alias{accessanalyzer_list_analyzed_resources}
\title{Retrieves a list of resources of the specified type that have been
analyzed by the specified external access analyzer}
\usage{
accessanalyzer_list_analyzed_resources(
  analyzerArn,
  resourceType = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{analyzerArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access-analyzer-getting-started.html#permission-resources}{ARN of the analyzer}
to retrieve a list of analyzed resources from.}

\item{resourceType}{The type of resource.}

\item{nextToken}{A token used for pagination of results returned.}

\item{maxResults}{The maximum number of results to return in the response.}
}
\description{
Retrieves a list of resources of the specified type that have been analyzed by the specified external access analyzer. This action is not supported for unused access analyzers.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_list_analyzed_resources/} for full documentation.
}
\keyword{internal}
