% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_list_archive_rules}
\alias{accessanalyzer_list_archive_rules}
\title{Retrieves a list of archive rules created for the specified analyzer}
\usage{
accessanalyzer_list_archive_rules(
  analyzerName,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{analyzerName}{[required] The name of the analyzer to retrieve rules from.}

\item{nextToken}{A token used for pagination of results returned.}

\item{maxResults}{The maximum number of results to return in the request.}
}
\description{
Retrieves a list of archive rules created for the specified analyzer.

See \url{https://paws-r.github.io/docs/accessanalyzer/list_archive_rules.html} for full documentation.
}
\keyword{internal}
