% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_list_findings}
\alias{accessanalyzer_list_findings}
\title{Retrieves a list of findings generated by the specified analyzer}
\usage{
accessanalyzer_list_findings(
  analyzerArn,
  filter = NULL,
  sort = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{analyzerArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access-analyzer-getting-started.html#permission-resources}{ARN of the analyzer}
to retrieve findings from.}

\item{filter}{A filter to match for the findings to return.}

\item{sort}{The sort order for the findings returned.}

\item{nextToken}{A token used for pagination of results returned.}

\item{maxResults}{The maximum number of results to return in the response.}
}
\description{
Retrieves a list of findings generated by the specified analyzer. ListFindings and ListFindingsV2 both use \code{access-analyzer:ListFindings} in the \code{Action} element of an IAM policy statement. You must have permission to perform the \code{access-analyzer:ListFindings} action.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_list_findings/} for full documentation.
}
\keyword{internal}
