% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_start_policy_generation}
\alias{accessanalyzer_start_policy_generation}
\title{Starts the policy generation request}
\usage{
accessanalyzer_start_policy_generation(
  policyGenerationDetails,
  cloudTrailDetails = NULL,
  clientToken = NULL
)
}
\arguments{
\item{policyGenerationDetails}{[required] Contains the ARN of the IAM entity (user or role) for which you are
generating a policy.}

\item{cloudTrailDetails}{A \code{CloudTrailDetails} object that contains details about a \code{Trail} that
you want to analyze to generate policies.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request. Idempotency ensures that an API request
completes only once. With an idempotent request, if the original request
completes successfully, the subsequent retries with the same client
token return the result from the original successful request and they
have no additional effect.

If you do not specify a client token, one is automatically generated by
the Amazon Web Services SDK.}
}
\description{
Starts the policy generation request.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_start_policy_generation/} for full documentation.
}
\keyword{internal}
