% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_tag_resource}
\alias{accessanalyzer_tag_resource}
\title{Adds a tag to the specified resource}
\usage{
accessanalyzer_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to add the tag to.}

\item{tags}{[required] The tags to add to the resource.}
}
\description{
Adds a tag to the specified resource.

See \url{https://paws-r.github.io/docs/accessanalyzer/tag_resource.html} for full documentation.
}
\keyword{internal}
