% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_untag_resource}
\alias{accessanalyzer_untag_resource}
\title{Removes a tag from the specified resource}
\usage{
accessanalyzer_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource to remove the tag from.}

\item{tagKeys}{[required] The key for the tag to add.}
}
\description{
Removes a tag from the specified resource.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_untag_resource/} for full documentation.
}
\keyword{internal}
