% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_update_analyzer}
\alias{accessanalyzer_update_analyzer}
\title{Modifies the configuration of an existing analyzer}
\usage{
accessanalyzer_update_analyzer(analyzerName, configuration = NULL)
}
\arguments{
\item{analyzerName}{[required] The name of the analyzer to modify.}

\item{configuration}{}
}
\description{
Modifies the configuration of an existing analyzer.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_update_analyzer/} for full documentation.
}
\keyword{internal}
