% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_operations.R
\name{account_get_primary_email_update_status}
\alias{account_get_primary_email_update_status}
\title{Retrieves the status of the most recent primary email update for the
specified account}
\usage{
account_get_primary_email_update_status(AccountId = NULL)
}
\arguments{
\item{AccountId}{Specifies the 12-digit account ID number of the Amazon Web Services account that you want to access or modify with this operation. To use this parameter, the caller must be an identity in the \href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_getting-started_concepts.html#account}{organization's management account} or a delegated administrator account. The specified account ID must be a member account in the same organization. The organization must have \href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_manage_org_support-all-features.html}{all features enabled}, and the organization must have \href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_integrate_services.html}{trusted access} enabled for the Account Management service, and optionally a \href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_getting-started_concepts.html#delegated-admin}{delegated admin} account assigned.

This operation can only be called from the management account or the delegated administrator account of an organization for a member account.

The management account can't specify its own \code{AccountId}.}
}
\description{
Retrieves the status of the most recent primary email update for the specified account. For complete details about how to update the primary email address, see \href{https://docs.aws.amazon.com/accounts/latest/reference/manage-acct-update-root-user-email.html}{Update the primary email address for your AWS account}.

See \url{https://www.paws-r-sdk.com/docs/account_get_primary_email_update_status/} for full documentation.
}
\keyword{internal}
