% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_operations.R
\name{account_verify_phone_number}
\alias{account_verify_phone_number}
\title{Verifies the phone number in the primary contact information of an
Amazon Web Services account by submitting the one-time passcode that
SendPhoneNumberVerification sent to that phone number}
\usage{
account_verify_phone_number(AccountId = NULL, Otp)
}
\arguments{
\item{AccountId}{Specifies the 12 digit account ID number of the Amazon Web Services account that you want to access or modify with this operation.

If you do not specify this parameter, it defaults to the Amazon Web Services account of the identity used to call the operation.

To use this parameter, the caller must be an identity in the \href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_getting-started_concepts.html#account}{organization's management account} or a delegated administrator account, and the specified account ID must be a member account in the same organization. The organization must have \href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_manage_org_support-all-features.html}{all features enabled}, and the organization must have \href{https://docs.aws.amazon.com/organizations/latest/userguide/services-that-can-integrate-account.html}{trusted access} enabled for the Account Management service, and optionally a \href{https://docs.aws.amazon.com/organizations/latest/userguide/orgs_getting-started_concepts.html#delegated-admin}{delegated administrator} account assigned.

The management account can't specify its own \code{AccountId}; it must call the operation in standalone context by not including the \code{AccountId} parameter.

To call this operation on an account that is not a member of an organization, then don't specify this parameter, and call the operation using an identity belonging to the account whose contacts you wish to retrieve or modify.}

\item{Otp}{[required] The one-time passcode sent to the phone number in the primary contact information by the \code{\link[=account_send_phone_number_verification]{send_phone_number_verification}} operation.}
}
\description{
Verifies the phone number in the primary contact information of an Amazon Web Services account by submitting the one-time passcode that \code{\link[=account_send_phone_number_verification]{send_phone_number_verification}} sent to that phone number.

See \url{https://www.paws-r-sdk.com/docs/account_verify_phone_number/} for full documentation.
}
\keyword{internal}
