% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_create_acme_domain_validation}
\alias{acm_create_acme_domain_validation}
\title{Creates a domain validation for an ACME endpoint}
\usage{
acm_create_acme_domain_validation(
  IdempotencyToken = NULL,
  AcmeEndpointArn,
  DomainName,
  PrevalidationOptions,
  Tags = NULL
)
}
\arguments{
\item{IdempotencyToken}{A unique, case-sensitive identifier to ensure idempotency of the request.}

\item{AcmeEndpointArn}{[required] The Amazon Resource Name (ARN) of the ACME endpoint.}

\item{DomainName}{[required] The domain name to validate.}

\item{PrevalidationOptions}{[required] The prevalidation options for the domain.}

\item{Tags}{One or more tags to associate with the domain validation.}
}
\description{
Creates a domain validation for an ACME endpoint. Domain validations authorize the endpoint to issue certificates for specified domain names. You configure prevalidation to prove domain ownership.

See \url{https://www.paws-r-sdk.com/docs/acm_create_acme_domain_validation/} for full documentation.
}
\keyword{internal}
