% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_create_acme_endpoint}
\alias{acm_create_acme_endpoint}
\title{Creates an ACME endpoint, which is a managed ACME server with a unique
endpoint URL}
\usage{
acm_create_acme_endpoint(
  IdempotencyToken = NULL,
  AuthorizationBehavior,
  Contact = NULL,
  CertificateAuthority,
  Tags = NULL,
  CertificateTags = NULL
)
}
\arguments{
\item{IdempotencyToken}{A unique, case-sensitive identifier to ensure idempotency of the request.}

\item{AuthorizationBehavior}{[required] The authorization behavior for the ACME endpoint.}

\item{Contact}{Specifies whether ACME clients must provide contact information during account registration.}

\item{CertificateAuthority}{[required] The type of certificate authority to use for issuing certificates through this ACME endpoint.}

\item{Tags}{One or more tags to associate with the ACME endpoint.}

\item{CertificateTags}{Tags to apply to certificates issued through this ACME endpoint.}
}
\description{
Creates an ACME endpoint, which is a managed ACME server with a unique endpoint URL. After creation, ACME clients can use the endpoint URL to automate certificate issuance using the ACME protocol.

See \url{https://www.paws-r-sdk.com/docs/acm_create_acme_endpoint/} for full documentation.
}
\keyword{internal}
