% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_create_acme_external_account_binding}
\alias{acm_create_acme_external_account_binding}
\title{Creates an external account binding (EAB) for an ACME endpoint}
\usage{
acm_create_acme_external_account_binding(
  IdempotencyToken = NULL,
  AcmeEndpointArn,
  RoleArn,
  Expiration = NULL,
  Tags = NULL
)
}
\arguments{
\item{IdempotencyToken}{A unique, case-sensitive identifier to ensure idempotency of the request.}

\item{AcmeEndpointArn}{[required] The Amazon Resource Name (ARN) of the ACME endpoint.}

\item{RoleArn}{[required] The Amazon Resource Name (ARN) of the IAM role to associate with the external account binding.}

\item{Expiration}{The expiration configuration for the external account binding.}

\item{Tags}{One or more tags to associate with the external account binding.}
}
\description{
Creates an external account binding (EAB) for an ACME endpoint. An EAB provides credentials that authorize an ACME client to register an account with the endpoint. Each EAB is associated with an IAM role that controls what certificate operations the ACME client can perform.

See \url{https://www.paws-r-sdk.com/docs/acm_create_acme_external_account_binding/} for full documentation.
}
\keyword{internal}
