% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_describe_acme_account}
\alias{acm_describe_acme_account}
\title{Returns detailed metadata about the specified ACME account, including
its status, public key thumbprint, and associated external account
binding}
\usage{
acm_describe_acme_account(AcmeEndpointArn, AccountUrl)
}
\arguments{
\item{AcmeEndpointArn}{[required] The Amazon Resource Name (ARN) of the ACME endpoint.}

\item{AccountUrl}{[required] The URL of the ACME account.}
}
\description{
Returns detailed metadata about the specified ACME account, including its status, public key thumbprint, and associated external account binding.

See \url{https://www.paws-r-sdk.com/docs/acm_describe_acme_account/} for full documentation.
}
\keyword{internal}
