% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_describe_acme_domain_validation}
\alias{acm_describe_acme_domain_validation}
\title{Returns detailed metadata about the specified domain validation,
including its status, domain scope, and DNS resource records required
for validation}
\usage{
acm_describe_acme_domain_validation(AcmeDomainValidationArn)
}
\arguments{
\item{AcmeDomainValidationArn}{[required] The Amazon Resource Name (ARN) of the ACME domain validation.}
}
\description{
Returns detailed metadata about the specified domain validation, including its status, domain scope, and DNS resource records required for validation.

See \url{https://www.paws-r-sdk.com/docs/acm_describe_acme_domain_validation/} for full documentation.
}
\keyword{internal}
