% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_describe_acme_endpoint}
\alias{acm_describe_acme_endpoint}
\title{Returns detailed metadata about the specified ACME endpoint, including
its status, URL, authorization behavior, and certificate authority
configuration}
\usage{
acm_describe_acme_endpoint(AcmeEndpointArn)
}
\arguments{
\item{AcmeEndpointArn}{[required] The Amazon Resource Name (ARN) of the ACME endpoint.}
}
\description{
Returns detailed metadata about the specified ACME endpoint, including its status, URL, authorization behavior, and certificate authority configuration.

See \url{https://www.paws-r-sdk.com/docs/acm_describe_acme_endpoint/} for full documentation.
}
\keyword{internal}
