% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_describe_acme_external_account_binding}
\alias{acm_describe_acme_external_account_binding}
\title{Returns detailed metadata about the specified external account binding,
including the associated IAM role, expiration time, and usage history}
\usage{
acm_describe_acme_external_account_binding(AcmeExternalAccountBindingArn)
}
\arguments{
\item{AcmeExternalAccountBindingArn}{[required] The Amazon Resource Name (ARN) of the ACME external account binding.}
}
\description{
Returns detailed metadata about the specified external account binding, including the associated IAM role, expiration time, and usage history.

See \url{https://www.paws-r-sdk.com/docs/acm_describe_acme_external_account_binding/} for full documentation.
}
\keyword{internal}
