% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_get_account_configuration}
\alias{acm_get_account_configuration}
\title{Returns the account configuration options associated with an AWS account}
\usage{
acm_get_account_configuration()
}
\value{
A list with the following syntax:\preformatted{list(
  ExpiryEvents = list(
    DaysBeforeExpiry = 123
  )
)
}
}
\description{
Returns the account configuration options associated with an AWS
account.
}
\keyword{internal}
