% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_get_acme_external_account_binding_credentials}
\alias{acm_get_acme_external_account_binding_credentials}
\title{Retrieves the key ID and MAC key credentials for an external account
binding}
\usage{
acm_get_acme_external_account_binding_credentials(
  AcmeExternalAccountBindingArn
)
}
\arguments{
\item{AcmeExternalAccountBindingArn}{[required] The Amazon Resource Name (ARN) of the ACME external account binding.}
}
\description{
Retrieves the key ID and MAC key credentials for an external account binding. These credentials are used by ACME clients during account registration to bind to the endpoint.

See \url{https://www.paws-r-sdk.com/docs/acm_get_acme_external_account_binding_credentials/} for full documentation.
}
\keyword{internal}
