% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_list_acme_external_account_bindings}
\alias{acm_list_acme_external_account_bindings}
\title{Retrieves a list of external account bindings for the specified ACME
endpoint}
\usage{
acm_list_acme_external_account_bindings(
  NextToken = NULL,
  MaxResults = NULL,
  AcmeEndpointArn
)
}
\arguments{
\item{NextToken}{A token for pagination.}

\item{MaxResults}{The maximum number of results to return.}

\item{AcmeEndpointArn}{[required] The Amazon Resource Name (ARN) of the ACME endpoint.}
}
\description{
Retrieves a list of external account bindings for the specified ACME endpoint.

See \url{https://www.paws-r-sdk.com/docs/acm_list_acme_external_account_bindings/} for full documentation.
}
\keyword{internal}
