% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_list_tags_for_resource}
\alias{acm_list_tags_for_resource}
\title{Lists the tags associated with an ACM resource}
\usage{
acm_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the ACM resource for which to list tags.}
}
\description{
Lists the tags associated with an ACM resource.

See \url{https://www.paws-r-sdk.com/docs/acm_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
