% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_revoke_acme_account}
\alias{acm_revoke_acme_account}
\title{Revokes an ACME account, preventing it from requesting or revoking
certificates}
\usage{
acm_revoke_acme_account(AcmeEndpointArn, AccountUrl)
}
\arguments{
\item{AcmeEndpointArn}{[required] The Amazon Resource Name (ARN) of the ACME endpoint.}

\item{AccountUrl}{[required] The URL of the ACME account to revoke.}
}
\description{
Revokes an ACME account, preventing it from requesting or revoking certificates. This operation is irreversible.

See \url{https://www.paws-r-sdk.com/docs/acm_revoke_acme_account/} for full documentation.
}
\keyword{internal}
