% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_revoke_acme_external_account_binding}
\alias{acm_revoke_acme_external_account_binding}
\title{Revokes an external account binding, preventing new ACME accounts from
being registered using this binding}
\usage{
acm_revoke_acme_external_account_binding(AcmeExternalAccountBindingArn)
}
\arguments{
\item{AcmeExternalAccountBindingArn}{[required] The Amazon Resource Name (ARN) of the ACME external account binding to revoke.}
}
\description{
Revokes an external account binding, preventing new ACME accounts from being registered using this binding. Existing ACME accounts that were previously registered using the binding are not affected and must be revoked separately.

See \url{https://www.paws-r-sdk.com/docs/acm_revoke_acme_external_account_binding/} for full documentation.
}
\keyword{internal}
