% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_revoke_certificate}
\alias{acm_revoke_certificate}
\title{Revokes a public ACM certificate}
\usage{
acm_revoke_certificate(CertificateArn, RevocationReason)
}
\arguments{
\item{CertificateArn}{[required] The Amazon Resource Name (ARN) of the public or private certificate that
will be revoked. The ARN must have the following form:

\code{arn:aws:acm:region:account:certificate/12345678-1234-1234-1234-123456789012}}

\item{RevocationReason}{[required] Specifies why you revoked the certificate.}
}
\description{
Revokes a public ACM certificate. You can only revoke certificates that have been previously exported.

See \url{https://www.paws-r-sdk.com/docs/acm_revoke_certificate/} for full documentation.
}
\keyword{internal}
