% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_tag_resource}
\alias{acm_tag_resource}
\title{Adds one or more tags to an ACM resource}
\usage{
acm_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the ACM resource to which the tag is to be applied.}

\item{Tags}{[required] The key-value pair that defines the tag to apply.}
}
\description{
Adds one or more tags to an ACM resource. Tags are labels that you can use to identify and organize your Amazon Web Services resources. Each tag consists of a \code{key} and an optional \code{value}.

See \url{https://www.paws-r-sdk.com/docs/acm_tag_resource/} for full documentation.
}
\keyword{internal}
