% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_untag_resource}
\alias{acm_untag_resource}
\title{Removes one or more tags from an ACM resource}
\usage{
acm_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the ACM resource from which the tag is to be removed.}

\item{TagKeys}{[required] The key of each tag to remove.}
}
\description{
Removes one or more tags from an ACM resource.

See \url{https://www.paws-r-sdk.com/docs/acm_untag_resource/} for full documentation.
}
\keyword{internal}
