% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_update_acme_domain_validation}
\alias{acm_update_acme_domain_validation}
\title{Updates the prevalidation configuration of an existing domain validation}
\usage{
acm_update_acme_domain_validation(
  AcmeDomainValidationArn,
  PrevalidationOptions = NULL
)
}
\arguments{
\item{AcmeDomainValidationArn}{[required] The Amazon Resource Name (ARN) of the ACME domain validation to update.}

\item{PrevalidationOptions}{The updated prevalidation options.}
}
\description{
Updates the prevalidation configuration of an existing domain validation.

See \url{https://www.paws-r-sdk.com/docs/acm_update_acme_domain_validation/} for full documentation.
}
\keyword{internal}
