% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_update_acme_endpoint}
\alias{acm_update_acme_endpoint}
\title{Updates the configuration of an existing ACME endpoint}
\usage{
acm_update_acme_endpoint(
  AcmeEndpointArn,
  AuthorizationBehavior = NULL,
  Contact = NULL,
  CertificateAuthority = NULL
)
}
\arguments{
\item{AcmeEndpointArn}{[required] The Amazon Resource Name (ARN) of the ACME endpoint to update.}

\item{AuthorizationBehavior}{The updated authorization behavior.}

\item{Contact}{The updated contact requirement.}

\item{CertificateAuthority}{The updated certificate authority configuration.}
}
\description{
Updates the configuration of an existing ACME endpoint. You can change the authorization behavior, contact requirement, or certificate authority settings.

See \url{https://www.paws-r-sdk.com/docs/acm_update_acme_endpoint/} for full documentation.
}
\keyword{internal}
