% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_update_certificate_options}
\alias{acm_update_certificate_options}
\title{Updates certificate options}
\usage{
acm_update_certificate_options(CertificateArn, Options)
}
\arguments{
\item{CertificateArn}{[required] ARN of the requested certificate to update. This must be of the form:

\code{arn:aws:acm:us-east-1:account:certificate/12345678-1234-1234-1234-123456789012 }}

\item{Options}{[required] Use to update the options for your certificate. Currently, you can change the domain validation method or specify whether to export your certificate. For more information about migrating from email to DNS validation, see \href{https://docs.aws.amazon.com/acm/latest/userguide/}{Migrate from email to DNS validation}.}
}
\description{
Updates certificate options. You can use this operation to change the domain validation method or specify whether to export your certificate. For more information, see \href{https://docs.aws.amazon.com/acm/latest/userguide/}{Migrate from email to DNS validation} and \href{https://docs.aws.amazon.com/acm/latest/userguide/acm-exportable-certificates.html}{Certificate Manager Exportable Managed Certificates}.

See \url{https://www.paws-r-sdk.com/docs/acm_update_certificate_options/} for full documentation.
}
\keyword{internal}
