% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acmpca_operations.R
\name{acmpca_delete_policy}
\alias{acmpca_delete_policy}
\title{Deletes the resource-based policy attached to a private CA}
\usage{
acmpca_delete_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Number (ARN) of the private CA that will have its
policy deleted. You can find the CA's ARN by calling the
\code{\link[=acmpca_list_certificate_authorities]{list_certificate_authorities}}
action. The ARN value must have the form
\verb{arn:aws:acm-pca:region:account:certificate-authority/01234567-89ab-cdef-0123-0123456789ab}.}
}
\value{
An empty list.
}
\description{
Deletes the resource-based policy attached to a private CA. Deletion
will remove any access that the policy has granted. If there is no
policy attached to the private CA, this action will return successful.

If you delete a policy that was applied through AWS Resource Access
Manager (RAM), the CA will be removed from all shares in which it was
included.

The AWS Certificate Manager Service Linked Role that the policy supports
is not affected when you delete the policy.

The current policy can be shown with \code{\link[=acmpca_get_policy]{get_policy}}
and updated with \code{\link[=acmpca_put_policy]{put_policy}}.

\strong{About Policies}
\itemize{
\item A policy grants access on a private CA to an AWS customer account,
to AWS Organizations, or to an AWS Organizations unit. Policies are
under the control of a CA administrator. For more information, see
\href{https://docs.aws.amazon.com/acm-pca/latest/userguide/pca-rbp.html}{Using a Resource Based Policy with ACM Private CA}.
\item A policy permits a user of AWS Certificate Manager (ACM) to issue
ACM certificates signed by a CA in another account.
\item For ACM to manage automatic renewal of these certificates, the ACM
user must configure a Service Linked Role (SLR). The SLR allows the
ACM service to assume the identity of the user, subject to
confirmation against the ACM Private CA policy. For more
information, see \href{https://docs.aws.amazon.com/acm/latest/userguide/acm-slr.html}{Using a Service Linked Role with ACM}.
\item Updates made in AWS Resource Manager (RAM) are reflected in
policies. For more information, see \href{https://docs.aws.amazon.com/acm-pca/latest/userguide/pca-ram.html}{Attach a Policy for Cross-Account Access}.
}
}
\section{Request syntax}{
\preformatted{svc$delete_policy(
  ResourceArn = "string"
)
}
}

\keyword{internal}
