% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acmpca_operations.R
\name{acmpca_describe_certificate_authority_audit_report}
\alias{acmpca_describe_certificate_authority_audit_report}
\title{Lists information about a specific audit report created by calling the
CreateCertificateAuthorityAuditReport action}
\usage{
acmpca_describe_certificate_authority_audit_report(
  CertificateAuthorityArn, AuditReportId)
}
\arguments{
\item{CertificateAuthorityArn}{[required] The Amazon Resource Name (ARN) of the private CA. This must be of the
form:

\code{arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 }.}

\item{AuditReportId}{[required] The report ID returned by calling the
\code{\link[=acmpca_create_certificate_authority_audit_report]{create_certificate_authority_audit_report}}
action.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AuditReportStatus = "CREATING"|"SUCCESS"|"FAILED",
  S3BucketName = "string",
  S3Key = "string",
  CreatedAt = as.POSIXct(
    "2015-01-01"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists information about a specific audit report created by calling the
\code{\link[=acmpca_create_certificate_authority_audit_report]{create_certificate_authority_audit_report}}
action. Audit information is created every time the certificate
authority (CA) private key is used. The private key is used when you
call the \code{\link[=acmpca_issue_certificate]{issue_certificate}} action or the
\code{\link[=acmpca_revoke_certificate]{revoke_certificate}} action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_certificate_authority_audit_report(
  CertificateAuthorityArn = "string",
  AuditReportId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
