% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_cancel_trained_model}
\alias{cleanroomsml_cancel_trained_model}
\title{Submits a request to cancel the trained model job}
\usage{
cleanroomsml_cancel_trained_model(membershipIdentifier, trainedModelArn)
}
\arguments{
\item{membershipIdentifier}{[required] The membership ID of the trained model job that you want to cancel.}

\item{trainedModelArn}{[required] The Amazon Resource Name (ARN) of the trained model job that you want to
cancel.}
}
\description{
Submits a request to cancel the trained model job.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_cancel_trained_model/} for full documentation.
}
\keyword{internal}
