% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_cancel_trained_model_inference_job}
\alias{cleanroomsml_cancel_trained_model_inference_job}
\title{Submits a request to cancel a trained model inference job}
\usage{
cleanroomsml_cancel_trained_model_inference_job(
  membershipIdentifier,
  trainedModelInferenceJobArn
)
}
\arguments{
\item{membershipIdentifier}{[required] The membership ID of the trained model inference job that you want to
cancel.}

\item{trainedModelInferenceJobArn}{[required] The Amazon Resource Name (ARN) of the trained model inference job that
you want to cancel.}
}
\description{
Submits a request to cancel a trained model inference job.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_cancel_trained_model_inference_job/} for full documentation.
}
\keyword{internal}
