% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_delete_audience_generation_job}
\alias{cleanroomsml_delete_audience_generation_job}
\title{Deletes the specified audience generation job, and removes all data
associated with the job}
\usage{
cleanroomsml_delete_audience_generation_job(audienceGenerationJobArn)
}
\arguments{
\item{audienceGenerationJobArn}{[required] The Amazon Resource Name (ARN) of the audience generation job that you
want to delete.}
}
\description{
Deletes the specified audience generation job, and removes all data associated with the job.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_delete_audience_generation_job/} for full documentation.
}
\keyword{internal}
