% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_delete_audience_model}
\alias{cleanroomsml_delete_audience_model}
\title{Specifies an audience model that you want to delete}
\usage{
cleanroomsml_delete_audience_model(audienceModelArn)
}
\arguments{
\item{audienceModelArn}{[required] The Amazon Resource Name (ARN) of the audience model that you want to
delete.}
}
\description{
Specifies an audience model that you want to delete. You can't delete an audience model if there are any configured audience models that depend on the audience model.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_delete_audience_model/} for full documentation.
}
\keyword{internal}
