% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_delete_configured_audience_model}
\alias{cleanroomsml_delete_configured_audience_model}
\title{Deletes the specified configured audience model}
\usage{
cleanroomsml_delete_configured_audience_model(configuredAudienceModelArn)
}
\arguments{
\item{configuredAudienceModelArn}{[required] The Amazon Resource Name (ARN) of the configured audience model that you
want to delete.}
}
\description{
Deletes the specified configured audience model. You can't delete a configured audience model if there are any lookalike models that use the configured audience model. If you delete a configured audience model, it will be removed from any collaborations that it is associated to.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_delete_configured_audience_model/} for full documentation.
}
\keyword{internal}
