% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_delete_configured_audience_model_policy}
\alias{cleanroomsml_delete_configured_audience_model_policy}
\title{Deletes the specified configured audience model policy}
\usage{
cleanroomsml_delete_configured_audience_model_policy(
  configuredAudienceModelArn
)
}
\arguments{
\item{configuredAudienceModelArn}{[required] The Amazon Resource Name (ARN) of the configured audience model policy
that you want to delete.}
}
\description{
Deletes the specified configured audience model policy.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_delete_configured_audience_model_policy/} for full documentation.
}
\keyword{internal}
