% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_delete_configured_model_algorithm}
\alias{cleanroomsml_delete_configured_model_algorithm}
\title{Deletes a configured model algorithm}
\usage{
cleanroomsml_delete_configured_model_algorithm(configuredModelAlgorithmArn)
}
\arguments{
\item{configuredModelAlgorithmArn}{[required] The Amazon Resource Name (ARN) of the configured model algorithm that
you want to delete.}
}
\description{
Deletes a configured model algorithm.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_delete_configured_model_algorithm/} for full documentation.
}
\keyword{internal}
