% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_delete_ml_configuration}
\alias{cleanroomsml_delete_ml_configuration}
\title{Deletes a ML modeling configuration}
\usage{
cleanroomsml_delete_ml_configuration(membershipIdentifier)
}
\arguments{
\item{membershipIdentifier}{[required] The membership ID of the of the member that is deleting the ML modeling
configuration.}
}
\description{
Deletes a ML modeling configuration.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_delete_ml_configuration/} for full documentation.
}
\keyword{internal}
