% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_delete_ml_input_channel_data}
\alias{cleanroomsml_delete_ml_input_channel_data}
\title{Provides the information necessary to delete an ML input channel}
\usage{
cleanroomsml_delete_ml_input_channel_data(
  mlInputChannelArn,
  membershipIdentifier
)
}
\arguments{
\item{mlInputChannelArn}{[required] The Amazon Resource Name (ARN) of the ML input channel that you want to
delete.}

\item{membershipIdentifier}{[required] The membership ID of the membership that contains the ML input channel
you want to delete.}
}
\description{
Provides the information necessary to delete an ML input channel.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_delete_ml_input_channel_data/} for full documentation.
}
\keyword{internal}
