% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_delete_trained_model_output}
\alias{cleanroomsml_delete_trained_model_output}
\title{Deletes the output of a trained model}
\usage{
cleanroomsml_delete_trained_model_output(trainedModelArn, membershipIdentifier)
}
\arguments{
\item{trainedModelArn}{[required] The Amazon Resource Name (ARN) of the trained model whose output you
want to delete.}

\item{membershipIdentifier}{[required] The membership ID of the member that is deleting the trained model
output.}
}
\description{
Deletes the output of a trained model.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_delete_trained_model_output/} for full documentation.
}
\keyword{internal}
