% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_delete_training_dataset}
\alias{cleanroomsml_delete_training_dataset}
\title{Specifies a training dataset that you want to delete}
\usage{
cleanroomsml_delete_training_dataset(trainingDatasetArn)
}
\arguments{
\item{trainingDatasetArn}{[required] The Amazon Resource Name (ARN) of the training dataset that you want to
delete.}
}
\description{
Specifies a training dataset that you want to delete. You can't delete a training dataset if there are any audience models that depend on the training dataset. In Clean Rooms ML, the \code{TrainingDataset} is metadata that points to a Glue table, which is read only during \code{AudienceModel} creation. This action deletes the metadata.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_delete_training_dataset/} for full documentation.
}
\keyword{internal}
