% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_get_audience_generation_job}
\alias{cleanroomsml_get_audience_generation_job}
\title{Returns information about an audience generation job}
\usage{
cleanroomsml_get_audience_generation_job(audienceGenerationJobArn)
}
\arguments{
\item{audienceGenerationJobArn}{[required] The Amazon Resource Name (ARN) of the audience generation job that you
are interested in.}
}
\description{
Returns information about an audience generation job.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_get_audience_generation_job/} for full documentation.
}
\keyword{internal}
