% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_get_audience_model}
\alias{cleanroomsml_get_audience_model}
\title{Returns information about an audience model}
\usage{
cleanroomsml_get_audience_model(audienceModelArn)
}
\arguments{
\item{audienceModelArn}{[required] The Amazon Resource Name (ARN) of the audience model that you are
interested in.}
}
\description{
Returns information about an audience model

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_get_audience_model/} for full documentation.
}
\keyword{internal}
