% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_get_collaboration_configured_model_algorithm_association}
\alias{cleanroomsml_get_collaboration_configured_model_algorithm_association}
\title{Returns information about the configured model algorithm association in
a collaboration}
\usage{
cleanroomsml_get_collaboration_configured_model_algorithm_association(
  configuredModelAlgorithmAssociationArn,
  collaborationIdentifier
)
}
\arguments{
\item{configuredModelAlgorithmAssociationArn}{[required] The Amazon Resource Name (ARN) of the configured model algorithm
association that you want to return information about.}

\item{collaborationIdentifier}{[required] The collaboration ID for the collaboration that contains the configured
model algorithm association that you want to return information about.}
}
\description{
Returns information about the configured model algorithm association in a collaboration.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_get_collaboration_configured_model_algorithm_association/} for full documentation.
}
\keyword{internal}
