% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_get_collaboration_ml_input_channel}
\alias{cleanroomsml_get_collaboration_ml_input_channel}
\title{Returns information about a specific ML input channel in a collaboration}
\usage{
cleanroomsml_get_collaboration_ml_input_channel(
  mlInputChannelArn,
  collaborationIdentifier
)
}
\arguments{
\item{mlInputChannelArn}{[required] The Amazon Resource Name (ARN) of the ML input channel that you want to
get.}

\item{collaborationIdentifier}{[required] The collaboration ID of the collaboration that contains the ML input
channel that you want to get.}
}
\description{
Returns information about a specific ML input channel in a collaboration.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_get_collaboration_ml_input_channel/} for full documentation.
}
\keyword{internal}
