% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_get_collaboration_trained_model}
\alias{cleanroomsml_get_collaboration_trained_model}
\title{Returns information about a trained model in a collaboration}
\usage{
cleanroomsml_get_collaboration_trained_model(
  trainedModelArn,
  collaborationIdentifier
)
}
\arguments{
\item{trainedModelArn}{[required] The Amazon Resource Name (ARN) of the trained model that you want to
return information about.}

\item{collaborationIdentifier}{[required] The collaboration ID that contains the trained model that you want to
return information about.}
}
\description{
Returns information about a trained model in a collaboration.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_get_collaboration_trained_model/} for full documentation.
}
\keyword{internal}
