% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_get_configured_model_algorithm}
\alias{cleanroomsml_get_configured_model_algorithm}
\title{Returns information about a configured model algorithm}
\usage{
cleanroomsml_get_configured_model_algorithm(configuredModelAlgorithmArn)
}
\arguments{
\item{configuredModelAlgorithmArn}{[required] The Amazon Resource Name (ARN) of the configured model algorithm that
you want to return information about.}
}
\description{
Returns information about a configured model algorithm.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_get_configured_model_algorithm/} for full documentation.
}
\keyword{internal}
