% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_get_ml_configuration}
\alias{cleanroomsml_get_ml_configuration}
\title{Returns information about a specific ML configuration}
\usage{
cleanroomsml_get_ml_configuration(membershipIdentifier)
}
\arguments{
\item{membershipIdentifier}{[required] The membership ID of the member that owns the ML configuration you want
to return information about.}
}
\description{
Returns information about a specific ML configuration.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_get_ml_configuration/} for full documentation.
}
\keyword{internal}
