% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_get_ml_input_channel}
\alias{cleanroomsml_get_ml_input_channel}
\title{Returns information about an ML input channel}
\usage{
cleanroomsml_get_ml_input_channel(mlInputChannelArn, membershipIdentifier)
}
\arguments{
\item{mlInputChannelArn}{[required] The Amazon Resource Name (ARN) of the ML input channel that you want to
get.}

\item{membershipIdentifier}{[required] The membership ID of the membership that contains the ML input channel
that you want to get.}
}
\description{
Returns information about an ML input channel.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_get_ml_input_channel/} for full documentation.
}
\keyword{internal}
