% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_get_trained_model}
\alias{cleanroomsml_get_trained_model}
\title{Returns information about a trained model}
\usage{
cleanroomsml_get_trained_model(trainedModelArn, membershipIdentifier)
}
\arguments{
\item{trainedModelArn}{[required] The Amazon Resource Name (ARN) of the trained model that you are
interested in.}

\item{membershipIdentifier}{[required] The membership ID of the member that created the trained model that you
are interested in.}
}
\description{
Returns information about a trained model.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_get_trained_model/} for full documentation.
}
\keyword{internal}
