% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_get_trained_model_inference_job}
\alias{cleanroomsml_get_trained_model_inference_job}
\title{Returns information about a trained model inference job}
\usage{
cleanroomsml_get_trained_model_inference_job(
  membershipIdentifier,
  trainedModelInferenceJobArn
)
}
\arguments{
\item{membershipIdentifier}{[required] Provides the membership ID of the membership that contains the trained
model inference job that you are interested in.}

\item{trainedModelInferenceJobArn}{[required] Provides the Amazon Resource Name (ARN) of the trained model inference
job that you are interested in.}
}
\description{
Returns information about a trained model inference job.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_get_trained_model_inference_job/} for full documentation.
}
\keyword{internal}
