% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_get_training_dataset}
\alias{cleanroomsml_get_training_dataset}
\title{Returns information about a training dataset}
\usage{
cleanroomsml_get_training_dataset(trainingDatasetArn)
}
\arguments{
\item{trainingDatasetArn}{[required] The Amazon Resource Name (ARN) of the training dataset that you are
interested in.}
}
\description{
Returns information about a training dataset.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_get_training_dataset/} for full documentation.
}
\keyword{internal}
