% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_list_audience_generation_jobs}
\alias{cleanroomsml_list_audience_generation_jobs}
\title{Returns a list of audience generation jobs}
\usage{
cleanroomsml_list_audience_generation_jobs(
  nextToken = NULL,
  maxResults = NULL,
  configuredAudienceModelArn = NULL,
  collaborationId = NULL
)
}
\arguments{
\item{nextToken}{The token value retrieved from a previous call to access the next page
of results.}

\item{maxResults}{The maximum size of the results that is returned per call.}

\item{configuredAudienceModelArn}{The Amazon Resource Name (ARN) of the configured audience model that was
used for the audience generation jobs that you are interested in.}

\item{collaborationId}{The identifier of the collaboration that contains the audience
generation jobs that you are interested in.}
}
\description{
Returns a list of audience generation jobs.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_list_audience_generation_jobs/} for full documentation.
}
\keyword{internal}
