% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_list_collaboration_configured_model_algorithm_associations}
\alias{cleanroomsml_list_collaboration_configured_model_algorithm_associations}
\title{Returns a list of the configured model algorithm associations in a
collaboration}
\usage{
cleanroomsml_list_collaboration_configured_model_algorithm_associations(
  nextToken = NULL,
  maxResults = NULL,
  collaborationIdentifier
)
}
\arguments{
\item{nextToken}{The token value retrieved from a previous call to access the next page
of results.}

\item{maxResults}{The maximum size of the results that is returned per call.}

\item{collaborationIdentifier}{[required] The collaboration ID of the collaboration that contains the configured
model algorithm associations that you are interested in.}
}
\description{
Returns a list of the configured model algorithm associations in a collaboration.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_list_collaboration_configured_model_algorithm_associations/} for full documentation.
}
\keyword{internal}
