% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_list_collaboration_trained_models}
\alias{cleanroomsml_list_collaboration_trained_models}
\title{Returns a list of the trained models in a collaboration}
\usage{
cleanroomsml_list_collaboration_trained_models(
  nextToken = NULL,
  maxResults = NULL,
  collaborationIdentifier
)
}
\arguments{
\item{nextToken}{The token value retrieved from a previous call to access the next page
of results.}

\item{maxResults}{The maximum size of the results that is returned per call.}

\item{collaborationIdentifier}{[required] The collaboration ID of the collaboration that contains the trained
models you are interested in.}
}
\description{
Returns a list of the trained models in a collaboration.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_list_collaboration_trained_models/} for full documentation.
}
\keyword{internal}
