% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_list_configured_model_algorithms}
\alias{cleanroomsml_list_configured_model_algorithms}
\title{Returns a list of configured model algorithms}
\usage{
cleanroomsml_list_configured_model_algorithms(
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{nextToken}{The token value retrieved from a previous call to access the next page
of results.}

\item{maxResults}{The maximum size of the results that is returned per call.}
}
\description{
Returns a list of configured model algorithms.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_list_configured_model_algorithms/} for full documentation.
}
\keyword{internal}
