% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_list_tags_for_resource}
\alias{cleanroomsml_list_tags_for_resource}
\title{Returns a list of tags for a provided resource}
\usage{
cleanroomsml_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you are interested
in.}
}
\description{
Returns a list of tags for a provided resource.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
