% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_list_trained_model_inference_jobs}
\alias{cleanroomsml_list_trained_model_inference_jobs}
\title{Returns a list of trained model inference jobs that match the request
parameters}
\usage{
cleanroomsml_list_trained_model_inference_jobs(
  nextToken = NULL,
  maxResults = NULL,
  membershipIdentifier,
  trainedModelArn = NULL
)
}
\arguments{
\item{nextToken}{The token value retrieved from a previous call to access the next page
of results.}

\item{maxResults}{The maximum size of the results that is returned per call.}

\item{membershipIdentifier}{[required] The membership}

\item{trainedModelArn}{The Amazon Resource Name (ARN) of a trained model that was used to
create the trained model inference jobs that you are interested in.}
}
\description{
Returns a list of trained model inference jobs that match the request parameters.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_list_trained_model_inference_jobs/} for full documentation.
}
\keyword{internal}
