% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_list_training_datasets}
\alias{cleanroomsml_list_training_datasets}
\title{Returns a list of training datasets}
\usage{
cleanroomsml_list_training_datasets(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{The token value retrieved from a previous call to access the next page
of results.}

\item{maxResults}{The maximum size of the results that is returned per call.}
}
\description{
Returns a list of training datasets.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_list_training_datasets/} for full documentation.
}
\keyword{internal}
