% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_put_configured_audience_model_policy}
\alias{cleanroomsml_put_configured_audience_model_policy}
\title{Create or update the resource policy for a configured audience model}
\usage{
cleanroomsml_put_configured_audience_model_policy(
  configuredAudienceModelArn,
  configuredAudienceModelPolicy,
  previousPolicyHash = NULL,
  policyExistenceCondition = NULL
)
}
\arguments{
\item{configuredAudienceModelArn}{[required] The Amazon Resource Name (ARN) of the configured audience model that the
resource policy will govern.}

\item{configuredAudienceModelPolicy}{[required] The IAM resource policy.}

\item{previousPolicyHash}{A cryptographic hash of the contents of the policy used to prevent
unexpected concurrent modification of the policy.}

\item{policyExistenceCondition}{Use this to prevent unexpected concurrent modification of the policy.}
}
\description{
Create or update the resource policy for a configured audience model.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_put_configured_audience_model_policy/} for full documentation.
}
\keyword{internal}
