% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_put_ml_configuration}
\alias{cleanroomsml_put_ml_configuration}
\title{Assigns information about an ML configuration}
\usage{
cleanroomsml_put_ml_configuration(membershipIdentifier, defaultOutputLocation)
}
\arguments{
\item{membershipIdentifier}{[required] The membership ID of the member that is being configured.}

\item{defaultOutputLocation}{[required] The default Amazon S3 location where ML output is stored for the
specified member.}
}
\description{
Assigns information about an ML configuration.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_put_ml_configuration/} for full documentation.
}
\keyword{internal}
