% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_start_audience_export_job}
\alias{cleanroomsml_start_audience_export_job}
\title{Export an audience of a specified size after you have generated an
audience}
\usage{
cleanroomsml_start_audience_export_job(
  name,
  audienceGenerationJobArn,
  audienceSize,
  description = NULL
)
}
\arguments{
\item{name}{[required] The name of the audience export job.}

\item{audienceGenerationJobArn}{[required] The Amazon Resource Name (ARN) of the audience generation job that you
want to export.}

\item{audienceSize}{[required]}

\item{description}{The description of the audience export job.}
}
\description{
Export an audience of a specified size after you have generated an audience.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_start_audience_export_job/} for full documentation.
}
\keyword{internal}
