% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_start_trained_model_export_job}
\alias{cleanroomsml_start_trained_model_export_job}
\title{Provides the information necessary to start a trained model export job}
\usage{
cleanroomsml_start_trained_model_export_job(
  name,
  trainedModelArn,
  trainedModelVersionIdentifier = NULL,
  membershipIdentifier,
  outputConfiguration,
  description = NULL
)
}
\arguments{
\item{name}{[required] The name of the trained model export job.}

\item{trainedModelArn}{[required] The Amazon Resource Name (ARN) of the trained model that you want to
export.}

\item{trainedModelVersionIdentifier}{The version identifier of the trained model to export. This specifies
which version of the trained model should be exported to the specified
destination.}

\item{membershipIdentifier}{[required] The membership ID of the member that is receiving the exported trained
model artifacts.}

\item{outputConfiguration}{[required] The output configuration information for the trained model export job.}

\item{description}{The description of the trained model export job.}
}
\description{
Provides the information necessary to start a trained model export job.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_start_trained_model_export_job/} for full documentation.
}
\keyword{internal}
