% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_untag_resource}
\alias{cleanroomsml_untag_resource}
\title{Removes metadata tags from a specified resource}
\usage{
cleanroomsml_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to remove
tags from.}

\item{tagKeys}{[required] The key values of tags that you want to remove.}
}
\description{
Removes metadata tags from a specified resource.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_untag_resource/} for full documentation.
}
\keyword{internal}
