% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanroomsml_operations.R
\name{cleanroomsml_update_configured_audience_model}
\alias{cleanroomsml_update_configured_audience_model}
\title{Provides the information necessary to update a configured audience model}
\usage{
cleanroomsml_update_configured_audience_model(
  configuredAudienceModelArn,
  outputConfig = NULL,
  audienceModelArn = NULL,
  sharedAudienceMetrics = NULL,
  minMatchingSeedSize = NULL,
  audienceSizeConfig = NULL,
  description = NULL
)
}
\arguments{
\item{configuredAudienceModelArn}{[required] The Amazon Resource Name (ARN) of the configured audience model that you
want to update.}

\item{outputConfig}{The new output configuration.}

\item{audienceModelArn}{The Amazon Resource Name (ARN) of the new audience model that you want
to use.}

\item{sharedAudienceMetrics}{The new value for whether to share audience metrics.}

\item{minMatchingSeedSize}{The minimum number of users from the seed audience that must match with
users in the training data of the audience model.}

\item{audienceSizeConfig}{The new audience size configuration.}

\item{description}{The new description of the configured audience model.}
}
\description{
Provides the information necessary to update a configured audience model. Updates that impact audience generation jobs take effect when a new job starts, but do not impact currently running jobs.

See \url{https://www.paws-r-sdk.com/docs/cleanroomsml_update_configured_audience_model/} for full documentation.
}
\keyword{internal}
