% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_attach_object}
\alias{clouddirectory_attach_object}
\title{Attaches an existing object to another object}
\usage{
clouddirectory_attach_object(DirectoryArn, ParentReference,
  ChildReference, LinkName)
}
\arguments{
\item{DirectoryArn}{[required] Amazon Resource Name (ARN) that is associated with the Directory where
both objects reside. For more information, see arns.}

\item{ParentReference}{[required] The parent object reference.}

\item{ChildReference}{[required] The child object reference to be attached to the object.}

\item{LinkName}{[required] The link name with which the child object is attached to the parent.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AttachedObjectIdentifier = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Attaches an existing object to another object. An object can be accessed
in two ways:
\enumerate{
\item Using the path
\item Using \code{ObjectIdentifier}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$attach_object(
  DirectoryArn = "string",
  ParentReference = list(
    Selector = "string"
  ),
  ChildReference = list(
    Selector = "string"
  ),
  LinkName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
