% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_delete_schema}
\alias{clouddirectory_delete_schema}
\title{Deletes a given schema}
\usage{
clouddirectory_delete_schema(SchemaArn)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) of the development schema. For more
information, see arns.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SchemaArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a given schema. Schemas in a development and published state can
only be deleted.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_schema(
  SchemaArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
