% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_delete_typed_link_facet}
\alias{clouddirectory_delete_typed_link_facet}
\title{Deletes a TypedLinkFacet}
\usage{
clouddirectory_delete_typed_link_facet(SchemaArn, Name)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) that is associated with the schema. For
more information, see arns.}

\item{Name}{[required] The unique name of the typed link facet.}
}
\description{
Deletes a TypedLinkFacet. For more information, see \href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/directory_objects_links.html#directory_objects_links_typedlink}{Typed Links}.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_delete_typed_link_facet/} for full documentation.
}
\keyword{internal}
