% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_detach_from_index}
\alias{clouddirectory_detach_from_index}
\title{Detaches the specified object from the specified index}
\usage{
clouddirectory_detach_from_index(DirectoryArn, IndexReference, TargetReference)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) of the directory the index and object
exist in.}

\item{IndexReference}{[required] A reference to the index object.}

\item{TargetReference}{[required] A reference to the object being detached from the index.}
}
\description{
Detaches the specified object from the specified index.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_detach_from_index/} for full documentation.
}
\keyword{internal}
