% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_detach_policy}
\alias{clouddirectory_detach_policy}
\title{Detaches a policy from an object}
\usage{
clouddirectory_detach_policy(DirectoryArn, PolicyReference,
  ObjectReference)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) that is associated with the
<span>Directory</span> where both objects reside. For more information,
see <span>arns</span>.}

\item{PolicyReference}{[required] Reference that identifies the policy object.}

\item{ObjectReference}{[required] Reference that identifies the object whose policy object will be
detached.}
}
\value{
An empty list.
}
\description{
Detaches a policy from an object.
}
\section{Request syntax}{
\preformatted{svc$detach_policy(
  DirectoryArn = "string",
  PolicyReference = list(
    Selector = "string"
  ),
  ObjectReference = list(
    Selector = "string"
  )
)
}
}

\keyword{internal}
