% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_detach_typed_link}
\alias{clouddirectory_detach_typed_link}
\title{Detaches a typed link from a specified source and target object}
\usage{
clouddirectory_detach_typed_link(DirectoryArn, TypedLinkSpecifier)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) of the directory where you want to detach
the typed link.}

\item{TypedLinkSpecifier}{[required] Used to accept a typed link specifier as input.}
}
\description{
Detaches a typed link from a specified source and target object. For more information, see \href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/directory_objects_links.html#directory_objects_links_typedlink}{Typed Links}.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_detach_typed_link/} for full documentation.
}
\keyword{internal}
