% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_disable_directory}
\alias{clouddirectory_disable_directory}
\title{Disables the specified directory}
\usage{
clouddirectory_disable_directory(DirectoryArn)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory to disable.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DirectoryArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Disables the specified directory. Disabled directories cannot be read or
written to. Only enabled directories can be disabled. Disabled
directories may be reenabled.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disable_directory(
  DirectoryArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
