% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_enable_directory}
\alias{clouddirectory_enable_directory}
\title{Enables the specified directory}
\usage{
clouddirectory_enable_directory(DirectoryArn)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory to enable.}
}
\description{
Enables the specified directory. Only disabled directories can be enabled. Once enabled, the directory can then be read and written to.

See \url{https://paws-r.github.io/docs/clouddirectory/enable_directory.html} for full documentation.
}
\keyword{internal}
