% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_get_object_information}
\alias{clouddirectory_get_object_information}
\title{Retrieves metadata about an object}
\usage{
clouddirectory_get_object_information(DirectoryArn, ObjectReference,
  ConsistencyLevel)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory being retrieved.}

\item{ObjectReference}{[required] A reference to the object.}

\item{ConsistencyLevel}{The consistency level at which to retrieve the object information.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SchemaFacets = list(
    list(
      SchemaArn = "string",
      FacetName = "string"
    )
  ),
  ObjectIdentifier = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves metadata about an object.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_object_information(
  DirectoryArn = "string",
  ObjectReference = list(
    Selector = "string"
  ),
  ConsistencyLevel = "SERIALIZABLE"|"EVENTUAL"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
