% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_get_typed_link_facet_information}
\alias{clouddirectory_get_typed_link_facet_information}
\title{Returns the identity attribute order for a specific TypedLinkFacet}
\usage{
clouddirectory_get_typed_link_facet_information(SchemaArn, Name)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) that is associated with the schema. For
more information, see arns.}

\item{Name}{[required] The unique name of the typed link facet.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  IdentityAttributeOrder = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Returns the identity attribute order for a specific TypedLinkFacet. For
more information, see \href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/directory_objects_links.html#directory_objects_links_typedlink}{Typed Links}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_typed_link_facet_information(
  SchemaArn = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
