% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_list_development_schema_arns}
\alias{clouddirectory_list_development_schema_arns}
\title{Retrieves each Amazon Resource Name (ARN) of schemas in the development
state}
\usage{
clouddirectory_list_development_schema_arns(NextToken, MaxResults)
}
\arguments{
\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\value{
A list with the following syntax:\preformatted{list(
  SchemaArns = list(
    "string"
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves each Amazon Resource Name (ARN) of schemas in the development
state.
}
\section{Request syntax}{
\preformatted{svc$list_development_schema_arns(
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
