% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_list_facet_names}
\alias{clouddirectory_list_facet_names}
\title{Retrieves the names of facets that exist in a schema}
\usage{
clouddirectory_list_facet_names(SchemaArn, NextToken, MaxResults)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) to retrieve facet names from.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FacetNames = list(
    "string"
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the names of facets that exist in a schema.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_facet_names(
  SchemaArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
