% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_list_managed_schema_arns}
\alias{clouddirectory_list_managed_schema_arns}
\title{Lists the major version families of each managed schema}
\usage{
clouddirectory_list_managed_schema_arns(
  SchemaArn = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{SchemaArn}{The response for ListManagedSchemaArns. When this parameter is used, all
minor version ARNs for a major version are listed.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\description{
Lists the major version families of each managed schema. If a major version ARN is provided as SchemaArn, the minor version revisions in that family are listed instead.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_list_managed_schema_arns/} for full documentation.
}
\keyword{internal}
