% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_list_outgoing_typed_links}
\alias{clouddirectory_list_outgoing_typed_links}
\title{Returns a paginated list of all the outgoing TypedLinkSpecifier
information for an object}
\usage{
clouddirectory_list_outgoing_typed_links(DirectoryArn, ObjectReference,
  FilterAttributeRanges, FilterTypedLink, NextToken, MaxResults,
  ConsistencyLevel)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) of the directory where you want to list
the typed links.}

\item{ObjectReference}{[required] A reference that identifies the object whose attributes will be listed.}

\item{FilterAttributeRanges}{Provides range filters for multiple attributes. When providing ranges to
typed link selection, any inexact ranges must be specified at the end.
Any attributes that do not have a range specified are presumed to match
the entire range.}

\item{FilterTypedLink}{Filters are interpreted in the order of the attributes defined on the
typed link facet, not the order they are supplied to any API calls.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}

\item{ConsistencyLevel}{The consistency level to execute the request at.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TypedLinkSpecifiers = list(
    list(
      TypedLinkFacet = list(
        SchemaArn = "string",
        TypedLinkName = "string"
      ),
      SourceObjectReference = list(
        Selector = "string"
      ),
      TargetObjectReference = list(
        Selector = "string"
      ),
      IdentityAttributeValues = list(
        list(
          AttributeName = "string",
          Value = list(
            StringValue = "string",
            BinaryValue = raw,
            BooleanValue = TRUE|FALSE,
            NumberValue = "string",
            DatetimeValue = as.POSIXct(
              "2015-01-01"
            )
          )
        )
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a paginated list of all the outgoing TypedLinkSpecifier
information for an object. It also supports filtering by typed link
facet and identity attributes. For more information, see \href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/directory_objects_links.html#directory_objects_links_typedlink}{Typed Links}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_outgoing_typed_links(
  DirectoryArn = "string",
  ObjectReference = list(
    Selector = "string"
  ),
  FilterAttributeRanges = list(
    list(
      AttributeName = "string",
      Range = list(
        StartMode = "FIRST"|"LAST"|"LAST_BEFORE_MISSING_VALUES"|"INCLUSIVE"|"EXCLUSIVE",
        StartValue = list(
          StringValue = "string",
          BinaryValue = raw,
          BooleanValue = TRUE|FALSE,
          NumberValue = "string",
          DatetimeValue = as.POSIXct(
            "2015-01-01"
          )
        ),
        EndMode = "FIRST"|"LAST"|"LAST_BEFORE_MISSING_VALUES"|"INCLUSIVE"|"EXCLUSIVE",
        EndValue = list(
          StringValue = "string",
          BinaryValue = raw,
          BooleanValue = TRUE|FALSE,
          NumberValue = "string",
          DatetimeValue = as.POSIXct(
            "2015-01-01"
          )
        )
      )
    )
  ),
  FilterTypedLink = list(
    SchemaArn = "string",
    TypedLinkName = "string"
  ),
  NextToken = "string",
  MaxResults = 123,
  ConsistencyLevel = "SERIALIZABLE"|"EVENTUAL"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
