% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_publish_schema}
\alias{clouddirectory_publish_schema}
\title{Publishes a development schema with a major version and a recommended
minor version}
\usage{
clouddirectory_publish_schema(DevelopmentSchemaArn, Version,
  MinorVersion, Name)
}
\arguments{
\item{DevelopmentSchemaArn}{[required] The Amazon Resource Name (ARN) that is associated with the development
schema. For more information, see arns.}

\item{Version}{[required] The major version under which the schema will be published. Schemas have
both a major and minor version associated with them.}

\item{MinorVersion}{The minor version under which the schema will be published. This
parameter is recommended. Schemas have both a major and minor version
associated with them.}

\item{Name}{The new name under which the schema will be published. If this is not
provided, the development schema is considered.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PublishedSchemaArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Publishes a development schema with a major version and a recommended
minor version.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$publish_schema(
  DevelopmentSchemaArn = "string",
  Version = "string",
  MinorVersion = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
