% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_remove_facet_from_object}
\alias{clouddirectory_remove_facet_from_object}
\title{Removes the specified facet from the specified object}
\usage{
clouddirectory_remove_facet_from_object(
  DirectoryArn,
  SchemaFacet,
  ObjectReference
)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory in which the object resides.}

\item{SchemaFacet}{[required] The facet to remove. See SchemaFacet for details.}

\item{ObjectReference}{[required] A reference to the object to remove the facet from.}
}
\description{
Removes the specified facet from the specified object.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_remove_facet_from_object/} for full documentation.
}
\keyword{internal}
