% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_untag_resource}
\alias{clouddirectory_untag_resource}
\title{An API operation for removing tags from a resource}
\usage{
clouddirectory_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource. Tagging is only
supported for directories.}

\item{TagKeys}{[required] Keys of the tag that need to be removed from the resource.}
}
\description{
An API operation for removing tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_untag_resource/} for full documentation.
}
\keyword{internal}
