% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_update_link_attributes}
\alias{clouddirectory_update_link_attributes}
\title{Updates a given typed link’s attributes}
\usage{
clouddirectory_update_link_attributes(
  DirectoryArn,
  TypedLinkSpecifier,
  AttributeUpdates
)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) that is associated with the Directory
where the updated typed link resides. For more information, see arns or
\href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/directory_objects_links.html#directory_objects_links_typedlink}{Typed Links}.}

\item{TypedLinkSpecifier}{[required] Allows a typed link specifier to be accepted as input.}

\item{AttributeUpdates}{[required] The attributes update structure.}
}
\description{
Updates a given typed link’s attributes. Attributes to be updated must not contribute to the typed link’s identity, as defined by its \code{IdentityAttributeOrder}.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_update_link_attributes/} for full documentation.
}
\keyword{internal}
