% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_update_object_attributes}
\alias{clouddirectory_update_object_attributes}
\title{Updates a given object's attributes}
\usage{
clouddirectory_update_object_attributes(
  DirectoryArn,
  ObjectReference,
  AttributeUpdates
)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) that is associated with the Directory
where the object resides. For more information, see arns.}

\item{ObjectReference}{[required] The reference that identifies the object.}

\item{AttributeUpdates}{[required] The attributes update structure.}
}
\description{
Updates a given object's attributes.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_update_object_attributes/} for full documentation.
}
\keyword{internal}
