% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_update_schema}
\alias{clouddirectory_update_schema}
\title{Updates the schema name with a new name}
\usage{
clouddirectory_update_schema(SchemaArn, Name)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) of the development schema. For more
information, see arns.}

\item{Name}{[required] The name of the schema.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SchemaArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates the schema name with a new name. Only development schema names
can be updated.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_schema(
  SchemaArn = "string",
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
