% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_add_tags_to_resource}
\alias{cloudhsm_add_tags_to_resource}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_add_tags_to_resource(ResourceArn, TagList)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the AWS CloudHSM resource to tag.}

\item{TagList}{[required] One or more tags.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

See \url{https://www.paws-r-sdk.com/docs/cloudhsm_add_tags_to_resource/} for full documentation.
}
\keyword{internal}
