% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_create_luna_client}
\alias{cloudhsm_create_luna_client}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_create_luna_client(Label = NULL, Certificate)
}
\arguments{
\item{Label}{The label for the client.}

\item{Certificate}{[required] The contents of a Base64-Encoded X.509 v3 certificate to be installed on
the HSMs used by this client.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

See \url{https://www.paws-r-sdk.com/docs/cloudhsm_create_luna_client/} for full documentation.
}
\keyword{internal}
