% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_describe_luna_client}
\alias{cloudhsm_describe_luna_client}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_describe_luna_client(ClientArn, CertificateFingerprint)
}
\arguments{
\item{ClientArn}{The ARN of the client.}

\item{CertificateFingerprint}{The certificate fingerprint.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more
information, see \href{http://aws.amazon.com/cloudhsm/faqs-classic/}{AWS CloudHSM Classic FAQs}, the \href{http://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide}, and the
\href{http://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.
}
\details{
\strong{For information about the current version of AWS CloudHSM}, see \href{http://aws.amazon.com/cloudhsm/}{AWS CloudHSM}, the \href{http://docs.aws.amazon.com/cloudhsm/latest/userguide/}{AWS CloudHSM User Guide}, and the
\href{http://docs.aws.amazon.com/cloudhsm/latest/APIReference/}{AWS CloudHSM API Reference}.

Retrieves information about an HSM client.
}
\section{Request syntax}{
\preformatted{svc$describe_luna_client(
  ClientArn = "string",
  CertificateFingerprint = "string"
)
}
}

\keyword{internal}
