% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_list_hsms}
\alias{cloudhsm_list_hsms}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_list_hsms(NextToken)
}
\arguments{
\item{NextToken}{The \code{NextToken} value from a previous call to
\code{\link[=cloudhsm_list_hsms]{list_hsms}}. Pass null if this is the first call.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  HsmList = list(
    "string"
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more
information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the \href{https://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide},
and the \href{https://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.

\strong{For information about the current version of AWS CloudHSM}, see \href{https://aws.amazon.com/cloudhsm/}{AWS CloudHSM}, the \href{https://docs.aws.amazon.com/cloudhsm/latest/userguide/}{AWS CloudHSM User Guide}, and the
\href{https://docs.aws.amazon.com/cloudhsm/latest/APIReference/}{AWS CloudHSM API Reference}.

Retrieves the identifiers of all of the HSMs provisioned for the current
customer.

This operation supports pagination with the use of the \code{NextToken}
member. If more results are available, the \code{NextToken} member of the
response contains a token that you pass in the next call to
\code{\link[=cloudhsm_list_hsms]{list_hsms}} to retrieve the next set of items.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_hsms(
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
