% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_modify_hapg}
\alias{cloudhsm_modify_hapg}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_modify_hapg(HapgArn, Label, PartitionSerialList)
}
\arguments{
\item{HapgArn}{[required] The ARN of the high-availability partition group to modify.}

\item{Label}{The new label for the high-availability partition group.}

\item{PartitionSerialList}{The list of partition serial numbers to make members of the
high-availability partition group.}
}
\value{
A list with the following syntax:\preformatted{list(
  HapgArn = "string"
)
}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more
information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the \href{https://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide},
and the \href{https://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.

\strong{For information about the current version of AWS CloudHSM}, see \href{https://aws.amazon.com/cloudhsm/}{AWS CloudHSM}, the \href{https://docs.aws.amazon.com/cloudhsm/latest/userguide/}{AWS CloudHSM User Guide}, and the
\href{https://docs.aws.amazon.com/cloudhsm/latest/APIReference/}{AWS CloudHSM API Reference}.

Modifies an existing high-availability partition group.
}
\section{Request syntax}{
\preformatted{svc$modify_hapg(
  HapgArn = "string",
  Label = "string",
  PartitionSerialList = list(
    "string"
  )
)
}
}

\keyword{internal}
