% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_modify_luna_client}
\alias{cloudhsm_modify_luna_client}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_modify_luna_client(ClientArn, Certificate)
}
\arguments{
\item{ClientArn}{[required] The ARN of the client.}

\item{Certificate}{[required] The new certificate for the client.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

See \url{https://www.paws-r-sdk.com/docs/cloudhsm_modify_luna_client/} for full documentation.
}
\keyword{internal}
