% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_remove_tags_from_resource}
\alias{cloudhsm_remove_tags_from_resource}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_remove_tags_from_resource(ResourceArn, TagKeyList)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the AWS CloudHSM resource.}

\item{TagKeyList}{[required] The tag key or keys to remove.

Specify only the tag key to remove (not the value). To overwrite the
value for an existing tag, use
\code{\link[=cloudhsm_add_tags_to_resource]{add_tags_to_resource}}.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more
information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the \href{https://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide},
and the \href{https://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.

\strong{For information about the current version of AWS CloudHSM}, see \href{https://aws.amazon.com/cloudhsm/}{AWS CloudHSM}, the \href{https://docs.aws.amazon.com/cloudhsm/latest/userguide/}{AWS CloudHSM User Guide}, and the
\href{https://docs.aws.amazon.com/cloudhsm/latest/APIReference/}{AWS CloudHSM API Reference}.

Removes one or more tags from the specified AWS CloudHSM resource.

To remove a tag, specify only the tag key to remove (not the value). To
overwrite the value for an existing tag, use
\code{\link[=cloudhsm_add_tags_to_resource]{add_tags_to_resource}}.
}
\section{Request syntax}{
\preformatted{svc$remove_tags_from_resource(
  ResourceArn = "string",
  TagKeyList = list(
    "string"
  )
)
}
}

\keyword{internal}
