% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_copy_backup_to_region}
\alias{cloudhsmv2_copy_backup_to_region}
\title{Copy an AWS CloudHSM cluster backup to a different region}
\usage{
cloudhsmv2_copy_backup_to_region(DestinationRegion, BackupId)
}
\arguments{
\item{DestinationRegion}{[required] The AWS region that will contain your copied CloudHSM cluster backup.}

\item{BackupId}{[required] The ID of the backup that will be copied to the destination region.}
}
\description{
Copy an AWS CloudHSM cluster backup to a different region.
}
\section{Request syntax}{
\preformatted{svc$copy_backup_to_region(
  DestinationRegion = "string",
  BackupId = "string"
)
}
}

\keyword{internal}
