% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_create_hsm}
\alias{cloudhsmv2_create_hsm}
\title{Creates a new hardware security module (HSM) in the specified AWS
CloudHSM cluster}
\usage{
cloudhsmv2_create_hsm(ClusterId, AvailabilityZone, IpAddress)
}
\arguments{
\item{ClusterId}{[required] The identifier (ID) of the HSM's cluster. To find the cluster ID, use
\code{\link[=cloudhsmv2_describe_clusters]{describe_clusters}}.}

\item{AvailabilityZone}{[required] The Availability Zone where you are creating the HSM. To find the
cluster's Availability Zones, use
\code{\link[=cloudhsmv2_describe_clusters]{describe_clusters}}.}

\item{IpAddress}{The HSM's IP address. If you specify an IP address, use an available
address from the subnet that maps to the Availability Zone where you are
creating the HSM. If you don't specify an IP address, one is chosen for
you from that subnet.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Hsm = list(
    AvailabilityZone = "string",
    ClusterId = "string",
    SubnetId = "string",
    EniId = "string",
    EniIp = "string",
    HsmId = "string",
    State = "CREATE_IN_PROGRESS"|"ACTIVE"|"DEGRADED"|"DELETE_IN_PROGRESS"|"DELETED",
    StateMessage = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a new hardware security module (HSM) in the specified AWS
CloudHSM cluster.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_hsm(
  ClusterId = "string",
  AvailabilityZone = "string",
  IpAddress = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
