% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_get_resource_policy}
\alias{cloudhsmv2_get_resource_policy}
\title{Retrieves the resource policy document attached to a given resource}
\usage{
cloudhsmv2_get_resource_policy(ResourceArn = NULL)
}
\arguments{
\item{ResourceArn}{Amazon Resource Name (ARN) of the resource to which a policy is
attached.}
}
\description{
Retrieves the resource policy document attached to a given resource.

See \url{https://www.paws-r-sdk.com/docs/cloudhsmv2_get_resource_policy/} for full documentation.
}
\keyword{internal}
