% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_modify_cluster}
\alias{cloudhsmv2_modify_cluster}
\title{Modifies CloudHSM cluster}
\usage{
cloudhsmv2_modify_cluster(
  HsmType = NULL,
  BackupRetentionPolicy = NULL,
  ClusterId
)
}
\arguments{
\item{HsmType}{The desired HSM type of the cluster.}

\item{BackupRetentionPolicy}{A policy that defines how the service retains backups.}

\item{ClusterId}{[required] The identifier (ID) of the cluster that you want to modify. To find the
cluster ID, use \code{\link[=cloudhsmv2_describe_clusters]{describe_clusters}}.}
}
\description{
Modifies CloudHSM cluster.

See \url{https://www.paws-r-sdk.com/docs/cloudhsmv2_modify_cluster/} for full documentation.
}
\keyword{internal}
