% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_put_resource_policy}
\alias{cloudhsmv2_put_resource_policy}
\title{Creates or updates an CloudHSM resource policy}
\usage{
cloudhsmv2_put_resource_policy(ResourceArn = NULL, Policy = NULL)
}
\arguments{
\item{ResourceArn}{Amazon Resource Name (ARN) of the resource to which you want to attach a
policy.}

\item{Policy}{The policy you want to associate with a resource.

For an example policy, see \href{https://docs.aws.amazon.com/cloudhsm/latest/userguide/sharing.html}{Working with shared backups}
in the CloudHSM User Guide}
}
\description{
Creates or updates an CloudHSM resource policy. A resource policy helps you to define the IAM entity (for example, an Amazon Web Services account) that can manage your CloudHSM resources. The following resources support CloudHSM resource policies:

See \url{https://www.paws-r-sdk.com/docs/cloudhsmv2_put_resource_policy/} for full documentation.
}
\keyword{internal}
