% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_restore_backup}
\alias{cloudhsmv2_restore_backup}
\title{Restores a specified AWS CloudHSM backup that is in the PENDING_DELETION
state}
\usage{
cloudhsmv2_restore_backup(BackupId)
}
\arguments{
\item{BackupId}{[required] The ID of the backup to be restored. To find the ID of a backup, use the
\code{\link[=cloudhsmv2_describe_backups]{describe_backups}} operation.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Backup = list(
    BackupId = "string",
    BackupState = "CREATE_IN_PROGRESS"|"READY"|"DELETED"|"PENDING_DELETION",
    ClusterId = "string",
    CreateTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    CopyTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    NeverExpires = TRUE|FALSE,
    SourceRegion = "string",
    SourceBackup = "string",
    SourceCluster = "string",
    DeleteTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    TagList = list(
      list(
        Key = "string",
        Value = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Restores a specified AWS CloudHSM backup that is in the
\code{PENDING_DELETION} state. For mor information on deleting a backup, see
\code{\link[=cloudhsmv2_delete_backup]{delete_backup}}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$restore_backup(
  BackupId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
