% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_tag_resource}
\alias{cloudhsmv2_tag_resource}
\title{Adds or overwrites one or more tags for the specified AWS CloudHSM
cluster}
\usage{
cloudhsmv2_tag_resource(ResourceId, TagList)
}
\arguments{
\item{ResourceId}{[required] The cluster identifier (ID) for the cluster that you are tagging. To
find the cluster ID, use
\code{\link[=cloudhsmv2_describe_clusters]{describe_clusters}}.}

\item{TagList}{[required] A list of one or more tags.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Adds or overwrites one or more tags for the specified AWS CloudHSM
cluster.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceId = "string",
  TagList = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
