% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_untag_resource}
\alias{cloudhsmv2_untag_resource}
\title{Removes the specified tag or tags from the specified AWS CloudHSM
cluster}
\usage{
cloudhsmv2_untag_resource(ResourceId, TagKeyList)
}
\arguments{
\item{ResourceId}{[required] The cluster identifier (ID) for the cluster whose tags you are removing.
To find the cluster ID, use
\code{\link[=cloudhsmv2_describe_clusters]{describe_clusters}}.}

\item{TagKeyList}{[required] A list of one or more tag keys for the tags that you are removing.
Specify only the tag keys, not the tag values.}
}
\value{
An empty list.
}
\description{
Removes the specified tag or tags from the specified AWS CloudHSM
cluster.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$untag_resource(
  ResourceId = "string",
  TagKeyList = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
