% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_delete_identities}
\alias{cognitoidentity_delete_identities}
\title{Deletes identities from an identity pool}
\usage{
cognitoidentity_delete_identities(IdentityIdsToDelete)
}
\arguments{
\item{IdentityIdsToDelete}{[required] A list of 1-60 identities that you want to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UnprocessedIdentityIds = list(
    list(
      IdentityId = "string",
      ErrorCode = "AccessDenied"|"InternalServerError"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes identities from an identity pool. You can specify a list of 1-60
identities that you want to delete.

You must use AWS Developer credentials to call this API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_identities(
  IdentityIdsToDelete = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
