% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_get_identity_pool_roles}
\alias{cognitoidentity_get_identity_pool_roles}
\title{Gets the roles for an identity pool}
\usage{
cognitoidentity_get_identity_pool_roles(IdentityPoolId)
}
\arguments{
\item{IdentityPoolId}{[required] An identity pool ID in the format REGION:GUID.}
}
\description{
Gets the roles for an identity pool.

You must use AWS Developer credentials to call this API.
}
\section{Request syntax}{
\preformatted{svc$get_identity_pool_roles(
  IdentityPoolId = "string"
)
}
}

\keyword{internal}
