% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_get_principal_tag_attribute_map}
\alias{cognitoidentity_get_principal_tag_attribute_map}
\title{Use GetPrincipalTagAttributeMap to list all mappings between
PrincipalTags and user attributes}
\usage{
cognitoidentity_get_principal_tag_attribute_map(IdentityPoolId,
  IdentityProviderName)
}
\arguments{
\item{IdentityPoolId}{[required] You can use this operation to get the ID of the Identity Pool you setup
attribute mappings for.}

\item{IdentityProviderName}{[required] You can use this operation to get the provider name.}
}
\value{
A list with the following syntax:\preformatted{list(
  IdentityPoolId = "string",
  IdentityProviderName = "string",
  UseDefaults = TRUE|FALSE,
  PrincipalTags = list(
    "string"
  )
)
}
}
\description{
Use
\code{\link[=cognitoidentity_get_principal_tag_attribute_map]{get_principal_tag_attribute_map}}
to list all mappings between \code{PrincipalTags} and user attributes.
}
\section{Request syntax}{
\preformatted{svc$get_principal_tag_attribute_map(
  IdentityPoolId = "string",
  IdentityProviderName = "string"
)
}
}

\keyword{internal}
