% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_list_identities}
\alias{cognitoidentity_list_identities}
\title{Lists the identities in an identity pool}
\usage{
cognitoidentity_list_identities(IdentityPoolId, MaxResults, NextToken,
  HideDisabled)
}
\arguments{
\item{IdentityPoolId}{[required] An identity pool ID in the format REGION:GUID.}

\item{MaxResults}{[required] The maximum number of identities to return.}

\item{NextToken}{A pagination token.}

\item{HideDisabled}{An optional boolean parameter that allows you to hide disabled
identities. If omitted, the ListIdentities API will include disabled
identities in the response.}
}
\value{
A list with the following syntax:\preformatted{list(
  IdentityPoolId = "string",
  Identities = list(
    list(
      IdentityId = "string",
      Logins = list(
        "string"
      ),
      CreationDate = as.POSIXct(
        "2015-01-01"
      ),
      LastModifiedDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists the identities in an identity pool.

You must use AWS Developer credentials to call this API.
}
\section{Request syntax}{
\preformatted{svc$list_identities(
  IdentityPoolId = "string",
  MaxResults = 123,
  NextToken = "string",
  HideDisabled = TRUE|FALSE
)
}
}

\keyword{internal}
