% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_set_principal_tag_attribute_map}
\alias{cognitoidentity_set_principal_tag_attribute_map}
\title{You can use this operation to use default (username and clientID)
attribute or custom attribute mappings}
\usage{
cognitoidentity_set_principal_tag_attribute_map(IdentityPoolId,
  IdentityProviderName, UseDefaults, PrincipalTags)
}
\arguments{
\item{IdentityPoolId}{[required] The ID of the Identity Pool you want to set attribute mappings for.}

\item{IdentityProviderName}{[required] The provider name you want to use for attribute mappings.}

\item{UseDefaults}{You can use this operation to use default (username and clientID)
attribute mappings.}

\item{PrincipalTags}{You can use this operation to add principal tags.}
}
\value{
A list with the following syntax:\preformatted{list(
  IdentityPoolId = "string",
  IdentityProviderName = "string",
  UseDefaults = TRUE|FALSE,
  PrincipalTags = list(
    "string"
  )
)
}
}
\description{
You can use this operation to use default (username and clientID)
attribute or custom attribute mappings.
}
\section{Request syntax}{
\preformatted{svc$set_principal_tag_attribute_map(
  IdentityPoolId = "string",
  IdentityProviderName = "string",
  UseDefaults = TRUE|FALSE,
  PrincipalTags = list(
    "string"
  )
)
}
}

\keyword{internal}
