% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_add_user_pool_client_secret}
\alias{cognitoidentityprovider_add_user_pool_client_secret}
\title{Creates a new client secret for an existing confidential user pool app
client}
\usage{
cognitoidentityprovider_add_user_pool_client_secret(
  UserPoolId,
  ClientId,
  ClientSecret = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that contains the app client.}

\item{ClientId}{[required] The ID of the app client for which you want to create a new secret.}

\item{ClientSecret}{The client secret value you want to use. If you don't provide this
parameter, Amazon Cognito generates a secure secret for you.}
}
\description{
Creates a new client secret for an existing confidential user pool app client. Supports up to 2 active secrets per app client for zero-downtime credential rotation workflows.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_add_user_pool_client_secret/} for full documentation.
}
\keyword{internal}
