% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_delete_software_token}
\alias{cognitoidentityprovider_admin_delete_software_token}
\title{Deletes a user's registered time-based one-time password (TOTP)
multi-factor authentication (MFA) factor, also known as a software token}
\usage{
cognitoidentityprovider_admin_delete_software_token(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to delete the user's software token.}

\item{Username}{[required] The name of the user that you want to query or modify. The value of this parameter is typically your user's username, but it can be any of their alias attributes. If \code{username} isn't an alias attribute in your user pool, this value must be the \code{sub} of a local user or the username of a user from a third-party IdP.}
}
\description{
Deletes a user's registered time-based one-time password (TOTP) multi-factor authentication (MFA) factor, also known as a software token. After this operation, the user can no longer sign in with TOTP MFA, and can register a new TOTP factor with \code{\link[=cognitoidentityprovider_associate_software_token]{associate_software_token}}. Use this operation when a user loses access to their TOTP-generating device, for example, a lost or reset phone, and needs to register a new one.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_delete_software_token/} for full documentation.
}
\keyword{internal}
