% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_get_user_auth_factors}
\alias{cognitoidentityprovider_admin_get_user_auth_factors}
\title{Lists the authentication options for a user in a user pool}
\usage{
cognitoidentityprovider_admin_get_user_auth_factors(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to get information about the user's authentication factors.}

\item{Username}{[required] The name of the user that you want to query or modify. The value of this parameter is typically your user's username, but it can be any of their alias attributes. If \code{username} isn't an alias attribute in your user pool, this value must be the \code{sub} of a local user or the username of a user from a third-party IdP.}
}
\description{
Lists the authentication options for a user in a user pool. Returns the following:

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_get_user_auth_factors/} for full documentation.
}
\keyword{internal}
