% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_set_user_mfa_preference}
\alias{cognitoidentityprovider_admin_set_user_mfa_preference}
\title{Sets the user's multi-factor authentication (MFA) preference}
\usage{
cognitoidentityprovider_admin_set_user_mfa_preference(SMSMfaSettings,
  SoftwareTokenMfaSettings, Username, UserPoolId)
}
\arguments{
\item{SMSMfaSettings}{The SMS text message MFA settings.}

\item{SoftwareTokenMfaSettings}{The time-based one-time password software token MFA settings.}

\item{Username}{[required] The user pool username or alias.}

\item{UserPoolId}{[required] The user pool ID.}
}
\description{
Sets the user's multi-factor authentication (MFA) preference.
}
\section{Request syntax}{
\preformatted{svc$admin_set_user_mfa_preference(
  SMSMfaSettings = list(
    Enabled = TRUE|FALSE,
    PreferredMfa = TRUE|FALSE
  ),
  SoftwareTokenMfaSettings = list(
    Enabled = TRUE|FALSE,
    PreferredMfa = TRUE|FALSE
  ),
  Username = "string",
  UserPoolId = "string"
)
}
}

\keyword{internal}
