% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_associate_software_token}
\alias{cognitoidentityprovider_associate_software_token}
\title{Begins setup of time-based one-time password (TOTP) multi-factor
authentication (MFA) for a user, with a unique private key that Amazon
Cognito generates and returns in the API response}
\usage{
cognitoidentityprovider_associate_software_token(
  AccessToken = NULL,
  Session = NULL
)
}
\arguments{
\item{AccessToken}{A valid access token that Amazon Cognito issued to the currently
signed-in user. Must include a scope claim for
\code{aws.cognito.signin.user.admin}.

You can provide either an access token or a session ID in the request.}

\item{Session}{The session identifier that maintains the state of authentication
requests and challenge responses. In
\code{\link[=cognitoidentityprovider_associate_software_token]{associate_software_token}},
this is the session ID from a successful sign-in. You can provide either
an access token or a session ID in the request.}
}
\description{
Begins setup of time-based one-time password (TOTP) multi-factor authentication (MFA) for a user, with a unique private key that Amazon Cognito generates and returns in the API response. You can authorize an \code{\link[=cognitoidentityprovider_associate_software_token]{associate_software_token}} request with either the user's access token, or a session string from a challenge response that you received from Amazon Cognito.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_associate_software_token/} for full documentation.
}
\keyword{internal}
