% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_create_terms}
\alias{cognitoidentityprovider_create_terms}
\title{Creates terms documents for the requested app client}
\usage{
cognitoidentityprovider_create_terms(
  UserPoolId,
  ClientId,
  TermsName,
  TermsSource,
  Enforcement,
  Links = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to create terms documents.}

\item{ClientId}{[required] The ID of the app client where you want to create terms documents. Must
be an app client in the requested user pool.}

\item{TermsName}{[required] A friendly name for the document that you want to create in the current
request. Must begin with \code{terms-of-use} or \code{privacy-policy} as
identification of the document type. Provide URLs for both
\code{terms-of-use} and \code{privacy-policy} in separate requests.}

\item{TermsSource}{[required] This parameter is reserved for future use and currently accepts only one
value.}

\item{Enforcement}{[required] This parameter is reserved for future use and currently accepts only one
value.}

\item{Links}{A map of URLs to languages. For each localized language that will view
the requested \code{TermsName}, assign a URL. A selection of
\code{cognito:default} displays for all languages that don't have a
language-specific URL.

For example,
\verb{"cognito:default": "https://terms.example.com", "cognito:spanish": "https://terms.example.com/es"}.}
}
\description{
Creates terms documents for the requested app client. When Terms and conditions and Privacy policy documents are configured, the app client displays links to them in the sign-up page of managed login for the app client.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_create_terms/} for full documentation.
}
\keyword{internal}
