% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_create_user_import_job}
\alias{cognitoidentityprovider_create_user_import_job}
\title{Creates a user import job}
\usage{
cognitoidentityprovider_create_user_import_job(
  JobName,
  UserPoolId,
  CloudWatchLogsRoleArn,
  PasswordHashingAlgorithm = NULL
)
}
\arguments{
\item{JobName}{[required] A friendly name for the user import job.}

\item{UserPoolId}{[required] The ID of the user pool that you want to import users into.}

\item{CloudWatchLogsRoleArn}{[required] You must specify an IAM role that has permission to log import-job results to Amazon CloudWatch Logs. This parameter is the ARN of that role.}

\item{PasswordHashingAlgorithm}{The password hashing algorithm used to generate the hashes in the CSV file for this import job.

Valid values: \code{BCRYPT} | \code{SCRYPT} | \code{ARGON2ID} | \code{PBKDF2_SHA256}}
}
\description{
Creates a user import job. You can import users into user pools from a comma-separated values (CSV) file without adding Amazon Cognito MAU costs to your Amazon Web Services bill.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_create_user_import_job/} for full documentation.
}
\keyword{internal}
