% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_create_user_pool_replica}
\alias{cognitoidentityprovider_create_user_pool_replica}
\title{Creates a replica of an existing user pool in a specified Amazon Web
Services Region}
\usage{
cognitoidentityprovider_create_user_pool_replica(
  UserPoolId,
  RegionName,
  UserPoolTags = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool to replicate.}

\item{RegionName}{[required] The Amazon Web Services Region where you want to create the replica user pool.}

\item{UserPoolTags}{A map of tags to assign to the replica user pool. Each tag consists of a key and an optional value, both of which you define. You can maintain tags independently on replica user pools.}
}
\description{
Creates a replica of an existing user pool in a specified Amazon Web Services Region. The replica enables multi-region replication for high availability and disaster recovery. To create a replica, you must have permissions to create user pools in the target Region.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_create_user_pool_replica/} for full documentation.
}
\keyword{internal}
