% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_managed_login_branding}
\alias{cognitoidentityprovider_delete_managed_login_branding}
\title{Deletes a managed login branding style}
\usage{
cognitoidentityprovider_delete_managed_login_branding(
  ManagedLoginBrandingId,
  UserPoolId
)
}
\arguments{
\item{ManagedLoginBrandingId}{[required] The ID of the managed login branding style that you want to delete.}

\item{UserPoolId}{[required] The ID of the user pool that contains the managed login branding style
that you want to delete.}
}
\description{
Deletes a managed login branding style. When you delete a style, you delete the branding association for an app client. When an app client doesn't have a style assigned, your managed login pages for that app client are nonfunctional until you create a new style or switch the domain branding version.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_delete_managed_login_branding/} for full documentation.
}
\keyword{internal}
