% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_terms}
\alias{cognitoidentityprovider_delete_terms}
\title{Deletes the terms documents with the requested ID from your app client}
\usage{
cognitoidentityprovider_delete_terms(TermsId, UserPoolId)
}
\arguments{
\item{TermsId}{[required] The ID of the terms documents that you want to delete.}

\item{UserPoolId}{[required] The ID of the user pool that contains the terms documents that you want
to delete.}
}
\description{
Deletes the terms documents with the requested ID from your app client.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_delete_terms/} for full documentation.
}
\keyword{internal}
