% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_user_pool}
\alias{cognitoidentityprovider_delete_user_pool}
\title{Deletes a user pool}
\usage{
cognitoidentityprovider_delete_user_pool(UserPoolId)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that you want to delete.}
}
\description{
Deletes a user pool. After you delete a user pool, users can no longer sign in to any associated applications.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_delete_user_pool/} for full documentation.
}
\keyword{internal}
