% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_user_pool_client_secret}
\alias{cognitoidentityprovider_delete_user_pool_client_secret}
\title{Deletes a specific client secret from a user pool app client}
\usage{
cognitoidentityprovider_delete_user_pool_client_secret(
  UserPoolId,
  ClientId,
  ClientSecretId
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that contains the app client.}

\item{ClientId}{[required] The ID of the app client from which you want to delete the secret.}

\item{ClientSecretId}{[required] The unique identifier of the client secret you want to delete.}
}
\description{
Deletes a specific client secret from a user pool app client. You cannot delete the last remaining secret for an app client.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_delete_user_pool_client_secret/} for full documentation.
}
\keyword{internal}
