% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_user_pool_replica}
\alias{cognitoidentityprovider_delete_user_pool_replica}
\title{Deletes a secondary replica user pool}
\usage{
cognitoidentityprovider_delete_user_pool_replica(UserPoolId, RegionName)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that contains the replica to delete.}

\item{RegionName}{[required] The Amazon Web Services Region of the replica to delete.}
}
\description{
Deletes a secondary replica user pool. You can only delete replicas that are in the INACTIVE status. This operation must be called from the primary Region.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_delete_user_pool_replica/} for full documentation.
}
\keyword{internal}
