% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_managed_login_branding_by_client}
\alias{cognitoidentityprovider_describe_managed_login_branding_by_client}
\title{Given the ID of a user pool app client, returns detailed information
about the style assigned to the app client}
\usage{
cognitoidentityprovider_describe_managed_login_branding_by_client(
  UserPoolId,
  ClientId,
  ReturnMergedResources = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that contains the app client where you want more
information about the managed login branding style.}

\item{ClientId}{[required] The app client that's assigned to the branding style that you want more
information about.}

\item{ReturnMergedResources}{When \code{true}, returns values for branding options that are unchanged from
Amazon Cognito defaults. When \code{false} or when you omit this parameter,
returns only values that you customized in your branding style.}
}
\description{
Given the ID of a user pool app client, returns detailed information about the style assigned to the app client.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_describe_managed_login_branding_by_client/} for full documentation.
}
\keyword{internal}
