% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_managed_login_branding}
\alias{cognitoidentityprovider_describe_managed_login_branding}
\title{Given the ID of a managed login branding style, returns detailed
information about the style}
\usage{
cognitoidentityprovider_describe_managed_login_branding(
  UserPoolId,
  ManagedLoginBrandingId,
  ReturnMergedResources = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that contains the managed login branding style
that you want to get information about.}

\item{ManagedLoginBrandingId}{[required] The ID of the managed login branding style that you want to get more
information about.}

\item{ReturnMergedResources}{When \code{true}, returns values for branding options that are unchanged from
Amazon Cognito defaults. When \code{false} or when you omit this parameter,
returns only values that you customized in your branding style.}
}
\description{
Given the ID of a managed login branding style, returns detailed information about the style.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_describe_managed_login_branding/} for full documentation.
}
\keyword{internal}
