% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_resource_server}
\alias{cognitoidentityprovider_describe_resource_server}
\title{Describes a resource server}
\usage{
cognitoidentityprovider_describe_resource_server(UserPoolId, Identifier)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool that hosts the resource server.}

\item{Identifier}{[required] The identifier for the resource server}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ResourceServer = list(
    UserPoolId = "string",
    Identifier = "string",
    Name = "string",
    Scopes = list(
      list(
        ScopeName = "string",
        ScopeDescription = "string"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes a resource server.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_resource_server(
  UserPoolId = "string",
  Identifier = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
