% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_terms}
\alias{cognitoidentityprovider_describe_terms}
\title{Returns details for the requested terms documents ID}
\usage{
cognitoidentityprovider_describe_terms(TermsId, UserPoolId)
}
\arguments{
\item{TermsId}{[required] The ID of the terms documents that you want to describe.}

\item{UserPoolId}{[required] The ID of the user pool that contains the terms documents that you want
to describe.}
}
\description{
Returns details for the requested terms documents ID. For more information, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-pools-managed-login.html#managed-login-terms-documents}{Terms documents}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_describe_terms/} for full documentation.
}
\keyword{internal}
