% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_terms_by_client}
\alias{cognitoidentityprovider_describe_terms_by_client}
\title{Returns details for the terms documents that are associated with an app
client, identified by the app client ID, user pool ID, and terms name}
\usage{
cognitoidentityprovider_describe_terms_by_client(
  ClientId,
  UserPoolId,
  TermsName
)
}
\arguments{
\item{ClientId}{[required] The ID of the app client that the terms documents are associated with.}

\item{UserPoolId}{[required] The ID of the user pool that contains the terms documents that you want to describe.}

\item{TermsName}{[required] The name of the terms documents that you want to describe.}
}
\description{
Returns details for the terms documents that are associated with an app client, identified by the app client ID, user pool ID, and terms name. For more information, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-pools-managed-login.html#managed-login-terms-documents}{Terms documents}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_describe_terms_by_client/} for full documentation.
}
\keyword{internal}
