% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_user_import_job}
\alias{cognitoidentityprovider_describe_user_import_job}
\title{Describes a user import job}
\usage{
cognitoidentityprovider_describe_user_import_job(UserPoolId, JobId)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that's associated with the import job.}

\item{JobId}{[required] The Id of the user import job that you want to describe.}
}
\description{
Describes a user import job. For more information about user CSV import, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-pools-using-import-tool.html}{Importing users from a CSV file}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_describe_user_import_job/} for full documentation.
}
\keyword{internal}
