% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_provisioned_limit}
\alias{cognitoidentityprovider_get_provisioned_limit}
\title{Returns the current provisioned limit for a specific API category}
\usage{
cognitoidentityprovider_get_provisioned_limit(LimitDefinition)
}
\arguments{
\item{LimitDefinition}{[required] The limit to retrieve. Specify the limit class and the attributes that identify the limit.}
}
\description{
Returns the current provisioned limit for a specific API category.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_get_provisioned_limit/} for full documentation.
}
\keyword{internal}
