% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_signing_certificate}
\alias{cognitoidentityprovider_get_signing_certificate}
\title{This method takes a user pool ID, and returns the signing certificate}
\usage{
cognitoidentityprovider_get_signing_certificate(UserPoolId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}
}
\description{
This method takes a user pool ID, and returns the signing certificate.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/get_signing_certificate.html} for full documentation.
}
\keyword{internal}
