% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_user_auth_factors}
\alias{cognitoidentityprovider_get_user_auth_factors}
\title{Lists the authentication options for the currently signed-in user}
\usage{
cognitoidentityprovider_get_user_auth_factors(AccessToken)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user whose
authentication factors you want to view.}
}
\description{
Lists the authentication options for the currently signed-in user. Returns the following:

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_get_user_auth_factors/} for full documentation.
}
\keyword{internal}
