% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_user_pool_mfa_config}
\alias{cognitoidentityprovider_get_user_pool_mfa_config}
\title{Gets the user pool multi-factor authentication (MFA) configuration}
\usage{
cognitoidentityprovider_get_user_pool_mfa_config(UserPoolId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SmsMfaConfiguration = list(
    SmsAuthenticationMessage = "string",
    SmsConfiguration = list(
      SnsCallerArn = "string",
      ExternalId = "string"
    )
  ),
  SoftwareTokenMfaConfiguration = list(
    Enabled = TRUE|FALSE
  ),
  MfaConfiguration = "OFF"|"ON"|"OPTIONAL"
)
}\if{html}{\out{</div>}}
}
\description{
Gets the user pool multi-factor authentication (MFA) configuration.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_user_pool_mfa_config(
  UserPoolId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
