% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_global_sign_out}
\alias{cognitoidentityprovider_global_sign_out}
\title{Signs out users from all devices}
\usage{
cognitoidentityprovider_global_sign_out(AccessToken)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user who you want
to sign out.}
}
\description{
Signs out users from all devices. It also invalidates all refresh tokens that Amazon Cognito has issued to a user. A user can still use a hosted UI cookie to retrieve new tokens for the duration of the 1-hour cookie validity period.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/global_sign_out.html} for full documentation.
}
\keyword{internal}
