% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_devices}
\alias{cognitoidentityprovider_list_devices}
\title{Lists the devices that Amazon Cognito has registered to the currently
signed-in user}
\usage{
cognitoidentityprovider_list_devices(
  AccessToken,
  Limit = NULL,
  PaginationToken = NULL
)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the currently
signed-in user. Must include a scope claim for
\code{aws.cognito.signin.user.admin}.}

\item{Limit}{The maximum number of devices that you want Amazon Cognito to return in
the response.}

\item{PaginationToken}{This API operation returns a limited number of results. The pagination
token is an identifier that you can present in an additional API request
with the same parameters. When you include the pagination token, Amazon
Cognito returns the next set of items after the current list. Subsequent
requests return a new pagination token. By use of this token, you can
paginate through the full list of items.}
}
\description{
Lists the devices that Amazon Cognito has registered to the currently signed-in user. For more information about device authentication, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/amazon-cognito-user-pools-device-tracking.html}{Working with user devices in your user pool}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_devices/} for full documentation.
}
\keyword{internal}
