% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_tags_for_resource}
\alias{cognitoidentityprovider_list_tags_for_resource}
\title{Lists the tags that are assigned to an Amazon Cognito user pool}
\usage{
cognitoidentityprovider_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the user pool that the tags are
assigned to.}
}
\description{
Lists the tags that are assigned to an Amazon Cognito user pool.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
