% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_terms}
\alias{cognitoidentityprovider_list_terms}
\title{Returns details about all terms documents for the requested user pool}
\usage{
cognitoidentityprovider_list_terms(
  UserPoolId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to list terms documents.}

\item{MaxResults}{The maximum number of terms documents that you want Amazon Cognito to
return in the response.}

\item{NextToken}{This API operation returns a limited number of results. The pagination
token is an identifier that you can present in an additional API request
with the same parameters. When you include the pagination token, Amazon
Cognito returns the next set of items after the current list. Subsequent
requests return a new pagination token. By use of this token, you can
paginate through the full list of items.}
}
\description{
Returns details about all terms documents for the requested user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_terms/} for full documentation.
}
\keyword{internal}
