% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_user_pool_client_secrets}
\alias{cognitoidentityprovider_list_user_pool_client_secrets}
\title{Lists all client secrets associated with a user pool app client}
\usage{
cognitoidentityprovider_list_user_pool_client_secrets(
  UserPoolId,
  ClientId,
  NextToken = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that contains the app client.}

\item{ClientId}{[required] The ID of the app client whose secrets you want to list.}

\item{NextToken}{This API operation returns a limited number of results. The pagination
token is an identifier that you can present in an additional API request
with the same parameters. When you include the pagination token, Amazon
Cognito returns the next set of items after the current list. Subsequent
requests return a new pagination token. By use of this token, you can
paginate through the full list of items.}
}
\description{
Lists all client secrets associated with a user pool app client. Returns metadata about the secrets. The response does not include pagination tokens as there are only 2 secrets at any given time and we return both with every ListUserPoolClientSecrets call. For security reasons, the response never reveals the actual secret value in ClientSecretValue.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_user_pool_client_secrets/} for full documentation.
}
\keyword{internal}
