% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_user_pool_clients}
\alias{cognitoidentityprovider_list_user_pool_clients}
\title{Given a user pool ID, lists app clients}
\usage{
cognitoidentityprovider_list_user_pool_clients(
  UserPoolId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to list user pool clients.}

\item{MaxResults}{The maximum number of app clients that you want Amazon Cognito to return
in the response.}

\item{NextToken}{This API operation returns a limited number of results. The pagination
token is an identifier that you can present in an additional API request
with the same parameters. When you include the pagination token, Amazon
Cognito returns the next set of items after the current list. Subsequent
requests return a new pagination token. By use of this token, you can
paginate through the full list of items.}
}
\description{
Given a user pool ID, lists app clients. App clients are sets of rules for the access that you want a user pool to grant to one application. For more information, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/user-pool-settings-client-apps.html}{App clients}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_user_pool_clients/} for full documentation.
}
\keyword{internal}
