% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_user_pool_replicas}
\alias{cognitoidentityprovider_list_user_pool_replicas}
\title{Lists all replicas for a user pool, including both primary and secondary
replicas}
\usage{
cognitoidentityprovider_list_user_pool_replicas(UserPoolId, NextToken = NULL)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool for which to list replicas.}

\item{NextToken}{A pagination token for retrieving the next page of results. If this parameter is omitted, the operation returns the first page of results.}
}
\description{
Lists all replicas for a user pool, including both primary and secondary replicas. We recommend using pagination to ensure that the operation returns quickly and successfully.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_user_pool_replicas/} for full documentation.
}
\keyword{internal}
