% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_web_authn_credentials}
\alias{cognitoidentityprovider_list_web_authn_credentials}
\title{Generates a list of the current user's registered passkey, or webauthN,
credentials}
\usage{
cognitoidentityprovider_list_web_authn_credentials(
  AccessToken,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user whose
registered passkeys you want to list.}

\item{NextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}

\item{MaxResults}{The maximum number of the user's passkey credentials that you want to
return.}
}
\description{
Generates a list of the current user's registered passkey, or webauthN, credentials.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_web_authn_credentials/} for full documentation.
}
\keyword{internal}
