% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_set_log_delivery_configuration}
\alias{cognitoidentityprovider_set_log_delivery_configuration}
\title{Sets up or modifies the logging configuration of a user pool}
\usage{
cognitoidentityprovider_set_log_delivery_configuration(
  UserPoolId,
  LogConfigurations
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to configure logging.}

\item{LogConfigurations}{[required] A collection of the logging configurations for a user pool.}
}
\description{
Sets up or modifies the logging configuration of a user pool. User pools can export user notification logs and advanced security features user activity logs.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_set_log_delivery_configuration/} for full documentation.
}
\keyword{internal}
