% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_set_user_settings}
\alias{cognitoidentityprovider_set_user_settings}
\title{This action is no longer supported}
\usage{
cognitoidentityprovider_set_user_settings(AccessToken, MFAOptions)
}
\arguments{
\item{AccessToken}{[required] The access token for the set user settings request.}

\item{MFAOptions}{[required] You can use this parameter only to set an SMS configuration that uses
SMS for delivery.}
}
\value{
An empty list.
}
\description{
\emph{This action is no longer supported.} You can use it to configure only
SMS MFA. You can't use it to configure TOTP software token MFA. To
configure either type of MFA, use
\code{\link[=cognitoidentityprovider_set_user_mfa_preference]{set_user_mfa_preference}}
instead.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$set_user_settings(
  AccessToken = "string",
  MFAOptions = list(
    list(
      DeliveryMedium = "SMS"|"EMAIL",
      AttributeName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
