% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_start_web_authn_registration}
\alias{cognitoidentityprovider_start_web_authn_registration}
\title{Requests credential creation options from your user pool for
registration of a passkey authenticator}
\usage{
cognitoidentityprovider_start_web_authn_registration(AccessToken)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user whose
passkey metadata you want to generate.}
}
\description{
Requests credential creation options from your user pool for registration of a passkey authenticator. Returns information about the user pool, the user profile, and authentication requirements. Users must provide this information in their request to enroll your application with their passkey provider.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_start_web_authn_registration/} for full documentation.
}
\keyword{internal}
