% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_stop_user_import_job}
\alias{cognitoidentityprovider_stop_user_import_job}
\title{Instructs your user pool to stop a running job that's importing users
from a CSV file that contains their usernames and attributes}
\usage{
cognitoidentityprovider_stop_user_import_job(UserPoolId, JobId)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that you want to stop.}

\item{JobId}{[required] The ID of a running user import job.}
}
\description{
Instructs your user pool to stop a running job that's importing users from a CSV file that contains their usernames and attributes. For more information about importing users from a CSV file, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-pools-using-import-tool.html}{Importing users from a CSV file}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_stop_user_import_job/} for full documentation.
}
\keyword{internal}
