% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_tag_resource}
\alias{cognitoidentityprovider_tag_resource}
\title{Assigns a set of tags to an Amazon Cognito user pool}
\usage{
cognitoidentityprovider_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the user pool to assign the tags to.}

\item{Tags}{[required] The tags to assign to the user pool.}
}
\description{
Assigns a set of tags to an Amazon Cognito user pool. A tag is a label that you can use to categorize and manage user pools in different ways, such as by purpose, owner, environment, or other criteria.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_tag_resource/} for full documentation.
}
\keyword{internal}
