% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_update_provisioned_limit}
\alias{cognitoidentityprovider_update_provisioned_limit}
\title{Sets the provisioned limit for a specific API category}
\usage{
cognitoidentityprovider_update_provisioned_limit(
  LimitDefinition,
  RequestedLimitValue
)
}
\arguments{
\item{LimitDefinition}{[required] The limit to update. Specify the limit class and the attributes that identify the limit.}

\item{RequestedLimitValue}{[required] The provisioned rate to set, in requests per second (RPS).}
}
\description{
Sets the provisioned limit for a specific API category. The value must be between the default limit and your account-level maximum limit in Service Quotas.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_update_provisioned_limit/} for full documentation.
}
\keyword{internal}
