% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_update_resource_server}
\alias{cognitoidentityprovider_update_resource_server}
\title{Updates the name and scopes of resource server}
\usage{
cognitoidentityprovider_update_resource_server(
  UserPoolId,
  Identifier,
  Name,
  Scopes = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{Identifier}{[required] The identifier for the resource server.}

\item{Name}{[required] The name of the resource server.}

\item{Scopes}{The scope values to be set for the resource server.}
}
\description{
Updates the name and scopes of resource server. All other fields are read-only.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_update_resource_server/} for full documentation.
}
\keyword{internal}
