% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_update_user_pool_domain}
\alias{cognitoidentityprovider_update_user_pool_domain}
\title{A user pool domain hosts managed login, an authorization server and web
server for authentication in your application}
\usage{
cognitoidentityprovider_update_user_pool_domain(
  Domain,
  UserPoolId,
  ManagedLoginVersion = NULL,
  CustomDomainConfig = NULL,
  Routing = NULL
)
}
\arguments{
\item{Domain}{[required] The name of the domain that you want to update. For custom domains, this is the fully-qualified domain name, for example \code{auth.example.com}. For prefix domains, this is the prefix alone, such as \code{myprefix}.}

\item{UserPoolId}{[required] The ID of the user pool that is associated with the domain you're updating.}

\item{ManagedLoginVersion}{A version number that indicates the state of managed login for your domain. Version \code{1} is hosted UI (classic). Version \code{2} is the newer managed login with the branding editor. For more information, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/cognito-user-pools-managed-login.html}{Managed login}.}

\item{CustomDomainConfig}{The configuration for a custom domain that hosts managed login for your application. In an \code{\link[=cognitoidentityprovider_update_user_pool_domain]{update_user_pool_domain}} request, this parameter specifies an SSL certificate for the managed login hosted webserver. The certificate must be an ACM ARN in \code{us-east-1}.

When you create a custom domain, the passkey RP ID defaults to the custom domain. If you had a prefix domain active, this will cause passkey integration for your prefix domain to stop working due to a mismatch in RP ID. To keep the prefix domain passkey integration working, you can explicitly set RP ID to the prefix domain.}

\item{Routing}{The routing configuration for the user pool domain. Specifies failover settings for multi-region deployments.}
}
\description{
A user pool domain hosts managed login, an authorization server and web server for authentication in your application. This operation updates the branding version for user pool domains between \code{1} for hosted UI (classic) and \code{2} for managed login. It also updates the SSL certificate for user pool custom domains.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_update_user_pool_domain/} for full documentation.
}
\keyword{internal}
