% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_update_user_pool_replica}
\alias{cognitoidentityprovider_update_user_pool_replica}
\title{Updates replica-specific settings for a user pool replica}
\usage{
cognitoidentityprovider_update_user_pool_replica(
  UserPoolId,
  RegionName,
  Status
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that contains the replica to update.}

\item{RegionName}{[required] The Amazon Web Services Region of the replica to update.}

\item{Status}{[required] The status to set for the replica. Valid values are ACTIVE and INACTIVE.}
}
\description{
Updates replica-specific settings for a user pool replica. You can modify the status to activate or deactivate the replica. This request can be made in both primary and secondary regions of the user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_update_user_pool_replica/} for full documentation.
}
\keyword{internal}
