% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_verify_software_token}
\alias{cognitoidentityprovider_verify_software_token}
\title{Registers the current user's time-based one-time password (TOTP)
authenticator with a code generated in their authenticator app from a
private key that's supplied by your user pool}
\usage{
cognitoidentityprovider_verify_software_token(
  AccessToken = NULL,
  Session = NULL,
  UserCode,
  FriendlyDeviceName = NULL
)
}
\arguments{
\item{AccessToken}{A valid access token that Amazon Cognito issued to the currently
signed-in user. Must include a scope claim for
\code{aws.cognito.signin.user.admin}.}

\item{Session}{The session ID from an
\code{\link[=cognitoidentityprovider_associate_software_token]{associate_software_token}}
request.}

\item{UserCode}{[required] A TOTP that the user generated in their configured authenticator app.}

\item{FriendlyDeviceName}{A friendly name for the device that's running the TOTP authenticator.}
}
\description{
Registers the current user's time-based one-time password (TOTP) authenticator with a code generated in their authenticator app from a private key that's supplied by your user pool. Marks the user's software token MFA status as "verified" if successful. The request takes an access token or a session string, but not both.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_verify_software_token/} for full documentation.
}
\keyword{internal}
