% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitosync_operations.R
\name{cognitosync_describe_identity_pool_usage}
\alias{cognitosync_describe_identity_pool_usage}
\title{Gets usage details (for example, data storage) about a particular
identity pool}
\usage{
cognitosync_describe_identity_pool_usage(IdentityPoolId)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example,
us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon
Cognito. GUID generation is unique within a region.}
}
\value{
A list with the following syntax:\preformatted{list(
  IdentityPoolUsage = list(
    IdentityPoolId = "string",
    SyncSessionsCount = 123,
    DataStorage = 123,
    LastModifiedDate = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Gets usage details (for example, data storage) about a particular
identity pool.

This API can only be called with developer credentials. You cannot call
this API with the temporary user credentials provided by Cognito
Identity.
}
\section{Request syntax}{
\preformatted{svc$describe_identity_pool_usage(
  IdentityPoolId = "string"
)
}
}

\keyword{internal}
