% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitosync_operations.R
\name{cognitosync_get_bulk_publish_details}
\alias{cognitosync_get_bulk_publish_details}
\title{Get the status of the last BulkPublish operation for an identity pool}
\usage{
cognitosync_get_bulk_publish_details(IdentityPoolId)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example,
us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon
Cognito. GUID generation is unique within a region.}
}
\value{
A list with the following syntax:\preformatted{list(
  IdentityPoolId = "string",
  BulkPublishStartTime = as.POSIXct(
    "2015-01-01"
  ),
  BulkPublishCompleteTime = as.POSIXct(
    "2015-01-01"
  ),
  BulkPublishStatus = "NOT_STARTED"|"IN_PROGRESS"|"FAILED"|"SUCCEEDED",
  FailureMessage = "string"
)
}
}
\description{
Get the status of the last BulkPublish operation for an identity pool.

This API can only be called with developer credentials. You cannot call
this API with the temporary user credentials provided by Cognito
Identity.
}
\section{Request syntax}{
\preformatted{svc$get_bulk_publish_details(
  IdentityPoolId = "string"
)
}
}

\keyword{internal}
