% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitosync_operations.R
\name{cognitosync_subscribe_to_dataset}
\alias{cognitosync_subscribe_to_dataset}
\title{Subscribes to receive notifications when a dataset is modified by
another device}
\usage{
cognitosync_subscribe_to_dataset(
  IdentityPoolId,
  IdentityId,
  DatasetName,
  DeviceId
)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example,
us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon
Cognito. The ID of the pool to which the identity belongs.}

\item{IdentityId}{[required] Unique ID for this identity.}

\item{DatasetName}{[required] The name of the dataset to subcribe to.}

\item{DeviceId}{[required] The unique ID generated for this device by Cognito.}
}
\description{
Subscribes to receive notifications when a dataset is modified by another device.

See \url{https://www.paws-r-sdk.com/docs/cognitosync_subscribe_to_dataset/} for full documentation.
}
\keyword{internal}
