% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitosync_operations.R
\name{cognitosync_unsubscribe_from_dataset}
\alias{cognitosync_unsubscribe_from_dataset}
\title{Unsubscribes from receiving notifications when a dataset is modified by
another device}
\usage{
cognitosync_unsubscribe_from_dataset(IdentityPoolId, IdentityId,
  DatasetName, DeviceId)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example,
us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon
Cognito. The ID of the pool to which this identity belongs.}

\item{IdentityId}{[required] Unique ID for this identity.}

\item{DatasetName}{[required] The name of the dataset from which to unsubcribe.}

\item{DeviceId}{[required] The unique ID generated for this device by Cognito.}
}
\value{
An empty list.
}
\description{
Unsubscribes from receiving notifications when a dataset is modified by
another device.

This API can only be called with temporary credentials provided by
Cognito Identity. You cannot call this API with developer credentials.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$unsubscribe_from_dataset(
  IdentityPoolId = "string",
  IdentityId = "string",
  DatasetName = "string",
  DeviceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
