% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_accept_invitation}
\alias{detective_accept_invitation}
\title{Accepts an invitation for the member account to contribute data to a
behavior graph}
\usage{
detective_accept_invitation(GraphArn)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph that the member account is accepting the
invitation for.

The member account status in the behavior graph must be \code{INVITED}.}
}
\description{
Accepts an invitation for the member account to contribute data to a behavior graph. This operation can only be called by an invited member account.

See \url{https://www.paws-r-sdk.com/docs/detective_accept_invitation/} for full documentation.
}
\keyword{internal}
