% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_batch_get_graph_member_datasources}
\alias{detective_batch_get_graph_member_datasources}
\title{Gets data source package information for the behavior graph}
\usage{
detective_batch_get_graph_member_datasources(GraphArn, AccountIds)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph.}

\item{AccountIds}{[required] The list of Amazon Web Services accounts to get data source package
information on.}
}
\description{
Gets data source package information for the behavior graph.

See \url{https://www.paws-r-sdk.com/docs/detective_batch_get_graph_member_datasources/} for full documentation.
}
\keyword{internal}
