% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_batch_get_membership_datasources}
\alias{detective_batch_get_membership_datasources}
\title{Gets information on the data source package history for an account}
\usage{
detective_batch_get_membership_datasources(GraphArns)
}
\arguments{
\item{GraphArns}{[required] The ARN of the behavior graph.}
}
\description{
Gets information on the data source package history for an account.

See \url{https://paws-r.github.io/docs/detective/batch_get_membership_datasources.html} for full documentation.
}
\keyword{internal}
