% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_delete_graph}
\alias{detective_delete_graph}
\title{Disables the specified behavior graph and queues it to be deleted}
\usage{
detective_delete_graph(GraphArn)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph to disable.}
}
\description{
Disables the specified behavior graph and queues it to be deleted. This operation removes the behavior graph from each member account's list of behavior graphs.

See \url{https://www.paws-r-sdk.com/docs/detective_delete_graph/} for full documentation.
}
\keyword{internal}
