% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_delete_members}
\alias{detective_delete_members}
\title{Removes the specified member accounts from the behavior graph}
\usage{
detective_delete_members(GraphArn, AccountIds)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph to remove members from.}

\item{AccountIds}{[required] The list of Amazon Web Services account identifiers for the member
accounts to remove from the behavior graph. You can remove up to 50
member accounts at a time.}
}
\description{
Removes the specified member accounts from the behavior graph. The removed accounts no longer contribute data to the behavior graph. This operation can only be called by the administrator account for the behavior graph.

See \url{https://www.paws-r-sdk.com/docs/detective_delete_members/} for full documentation.
}
\keyword{internal}
