% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_describe_organization_configuration}
\alias{detective_describe_organization_configuration}
\title{Returns information about the configuration for the organization
behavior graph}
\usage{
detective_describe_organization_configuration(GraphArn)
}
\arguments{
\item{GraphArn}{[required] The ARN of the organization behavior graph.}
}
\description{
Returns information about the configuration for the organization behavior graph. Currently indicates whether to automatically enable new organization accounts as member accounts.

See \url{https://www.paws-r-sdk.com/docs/detective_describe_organization_configuration/} for full documentation.
}
\keyword{internal}
