% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_disassociate_membership}
\alias{detective_disassociate_membership}
\title{Removes the member account from the specified behavior graph}
\usage{
detective_disassociate_membership(GraphArn)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph to remove the member account from.

The member account's member status in the behavior graph must be
\code{ENABLED}.}
}
\description{
Removes the member account from the specified behavior graph. This operation can only be called by an invited member account that has the \code{ENABLED} status.

See \url{https://www.paws-r-sdk.com/docs/detective_disassociate_membership/} for full documentation.
}
\keyword{internal}
