% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_enable_organization_admin_account}
\alias{detective_enable_organization_admin_account}
\title{Designates the Detective administrator account for the organization in
the current Region}
\usage{
detective_enable_organization_admin_account(AccountId)
}
\arguments{
\item{AccountId}{[required] The Amazon Web Services account identifier of the account to designate
as the Detective administrator account for the organization.}
}
\description{
Designates the Detective administrator account for the organization in the current Region.

See \url{https://www.paws-r-sdk.com/docs/detective_enable_organization_admin_account/} for full documentation.
}
\keyword{internal}
