% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_get_investigation}
\alias{detective_get_investigation}
\title{Returns the investigation results of an investigation for a behavior
graph}
\usage{
detective_get_investigation(GraphArn, InvestigationId)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph.}

\item{InvestigationId}{[required] The investigation ID of the investigation report.}
}
\description{
Returns the investigation results of an investigation for a behavior graph.

See \url{https://www.paws-r-sdk.com/docs/detective_get_investigation/} for full documentation.
}
\keyword{internal}
