% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_list_datasource_packages}
\alias{detective_list_datasource_packages}
\title{Lists data source packages in the behavior graph}
\usage{
detective_list_datasource_packages(
  GraphArn,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph.}

\item{NextToken}{For requests to get the next page of results, the pagination token that
was returned with the previous set of results. The initial request does
not include a pagination token.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Lists data source packages in the behavior graph.

See \url{https://www.paws-r-sdk.com/docs/detective_list_datasource_packages/} for full documentation.
}
\keyword{internal}
