% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_list_graphs}
\alias{detective_list_graphs}
\title{Returns the list of behavior graphs that the calling account is an
administrator account of}
\usage{
detective_list_graphs(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{For requests to get the next page of results, the pagination token that
was returned with the previous set of results. The initial request does
not include a pagination token.}

\item{MaxResults}{The maximum number of graphs to return at a time. The total must be less
than the overall limit on the number of results to return, which is
currently 200.}
}
\description{
Returns the list of behavior graphs that the calling account is an administrator account of. This operation can only be called by an administrator account.

See \url{https://www.paws-r-sdk.com/docs/detective_list_graphs/} for full documentation.
}
\keyword{internal}
