% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_list_investigations}
\alias{detective_list_investigations}
\title{List all Investigations}
\usage{
detective_list_investigations(
  GraphArn,
  NextToken = NULL,
  MaxResults = NULL,
  FilterCriteria = NULL,
  SortCriteria = NULL
)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph.}

\item{NextToken}{List if there are more results available. The value of nextToken is a
unique pagination token for each page. Repeat the call using the
returned token to retrieve the next page. Keep all other arguments
unchanged.

Each pagination token expires after 24 hours. Using an expired
pagination token will return a Validation Exception error.}

\item{MaxResults}{List the maximum number of investigations in a page.}

\item{FilterCriteria}{Filter the investigation results based on a criteria.}

\item{SortCriteria}{Sorts the investigation results based on a criteria.}
}
\description{
List all Investigations.

See \url{https://www.paws-r-sdk.com/docs/detective_list_investigations/} for full documentation.
}
\keyword{internal}
