% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_list_members}
\alias{detective_list_members}
\title{Retrieves the list of member accounts for a behavior graph}
\usage{
detective_list_members(GraphArn, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph for which to retrieve the list of member
accounts.}

\item{NextToken}{For requests to retrieve the next page of member account results, the
pagination token that was returned with the previous page of results.
The initial request does not include a pagination token.}

\item{MaxResults}{The maximum number of member accounts to include in the response. The
total must be less than the overall limit on the number of results to
return, which is currently 200.}
}
\description{
Retrieves the list of member accounts for a behavior graph.

See \url{https://www.paws-r-sdk.com/docs/detective_list_members/} for full documentation.
}
\keyword{internal}
