% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_list_organization_admin_accounts}
\alias{detective_list_organization_admin_accounts}
\title{Returns information about the Detective administrator account for an
organization}
\usage{
detective_list_organization_admin_accounts(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{For requests to get the next page of results, the pagination token that
was returned with the previous set of results. The initial request does
not include a pagination token.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Returns information about the Detective administrator account for an organization. Can only be called by the organization management account.

See \url{https://www.paws-r-sdk.com/docs/detective_list_organization_admin_accounts/} for full documentation.
}
\keyword{internal}
