% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_reject_invitation}
\alias{detective_reject_invitation}
\title{Rejects an invitation to contribute the account data to a behavior graph}
\usage{
detective_reject_invitation(GraphArn)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph to reject the invitation to.

The member account's current member status in the behavior graph must be
\code{INVITED}.}
}
\description{
Rejects an invitation to contribute the account data to a behavior graph. This operation must be called by an invited member account that has the \code{INVITED} status.

See \url{https://www.paws-r-sdk.com/docs/detective_reject_invitation/} for full documentation.
}
\keyword{internal}
