% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_untag_resource}
\alias{detective_untag_resource}
\title{Removes tags from a behavior graph}
\usage{
detective_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the behavior graph to remove the tags from.}

\item{TagKeys}{[required] The tag keys of the tags to remove from the behavior graph. You can
remove up to 50 tags at a time.}
}
\description{
Removes tags from a behavior graph.

See \url{https://www.paws-r-sdk.com/docs/detective_untag_resource/} for full documentation.
}
\keyword{internal}
