% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_update_datasource_packages}
\alias{detective_update_datasource_packages}
\title{Starts a data source packages for the behavior graph}
\usage{
detective_update_datasource_packages(GraphArn, DatasourcePackages)
}
\arguments{
\item{GraphArn}{[required] The ARN of the behavior graph.}

\item{DatasourcePackages}{[required] The data source package start for the behavior graph.}
}
\description{
Starts a data source packages for the behavior graph.

See \url{https://www.paws-r-sdk.com/docs/detective_update_datasource_packages/} for full documentation.
}
\keyword{internal}
