% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_update_investigation_state}
\alias{detective_update_investigation_state}
\title{Updates the state of an investigation}
\usage{
detective_update_investigation_state(GraphArn, InvestigationId, State)
}
\arguments{
\item{GraphArn}{[required] The Amazon Resource Name (ARN) of the behavior graph.}

\item{InvestigationId}{[required] The investigation ID of the investigation report.}

\item{State}{[required] The current state of the investigation. An archived investigation
indicates you have completed reviewing the investigation.}
}
\description{
Updates the state of an investigation.

See \url{https://www.paws-r-sdk.com/docs/detective_update_investigation_state/} for full documentation.
}
\keyword{internal}
