% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_update_organization_configuration}
\alias{detective_update_organization_configuration}
\title{Updates the configuration for the Organizations integration in the
current Region}
\usage{
detective_update_organization_configuration(GraphArn, AutoEnable = NULL)
}
\arguments{
\item{GraphArn}{[required] The ARN of the organization behavior graph.}

\item{AutoEnable}{Indicates whether to automatically enable new organization accounts as
member accounts in the organization behavior graph.}
}
\description{
Updates the configuration for the Organizations integration in the current Region. Can only be called by the Detective administrator account for the organization.

See \url{https://www.paws-r-sdk.com/docs/detective_update_organization_configuration/} for full documentation.
}
\keyword{internal}
