% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_create_alias}
\alias{directoryservice_create_alias}
\title{Creates an alias for a directory and assigns the alias to the directory}
\usage{
directoryservice_create_alias(DirectoryId, Alias)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory for which to create the alias.}

\item{Alias}{[required] The requested alias.

The alias must be unique amongst all aliases in Amazon Web Services.
This operation throws an \code{EntityAlreadyExistsException} error if the
alias already exists.}
}
\description{
Creates an alias for a directory and assigns the alias to the directory. The alias is used to construct the access URL for the directory, such as \verb{http://<alias>.awsapps.com}.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_create_alias/} for full documentation.
}
\keyword{internal}
