% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_create_log_subscription}
\alias{directoryservice_create_log_subscription}
\title{Creates a subscription to forward real-time Directory Service domain
controller security logs to the specified Amazon CloudWatch log group in
your Amazon Web Services account}
\usage{
directoryservice_create_log_subscription(DirectoryId, LogGroupName)
}
\arguments{
\item{DirectoryId}{[required] Identifier of the directory to which you want to subscribe and receive
real-time logs to your specified CloudWatch log group.}

\item{LogGroupName}{[required] The name of the CloudWatch log group where the real-time domain
controller logs are forwarded.}
}
\description{
Creates a subscription to forward real-time Directory Service domain controller security logs to the specified Amazon CloudWatch log group in your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_create_log_subscription/} for full documentation.
}
\keyword{internal}
