% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_delete_ad_assessment}
\alias{directoryservice_delete_ad_assessment}
\title{Deletes a directory assessment and all associated data}
\usage{
directoryservice_delete_ad_assessment(AssessmentId)
}
\arguments{
\item{AssessmentId}{[required] The unique identifier of the directory assessment to delete.}
}
\description{
Deletes a directory assessment and all associated data. This operation permanently removes the assessment results, validation reports, and configuration information.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_delete_ad_assessment/} for full documentation.
}
\keyword{internal}
