% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_delete_directory}
\alias{directoryservice_delete_directory}
\title{Deletes an AWS Directory Service directory}
\usage{
directoryservice_delete_directory(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DirectoryId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes an AWS Directory Service directory.

Before you call \code{\link[=directoryservice_delete_directory]{delete_directory}},
ensure that all of the required permissions have been explicitly granted
through a policy. For details about what permissions are required to run
the \code{\link[=directoryservice_delete_directory]{delete_directory}} operation,
see \href{https://docs.aws.amazon.com/directoryservice/latest/admin-guide/UsingWithDS_IAM_ResourcePermissions.html}{AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_directory(
  DirectoryId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
