% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_delete_log_subscription}
\alias{directoryservice_delete_log_subscription}
\title{Deletes the specified log subscription}
\usage{
directoryservice_delete_log_subscription(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] Identifier of the directory whose log subscription you want to delete.}
}
\description{
Deletes the specified log subscription.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_delete_log_subscription/} for full documentation.
}
\keyword{internal}
