% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_delete_snapshot}
\alias{directoryservice_delete_snapshot}
\title{Deletes a directory snapshot}
\usage{
directoryservice_delete_snapshot(SnapshotId)
}
\arguments{
\item{SnapshotId}{[required] The identifier of the directory snapshot to be deleted.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SnapshotId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a directory snapshot.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_snapshot(
  SnapshotId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
