% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_delete_trust}
\alias{directoryservice_delete_trust}
\title{Deletes an existing trust relationship between your AWS Managed
Microsoft AD directory and an external domain}
\usage{
directoryservice_delete_trust(TrustId,
  DeleteAssociatedConditionalForwarder)
}
\arguments{
\item{TrustId}{[required] The Trust ID of the trust relationship to be deleted.}

\item{DeleteAssociatedConditionalForwarder}{Delete a conditional forwarder as part of a DeleteTrustRequest.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TrustId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes an existing trust relationship between your AWS Managed
Microsoft AD directory and an external domain.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_trust(
  TrustId = "string",
  DeleteAssociatedConditionalForwarder = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
