% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_deregister_certificate}
\alias{directoryservice_deregister_certificate}
\title{Deletes from the system the certificate that was registered for secure
LDAP or client certificate authentication}
\usage{
directoryservice_deregister_certificate(DirectoryId, CertificateId)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{CertificateId}{[required] The identifier of the certificate.}
}
\value{
An empty list.
}
\description{
Deletes from the system the certificate that was registered for secure
LDAP or client certificate authentication.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$deregister_certificate(
  DirectoryId = "string",
  CertificateId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
