% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_deregister_event_topic}
\alias{directoryservice_deregister_event_topic}
\title{Removes the specified directory as a publisher to the specified Amazon
SNS topic}
\usage{
directoryservice_deregister_event_topic(DirectoryId, TopicName)
}
\arguments{
\item{DirectoryId}{[required] The Directory ID to remove as a publisher. This directory will no longer
send messages to the specified Amazon SNS topic.}

\item{TopicName}{[required] The name of the Amazon SNS topic from which to remove the directory as a
publisher.}
}
\description{
Removes the specified directory as a publisher to the specified Amazon SNS topic.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_deregister_event_topic/} for full documentation.
}
\keyword{internal}
