% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_ca_enrollment_policy}
\alias{directoryservice_describe_ca_enrollment_policy}
\title{Retrieves detailed information about the certificate authority (CA)
enrollment policy for the specified directory}
\usage{
directoryservice_describe_ca_enrollment_policy(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory for which to retrieve the CA enrollment
policy information.}
}
\description{
Retrieves detailed information about the certificate authority (CA) enrollment policy for the specified directory. This policy determines how client certificates are automatically enrolled and managed through Amazon Web Services Private Certificate Authority.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_describe_ca_enrollment_policy/} for full documentation.
}
\keyword{internal}
