% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_directory_data_access}
\alias{directoryservice_describe_directory_data_access}
\title{Obtains status of directory data access enablement through the Directory
Service Data API for the specified directory}
\usage{
directoryservice_describe_directory_data_access(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] The directory identifier.}
}
\description{
Obtains status of directory data access enablement through the Directory Service Data API for the specified directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_describe_directory_data_access/} for full documentation.
}
\keyword{internal}
