% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_settings}
\alias{directoryservice_describe_settings}
\title{Retrieves information about the configurable settings for the specified
directory}
\usage{
directoryservice_describe_settings(
  DirectoryId,
  Status = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory for which to retrieve information.}

\item{Status}{The status of the directory settings for which to retrieve information.}

\item{NextToken}{The \code{DescribeSettingsResult.NextToken} value from a previous call to
\code{\link[=directoryservice_describe_settings]{describe_settings}}. Pass null if
this is the first call.}
}
\description{
Retrieves information about the configurable settings for the specified directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_describe_settings/} for full documentation.
}
\keyword{internal}
