% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_shared_directories}
\alias{directoryservice_describe_shared_directories}
\title{Returns the shared directories in your account}
\usage{
directoryservice_describe_shared_directories(OwnerDirectoryId,
  SharedDirectoryIds, NextToken, Limit)
}
\arguments{
\item{OwnerDirectoryId}{[required] Returns the identifier of the directory in the directory owner account.}

\item{SharedDirectoryIds}{A list of identifiers of all shared directories in your account.}

\item{NextToken}{The \code{DescribeSharedDirectoriesResult.NextToken} value from a previous
call to
\code{\link[=directoryservice_describe_shared_directories]{describe_shared_directories}}.
Pass null if this is the first call.}

\item{Limit}{The number of shared directories to return in the response object.}
}
\value{
A list with the following syntax:\preformatted{list(
  SharedDirectories = list(
    list(
      OwnerAccountId = "string",
      OwnerDirectoryId = "string",
      ShareMethod = "ORGANIZATIONS"|"HANDSHAKE",
      SharedAccountId = "string",
      SharedDirectoryId = "string",
      ShareStatus = "Shared"|"PendingAcceptance"|"Rejected"|"Rejecting"|"RejectFailed"|"Sharing"|"ShareFailed"|"Deleted"|"Deleting",
      ShareNotes = "string",
      CreatedDateTime = as.POSIXct(
        "2015-01-01"
      ),
      LastUpdatedDateTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Returns the shared directories in your account.
}
\section{Request syntax}{
\preformatted{svc$describe_shared_directories(
  OwnerDirectoryId = "string",
  SharedDirectoryIds = list(
    "string"
  ),
  NextToken = "string",
  Limit = 123
)
}
}

\keyword{internal}
