% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_update_directory}
\alias{directoryservice_describe_update_directory}
\title{Describes the updates of a directory for a particular update type}
\usage{
directoryservice_describe_update_directory(
  DirectoryId,
  UpdateType,
  RegionName = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DirectoryId}{[required] The unique identifier of the directory.}

\item{UpdateType}{[required] The type of updates you want to describe for the directory.}

\item{RegionName}{The name of the Region.}

\item{NextToken}{The \code{DescribeUpdateDirectoryResult}. NextToken value from a previous
call to
\code{\link[=directoryservice_describe_update_directory]{describe_update_directory}}.
Pass null if this is the first call.}
}
\description{
Describes the updates of a directory for a particular update type.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_describe_update_directory/} for full documentation.
}
\keyword{internal}
