% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_disable_ca_enrollment_policy}
\alias{directoryservice_disable_ca_enrollment_policy}
\title{Disables the certificate authority (CA) enrollment policy for the
specified directory}
\usage{
directoryservice_disable_ca_enrollment_policy(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory for which to disable the CA enrollment
policy.}
}
\description{
Disables the certificate authority (CA) enrollment policy for the specified directory. This stops automatic certificate enrollment and management for domain-joined clients, but does not affect existing certificates.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_disable_ca_enrollment_policy/} for full documentation.
}
\keyword{internal}
