% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_disable_client_authentication}
\alias{directoryservice_disable_client_authentication}
\title{Disables alternative client authentication methods for the specified
directory}
\usage{
directoryservice_disable_client_authentication(DirectoryId, Type)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory}

\item{Type}{[required] The type of client authentication to disable. Currently, only the
parameter, \code{SmartCard} is supported.}
}
\description{
Disables alternative client authentication methods for the specified directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_disable_client_authentication/} for full documentation.
}
\keyword{internal}
