% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_enable_client_authentication}
\alias{directoryservice_enable_client_authentication}
\title{Enables alternative client authentication methods for the specified
directory}
\usage{
directoryservice_enable_client_authentication(DirectoryId, Type)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the specified directory.}

\item{Type}{[required] The type of client authentication to enable. Currently only the value
\code{SmartCard} is supported. Smart card authentication in AD Connector
requires that you enable Kerberos Constrained Delegation for the Service
User to the LDAP service in the on-premises AD.}
}
\value{
An empty list.
}
\description{
Enables alternative client authentication methods for the specified
directory.
}
\section{Request syntax}{
\preformatted{svc$enable_client_authentication(
  DirectoryId = "string",
  Type = "SmartCard"
)
}
}

\keyword{internal}
