% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_enable_directory_data_access}
\alias{directoryservice_enable_directory_data_access}
\title{Enables access to directory data via the Directory Service Data API for
the specified directory}
\usage{
directoryservice_enable_directory_data_access(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] The directory identifier.}
}
\description{
Enables access to directory data via the Directory Service Data API for the specified directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_enable_directory_data_access/} for full documentation.
}
\keyword{internal}
