% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_enable_ldaps}
\alias{directoryservice_enable_ldaps}
\title{Activates the switch for the specific directory to always use LDAP
secure calls}
\usage{
directoryservice_enable_ldaps(DirectoryId, Type)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{Type}{The type of LDAP security the customer wants to enable. The security can
be either server or client, but currently only the default \code{Client} is
supported.}
}
\description{
Activates the switch for the specific directory to always use LDAP
secure calls.
}
\section{Request syntax}{
\preformatted{svc$enable_ldaps(
  DirectoryId = "string",
  Type = "Client"
)
}
}

\keyword{internal}
