% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_list_ad_assessments}
\alias{directoryservice_list_ad_assessments}
\title{Retrieves a list of directory assessments for the specified directory or
all assessments in your account}
\usage{
directoryservice_list_ad_assessments(
  DirectoryId = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{DirectoryId}{The identifier of the directory for which to list assessments. If not
specified, all assessments in your account are returned.}

\item{NextToken}{The pagination token from a previous request to
\code{\link[=directoryservice_list_ad_assessments]{list_ad_assessments}}. Pass null
if this is the first request.}

\item{Limit}{The maximum number of assessment summaries to return.}
}
\description{
Retrieves a list of directory assessments for the specified directory or all assessments in your account. Use this operation to monitor assessment status and manage multiple assessments.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_list_ad_assessments/} for full documentation.
}
\keyword{internal}
