% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_list_certificates}
\alias{directoryservice_list_certificates}
\title{For the specified directory, lists all the certificates registered for a
secure LDAP or client certificate authentication}
\usage{
directoryservice_list_certificates(DirectoryId, NextToken = NULL, Limit = NULL)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{NextToken}{A token for requesting another page of certificates if the \code{NextToken}
response element indicates that more certificates are available. Use the
value of the returned \code{NextToken} element in your request until the
token comes back as \code{null}. Pass \code{null} if this is the first call.}

\item{Limit}{The number of items that should show up on one page}
}
\description{
For the specified directory, lists all the certificates registered for a secure LDAP or client certificate authentication.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_list_certificates/} for full documentation.
}
\keyword{internal}
