% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_list_tags_for_resource}
\alias{directoryservice_list_tags_for_resource}
\title{Lists all tags on a directory}
\usage{
directoryservice_list_tags_for_resource(ResourceId, NextToken, Limit)
}
\arguments{
\item{ResourceId}{[required] Identifier (ID) of the directory for which you want to retrieve tags.}

\item{NextToken}{Reserved for future use.}

\item{Limit}{Reserved for future use.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all tags on a directory.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  ResourceId = "string",
  NextToken = "string",
  Limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
