% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_register_certificate}
\alias{directoryservice_register_certificate}
\title{Registers a certificate for a secure LDAP or client certificate
authentication}
\usage{
directoryservice_register_certificate(DirectoryId, CertificateData,
  Type, ClientCertAuthSettings)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{CertificateData}{[required] The certificate PEM string that needs to be registered.}

\item{Type}{The function that the registered certificate performs. Valid values
include \code{ClientLDAPS} or \code{ClientCertAuth}. The default value is
\code{ClientLDAPS}.}

\item{ClientCertAuthSettings}{A \code{ClientCertAuthSettings} object that contains client certificate
authentication settings.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CertificateId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Registers a certificate for a secure LDAP or client certificate
authentication.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$register_certificate(
  DirectoryId = "string",
  CertificateData = "string",
  Type = "ClientCertAuth"|"ClientLDAPS",
  ClientCertAuthSettings = list(
    OCSPUrl = "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
