% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_remove_ip_routes}
\alias{directoryservice_remove_ip_routes}
\title{Removes IP address blocks from a directory}
\usage{
directoryservice_remove_ip_routes(
  DirectoryId,
  CidrIps = NULL,
  CidrIpv6s = NULL
)
}
\arguments{
\item{DirectoryId}{[required] Identifier (ID) of the directory from which you want to remove the IP
addresses.}

\item{CidrIps}{IP address blocks that you want to remove.}

\item{CidrIpv6s}{IPv6 address blocks that you want to remove.}
}
\description{
Removes IP address blocks from a directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_remove_ip_routes/} for full documentation.
}
\keyword{internal}
