% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_start_ad_assessment}
\alias{directoryservice_start_ad_assessment}
\title{Initiates a directory assessment to validate your self-managed AD
environment for hybrid domain join}
\usage{
directoryservice_start_ad_assessment(
  AssessmentConfiguration = NULL,
  DirectoryId = NULL
)
}
\arguments{
\item{AssessmentConfiguration}{Configuration parameters for the directory assessment, including DNS
server information, domain name, Amazon VPC subnet, and Amazon Web
Services System Manager managed node details.}

\item{DirectoryId}{The identifier of the directory for which to perform the assessment.
This should be an existing directory. If the assessment is not for an
existing directory, this parameter should be omitted.}
}
\description{
Initiates a directory assessment to validate your self-managed AD environment for hybrid domain join. The assessment checks compatibility and connectivity of the self-managed AD environment.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_start_ad_assessment/} for full documentation.
}
\keyword{internal}
