% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_unshare_directory}
\alias{directoryservice_unshare_directory}
\title{Stops the directory sharing between the directory owner and consumer
accounts}
\usage{
directoryservice_unshare_directory(DirectoryId, UnshareTarget)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the AWS Managed Microsoft AD directory that you want
to stop sharing.}

\item{UnshareTarget}{[required] Identifier for the directory consumer account with whom the directory
has to be unshared.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SharedDirectoryId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Stops the directory sharing between the directory owner and consumer
accounts.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$unshare_directory(
  DirectoryId = "string",
  UnshareTarget = list(
    Id = "string",
    Type = "ACCOUNT"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
