% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_update_directory_setup}
\alias{directoryservice_update_directory_setup}
\title{Updates the directory for a particular update type}
\usage{
directoryservice_update_directory_setup(
  DirectoryId,
  UpdateType,
  OSUpdateSettings = NULL,
  CreateSnapshotBeforeUpdate = NULL
)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory on which you want to perform the update.}

\item{UpdateType}{[required] The type of update that needs to be performed on the directory. For
example, OS.}

\item{OSUpdateSettings}{The settings for the OS update that needs to be performed on the
directory.}

\item{CreateSnapshotBeforeUpdate}{The boolean that specifies if a snapshot for the directory needs to be
taken before updating the directory.}
}
\description{
Updates the directory for a particular update type.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_update_directory_setup/} for full documentation.
}
\keyword{internal}
