% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_update_hybrid_ad}
\alias{directoryservice_update_hybrid_ad}
\title{Updates the configuration of an existing hybrid directory}
\usage{
directoryservice_update_hybrid_ad(
  DirectoryId,
  HybridAdministratorAccountUpdate = NULL,
  SelfManagedInstancesSettings = NULL
)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the hybrid directory to update.}

\item{HybridAdministratorAccountUpdate}{We create a hybrid directory administrator account when we create a
hybrid directory. Use \code{HybridAdministratorAccountUpdate} to recover the
hybrid directory administrator account if you have deleted it.

To recover your hybrid directory administrator account, we need
temporary access to a user in your self-managed AD with administrator
permissions in the form of a secret from Amazon Web Services Secrets
Manager. We use these credentials once during recovery and don't store
them.

If your hybrid directory administrator account exists, then you don’t
need to use \code{HybridAdministratorAccountUpdate}, even if you have updated
your self-managed AD administrator user.}

\item{SelfManagedInstancesSettings}{Updates to the self-managed AD configuration, including DNS server IP
addresses and Amazon Web Services System Manager managed node
identifiers.}
}
\description{
Updates the configuration of an existing hybrid directory. You can recover hybrid directory administrator account or modify self-managed instance settings.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_update_hybrid_ad/} for full documentation.
}
\keyword{internal}
