% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_update_radius}
\alias{directoryservice_update_radius}
\title{Updates the Remote Authentication Dial In User Service (RADIUS) server
information for an AD Connector or Microsoft AD directory}
\usage{
directoryservice_update_radius(DirectoryId, RadiusSettings)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory for which to update the RADIUS server
information.}

\item{RadiusSettings}{[required] A RadiusSettings object that contains information about the RADIUS
server.}
}
\value{
An empty list.
}
\description{
Updates the Remote Authentication Dial In User Service (RADIUS) server
information for an AD Connector or Microsoft AD directory.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_radius(
  DirectoryId = "string",
  RadiusSettings = list(
    RadiusServers = list(
      "string"
    ),
    RadiusPort = 123,
    RadiusTimeout = 123,
    RadiusRetries = 123,
    SharedSecret = "string",
    AuthenticationProtocol = "PAP"|"CHAP"|"MS-CHAPv1"|"MS-CHAPv2",
    DisplayLabel = "string",
    UseSameUsername = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
