% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_update_trust}
\alias{directoryservice_update_trust}
\title{Updates the trust that has been set up between your AWS Managed
Microsoft AD directory and an on-premises Active Directory}
\usage{
directoryservice_update_trust(TrustId, SelectiveAuth)
}
\arguments{
\item{TrustId}{[required] Identifier of the trust relationship.}

\item{SelectiveAuth}{Updates selective authentication for the trust.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  RequestId = "string",
  TrustId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Updates the trust that has been set up between your AWS Managed
Microsoft AD directory and an on-premises Active Directory.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_trust(
  TrustId = "string",
  SelectiveAuth = "Enabled"|"Disabled"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
