% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_verify_trust}
\alias{directoryservice_verify_trust}
\title{AWS Directory Service for Microsoft Active Directory allows you to
configure and verify trust relationships}
\usage{
directoryservice_verify_trust(TrustId)
}
\arguments{
\item{TrustId}{[required] The unique Trust ID of the trust relationship to verify.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  TrustId = "string"
)
}\if{html}{\out{</div>}}
}
\description{
AWS Directory Service for Microsoft Active Directory allows you to
configure and verify trust relationships.

This action verifies a trust relationship between your AWS Managed
Microsoft AD directory and an external domain.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$verify_trust(
  TrustId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
