% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_associate_third_party_firewall}
\alias{fms_associate_third_party_firewall}
\title{Sets the Firewall Manager policy administrator as a tenant administrator
of a third-party firewall service}
\usage{
fms_associate_third_party_firewall(ThirdPartyFirewall)
}
\arguments{
\item{ThirdPartyFirewall}{[required] The name of the third-party firewall vendor.}
}
\description{
Sets the Firewall Manager policy administrator as a tenant administrator of a third-party firewall service. A tenant is an instance of the third-party firewall service that's associated with your Amazon Web Services customer account.

See \url{https://www.paws-r-sdk.com/docs/fms_associate_third_party_firewall/} for full documentation.
}
\keyword{internal}
