% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_batch_associate_resource}
\alias{fms_batch_associate_resource}
\title{Associate resources to a Firewall Manager resource set}
\usage{
fms_batch_associate_resource(ResourceSetIdentifier, Items)
}
\arguments{
\item{ResourceSetIdentifier}{[required] A unique identifier for the resource set, used in a request to refer to
the resource set.}

\item{Items}{[required] The uniform resource identifiers (URIs) of resources that should be
associated to the resource set. The URIs must be Amazon Resource Names
(ARNs).}
}
\description{
Associate resources to a Firewall Manager resource set.

See \url{https://www.paws-r-sdk.com/docs/fms_batch_associate_resource/} for full documentation.
}
\keyword{internal}
