% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_batch_disassociate_resource}
\alias{fms_batch_disassociate_resource}
\title{Disassociates resources from a Firewall Manager resource set}
\usage{
fms_batch_disassociate_resource(ResourceSetIdentifier, Items)
}
\arguments{
\item{ResourceSetIdentifier}{[required] A unique identifier for the resource set, used in a request to refer to
the resource set.}

\item{Items}{[required] The uniform resource identifiers (URI) of resources that should be
disassociated from the resource set. The URIs must be Amazon Resource
Names (ARNs).}
}
\description{
Disassociates resources from a Firewall Manager resource set.

See \url{https://www.paws-r-sdk.com/docs/fms_batch_disassociate_resource/} for full documentation.
}
\keyword{internal}
