% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_delete_notification_channel}
\alias{fms_delete_notification_channel}
\title{Deletes an AWS Firewall Manager association with the IAM role and the
Amazon Simple Notification Service (SNS) topic that is used to record
AWS Firewall Manager SNS logs}
\usage{
fms_delete_notification_channel()
}
\value{
An empty list.
}
\description{
Deletes an AWS Firewall Manager association with the IAM role and the
Amazon Simple Notification Service (SNS) topic that is used to record
AWS Firewall Manager SNS logs.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_notification_channel()
}\if{html}{\out{</div>}}
}

\keyword{internal}
