% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_delete_resource_set}
\alias{fms_delete_resource_set}
\title{Deletes the specified ResourceSet}
\usage{
fms_delete_resource_set(Identifier)
}
\arguments{
\item{Identifier}{[required] A unique identifier for the resource set, used in a request to refer to
the resource set.}
}
\description{
Deletes the specified ResourceSet.

See \url{https://www.paws-r-sdk.com/docs/fms_delete_resource_set/} for full documentation.
}
\keyword{internal}
