% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_disassociate_third_party_firewall}
\alias{fms_disassociate_third_party_firewall}
\title{Disassociates a Firewall Manager policy administrator from a third-party
firewall tenant}
\usage{
fms_disassociate_third_party_firewall(ThirdPartyFirewall)
}
\arguments{
\item{ThirdPartyFirewall}{[required] The name of the third-party firewall vendor.}
}
\description{
Disassociates a Firewall Manager policy administrator from a third-party firewall tenant. When you call \code{\link[=fms_disassociate_third_party_firewall]{disassociate_third_party_firewall}}, the third-party firewall vendor deletes all of the firewalls that are associated with the account.

See \url{https://www.paws-r-sdk.com/docs/fms_disassociate_third_party_firewall/} for full documentation.
}
\keyword{internal}
