% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_admin_scope}
\alias{fms_get_admin_scope}
\title{Returns information about the specified account's administrative scope}
\usage{
fms_get_admin_scope(AdminAccount)
}
\arguments{
\item{AdminAccount}{[required] The administator account that you want to get the details for.}
}
\description{
Returns information about the specified account's administrative scope. The admistrative scope defines the resources that an Firewall Manager administrator can manage.

See \url{https://paws-r.github.io/docs/fms/get_admin_scope.html} for full documentation.
}
\keyword{internal}
