% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_compliance_detail}
\alias{fms_get_compliance_detail}
\title{Returns detailed compliance information about the specified member
account}
\usage{
fms_get_compliance_detail(PolicyId, MemberAccount)
}
\arguments{
\item{PolicyId}{[required] The ID of the policy that you want to get the details for. \code{PolicyId} is
returned by \code{\link[=fms_put_policy]{put_policy}} and by
\code{\link[=fms_list_policies]{list_policies}}.}

\item{MemberAccount}{[required] The AWS account that owns the resources that you want to get the details
for.}
}
\value{
A list with the following syntax:\preformatted{list(
  PolicyComplianceDetail = list(
    PolicyOwner = "string",
    PolicyId = "string",
    MemberAccount = "string",
    Violators = list(
      list(
        ResourceId = "string",
        ViolationReason = "WEB_ACL_MISSING_RULE_GROUP"|"RESOURCE_MISSING_WEB_ACL"|"RESOURCE_INCORRECT_WEB_ACL"|"RESOURCE_MISSING_SHIELD_PROTECTION"|"RESOURCE_MISSING_WEB_ACL_OR_SHIELD_PROTECTION"|"RESOURCE_MISSING_SECURITY_GROUP"|"RESOURCE_VIOLATES_AUDIT_SECURITY_GROUP"|"SECURITY_GROUP_UNUSED"|"SECURITY_GROUP_REDUNDANT"|"MISSING_FIREWALL"|"MISSING_FIREWALL_SUBNET_IN_AZ"|"MISSING_EXPECTED_ROUTE_TABLE"|"NETWORK_FIREWALL_POLICY_MODIFIED",
        ResourceType = "string"
      )
    ),
    EvaluationLimitExceeded = TRUE|FALSE,
    ExpiredAt = as.POSIXct(
      "2015-01-01"
    ),
    IssueInfoMap = list(
      "string"
    )
  )
)
}
}
\description{
Returns detailed compliance information about the specified member
account. Details include resources that are in and out of compliance
with the specified policy. Resources are considered noncompliant for AWS
WAF and Shield Advanced policies if the specified policy has not been
applied to them. Resources are considered noncompliant for security
group policies if they are in scope of the policy, they violate one or
more of the policy rules, and remediation is disabled or not possible.
Resources are considered noncompliant for Network Firewall policies if a
firewall is missing in the VPC, if the firewall endpoint isn't set up in
an expected Availability Zone and subnet, if a subnet created by the
Firewall Manager doesn't have the expected route table, and for
modifications to a firewall policy that violate the Firewall Manager
policy's rules.
}
\section{Request syntax}{
\preformatted{svc$get_compliance_detail(
  PolicyId = "string",
  MemberAccount = "string"
)
}
}

\keyword{internal}
