% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_protection_status}
\alias{fms_get_protection_status}
\title{If you created a Shield Advanced policy, returns policy-level attack
summary information in the event of a potential DDoS attack}
\usage{
fms_get_protection_status(PolicyId, MemberAccountId, StartTime, EndTime,
  NextToken, MaxResults)
}
\arguments{
\item{PolicyId}{[required] The ID of the policy for which you want to get the attack information.}

\item{MemberAccountId}{The AWS account that is in scope of the policy that you want to get the
details for.}

\item{StartTime}{The start of the time period to query for the attacks. This is a
\code{timestamp} type. The sample request above indicates a number type
because the default used by AWS Firewall Manager is Unix time in
seconds. However, any valid \code{timestamp} format is allowed.}

\item{EndTime}{The end of the time period to query for the attacks. This is a
\code{timestamp} type. The sample request above indicates a number type
because the default used by AWS Firewall Manager is Unix time in
seconds. However, any valid \code{timestamp} format is allowed.}

\item{NextToken}{If you specify a value for \code{MaxResults} and you have more objects than
the number that you specify for \code{MaxResults}, AWS Firewall Manager
returns a \code{NextToken} value in the response that allows you to list
another group of objects. For the second and subsequent
\code{GetProtectionStatus} requests, specify the value of \code{NextToken} from
the previous response to get information about another batch of objects.}

\item{MaxResults}{Specifies the number of objects that you want AWS Firewall Manager to
return for this request. If you have more objects than the number that
you specify for \code{MaxResults}, the response includes a \code{NextToken} value
that you can use to get another batch of objects.}
}
\description{
If you created a Shield Advanced policy, returns policy-level attack
summary information in the event of a potential DDoS attack.
}
\section{Request syntax}{
\preformatted{svc$get_protection_status(
  PolicyId = "string",
  MemberAccountId = "string",
  StartTime = as.POSIXct(
    "2015-01-01"
  ),
  EndTime = as.POSIXct(
    "2015-01-01"
  ),
  NextToken = "string",
  MaxResults = 123
)
}
}

\keyword{internal}
