% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_protocols_list}
\alias{fms_get_protocols_list}
\title{Returns information about the specified AWS Firewall Manager protocols
list}
\usage{
fms_get_protocols_list(ListId, DefaultList)
}
\arguments{
\item{ListId}{[required] The ID of the AWS Firewall Manager protocols list that you want the
details for.}

\item{DefaultList}{Specifies whether the list to retrieve is a default list owned by AWS
Firewall Manager.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ProtocolsList = list(
    ListId = "string",
    ListName = "string",
    ListUpdateToken = "string",
    CreateTime = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdateTime = as.POSIXct(
      "2015-01-01"
    ),
    ProtocolsList = list(
      "string"
    ),
    PreviousProtocolsList = list(
      list(
        "string"
      )
    )
  ),
  ProtocolsListArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about the specified AWS Firewall Manager protocols
list.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_protocols_list(
  ListId = "string",
  DefaultList = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
