% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_resource_set}
\alias{fms_get_resource_set}
\title{Gets information about a specific resource set}
\usage{
fms_get_resource_set(Identifier)
}
\arguments{
\item{Identifier}{[required] A unique identifier for the resource set, used in a request to refer to
the resource set.}
}
\description{
Gets information about a specific resource set.

See \url{https://www.paws-r-sdk.com/docs/fms_get_resource_set/} for full documentation.
}
\keyword{internal}
