% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_third_party_firewall_association_status}
\alias{fms_get_third_party_firewall_association_status}
\title{The onboarding status of a Firewall Manager admin account to third-party
firewall vendor tenant}
\usage{
fms_get_third_party_firewall_association_status(ThirdPartyFirewall)
}
\arguments{
\item{ThirdPartyFirewall}{[required] The name of the third-party firewall vendor.}
}
\description{
The onboarding status of a Firewall Manager admin account to third-party firewall vendor tenant.

See \url{https://www.paws-r-sdk.com/docs/fms_get_third_party_firewall_association_status/} for full documentation.
}
\keyword{internal}
