% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_put_protocols_list}
\alias{fms_put_protocols_list}
\title{Creates an Firewall Manager protocols list}
\usage{
fms_put_protocols_list(ProtocolsList, TagList = NULL)
}
\arguments{
\item{ProtocolsList}{[required] The details of the Firewall Manager protocols list to be created.}

\item{TagList}{The tags associated with the resource.}
}
\description{
Creates an Firewall Manager protocols list.

See \url{https://www.paws-r-sdk.com/docs/fms_put_protocols_list/} for full documentation.
}
\keyword{internal}
