% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_put_resource_set}
\alias{fms_put_resource_set}
\title{Creates the resource set}
\usage{
fms_put_resource_set(ResourceSet, TagList = NULL)
}
\arguments{
\item{ResourceSet}{[required] Details about the resource set to be created or updated.\>}

\item{TagList}{Retrieves the tags associated with the specified resource set. Tags are
key:value pairs that you can use to categorize and manage your
resources, for purposes like billing. For example, you might set the tag
key to "customer" and the value to the customer name or ID. You can
specify one or more tags to add to each Amazon Web Services resource, up
to 50 tags for a resource.}
}
\description{
Creates the resource set.

See \url{https://www.paws-r-sdk.com/docs/fms_put_resource_set/} for full documentation.
}
\keyword{internal}
