% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_service.R
\name{guardduty}
\alias{guardduty}
\title{Amazon GuardDuty}
\usage{
guardduty(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
\item{\strong{use_dual_stack}: Set this to \code{true} to use the dualstack (IPv4 and IPv6) endpoint for a service, where available, falling back to the regular endpoint when it isn't. Defaults to the \code{AWS_USE_DUALSTACK_ENDPOINT} environment variable when unset.}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Amazon GuardDuty is a continuous security monitoring service that analyzes and processes the following foundational data sources - VPC flow logs, Amazon Web Services CloudTrail management event logs, CloudTrail S3 data event logs, EKS audit logs, DNS logs, Amazon EBS volume data, runtime activity belonging to container workloads, such as Amazon EKS, Amazon ECS (including Amazon Web Services Fargate), and Amazon EC2 instances. It uses threat intelligence feeds, such as lists of malicious IPs and domains, and machine learning to identify unexpected, potentially unauthorized, and malicious activity within your Amazon Web Services environment. This can include issues like escalations of privileges, uses of exposed credentials, or communication with malicious IPs, domains, or presence of malware on your Amazon EC2 instances and container workloads. For example, GuardDuty can detect compromised EC2 instances and container workloads serving malware, or mining bitcoin.

GuardDuty also monitors Amazon Web Services account access behavior for signs of compromise, such as unauthorized infrastructure deployments like EC2 instances deployed in a Region that has never been used, or unusual API calls like a password policy change to reduce password strength.

GuardDuty informs you about the status of your Amazon Web Services environment by producing security findings that you can view in the GuardDuty console or through Amazon EventBridge. For more information, see the \emph{\href{https://docs.aws.amazon.com/guardduty/latest/ug/what-is-guardduty.html}{Amazon GuardDuty User Guide}} .
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- guardduty(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string",
    use_dual_stack = "logical"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=guardduty_accept_administrator_invitation]{accept_administrator_invitation} \tab Accepts the invitation to be a member account and get monitored by a GuardDuty administrator account that sent the invitation\cr
\link[=guardduty_accept_invitation]{accept_invitation} \tab Accepts the invitation to be monitored by a GuardDuty administrator account\cr
\link[=guardduty_archive_findings]{archive_findings} \tab Archives GuardDuty findings that are specified by the list of finding IDs\cr
\link[=guardduty_create_custom_detection_rule_association]{create_custom_detection_rule_association} \tab Enables a custom detection rule for your account by creating an association\cr
\link[=guardduty_create_custom_detection_rule_org_configuration]{create_custom_detection_rule_org_configuration} \tab Creates an organization-level configuration that enables a custom detection rule across your organization\cr
\link[=guardduty_create_detector]{create_detector} \tab Creates a single GuardDuty detector\cr
\link[=guardduty_create_filter]{create_filter} \tab Creates a filter using the specified finding criteria\cr
\link[=guardduty_create_investigation]{create_investigation} \tab This API is currently available as a preview\cr
\link[=guardduty_create_ip_set]{create_ip_set} \tab Creates a new IPSet, which is called a trusted IP list in the console user interface\cr
\link[=guardduty_create_malware_protection_plan]{create_malware_protection_plan} \tab Creates a new Malware Protection plan for the protected resource\cr
\link[=guardduty_create_members]{create_members} \tab Creates member accounts of the current Amazon Web Services account by specifying a list of Amazon Web Services account IDs\cr
\link[=guardduty_create_publishing_destination]{create_publishing_destination} \tab Creates a publishing destination where you can export your GuardDuty findings\cr
\link[=guardduty_create_sample_findings]{create_sample_findings} \tab Generates sample findings of types specified by the list of finding types\cr
\link[=guardduty_create_threat_entity_set]{create_threat_entity_set} \tab Creates a new threat entity set\cr
\link[=guardduty_create_threat_intel_set]{create_threat_intel_set} \tab Creates a new ThreatIntelSet\cr
\link[=guardduty_create_trusted_entity_set]{create_trusted_entity_set} \tab Creates a new trusted entity set\cr
\link[=guardduty_decline_invitations]{decline_invitations} \tab Declines invitations sent to the current member account by Amazon Web Services accounts specified by their account IDs\cr
\link[=guardduty_delete_custom_detection_rule_association]{delete_custom_detection_rule_association} \tab Disables a custom detection rule by deleting its association\cr
\link[=guardduty_delete_custom_detection_rule_org_configuration]{delete_custom_detection_rule_org_configuration} \tab Deletes the organization-level configuration for a custom detection rule\cr
\link[=guardduty_delete_detector]{delete_detector} \tab Deletes an Amazon GuardDuty detector that is specified by the detector ID\cr
\link[=guardduty_delete_filter]{delete_filter} \tab Deletes the filter specified by the filter name\cr
\link[=guardduty_delete_invitations]{delete_invitations} \tab Deletes invitations sent to the current member account by Amazon Web Services accounts specified by their account IDs\cr
\link[=guardduty_delete_ip_set]{delete_ip_set} \tab Deletes the IPSet specified by the ipSetId\cr
\link[=guardduty_delete_malware_protection_plan]{delete_malware_protection_plan} \tab Deletes the Malware Protection plan ID associated with the Malware Protection plan resource\cr
\link[=guardduty_delete_members]{delete_members} \tab Deletes GuardDuty member accounts (to the current GuardDuty administrator account) specified by the account IDs\cr
\link[=guardduty_delete_publishing_destination]{delete_publishing_destination} \tab Deletes the publishing definition with the specified destinationId\cr
\link[=guardduty_delete_threat_entity_set]{delete_threat_entity_set} \tab Deletes the threat entity set that is associated with the specified threatEntitySetId\cr
\link[=guardduty_delete_threat_intel_set]{delete_threat_intel_set} \tab Deletes the ThreatIntelSet specified by the ThreatIntelSet ID\cr
\link[=guardduty_delete_trusted_entity_set]{delete_trusted_entity_set} \tab Deletes the trusted entity set that is associated with the specified trustedEntitySetId\cr
\link[=guardduty_describe_malware_scans]{describe_malware_scans} \tab Returns a list of malware scans\cr
\link[=guardduty_describe_organization_configuration]{describe_organization_configuration} \tab Returns information about the account selected as the delegated administrator for GuardDuty\cr
\link[=guardduty_describe_publishing_destination]{describe_publishing_destination} \tab Returns information about the publishing destination specified by the provided destinationId\cr
\link[=guardduty_disable_organization_admin_account]{disable_organization_admin_account} \tab Removes the existing GuardDuty delegated administrator of the organization\cr
\link[=guardduty_disassociate_from_administrator_account]{disassociate_from_administrator_account} \tab Disassociates the current GuardDuty member account from its administrator account\cr
\link[=guardduty_disassociate_from_master_account]{disassociate_from_master_account} \tab Disassociates the current GuardDuty member account from its administrator account\cr
\link[=guardduty_disassociate_members]{disassociate_members} \tab Disassociates GuardDuty member accounts (from the current administrator account) specified by the account IDs\cr
\link[=guardduty_enable_organization_admin_account]{enable_organization_admin_account} \tab Designates an Amazon Web Services account within the organization as your GuardDuty delegated administrator\cr
\link[=guardduty_get_administrator_account]{get_administrator_account} \tab Provides the details of the GuardDuty administrator account associated with the current GuardDuty member account\cr
\link[=guardduty_get_coverage_statistics]{get_coverage_statistics} \tab Retrieves aggregated statistics for your account\cr
\link[=guardduty_get_custom_detection_rule]{get_custom_detection_rule} \tab Returns details for a custom detection rule in GuardDuty, including its detection logic\cr
\link[=guardduty_get_custom_detection_rule_association]{get_custom_detection_rule_association} \tab Returns details for a custom detection rule association\cr
\link[=guardduty_get_custom_detection_rule_org_configuration]{get_custom_detection_rule_org_configuration} \tab Returns the organization-level configuration for a custom detection rule\cr
\link[=guardduty_get_detector]{get_detector} \tab Retrieves a GuardDuty detector specified by the detectorId\cr
\link[=guardduty_get_filter]{get_filter} \tab Returns the details of the filter specified by the filter name\cr
\link[=guardduty_get_findings]{get_findings} \tab Describes Amazon GuardDuty findings specified by finding IDs\cr
\link[=guardduty_get_findings_statistics]{get_findings_statistics} \tab Lists GuardDuty findings statistics for the specified detector ID\cr
\link[=guardduty_get_investigation]{get_investigation} \tab This API is currently available as a preview\cr
\link[=guardduty_get_invitations_count]{get_invitations_count} \tab Returns the count of all GuardDuty membership invitations that were sent to the current member account except the currently accepted invitation\cr
\link[=guardduty_get_ip_set]{get_ip_set} \tab Retrieves the IPSet specified by the ipSetId\cr
\link[=guardduty_get_malware_protection_plan]{get_malware_protection_plan} \tab Retrieves the Malware Protection plan details associated with a Malware Protection plan ID\cr
\link[=guardduty_get_malware_scan]{get_malware_scan} \tab Retrieves the detailed information for a specific malware scan\cr
\link[=guardduty_get_malware_scan_settings]{get_malware_scan_settings} \tab Returns the details of the malware scan settings\cr
\link[=guardduty_get_master_account]{get_master_account} \tab Provides the details for the GuardDuty administrator account associated with the current GuardDuty member account\cr
\link[=guardduty_get_member_detectors]{get_member_detectors} \tab Describes which data sources are enabled for the member account's detector\cr
\link[=guardduty_get_members]{get_members} \tab Retrieves GuardDuty member accounts (of the current GuardDuty administrator account) specified by the account IDs\cr
\link[=guardduty_get_organization_statistics]{get_organization_statistics} \tab Retrieves how many active member accounts have each feature enabled within GuardDuty\cr
\link[=guardduty_get_remaining_free_trial_days]{get_remaining_free_trial_days} \tab Provides the number of days left for each data source used in the free trial period\cr
\link[=guardduty_get_threat_entity_set]{get_threat_entity_set} \tab Retrieves the threat entity set associated with the specified threatEntitySetId\cr
\link[=guardduty_get_threat_intel_set]{get_threat_intel_set} \tab Retrieves the ThreatIntelSet that is specified by the ThreatIntelSet ID\cr
\link[=guardduty_get_trusted_entity_set]{get_trusted_entity_set} \tab Retrieves the trusted entity set associated with the specified trustedEntitySetId\cr
\link[=guardduty_get_usage_statistics]{get_usage_statistics} \tab Lists Amazon GuardDuty usage statistics over the last 30 days for the specified detector ID\cr
\link[=guardduty_invite_members]{invite_members} \tab Invites Amazon Web Services accounts to become members of an organization administered by the Amazon Web Services account that invokes this API\cr
\link[=guardduty_list_coverage]{list_coverage} \tab Lists coverage details for your GuardDuty account\cr
\link[=guardduty_list_custom_detection_rule_associations]{list_custom_detection_rule_associations} \tab Returns all custom detection rule associations for your account\cr
\link[=guardduty_list_custom_detection_rule_org_configurations]{list_custom_detection_rule_org_configurations} \tab Returns all organization-level configurations for custom detection rules\cr
\link[=guardduty_list_custom_detection_rules]{list_custom_detection_rules} \tab Returns all available custom detection rules in GuardDuty\cr
\link[=guardduty_list_detectors]{list_detectors} \tab Lists detectorIds of all the existing Amazon GuardDuty detector resources\cr
\link[=guardduty_list_filters]{list_filters} \tab Returns a paginated list of the current filters\cr
\link[=guardduty_list_findings]{list_findings} \tab Lists GuardDuty findings for the specified detector ID\cr
\link[=guardduty_list_investigations]{list_investigations} \tab This API is currently available as a preview\cr
\link[=guardduty_list_invitations]{list_invitations} \tab Lists all GuardDuty membership invitations that were sent to the current Amazon Web Services account\cr
\link[=guardduty_list_ip_sets]{list_ip_sets} \tab Lists the IPSets of the GuardDuty service specified by the detector ID\cr
\link[=guardduty_list_malware_protection_plans]{list_malware_protection_plans} \tab Lists the Malware Protection plan IDs associated with the protected resources in your Amazon Web Services account\cr
\link[=guardduty_list_malware_scans]{list_malware_scans} \tab Returns a list of malware scans\cr
\link[=guardduty_list_members]{list_members} \tab Lists details about all member accounts for the current GuardDuty administrator account\cr
\link[=guardduty_list_organization_admin_accounts]{list_organization_admin_accounts} \tab Lists the accounts designated as GuardDuty delegated administrators\cr
\link[=guardduty_list_publishing_destinations]{list_publishing_destinations} \tab Returns a list of publishing destinations associated with the specified detectorId\cr
\link[=guardduty_list_tags_for_resource]{list_tags_for_resource} \tab Lists tags for a resource\cr
\link[=guardduty_list_threat_entity_sets]{list_threat_entity_sets} \tab Lists the threat entity sets associated with the specified GuardDuty detector ID\cr
\link[=guardduty_list_threat_intel_sets]{list_threat_intel_sets} \tab Lists the ThreatIntelSets of the GuardDuty service specified by the detector ID\cr
\link[=guardduty_list_trusted_entity_sets]{list_trusted_entity_sets} \tab Lists the trusted entity sets associated with the specified GuardDuty detector ID\cr
\link[=guardduty_send_object_malware_scan]{send_object_malware_scan} \tab Initiates a malware scan for a specific S3 object\cr
\link[=guardduty_start_malware_scan]{start_malware_scan} \tab Initiates the malware scan\cr
\link[=guardduty_start_monitoring_members]{start_monitoring_members} \tab Turns on GuardDuty monitoring of the specified member accounts\cr
\link[=guardduty_stop_monitoring_members]{stop_monitoring_members} \tab Stops GuardDuty monitoring for the specified member accounts\cr
\link[=guardduty_tag_resource]{tag_resource} \tab Adds tags to a resource\cr
\link[=guardduty_unarchive_findings]{unarchive_findings} \tab Unarchives GuardDuty findings specified by the findingIds\cr
\link[=guardduty_untag_resource]{untag_resource} \tab Removes tags from a resource\cr
\link[=guardduty_update_custom_detection_rule_association]{update_custom_detection_rule_association} \tab Updates the mode of an existing custom detection rule association\cr
\link[=guardduty_update_custom_detection_rule_org_configuration]{update_custom_detection_rule_org_configuration} \tab Updates the organization-level configuration for a custom detection rule, including the mode and include/exclude account lists\cr
\link[=guardduty_update_detector]{update_detector} \tab Updates the GuardDuty detector specified by the detector ID\cr
\link[=guardduty_update_filter]{update_filter} \tab Updates the filter specified by the filter name\cr
\link[=guardduty_update_findings_feedback]{update_findings_feedback} \tab Marks the specified GuardDuty findings as useful or not useful\cr
\link[=guardduty_update_ip_set]{update_ip_set} \tab Updates the IPSet specified by the IPSet ID\cr
\link[=guardduty_update_malware_protection_plan]{update_malware_protection_plan} \tab Updates an existing Malware Protection plan resource\cr
\link[=guardduty_update_malware_scan_settings]{update_malware_scan_settings} \tab Updates the malware scan settings\cr
\link[=guardduty_update_member_detectors]{update_member_detectors} \tab Contains information on member accounts to be updated\cr
\link[=guardduty_update_organization_configuration]{update_organization_configuration} \tab Configures the delegated administrator account with the provided values\cr
\link[=guardduty_update_publishing_destination]{update_publishing_destination} \tab Updates information about the publishing destination specified by the destinationId\cr
\link[=guardduty_update_threat_entity_set]{update_threat_entity_set} \tab Updates the threat entity set associated with the specified threatEntitySetId\cr
\link[=guardduty_update_threat_intel_set]{update_threat_intel_set} \tab Updates the ThreatIntelSet specified by the ThreatIntelSet ID\cr
\link[=guardduty_update_trusted_entity_set]{update_trusted_entity_set} \tab Updates the trusted entity set associated with the specified trustedEntitySetId
}
}

\examples{
\dontrun{
svc <- guardduty()
svc$accept_administrator_invitation(
  Foo = 123
)
}

}
