% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_accept_administrator_invitation}
\alias{guardduty_accept_administrator_invitation}
\title{Accepts the invitation to be a member account and get monitored by a
GuardDuty administrator account that sent the invitation}
\usage{
guardduty_accept_administrator_invitation(
  DetectorId,
  AdministratorId,
  InvitationId
)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty member account.}

\item{AdministratorId}{[required] The account ID of the GuardDuty administrator account whose invitation
you're accepting.}

\item{InvitationId}{[required] The value that is used to validate the administrator account to the
member account.}
}
\description{
Accepts the invitation to be a member account and get monitored by a GuardDuty administrator account that sent the invitation.

See \url{https://www.paws-r-sdk.com/docs/guardduty_accept_administrator_invitation/} for full documentation.
}
\keyword{internal}
