% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_create_custom_detection_rule_association}
\alias{guardduty_create_custom_detection_rule_association}
\title{Enables a custom detection rule for your account by creating an
association}
\usage{
guardduty_create_custom_detection_rule_association(
  RuleId,
  Mode,
  ClientToken = NULL,
  Tags = NULL
)
}
\arguments{
\item{RuleId}{[required] The unique identifier for the custom detection rule.}

\item{Mode}{[required] The rule execution mode. Valid values: \code{LIVE} | \code{DRY_RUN}.}

\item{ClientToken}{A unique, case-sensitive identifier to ensure that the operation completes no more than one time. Maximum 64 characters.}

\item{Tags}{The tags to be added to the new custom detection rule association resource.}
}
\description{
Enables a custom detection rule for your account by creating an association. You specify the rule and the mode in which it operates.

See \url{https://www.paws-r-sdk.com/docs/guardduty_create_custom_detection_rule_association/} for full documentation.
}
\keyword{internal}
