% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_create_custom_detection_rule_org_configuration}
\alias{guardduty_create_custom_detection_rule_org_configuration}
\title{Creates an organization-level configuration that enables a custom
detection rule across your organization}
\usage{
guardduty_create_custom_detection_rule_org_configuration(
  RuleId,
  Mode,
  IncludeAccountIds = NULL,
  ExcludeAccountIds = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{RuleId}{[required] The unique identifier for the custom detection rule.}

\item{Mode}{[required] The execution mode of the organization configuration. Valid values: \code{LIVE} | \code{DRY_RUN}.}

\item{IncludeAccountIds}{The account IDs to include in the organization configuration. Mutually exclusive with \code{ExcludeAccountIds}.}

\item{ExcludeAccountIds}{The account IDs to exclude from the organization configuration. Mutually exclusive with \code{IncludeAccountIds}.}

\item{ClientToken}{A unique, case-sensitive identifier to ensure that the operation completes no more than one time.}
}
\description{
Creates an organization-level configuration that enables a custom detection rule across your organization. This operation is available only to the delegated administrator account.

See \url{https://www.paws-r-sdk.com/docs/guardduty_create_custom_detection_rule_org_configuration/} for full documentation.
}
\keyword{internal}
