% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_create_investigation}
\alias{guardduty_create_investigation}
\title{This API is currently available as a preview}
\usage{
guardduty_create_investigation(DetectorId, TriggerPrompt, ClientToken = NULL)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the GuardDuty detector for the account in which the investigation is created.

To find the \code{detectorId} in the current Region, see the Settings page in the GuardDuty console, or run the \code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{TriggerPrompt}{[required] A natural-language description of what to investigate. For example:
\itemize{
\item \code{"Investigate finding 1ab2c3d4e5f6a7b8c9d0e1f2a3b4c5d6 in account 123456789012"}
\item \code{"Analyze findings in account with id 123456789012"}
\item \code{"Analyze findings in my organization"}
}}

\item{ClientToken}{The idempotency token for the create request.}
}
\description{
This API is currently available as a preview. During the preview, you can initiate up to 10 investigations per account per day, with a total limit of 100 investigations per account. This feature is available in the following Amazon Web Services Regions: US East (N. Virginia), US East (Ohio), US West (Oregon), Canada (Central), Europe (Frankfurt), Europe (Ireland), Europe (London), Europe (Paris), Europe (Stockholm), and Asia Pacific (Tokyo).

See \url{https://www.paws-r-sdk.com/docs/guardduty_create_investigation/} for full documentation.
}
\keyword{internal}
