% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_create_sample_findings}
\alias{guardduty_create_sample_findings}
\title{Generates sample findings of types specified by the list of finding
types}
\usage{
guardduty_create_sample_findings(DetectorId, FindingTypes = NULL)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector for which you need to create sample findings.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{FindingTypes}{The types of sample findings to generate.}
}
\description{
Generates sample findings of types specified by the list of finding types. If 'NULL' is specified for \code{findingTypes}, the API generates sample findings of all supported finding types.

See \url{https://www.paws-r-sdk.com/docs/guardduty_create_sample_findings/} for full documentation.
}
\keyword{internal}
