% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_create_threat_entity_set}
\alias{guardduty_create_threat_entity_set}
\title{Creates a new threat entity set}
\usage{
guardduty_create_threat_entity_set(
  DetectorId,
  Name,
  Format,
  Location,
  ExpectedBucketOwner = NULL,
  Activate,
  ClientToken = NULL,
  Tags = NULL
)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty account for which you
want to create a threat entity set.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{Name}{[required] A user-friendly name to identify the threat entity set.

The name of your list can include lowercase letters, uppercase letters,
numbers, dash (-), and underscore (_).}

\item{Format}{[required] The format of the file that contains the threat entity set.}

\item{Location}{[required] The URI of the file that contains the threat entity set. The format of
the \code{Location} URL must be a valid Amazon S3 URL format. Invalid URL
formats will result in an error, regardless of whether you activate the
entity set or not. For more information about format of the location
URLs, see \href{https://docs.aws.amazon.com/guardduty/latest/ug/guardduty-lists-create-activate.html}{Format of location URL under Step 2: Adding trusted or threat intelligence data}
in the \emph{Amazon GuardDuty User Guide}.}

\item{ExpectedBucketOwner}{The Amazon Web Services account ID that owns the Amazon S3 bucket
specified in the \strong{location} parameter.}

\item{Activate}{[required] A boolean value that indicates whether GuardDuty should start using the
uploaded threat entity set to generate findings.}

\item{ClientToken}{The idempotency token for the create request.}

\item{Tags}{The tags to be added to a new threat entity set resource.}
}
\description{
Creates a new threat entity set. In a threat entity set, you can provide known malicious IP addresses and domains for your Amazon Web Services environment. GuardDuty generates findings based on the entries in the threat entity sets. Only users of the administrator account can manage entity sets, which automatically apply to member accounts.

See \url{https://www.paws-r-sdk.com/docs/guardduty_create_threat_entity_set/} for full documentation.
}
\keyword{internal}
