% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_custom_detection_rule_org_configuration}
\alias{guardduty_delete_custom_detection_rule_org_configuration}
\title{Deletes the organization-level configuration for a custom detection rule}
\usage{
guardduty_delete_custom_detection_rule_org_configuration(RuleId, Mode)
}
\arguments{
\item{RuleId}{[required] The unique identifier for the custom detection rule.}

\item{Mode}{[required] The execution mode of the organization configuration to delete. Valid values: \code{LIVE} | \code{DRY_RUN}.}
}
\description{
Deletes the organization-level configuration for a custom detection rule. This operation is available only to the delegated administrator account.

See \url{https://www.paws-r-sdk.com/docs/guardduty_delete_custom_detection_rule_org_configuration/} for full documentation.
}
\keyword{internal}
