% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_invitations}
\alias{guardduty_delete_invitations}
\title{Deletes invitations sent to the current member account by Amazon Web
Services accounts specified by their account IDs}
\usage{
guardduty_delete_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{[required] A list of account IDs of the Amazon Web Services accounts that sent
invitations to the current member account that you want to delete
invitations from.}
}
\description{
Deletes invitations sent to the current member account by Amazon Web Services accounts specified by their account IDs.

See \url{https://www.paws-r-sdk.com/docs/guardduty_delete_invitations/} for full documentation.
}
\keyword{internal}
