% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_malware_protection_plan}
\alias{guardduty_delete_malware_protection_plan}
\title{Deletes the Malware Protection plan ID associated with the Malware
Protection plan resource}
\usage{
guardduty_delete_malware_protection_plan(MalwareProtectionPlanId)
}
\arguments{
\item{MalwareProtectionPlanId}{[required] A unique identifier associated with Malware Protection plan resource.}
}
\description{
Deletes the Malware Protection plan ID associated with the Malware Protection plan resource. Use this API only when you no longer want to protect the resource associated with this Malware Protection plan ID.

See \url{https://www.paws-r-sdk.com/docs/guardduty_delete_malware_protection_plan/} for full documentation.
}
\keyword{internal}
