% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_threat_intel_set}
\alias{guardduty_delete_threat_intel_set}
\title{Deletes the ThreatIntelSet specified by the ThreatIntelSet ID}
\usage{
guardduty_delete_threat_intel_set(DetectorId, ThreatIntelSetId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that the threatIntelSet is associated
with.}

\item{ThreatIntelSetId}{[required] The unique ID of the threatIntelSet that you want to delete.}
}
\description{
Deletes the ThreatIntelSet specified by the ThreatIntelSet ID.

See \url{https://www.paws-r-sdk.com/docs/guardduty_delete_threat_intel_set/} for full documentation.
}
\keyword{internal}
