% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_trusted_entity_set}
\alias{guardduty_delete_trusted_entity_set}
\title{Deletes the trusted entity set that is associated with the specified
trustedEntitySetId}
\usage{
guardduty_delete_trusted_entity_set(DetectorId, TrustedEntitySetId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector associated with the trusted entity set
resource.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{TrustedEntitySetId}{[required] The unique ID that helps GuardDuty identify which trusted entity set
needs to be deleted.}
}
\description{
Deletes the trusted entity set that is associated with the specified \code{trustedEntitySetId}.

See \url{https://www.paws-r-sdk.com/docs/guardduty_delete_trusted_entity_set/} for full documentation.
}
\keyword{internal}
