% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_disable_organization_admin_account}
\alias{guardduty_disable_organization_admin_account}
\title{Disables an Amazon Web Services account within the Organization as the
GuardDuty delegated administrator}
\usage{
guardduty_disable_organization_admin_account(AdminAccountId)
}
\arguments{
\item{AdminAccountId}{[required] The Amazon Web Services Account ID for the organizations account to be
disabled as a GuardDuty delegated administrator.}
}
\description{
Disables an Amazon Web Services account within the Organization as the GuardDuty delegated administrator.

See \url{https://www.paws-r-sdk.com/docs/guardduty_disable_organization_admin_account/} for full documentation.
}
\keyword{internal}
