% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_disassociate_members}
\alias{guardduty_disassociate_members}
\title{Disassociates GuardDuty member accounts (to the current GuardDuty
administrator account) specified by the account IDs}
\usage{
guardduty_disassociate_members(DetectorId, AccountIds)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty account whose members you
want to disassociate from the administrator account.}

\item{AccountIds}{[required] A list of account IDs of the GuardDuty member accounts that you want to
disassociate from the administrator account.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UnprocessedAccounts = list(
    list(
      AccountId = "string",
      Result = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Disassociates GuardDuty member accounts (to the current GuardDuty
administrator account) specified by the account IDs.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disassociate_members(
  DetectorId = "string",
  AccountIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
