% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_enable_organization_admin_account}
\alias{guardduty_enable_organization_admin_account}
\title{Enables an Amazon Web Services account within the organization as the
GuardDuty delegated administrator}
\usage{
guardduty_enable_organization_admin_account(AdminAccountId)
}
\arguments{
\item{AdminAccountId}{[required] The Amazon Web Services Account ID for the organization account to be
enabled as a GuardDuty delegated administrator.}
}
\description{
Enables an Amazon Web Services account within the organization as the GuardDuty delegated administrator.

See \url{https://www.paws-r-sdk.com/docs/guardduty_enable_organization_admin_account/} for full documentation.
}
\keyword{internal}
