% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_coverage_statistics}
\alias{guardduty_get_coverage_statistics}
\title{Retrieves aggregated statistics for your account}
\usage{
guardduty_get_coverage_statistics(
  DetectorId,
  FilterCriteria = NULL,
  StatisticsType
)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the GuardDuty detector associated to the coverage
statistics.}

\item{FilterCriteria}{Represents the criteria used to filter the coverage statistics}

\item{StatisticsType}{[required] Represents the statistics type used to aggregate the coverage details.}
}
\description{
Retrieves aggregated statistics for your account. If you are a GuardDuty administrator, you can retrieve the statistics for all the resources associated with the active member accounts in your organization who have enabled Runtime Monitoring and have the GuardDuty security agent running on their resources.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_coverage_statistics/} for full documentation.
}
\keyword{internal}
