% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_custom_detection_rule}
\alias{guardduty_get_custom_detection_rule}
\title{Returns details for a custom detection rule in GuardDuty, including its
detection logic}
\usage{
guardduty_get_custom_detection_rule(RuleId)
}
\arguments{
\item{RuleId}{[required] The unique identifier for the custom detection rule.}
}
\description{
Returns details for a custom detection rule in GuardDuty, including its detection logic.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_custom_detection_rule/} for full documentation.
}
\keyword{internal}
