% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_custom_detection_rule_association}
\alias{guardduty_get_custom_detection_rule_association}
\title{Returns details for a custom detection rule association}
\usage{
guardduty_get_custom_detection_rule_association(RuleId, AssociationId)
}
\arguments{
\item{RuleId}{[required] The unique identifier for the custom detection rule.}

\item{AssociationId}{[required] The unique identifier for the association.}
}
\description{
Returns details for a custom detection rule association.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_custom_detection_rule_association/} for full documentation.
}
\keyword{internal}
