% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_investigation}
\alias{guardduty_get_investigation}
\title{This API is currently available as a preview}
\usage{
guardduty_get_investigation(DetectorId, InvestigationId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the GuardDuty detector associated with the investigation.

To find the \code{detectorId} in the current Region, see the Settings page in the GuardDuty console, or run the \code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{InvestigationId}{[required] The unique identifier of the investigation to retrieve.}
}
\description{
This API is currently available as a preview. This feature is available in the following Amazon Web Services Regions: US East (N. Virginia), US East (Ohio), US West (Oregon), Canada (Central), Europe (Frankfurt), Europe (Ireland), Europe (London), Europe (Paris), Europe (Stockholm), and Asia Pacific (Tokyo).

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_investigation/} for full documentation.
}
\keyword{internal}
