% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_invitations_count}
\alias{guardduty_get_invitations_count}
\title{Returns the count of all GuardDuty membership invitations that were sent
to the current member account except the currently accepted invitation}
\usage{
guardduty_get_invitations_count()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  InvitationsCount = 123
)
}\if{html}{\out{</div>}}
}
\description{
Returns the count of all GuardDuty membership invitations that were sent
to the current member account except the currently accepted invitation.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_invitations_count()
}\if{html}{\out{</div>}}
}

\keyword{internal}
