% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_malware_protection_plan}
\alias{guardduty_get_malware_protection_plan}
\title{Retrieves the Malware Protection plan details associated with a Malware
Protection plan ID}
\usage{
guardduty_get_malware_protection_plan(MalwareProtectionPlanId)
}
\arguments{
\item{MalwareProtectionPlanId}{[required] A unique identifier associated with Malware Protection plan resource.}
}
\description{
Retrieves the Malware Protection plan details associated with a Malware Protection plan ID.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_malware_protection_plan/} for full documentation.
}
\keyword{internal}
