% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_malware_scan}
\alias{guardduty_get_malware_scan}
\title{Retrieves the detailed information for a specific malware scan}
\usage{
guardduty_get_malware_scan(ScanId)
}
\arguments{
\item{ScanId}{[required] A unique identifier that gets generated when you invoke the API without
any error. Each malware scan has a corresponding scan ID. Using this
scan ID, you can monitor the status of your malware scan.}
}
\description{
Retrieves the detailed information for a specific malware scan. Each member account can view the malware scan details for their own account. An administrator can view malware scan details for all accounts in the organization.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_malware_scan/} for full documentation.
}
\keyword{internal}
