% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_malware_scan_settings}
\alias{guardduty_get_malware_scan_settings}
\title{Returns the details of the malware scan settings}
\usage{
guardduty_get_malware_scan_settings(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that is associated with this scan.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}
}
\description{
Returns the details of the malware scan settings.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_malware_scan_settings/} for full documentation.
}
\keyword{internal}
