% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_member_detectors}
\alias{guardduty_get_member_detectors}
\title{Describes which data sources are enabled for the member account's
detector}
\usage{
guardduty_get_member_detectors(DetectorId, AccountIds)
}
\arguments{
\item{DetectorId}{[required] The detector ID for the administrator account.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{AccountIds}{[required] A list of member account IDs.}
}
\description{
Describes which data sources are enabled for the member account's detector.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_member_detectors/} for full documentation.
}
\keyword{internal}
