% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_organization_statistics}
\alias{guardduty_get_organization_statistics}
\title{Retrieves how many active member accounts have each feature enabled
within GuardDuty}
\usage{
guardduty_get_organization_statistics()
}
\description{
Retrieves how many active member accounts have each feature enabled within GuardDuty. Only a delegated GuardDuty administrator of an organization can run this API.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_organization_statistics/} for full documentation.
}
\keyword{internal}
