% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_threat_intel_set}
\alias{guardduty_get_threat_intel_set}
\title{Retrieves the ThreatIntelSet that is specified by the ThreatIntelSet ID}
\usage{
guardduty_get_threat_intel_set(DetectorId, ThreatIntelSetId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that the threatIntelSet is associated
with.}

\item{ThreatIntelSetId}{[required] The unique ID of the threatIntelSet that you want to get.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Name = "string",
  Format = "TXT"|"STIX"|"OTX_CSV"|"ALIEN_VAULT"|"PROOF_POINT"|"FIRE_EYE",
  Location = "string",
  Status = "INACTIVE"|"ACTIVATING"|"ACTIVE"|"DEACTIVATING"|"ERROR"|"DELETE_PENDING"|"DELETED",
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the ThreatIntelSet that is specified by the ThreatIntelSet ID.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_threat_intel_set(
  DetectorId = "string",
  ThreatIntelSetId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
