% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_trusted_entity_set}
\alias{guardduty_get_trusted_entity_set}
\title{Retrieves the trusted entity set associated with the specified
trustedEntitySetId}
\usage{
guardduty_get_trusted_entity_set(DetectorId, TrustedEntitySetId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the GuardDuty detector associated with this trusted
entity set.}

\item{TrustedEntitySetId}{[required] The unique ID that helps GuardDuty identify the trusted entity set.}
}
\description{
Retrieves the trusted entity set associated with the specified \code{trustedEntitySetId}.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_trusted_entity_set/} for full documentation.
}
\keyword{internal}
