% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_coverage}
\alias{guardduty_list_coverage}
\title{Lists coverage details for your GuardDuty account}
\usage{
guardduty_list_coverage(
  DetectorId,
  NextToken = NULL,
  MaxResults = NULL,
  FilterCriteria = NULL,
  SortCriteria = NULL
)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector whose coverage details you want to
retrieve.}

\item{NextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. For subsequent calls, use the NextToken value returned from the
previous request to continue listing results after the first page.}

\item{MaxResults}{The maximum number of results to return in the response.}

\item{FilterCriteria}{Represents the criteria used to filter the coverage details.}

\item{SortCriteria}{Represents the criteria used to sort the coverage details.}
}
\description{
Lists coverage details for your GuardDuty account. If you're a GuardDuty administrator, you can retrieve all resources associated with the active member accounts in your organization.

See \url{https://www.paws-r-sdk.com/docs/guardduty_list_coverage/} for full documentation.
}
\keyword{internal}
