% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_custom_detection_rule_associations}
\alias{guardduty_list_custom_detection_rule_associations}
\title{Returns all custom detection rule associations for your account}
\usage{
guardduty_list_custom_detection_rule_associations(
  MaxResults = NULL,
  NextToken = NULL,
  RuleId = NULL,
  Mode = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in a single page. Minimum value of 1, maximum value of 100.}

\item{NextToken}{A pagination token from a previous response. Use this token to retrieve the next page of results.}

\item{RuleId}{The unique identifier for the custom detection rule to filter associations by.}

\item{Mode}{The rule execution mode to filter associations by.}
}
\description{
Returns all custom detection rule associations for your account. You can filter by rule ID and mode.

See \url{https://www.paws-r-sdk.com/docs/guardduty_list_custom_detection_rule_associations/} for full documentation.
}
\keyword{internal}
