% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_custom_detection_rule_org_configurations}
\alias{guardduty_list_custom_detection_rule_org_configurations}
\title{Returns all organization-level configurations for custom detection rules}
\usage{
guardduty_list_custom_detection_rule_org_configurations(
  MaxResults = NULL,
  NextToken = NULL,
  Status = NULL
)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in a single page. Minimum value of 1, maximum value of 100.}

\item{NextToken}{A pagination token from a previous response. Use this token to retrieve the next page of results.}

\item{Status}{The configuration status to filter by.}
}
\description{
Returns all organization-level configurations for custom detection rules. You can filter the results by status.

See \url{https://www.paws-r-sdk.com/docs/guardduty_list_custom_detection_rule_org_configurations/} for full documentation.
}
\keyword{internal}
