% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_investigations}
\alias{guardduty_list_investigations}
\title{This API is currently available as a preview}
\usage{
guardduty_list_investigations(
  DetectorId,
  SortCriteria = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the GuardDuty detector whose investigations you want to list.

To find the \code{detectorId} in the current Region, see the Settings page in the GuardDuty console, or run the \code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{SortCriteria}{Represents the criteria used for sorting investigations.}

\item{MaxResults}{You can use this parameter to indicate the maximum number of items you want in the response. The default value is 50.}

\item{NextToken}{You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.}
}
\description{
This API is currently available as a preview. This feature is available in the following Amazon Web Services Regions: US East (N. Virginia), US East (Ohio), US West (Oregon), Canada (Central), Europe (Frankfurt), Europe (Ireland), Europe (London), Europe (Paris), Europe (Stockholm), and Asia Pacific (Tokyo).

See \url{https://www.paws-r-sdk.com/docs/guardduty_list_investigations/} for full documentation.
}
\keyword{internal}
