% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_ip_sets}
\alias{guardduty_list_ip_sets}
\title{Lists the IPSets of the GuardDuty service specified by the detector ID}
\usage{
guardduty_list_ip_sets(DetectorId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that the IPSet is associated with.}

\item{MaxResults}{You can use this parameter to indicate the maximum number of items you
want in the response. The default value is 50. The maximum value is 50.}

\item{NextToken}{You can use this parameter when paginating results. Set the value of
this parameter to null on your first call to the list action. For
subsequent calls to the action, fill nextToken in the request with the
value of NextToken from the previous response to continue listing data.}
}
\description{
Lists the IPSets of the GuardDuty service specified by the detector ID. If you use this operation from a member account, the IPSets returned are the IPSets from the associated administrator account.

See \url{https://www.paws-r-sdk.com/docs/guardduty_list_ip_sets/} for full documentation.
}
\keyword{internal}
