% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_malware_protection_plans}
\alias{guardduty_list_malware_protection_plans}
\title{Lists the Malware Protection plan IDs associated with the protected
resources in your Amazon Web Services account}
\usage{
guardduty_list_malware_protection_plans(NextToken = NULL)
}
\arguments{
\item{NextToken}{You can use this parameter when paginating results. Set the value of
this parameter to null on your first call to the list action. For
subsequent calls to the action, fill nextToken in the request with the
value of \code{NextToken} from the previous response to continue listing
data. The default page size is 100 plans.}
}
\description{
Lists the Malware Protection plan IDs associated with the protected resources in your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/guardduty_list_malware_protection_plans/} for full documentation.
}
\keyword{internal}
