% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_malware_scans}
\alias{guardduty_list_malware_scans}
\title{Returns a list of malware scans}
\usage{
guardduty_list_malware_scans(
  MaxResults = NULL,
  NextToken = NULL,
  FilterCriteria = NULL,
  SortCriteria = NULL
)
}
\arguments{
\item{MaxResults}{You can use this parameter to indicate the maximum number of items that
you want in the response. The default value is 50. The maximum value is
50.}

\item{NextToken}{You can use this parameter when paginating results. Set the value of
this parameter to null on your first call to the list action. For
subsequent calls to the action, fill nextToken in the request with the
value of NextToken from the previous response to continue listing
results.}

\item{FilterCriteria}{Represents the criteria used to filter the malware scan entries.}

\item{SortCriteria}{Represents the criteria used for sorting malware scan entries.}
}
\description{
Returns a list of malware scans. Each member account can view the malware scans for their own accounts. An administrator can view the malware scans for all of its members' accounts.

See \url{https://www.paws-r-sdk.com/docs/guardduty_list_malware_scans/} for full documentation.
}
\keyword{internal}
