% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_organization_admin_accounts}
\alias{guardduty_list_organization_admin_accounts}
\title{Lists the accounts configured as GuardDuty delegated administrators}
\usage{
guardduty_list_organization_admin_accounts(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in the response.}

\item{NextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. For subsequent calls, use the \code{NextToken} value returned from
the previous request to continue listing results after the first page.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AdminAccounts = list(
    list(
      AdminAccountId = "string",
      AdminStatus = "ENABLED"|"DISABLE_IN_PROGRESS"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists the accounts configured as GuardDuty delegated administrators.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_organization_admin_accounts(
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
