% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_publishing_destinations}
\alias{guardduty_list_publishing_destinations}
\title{Returns a list of publishing destinations associated with the specified
detectorId}
\usage{
guardduty_list_publishing_destinations(
  DetectorId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DetectorId}{[required] The detector ID for which you want to retrieve the publishing
destination.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{MaxResults}{The maximum number of results to return in the response.}

\item{NextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. For subsequent calls, use the \code{NextToken} value returned from
the previous request to continue listing results after the first page.}
}
\description{
Returns a list of publishing destinations associated with the specified \code{detectorId}.

See \url{https://www.paws-r-sdk.com/docs/guardduty_list_publishing_destinations/} for full documentation.
}
\keyword{internal}
