% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_tags_for_resource}
\alias{guardduty_list_tags_for_resource}
\title{Lists tags for a resource}
\usage{
guardduty_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the given GuardDuty resource.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists tags for a resource. Tagging is currently supported for detectors,
finding filters, IP sets, and threat intel sets, with a limit of 50 tags
per resource. When invoked, this operation returns all assigned tags for
a given resource.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
