% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_threat_entity_sets}
\alias{guardduty_list_threat_entity_sets}
\title{Lists the threat entity sets associated with the specified GuardDuty
detector ID}
\usage{
guardduty_list_threat_entity_sets(
  DetectorId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the GuardDuty detector that is associated with this
threat entity set.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{MaxResults}{You can use this parameter to indicate the maximum number of items you
want in the response. The default value is 50.}

\item{NextToken}{You can use this parameter when paginating results. Set the value of
this parameter to null on your first call to the list action. For
subsequent calls to the action, fill nextToken in the request with the
value of NextToken from the previous response to continue listing data.}
}
\description{
Lists the threat entity sets associated with the specified GuardDuty detector ID. If you use this operation from a member account, the threat entity sets that are returned as a response, belong to the administrator account.

See \url{https://www.paws-r-sdk.com/docs/guardduty_list_threat_entity_sets/} for full documentation.
}
\keyword{internal}
