% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_send_object_malware_scan}
\alias{guardduty_send_object_malware_scan}
\title{Initiates a malware scan for a specific S3 object}
\usage{
guardduty_send_object_malware_scan(S3Object = NULL)
}
\arguments{
\item{S3Object}{The S3 object information for the object you want to scan. The bucket
must have a Malware Protection plan configured to use this API.}
}
\description{
Initiates a malware scan for a specific S3 object. This API allows you to perform on-demand malware scanning of individual objects in S3 buckets that have Malware Protection for S3 enabled.

See \url{https://www.paws-r-sdk.com/docs/guardduty_send_object_malware_scan/} for full documentation.
}
\keyword{internal}
