% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_start_malware_scan}
\alias{guardduty_start_malware_scan}
\title{Initiates the malware scan}
\usage{
guardduty_start_malware_scan(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] Amazon Resource Name (ARN) of the resource for which you invoked the
API.}
}
\description{
Initiates the malware scan. Invoking this API will automatically create the \href{https://docs.aws.amazon.com/guardduty/latest/ug/slr-permissions-malware-protection.html}{Service-linked role} in the corresponding account.

See \url{https://paws-r.github.io/docs/guardduty/start_malware_scan.html} for full documentation.
}
\keyword{internal}
