% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_stop_monitoring_members}
\alias{guardduty_stop_monitoring_members}
\title{Stops GuardDuty monitoring for the specified member accounts}
\usage{
guardduty_stop_monitoring_members(DetectorId, AccountIds)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector associated with the GuardDuty
administrator account that is monitoring member accounts.}

\item{AccountIds}{[required] A list of account IDs for the member accounts to stop monitoring.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UnprocessedAccounts = list(
    list(
      AccountId = "string",
      Result = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Stops GuardDuty monitoring for the specified member accounts. Use the
\code{\link[=guardduty_start_monitoring_members]{start_monitoring_members}}
operation to restart monitoring for those accounts.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$stop_monitoring_members(
  DetectorId = "string",
  AccountIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
