% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_tag_resource}
\alias{guardduty_tag_resource}
\title{Adds tags to a resource}
\usage{
guardduty_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the GuardDuty resource to apply a tag
to.}

\item{Tags}{[required] The tags to be added to a resource.}
}
\description{
Adds tags to a resource.

See \url{https://www.paws-r-sdk.com/docs/guardduty_tag_resource/} for full documentation.
}
\keyword{internal}
