% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_custom_detection_rule_association}
\alias{guardduty_update_custom_detection_rule_association}
\title{Updates the mode of an existing custom detection rule association}
\usage{
guardduty_update_custom_detection_rule_association(RuleId, AssociationId, Mode)
}
\arguments{
\item{RuleId}{[required] The unique identifier for the custom detection rule.}

\item{AssociationId}{[required] The unique identifier for the association to update.}

\item{Mode}{[required] The rule execution mode. Valid values: \code{LIVE} | \code{DRY_RUN}.}
}
\description{
Updates the mode of an existing custom detection rule association.

See \url{https://www.paws-r-sdk.com/docs/guardduty_update_custom_detection_rule_association/} for full documentation.
}
\keyword{internal}
