% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_detector}
\alias{guardduty_update_detector}
\title{Updates the Amazon GuardDuty detector specified by the detectorId}
\usage{
guardduty_update_detector(
  DetectorId,
  Enable = NULL,
  FindingPublishingFrequency = NULL,
  DataSources = NULL
)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector to update.}

\item{Enable}{Specifies whether the detector is enabled or not enabled.}

\item{FindingPublishingFrequency}{An enum value that specifies how frequently findings are exported, such
as to CloudWatch Events.}

\item{DataSources}{Describes which data sources will be updated.}
}
\description{
Updates the Amazon GuardDuty detector specified by the detectorId.

See \url{https://paws-r.github.io/docs/guardduty/update_detector.html} for full documentation.
}
\keyword{internal}
