% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_findings_feedback}
\alias{guardduty_update_findings_feedback}
\title{Marks the specified GuardDuty findings as useful or not useful}
\usage{
guardduty_update_findings_feedback(
  DetectorId,
  FindingIds,
  Feedback,
  Comments = NULL
)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector that is associated with the findings for which
you want to update the feedback.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{FindingIds}{[required] The IDs of the findings that you want to mark as useful or not useful.}

\item{Feedback}{[required] The feedback for the finding.}

\item{Comments}{Additional feedback about the GuardDuty findings.}
}
\description{
Marks the specified GuardDuty findings as useful or not useful.

See \url{https://www.paws-r-sdk.com/docs/guardduty_update_findings_feedback/} for full documentation.
}
\keyword{internal}
