% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_ip_set}
\alias{guardduty_update_ip_set}
\title{Updates the IPSet specified by the IPSet ID}
\usage{
guardduty_update_ip_set(DetectorId, IpSetId, Name, Location, Activate)
}
\arguments{
\item{DetectorId}{[required] The detectorID that specifies the GuardDuty service whose IPSet you want
to update.}

\item{IpSetId}{[required] The unique ID that specifies the IPSet that you want to update.}

\item{Name}{The unique ID that specifies the IPSet that you want to update.}

\item{Location}{The updated URI of the file that contains the IPSet. For example:
https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.}

\item{Activate}{The updated Boolean value that specifies whether the IPSet is active or
not.}
}
\value{
An empty list.
}
\description{
Updates the IPSet specified by the IPSet ID.
}
\section{Request syntax}{
\preformatted{svc$update_ip_set(
  DetectorId = "string",
  IpSetId = "string",
  Name = "string",
  Location = "string",
  Activate = TRUE|FALSE
)
}
}

\keyword{internal}
