% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_malware_protection_plan}
\alias{guardduty_update_malware_protection_plan}
\title{Updates an existing Malware Protection plan resource}
\usage{
guardduty_update_malware_protection_plan(
  MalwareProtectionPlanId,
  Role = NULL,
  Actions = NULL,
  ProtectedResource = NULL
)
}
\arguments{
\item{MalwareProtectionPlanId}{[required] A unique identifier associated with the Malware Protection plan.}

\item{Role}{Amazon Resource Name (ARN) of the IAM role with permissions to scan and
add tags to the associated protected resource.}

\item{Actions}{Information about whether the tags will be added to the S3 object after
scanning.}

\item{ProtectedResource}{Information about the protected resource that is associated with the
created Malware Protection plan. Presently, \code{S3Bucket} is the only
supported protected resource.}
}
\description{
Updates an existing Malware Protection plan resource.

See \url{https://www.paws-r-sdk.com/docs/guardduty_update_malware_protection_plan/} for full documentation.
}
\keyword{internal}
