% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_malware_scan_settings}
\alias{guardduty_update_malware_scan_settings}
\title{Updates the malware scan settings}
\usage{
guardduty_update_malware_scan_settings(
  DetectorId,
  ScanResourceCriteria = NULL,
  EbsSnapshotPreservation = NULL
)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that specifies the GuardDuty service where
you want to update scan settings.}

\item{ScanResourceCriteria}{Represents the criteria to be used in the filter for selecting resources
to scan.}

\item{EbsSnapshotPreservation}{An enum value representing possible snapshot preservations.}
}
\description{
Updates the malware scan settings.

See \url{https://paws-r.github.io/docs/guardduty/update_malware_scan_settings.html} for full documentation.
}
\keyword{internal}
