% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_member_detectors}
\alias{guardduty_update_member_detectors}
\title{Contains information on member accounts to be updated}
\usage{
guardduty_update_member_detectors(
  DetectorId,
  AccountIds,
  DataSources = NULL,
  Features = NULL
)
}
\arguments{
\item{DetectorId}{[required] The detector ID of the administrator account.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{AccountIds}{[required] A list of member account IDs to be updated.}

\item{DataSources}{Describes which data sources will be updated.}

\item{Features}{A list of features that will be updated for the specified member
accounts.}
}
\description{
Contains information on member accounts to be updated.

See \url{https://www.paws-r-sdk.com/docs/guardduty_update_member_detectors/} for full documentation.
}
\keyword{internal}
