% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_threat_entity_set}
\alias{guardduty_update_threat_entity_set}
\title{Updates the threat entity set associated with the specified
threatEntitySetId}
\usage{
guardduty_update_threat_entity_set(
  DetectorId,
  ThreatEntitySetId,
  Name = NULL,
  Location = NULL,
  ExpectedBucketOwner = NULL,
  Activate = NULL
)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the GuardDuty detector associated with the threat
entity set that you want to update.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{ThreatEntitySetId}{[required] The ID returned by GuardDuty after updating the threat entity set
resource.}

\item{Name}{A user-friendly name to identify the trusted entity set.

The name of your list can include lowercase letters, uppercase letters,
numbers, dash (-), and underscore (_).}

\item{Location}{The URI of the file that contains the trusted entity set.}

\item{ExpectedBucketOwner}{The Amazon Web Services account ID that owns the Amazon S3 bucket
specified in the \strong{location} parameter.}

\item{Activate}{A boolean value that indicates whether GuardDuty is to start using this
updated threat entity set. After you update an entity set, you will need
to activate it again. It might take up to 15 minutes for the updated
entity set to be effective.}
}
\description{
Updates the threat entity set associated with the specified \code{threatEntitySetId}.

See \url{https://www.paws-r-sdk.com/docs/guardduty_update_threat_entity_set/} for full documentation.
}
\keyword{internal}
