% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_accept_delegation_request}
\alias{iam_accept_delegation_request}
\title{Accepts a delegation request, granting the requested temporary access}
\usage{
iam_accept_delegation_request(DelegationRequestId)
}
\arguments{
\item{DelegationRequestId}{[required] The unique identifier of the delegation request to accept.}
}
\description{
Accepts a delegation request, granting the requested temporary access.

See \url{https://www.paws-r-sdk.com/docs/iam_accept_delegation_request/} for full documentation.
}
\keyword{internal}
