% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_acquire_role}
\alias{iam_acquire_role}
\title{Creates an IAM role from the specified role template}
\usage{
iam_acquire_role(
  TemplateArn,
  TemplateMinorVersion = NULL,
  ReplacementValues = NULL
)
}
\arguments{
\item{TemplateArn}{[required] The Amazon Resource Name (ARN) of the role template to create the role from.

For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html?refid=e072b57e-c748-4712-a675-b2b5d0ef7f14&trkcampaign=tw-webinar-series}{Amazon Resource Names (ARNs)} in the \emph{Amazon Web Services General Reference}.}

\item{TemplateMinorVersion}{The minor version of the role template to use. If you do not specify a minor version, the service uses the template's default minor version.}

\item{ReplacementValues}{A map of values to substitute for the parameters that are defined in the role template version. Each key is a parameter name from the template, and each value is a structure that contains the replacement values for that parameter.}
}
\description{
Creates an IAM role from the specified role template. The new role takes its configuration—including its name, path, trust policy, inline and managed policies, permissions boundary, tags, and maximum session duration—from the role template version that you specify. For more information about roles, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles.html}{IAM roles} in the \emph{IAM User Guide}.

See \url{https://www.paws-r-sdk.com/docs/iam_acquire_role/} for full documentation.
}
\keyword{internal}
