% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_add_user_to_group}
\alias{iam_add_user_to_group}
\title{Adds the specified user to the specified group}
\usage{
iam_add_user_to_group(GroupName, UserName)
}
\arguments{
\item{GroupName}{[required] The name of the group to update.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: \\_+=,.@-}

\item{UserName}{[required] The name of the user to add.

This parameter allows (through its \href{http://wikipedia.org/wiki/regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: \\_+=,.@-}
}
\description{
Adds the specified user to the specified group.
}
\section{Request syntax}{
\preformatted{svc$add_user_to_group(
  GroupName = "string",
  UserName = "string"
)
}
}

\examples{
# The following command adds an IAM user named Bob to the IAM group named
# Admins:
\dontrun{svc$add_user_to_group(
  GroupName = "Admins",
  UserName = "Bob"
)}

}
\keyword{internal}
