% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_associate_delegation_request}
\alias{iam_associate_delegation_request}
\title{Associates a delegation request with the current identity}
\usage{
iam_associate_delegation_request(DelegationRequestId)
}
\arguments{
\item{DelegationRequestId}{[required] The unique identifier of the delegation request to associate.}
}
\description{
Associates a delegation request with the current identity.

See \url{https://www.paws-r-sdk.com/docs/iam_associate_delegation_request/} for full documentation.
}
\keyword{internal}
