% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_change_password}
\alias{iam_change_password}
\title{Changes the password of the IAM user who is calling this operation}
\usage{
iam_change_password(OldPassword, NewPassword)
}
\arguments{
\item{OldPassword}{[required] The IAM user's current password.}

\item{NewPassword}{[required] The new password. The new password must conform to the AWS account's
password policy, if one exists.

The \href{https://en.wikipedia.org/wiki/Regex}{regex pattern} that is used to
validate this parameter is a string of characters. That string can
include almost any printable ASCII character from the space (\verb{\\u0020})
through the end of the ASCII character range (\verb{\\u00FF}). You can also
include the tab (\verb{\\u0009}), line feed (\verb{\\u000A}), and carriage return
(\verb{\\u000D}) characters. Any of these characters are valid in a password.
However, many tools, such as the AWS Management Console, might restrict
the ability to type certain characters because they have special meaning
within that tool.}
}
\value{
An empty list.
}
\description{
Changes the password of the IAM user who is calling this operation. This
operation can be performed using the AWS CLI, the AWS API, or the \strong{My
Security Credentials} page in the AWS Management Console. The AWS
account root user password is not affected by this operation.

Use \code{\link[=iam_update_login_profile]{update_login_profile}} to use the AWS
CLI, the AWS API, or the \strong{Users} page in the IAM console to change the
password for any IAM user. For more information about modifying
passwords, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_passwords_admin-change-user.html}{Managing passwords}
in the \emph{IAM User Guide}.
}
\section{Request syntax}{
\preformatted{svc$change_password(
  OldPassword = "string",
  NewPassword = "string"
)
}
}

\examples{
\dontrun{
# The following command changes the password for the current IAM user.
svc$change_password(
  NewPassword = "]35d/\{pB9Fo9wJ",
  OldPassword = "3s0K_;xh4~8XXI"
)
}

}
\keyword{internal}
