% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_create_service_specific_credential}
\alias{iam_create_service_specific_credential}
\title{Generates a set of credentials consisting of a user name and password
that can be used to access the service specified in the request}
\usage{
iam_create_service_specific_credential(
  UserName,
  ServiceName,
  CredentialAgeDays = NULL
)
}
\arguments{
\item{UserName}{[required] The name of the IAM user that is to be associated with the credentials. The new service-specific credentials have the same permissions as the associated user except that they can be used only to access the specified service.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-}

\item{ServiceName}{[required] The name of the Amazon Web Services service that is to be associated with the credentials. The service you specify here is the only service that can be accessed using these credentials.}

\item{CredentialAgeDays}{The number of days until the service specific credential expires. This field is only valid for services that support long-term API keys and must be a positive integer. When not specified, the credential will not expire.

To see which services support long-term API keys, refer to \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_api_keys_for_aws_services.html}{API keys for Amazon Web Services services} in the \emph{IAM User Guide}.}
}
\description{
Generates a set of credentials consisting of a user name and password that can be used to access the service specified in the request. These credentials are generated by IAM, and can be used only for the specified service.

See \url{https://www.paws-r-sdk.com/docs/iam_create_service_specific_credential/} for full documentation.
}
\keyword{internal}
