% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_group}
\alias{iam_delete_group}
\title{Deletes the specified IAM group}
\usage{
iam_delete_group(GroupName)
}
\arguments{
\item{GroupName}{[required] The name of the IAM group to delete.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Deletes the specified IAM group. The group must not contain any users or have any attached policies.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_group/} for full documentation.
}
\keyword{internal}
