% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_login_profile}
\alias{iam_delete_login_profile}
\title{Deletes the password for the specified IAM user, which terminates the
user's ability to access AWS services through the AWS Management Console}
\usage{
iam_delete_login_profile(UserName)
}
\arguments{
\item{UserName}{[required] The name of the user whose password you want to delete.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\value{
An empty list.
}
\description{
Deletes the password for the specified IAM user, which terminates the
user's ability to access AWS services through the AWS Management
Console.

You can use the AWS CLI, the AWS API, or the \strong{Users} page in the IAM
console to delete a password for any IAM user. You can use
\code{\link[=iam_change_password]{change_password}} to update, but not delete, your
own password in the \strong{My Security Credentials} page in the AWS
Management Console.

Deleting a user's password does not prevent a user from accessing AWS
through the command line interface or the API. To prevent all user
access, you must also either make any access keys inactive or delete
them. For more information about making keys inactive or deleting them,
see \code{\link[=iam_update_access_key]{update_access_key}} and
\code{\link[=iam_delete_access_key]{delete_access_key}}.
}
\section{Request syntax}{
\preformatted{svc$delete_login_profile(
  UserName = "string"
)
}
}

\examples{
\dontrun{
# The following command deletes the password for the IAM user named Bob.
svc$delete_login_profile(
  UserName = "Bob"
)
}

}
\keyword{internal}
