% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_role_permissions_boundary}
\alias{iam_delete_role_permissions_boundary}
\title{Deletes the permissions boundary for the specified IAM role}
\usage{
iam_delete_role_permissions_boundary(RoleName)
}
\arguments{
\item{RoleName}{[required] The name (friendly name, not ARN) of the IAM role from which you want to
remove the permissions boundary.}
}
\value{
An empty list.
}
\description{
Deletes the permissions boundary for the specified IAM role.

Deleting the permissions boundary for a role might increase its
permissions. For example, it might allow anyone who assumes the role to
perform all the actions granted in its permissions policies.
}
\section{Request syntax}{
\preformatted{svc$delete_role_permissions_boundary(
  RoleName = "string"
)
}
}

\keyword{internal}
