% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_saml_provider}
\alias{iam_delete_saml_provider}
\title{Deletes a SAML provider resource in IAM}
\usage{
iam_delete_saml_provider(SAMLProviderArn)
}
\arguments{
\item{SAMLProviderArn}{[required] The Amazon Resource Name (ARN) of the SAML provider to delete.}
}
\description{
Deletes a SAML provider resource in IAM.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_saml_provider/} for full documentation.
}
\keyword{internal}
