% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_user_permissions_boundary}
\alias{iam_delete_user_permissions_boundary}
\title{Deletes the permissions boundary for the specified IAM user}
\usage{
iam_delete_user_permissions_boundary(UserName)
}
\arguments{
\item{UserName}{[required] The name (friendly name, not ARN) of the IAM user from which you want to
remove the permissions boundary.}
}
\value{
An empty list.
}
\description{
Deletes the permissions boundary for the specified IAM user.

Deleting the permissions boundary for a user might increase its
permissions by allowing the user to perform all the actions granted in
its permissions policies.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_user_permissions_boundary(
  UserName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
