% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_virtual_mfa_device}
\alias{iam_delete_virtual_mfa_device}
\title{Deletes a virtual MFA device}
\usage{
iam_delete_virtual_mfa_device(SerialNumber)
}
\arguments{
\item{SerialNumber}{[required] The serial number that uniquely identifies the MFA device. For virtual
MFA devices, the serial number is the same as the ARN.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: =,.@:/-}
}
\description{
Deletes a virtual MFA device.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_virtual_mfa_device/} for full documentation.
}
\keyword{internal}
