% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_disable_organizations_root_credentials_management}
\alias{iam_disable_organizations_root_credentials_management}
\title{Disables the management of privileged root user credentials across
member accounts in your organization}
\usage{
iam_disable_organizations_root_credentials_management()
}
\description{
Disables the management of privileged root user credentials across member accounts in your organization. When you disable this feature, the management account and the delegated administrator for IAM can no longer manage root user credentials for member accounts in your organization.

See \url{https://www.paws-r-sdk.com/docs/iam_disable_organizations_root_credentials_management/} for full documentation.
}
\keyword{internal}
