% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_disable_organizations_root_sessions}
\alias{iam_disable_organizations_root_sessions}
\title{Disables root user sessions for privileged tasks across member accounts
in your organization}
\usage{
iam_disable_organizations_root_sessions()
}
\description{
Disables root user sessions for privileged tasks across member accounts in your organization. When you disable this feature, the management account and the delegated administrator for IAM can no longer perform privileged tasks on member accounts in your organization.

See \url{https://www.paws-r-sdk.com/docs/iam_disable_organizations_root_sessions/} for full documentation.
}
\keyword{internal}
