% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_disable_outbound_web_identity_federation}
\alias{iam_disable_outbound_web_identity_federation}
\title{Disables the outbound identity federation feature for your Amazon Web
Services account}
\usage{
iam_disable_outbound_web_identity_federation()
}
\description{
Disables the outbound identity federation feature for your Amazon Web Services account. When disabled, IAM principals in the account cannot use the \code{GetWebIdentityToken} API to obtain JSON Web Tokens (JWTs) for authentication with external services. This operation does not affect tokens that were issued before the feature was disabled.

See \url{https://www.paws-r-sdk.com/docs/iam_disable_outbound_web_identity_federation/} for full documentation.
}
\keyword{internal}
