% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_enable_organizations_root_credentials_management}
\alias{iam_enable_organizations_root_credentials_management}
\title{Enables the management of privileged root user credentials across member
accounts in your organization}
\usage{
iam_enable_organizations_root_credentials_management()
}
\description{
Enables the management of privileged root user credentials across member accounts in your organization. When you enable root credentials management for \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_root-user.html#id_root-user-access-management}{centralized root access}, the management account and the delegated administrator for IAM can manage root user credentials for member accounts in your organization.

See \url{https://www.paws-r-sdk.com/docs/iam_enable_organizations_root_credentials_management/} for full documentation.
}
\keyword{internal}
