% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_enable_organizations_root_sessions}
\alias{iam_enable_organizations_root_sessions}
\title{Allows the management account or delegated administrator to perform
privileged tasks on member accounts in your organization}
\usage{
iam_enable_organizations_root_sessions()
}
\description{
Allows the management account or delegated administrator to perform privileged tasks on member accounts in your organization. For more information, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_root-user.html#id_root-user-access-management}{Centrally manage root access for member accounts} in the \emph{Identity and Access Management User Guide}.

See \url{https://www.paws-r-sdk.com/docs/iam_enable_organizations_root_sessions/} for full documentation.
}
\keyword{internal}
