% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_enable_outbound_web_identity_federation}
\alias{iam_enable_outbound_web_identity_federation}
\title{Enables the outbound identity federation feature for your Amazon Web
Services account}
\usage{
iam_enable_outbound_web_identity_federation()
}
\description{
Enables the outbound identity federation feature for your Amazon Web Services account. When enabled, IAM principals in your account can use the \code{GetWebIdentityToken} API to obtain JSON Web Tokens (JWTs) for secure authentication with external services. This operation also generates a unique issuer URL for your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/iam_enable_outbound_web_identity_federation/} for full documentation.
}
\keyword{internal}
