% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_generate_credential_report}
\alias{iam_generate_credential_report}
\title{Generates a credential report for the AWS account}
\usage{
iam_generate_credential_report()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  State = "STARTED"|"INPROGRESS"|"COMPLETE",
  Description = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Generates a credential report for the AWS account. For more information
about the credential report, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_getting-report.html}{Getting Credential Reports}
in the \emph{IAM User Guide}.
}
\keyword{internal}
