% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_access_key_last_used}
\alias{iam_get_access_key_last_used}
\title{Retrieves information about when the specified access key was last used}
\usage{
iam_get_access_key_last_used(AccessKeyId)
}
\arguments{
\item{AccessKeyId}{[required] The identifier of an access key.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
that can consist of any upper or lowercased letter or digit.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UserName = "string",
  AccessKeyLastUsed = list(
    LastUsedDate = as.POSIXct(
      "2015-01-01"
    ),
    ServiceName = "string",
    Region = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves information about when the specified access key was last used.
The information includes the date and time of last use, along with the
AWS service and Region that were specified in the last request made with
that key.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_access_key_last_used(
  AccessKeyId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
