% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_account_properties}
\alias{iam_get_account_properties}
\title{Retrieves the account-level properties for the caller's Amazon Web
Services account}
\usage{
iam_get_account_properties()
}
\description{
Retrieves the account-level properties for the caller's Amazon Web Services account. Account properties are configuration settings that control account-wide IAM features such as Role Manager.

See \url{https://www.paws-r-sdk.com/docs/iam_get_account_properties/} for full documentation.
}
\keyword{internal}
