% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_delegation_request}
\alias{iam_get_delegation_request}
\title{Retrieves information about a specific delegation request}
\usage{
iam_get_delegation_request(
  DelegationRequestId,
  DelegationPermissionCheck = NULL
)
}
\arguments{
\item{DelegationRequestId}{[required] The unique identifier of the delegation request to retrieve.}

\item{DelegationPermissionCheck}{Specifies whether to perform a permission check for the delegation
request.

If set to true, the
\code{\link[=iam_get_delegation_request]{get_delegation_request}} API call will
start a permission check process. This process calculates whether the
caller has sufficient permissions to cover the asks from this delegation
request.

Setting this parameter to true does not guarantee an answer in the
response. See the \code{PermissionCheckStatus} and the
\code{PermissionCheckResult} response attributes for further details.}
}
\description{
Retrieves information about a specific delegation request.

See \url{https://www.paws-r-sdk.com/docs/iam_get_delegation_request/} for full documentation.
}
\keyword{internal}
