% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_mfa_device}
\alias{iam_get_mfa_device}
\title{Retrieves information about an MFA device for a specified user}
\usage{
iam_get_mfa_device(SerialNumber, UserName = NULL)
}
\arguments{
\item{SerialNumber}{[required] Serial number that uniquely identifies the MFA device. For this API, we
only accept FIDO security key
\href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{ARNs}.}

\item{UserName}{The friendly name identifying the user.}
}
\description{
Retrieves information about an MFA device for a specified user.

See \url{https://www.paws-r-sdk.com/docs/iam_get_mfa_device/} for full documentation.
}
\keyword{internal}
