% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_outbound_web_identity_federation_info}
\alias{iam_get_outbound_web_identity_federation_info}
\title{Retrieves the configuration information for the outbound identity
federation feature in your Amazon Web Services account}
\usage{
iam_get_outbound_web_identity_federation_info()
}
\description{
Retrieves the configuration information for the outbound identity federation feature in your Amazon Web Services account. The response includes the unique issuer URL for your Amazon Web Services account and the current enabled/disabled status of the feature. Use this operation to obtain the issuer URL that you need to configure trust relationships with external services.

See \url{https://www.paws-r-sdk.com/docs/iam_get_outbound_web_identity_federation_info/} for full documentation.
}
\keyword{internal}
