% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_service_linked_role_deletion_status}
\alias{iam_get_service_linked_role_deletion_status}
\title{Retrieves the status of your service-linked role deletion}
\usage{
iam_get_service_linked_role_deletion_status(DeletionTaskId)
}
\arguments{
\item{DeletionTaskId}{[required] The deletion task identifier. This identifier is returned by the
\code{\link[=iam_delete_service_linked_role]{delete_service_linked_role}} operation
in the format
\verb{task/aws-service-role/<service-principal-name>/<role-name>/<task-uuid>}.}
}
\description{
Retrieves the status of your service-linked role deletion. After you use \code{\link[=iam_delete_service_linked_role]{delete_service_linked_role}} to submit a service-linked role for deletion, you can use the \code{DeletionTaskId} parameter in \code{\link[=iam_get_service_linked_role_deletion_status]{get_service_linked_role_deletion_status}} to check the status of the deletion. If the deletion fails, this operation returns the reason that it failed, if that information is returned by the service.

See \url{https://www.paws-r-sdk.com/docs/iam_get_service_linked_role_deletion_status/} for full documentation.
}
\keyword{internal}
