% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_user_policy}
\alias{iam_get_user_policy}
\title{Retrieves the specified inline policy document that is embedded in the
specified IAM user}
\usage{
iam_get_user_policy(UserName, PolicyName)
}
\arguments{
\item{UserName}{[required] The name of the user who the policy is associated with.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}

\item{PolicyName}{[required] The name of the policy document to get.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Retrieves the specified inline policy document that is embedded in the specified IAM user.

See \url{https://www.paws-r-sdk.com/docs/iam_get_user_policy/} for full documentation.
}
\keyword{internal}
