% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_list_organizations_features}
\alias{iam_list_organizations_features}
\title{Lists the centralized root access features enabled for your organization}
\usage{
iam_list_organizations_features()
}
\description{
Lists the centralized root access features enabled for your organization. For more information, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_root-user.html#id_root-user-access-management}{Centrally manage root access for member accounts}.

See \url{https://www.paws-r-sdk.com/docs/iam_list_organizations_features/} for full documentation.
}
\keyword{internal}
