% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_reject_delegation_request}
\alias{iam_reject_delegation_request}
\title{Rejects a delegation request, denying the requested temporary access}
\usage{
iam_reject_delegation_request(DelegationRequestId, Notes = NULL)
}
\arguments{
\item{DelegationRequestId}{[required] The unique identifier of the delegation request to reject.}

\item{Notes}{Optional notes explaining the reason for rejecting the delegation
request.}
}
\description{
Rejects a delegation request, denying the requested temporary access.

See \url{https://www.paws-r-sdk.com/docs/iam_reject_delegation_request/} for full documentation.
}
\keyword{internal}
