% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_send_delegation_token}
\alias{iam_send_delegation_token}
\title{Sends the exchange token for an accepted delegation request}
\usage{
iam_send_delegation_token(DelegationRequestId)
}
\arguments{
\item{DelegationRequestId}{[required] The unique identifier of the delegation request for which to send the
token.}
}
\description{
Sends the exchange token for an accepted delegation request.

See \url{https://www.paws-r-sdk.com/docs/iam_send_delegation_token/} for full documentation.
}
\keyword{internal}
