% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_tag_role}
\alias{iam_tag_role}
\title{Adds one or more tags to an IAM role}
\usage{
iam_tag_role(RoleName, Tags)
}
\arguments{
\item{RoleName}{[required] The name of the IAM role to which you want to add tags.

This parameter accepts (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
that consist of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}

\item{Tags}{[required] The list of tags that you want to attach to the IAM role. Each tag
consists of a key name and an associated value.}
}
\value{
An empty list.
}
\description{
Adds one or more tags to an IAM role. The role can be a regular role or
a service-linked role. If a tag with the same key name already exists,
then that tag is overwritten with the new value.

A tag consists of a key name and an associated value. By assigning tags
to your resources, you can do the following:
\itemize{
\item \strong{Administrative grouping and discovery} - Attach tags to resources
to aid in organization and search. For example, you could search for
all resources with the key name \emph{Project} and the value
\emph{MyImportantProject}. Or search for all resources with the key name
\emph{Cost Center} and the value \emph{41200}.
\item \strong{Access control} - Include tags in IAM user-based and
resource-based policies. You can use tags to restrict access to only
an IAM role that has a specified tag attached. You can also restrict
access to only those resources that have a certain tag attached. For
examples of policies that show how to use tags to control access,
see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access_tags.html}{Control access using IAM tags}
in the \emph{IAM User Guide}.
\item \strong{Cost allocation} - Use tags to help track which individuals and
teams are using which AWS resources.
\item If any one of the tags is invalid or if you exceed the allowed
maximum number of tags, then the entire request fails and the
resource is not created. For more information about tagging, see
\href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_tags.html}{Tagging IAM resources}
in the \emph{IAM User Guide}.
\item AWS always interprets the tag \code{Value} as a single string. If you
need to store an array, you can store comma-separated values in the
string. However, you must interpret the value in your code.
}

For more information about tagging, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_tags.html}{Tagging IAM identities}
in the \emph{IAM User Guide}.
}
\section{Request syntax}{
\preformatted{svc$tag_role(
  RoleName = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\examples{
\dontrun{
# The following example shows how to add tags to an existing role.
svc$tag_role(
  RoleName = "taggedrole",
  Tags = list(
    list(
      Key = "Dept",
      Value = "Accounting"
    ),
    list(
      Key = "CostCenter",
      Value = "12345"
    )
  )
)
}

}
\keyword{internal}
