% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_tag_saml_provider}
\alias{iam_tag_saml_provider}
\title{Adds one or more tags to a Security Assertion Markup Language (SAML)
identity provider}
\usage{
iam_tag_saml_provider(SAMLProviderArn, Tags)
}
\arguments{
\item{SAMLProviderArn}{[required] The ARN of the SAML identity provider in IAM to which you want to add
tags.

This parameter accepts (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
that consist of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: =,.@-}

\item{Tags}{[required] The list of tags that you want to attach to the SAML identity provider
in IAM. Each tag consists of a key name and an associated value.}
}
\value{
An empty list.
}
\description{
Adds one or more tags to a Security Assertion Markup Language (SAML)
identity provider. For more information about these providers, see
\href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_providers_saml.html}{About SAML 2.0-based federation}
. If a tag with the same key name already exists, then that tag is
overwritten with the new value.

A tag consists of a key name and an associated value. By assigning tags
to your resources, you can do the following:
\itemize{
\item \strong{Administrative grouping and discovery} - Attach tags to resources
to aid in organization and search. For example, you could search for
all resources with the key name \emph{Project} and the value
\emph{MyImportantProject}. Or search for all resources with the key name
\emph{Cost Center} and the value \emph{41200}.
\item \strong{Access control} - Include tags in IAM user-based and
resource-based policies. You can use tags to restrict access to only
a SAML identity provider that has a specified tag attached. For
examples of policies that show how to use tags to control access,
see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access_tags.html}{Control access using IAM tags}
in the \emph{IAM User Guide}.
\item If any one of the tags is invalid or if you exceed the allowed
maximum number of tags, then the entire request fails and the
resource is not created. For more information about tagging, see
\href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_tags.html}{Tagging IAM resources}
in the \emph{IAM User Guide}.
\item AWS always interprets the tag \code{Value} as a single string. If you
need to store an array, you can store comma-separated values in the
string. However, you must interpret the value in your code.
}
}
\section{Request syntax}{
\preformatted{svc$tag_saml_provider(
  SAMLProviderArn = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
