% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_untag_open_id_connect_provider}
\alias{iam_untag_open_id_connect_provider}
\title{Removes the specified tags from the specified OpenID Connect
(OIDC)-compatible identity provider in IAM}
\usage{
iam_untag_open_id_connect_provider(OpenIDConnectProviderArn, TagKeys)
}
\arguments{
\item{OpenIDConnectProviderArn}{[required] The ARN of the OIDC provider in IAM from which you want to remove tags.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}

\item{TagKeys}{[required] A list of key names as a simple array of strings. The tags with matching
keys are removed from the specified OIDC provider.}
}
\description{
Removes the specified tags from the specified OpenID Connect (OIDC)-compatible identity provider in IAM. For more information about OIDC providers, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_providers_oidc.html}{About web identity federation}. For more information about tagging, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_tags.html}{Tagging IAM resources} in the \emph{IAM User Guide}.

See \url{https://www.paws-r-sdk.com/docs/iam_untag_open_id_connect_provider/} for full documentation.
}
\keyword{internal}
