% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_untag_policy}
\alias{iam_untag_policy}
\title{Removes the specified tags from the customer managed policy}
\usage{
iam_untag_policy(PolicyArn, TagKeys)
}
\arguments{
\item{PolicyArn}{[required] The ARN of the IAM customer managed policy from which you want to remove
tags.

This parameter accepts (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
that consist of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: =,.@-}

\item{TagKeys}{[required] A list of key names as a simple array of strings. The tags with matching
keys are removed from the specified policy.}
}
\value{
An empty list.
}
\description{
Removes the specified tags from the customer managed policy. For more
information about tagging, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_tags.html}{Tagging IAM resources}
in the \emph{IAM User Guide}.
}
\section{Request syntax}{
\preformatted{svc$untag_policy(
  PolicyArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
