% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_untag_server_certificate}
\alias{iam_untag_server_certificate}
\title{Removes the specified tags from the IAM server certificate}
\usage{
iam_untag_server_certificate(ServerCertificateName, TagKeys)
}
\arguments{
\item{ServerCertificateName}{[required] The name of the IAM server certificate from which you want to remove
tags.

This parameter accepts (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
that consist of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: =,.@-}

\item{TagKeys}{[required] A list of key names as a simple array of strings. The tags with matching
keys are removed from the specified IAM server certificate.}
}
\value{
An empty list.
}
\description{
Removes the specified tags from the IAM server certificate. For more
information about tagging, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_tags.html}{Tagging IAM resources}
in the \emph{IAM User Guide}.

For certificates in a Region supported by AWS Certificate Manager (ACM),
we recommend that you don't use IAM server certificates. Instead, use
ACM to provision, manage, and deploy your server certificates. For more
information about IAM server certificates, \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_server-certs.html}{Working with server certificates}
in the \emph{IAM User Guide}.
}
\section{Request syntax}{
\preformatted{svc$untag_server_certificate(
  ServerCertificateName = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
