% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_update_delegation_request}
\alias{iam_update_delegation_request}
\title{Updates an existing delegation request with additional information}
\usage{
iam_update_delegation_request(DelegationRequestId, Notes = NULL)
}
\arguments{
\item{DelegationRequestId}{[required] The unique identifier of the delegation request to update.}

\item{Notes}{Additional notes or comments to add to the delegation request.}
}
\description{
Updates an existing delegation request with additional information. When the delegation request is updated, it reaches the \code{PENDING_APPROVAL} state.

See \url{https://www.paws-r-sdk.com/docs/iam_update_delegation_request/} for full documentation.
}
\keyword{internal}
