% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_delete_attribute_mapping}
\alias{iamrolesanywhere_delete_attribute_mapping}
\title{Delete an entry from the attribute mapping rules enforced by a given
profile}
\usage{
iamrolesanywhere_delete_attribute_mapping(
  certificateField,
  profileId,
  specifiers = NULL
)
}
\arguments{
\item{certificateField}{[required] Fields (x509Subject, x509Issuer and x509SAN) within X.509 certificates.}

\item{profileId}{[required] The unique identifier of the profile.}

\item{specifiers}{A list of specifiers of a certificate field; for example, CN, OU, UID
from a Subject.}
}
\description{
Delete an entry from the attribute mapping rules enforced by a given profile.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_delete_attribute_mapping/} for full documentation.
}
\keyword{internal}
