% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_delete_crl}
\alias{iamrolesanywhere_delete_crl}
\title{Deletes a certificate revocation list (CRL)}
\usage{
iamrolesanywhere_delete_crl(crlId)
}
\arguments{
\item{crlId}{[required] The unique identifier of the certificate revocation list (CRL).}
}
\description{
Deletes a certificate revocation list (CRL).

See \url{https://paws-r.github.io/docs/iamrolesanywhere/delete_crl.html} for full documentation.
}
\keyword{internal}
