% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_disable_crl}
\alias{iamrolesanywhere_disable_crl}
\title{Disables a certificate revocation list (CRL)}
\usage{
iamrolesanywhere_disable_crl(crlId)
}
\arguments{
\item{crlId}{[required] The unique identifier of the certificate revocation list (CRL).}
}
\description{
Disables a certificate revocation list (CRL).

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_disable_crl/} for full documentation.
}
\keyword{internal}
