% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_disable_profile}
\alias{iamrolesanywhere_disable_profile}
\title{Disables a profile}
\usage{
iamrolesanywhere_disable_profile(profileId)
}
\arguments{
\item{profileId}{[required] The unique identifier of the profile.}
}
\description{
Disables a profile. When disabled, \href{https://docs.aws.amazon.com/rolesanywhere/latest/APIReference/}{CreateSession} requests with this profile fail.

See \url{https://paws-r.github.io/docs/iamrolesanywhere/disable_profile.html} for full documentation.
}
\keyword{internal}
