% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_disable_trust_anchor}
\alias{iamrolesanywhere_disable_trust_anchor}
\title{Disables a trust anchor}
\usage{
iamrolesanywhere_disable_trust_anchor(trustAnchorId)
}
\arguments{
\item{trustAnchorId}{[required] The unique identifier of the trust anchor.}
}
\description{
Disables a trust anchor. When disabled, temporary credential requests specifying this trust anchor are unauthorized.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_disable_trust_anchor/} for full documentation.
}
\keyword{internal}
