% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_enable_crl}
\alias{iamrolesanywhere_enable_crl}
\title{Enables a certificate revocation list (CRL)}
\usage{
iamrolesanywhere_enable_crl(crlId)
}
\arguments{
\item{crlId}{[required] The unique identifier of the certificate revocation list (CRL).}
}
\description{
Enables a certificate revocation list (CRL). When enabled, certificates stored in the CRL are unauthorized to receive session credentials.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_enable_crl/} for full documentation.
}
\keyword{internal}
