% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_enable_profile}
\alias{iamrolesanywhere_enable_profile}
\title{Enables temporary credential requests for a profile}
\usage{
iamrolesanywhere_enable_profile(profileId)
}
\arguments{
\item{profileId}{[required] The unique identifier of the profile.}
}
\description{
Enables temporary credential requests for a profile.

See \url{https://paws-r.github.io/docs/iamrolesanywhere/enable_profile.html} for full documentation.
}
\keyword{internal}
