% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_enable_trust_anchor}
\alias{iamrolesanywhere_enable_trust_anchor}
\title{Enables a trust anchor}
\usage{
iamrolesanywhere_enable_trust_anchor(trustAnchorId)
}
\arguments{
\item{trustAnchorId}{[required] The unique identifier of the trust anchor.}
}
\description{
Enables a trust anchor. When enabled, certificates in the trust anchor chain are authorized for trust validation.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_enable_trust_anchor/} for full documentation.
}
\keyword{internal}
