% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_get_profile}
\alias{iamrolesanywhere_get_profile}
\title{Gets a profile}
\usage{
iamrolesanywhere_get_profile(profileId)
}
\arguments{
\item{profileId}{[required] The unique identifier of the profile.}
}
\description{
Gets a profile.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_get_profile/} for full documentation.
}
\keyword{internal}
