% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_get_subject}
\alias{iamrolesanywhere_get_subject}
\title{Gets a Subject}
\usage{
iamrolesanywhere_get_subject(subjectId)
}
\arguments{
\item{subjectId}{[required] The unique identifier of the subject.}
}
\description{
Gets a Subject. A Subject associates a certificate identity with authentication attempts by CreateSession. The Subject resources stores audit information such as status of the last authentication attempt, the certificate data used in the attempt, and the last time the associated identity attempted authentication.

See \url{https://paws-r.github.io/docs/iamrolesanywhere/get_subject.html} for full documentation.
}
\keyword{internal}
