% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_get_trust_anchor}
\alias{iamrolesanywhere_get_trust_anchor}
\title{Gets a trust anchor}
\usage{
iamrolesanywhere_get_trust_anchor(trustAnchorId)
}
\arguments{
\item{trustAnchorId}{[required] The unique identifier of the trust anchor.}
}
\description{
Gets a trust anchor.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_get_trust_anchor/} for full documentation.
}
\keyword{internal}
