% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_put_attribute_mapping}
\alias{iamrolesanywhere_put_attribute_mapping}
\title{Put an entry in the attribute mapping rules that will be enforced by a
given profile}
\usage{
iamrolesanywhere_put_attribute_mapping(
  profileId,
  certificateField,
  mappingRules
)
}
\arguments{
\item{profileId}{[required] The unique identifier of the profile.}

\item{certificateField}{[required] Fields (x509Subject, x509Issuer and x509SAN) within X.509 certificates.}

\item{mappingRules}{[required] A list of mapping entries for every supported specifier or sub-field.}
}
\description{
Put an entry in the attribute mapping rules that will be enforced by a given profile. A mapping specifies a certificate field and one or more specifiers that have contextual meanings.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_put_attribute_mapping/} for full documentation.
}
\keyword{internal}
