% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_put_notification_settings}
\alias{iamrolesanywhere_put_notification_settings}
\title{Attaches a list of notification settings to a trust anchor}
\usage{
iamrolesanywhere_put_notification_settings(trustAnchorId, notificationSettings)
}
\arguments{
\item{trustAnchorId}{[required] The unique identifier of the trust anchor.}

\item{notificationSettings}{[required] A list of notification settings to be associated to the trust anchor.}
}
\description{
Attaches a list of \emph{notification settings} to a trust anchor.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_put_notification_settings/} for full documentation.
}
\keyword{internal}
