% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_reset_notification_settings}
\alias{iamrolesanywhere_reset_notification_settings}
\title{Resets the custom notification setting to IAM Roles Anywhere default
setting}
\usage{
iamrolesanywhere_reset_notification_settings(
  notificationSettingKeys,
  trustAnchorId
)
}
\arguments{
\item{notificationSettingKeys}{[required] A list of notification setting keys to reset. A notification setting key
includes the event and the channel.}

\item{trustAnchorId}{[required] The unique identifier of the trust anchor.}
}
\description{
Resets the \emph{custom notification setting} to IAM Roles Anywhere default setting.

See \url{https://paws-r.github.io/docs/iamrolesanywhere/reset_notification_settings.html} for full documentation.
}
\keyword{internal}
