% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_tag_resource}
\alias{iamrolesanywhere_tag_resource}
\title{Attaches tags to a resource}
\usage{
iamrolesanywhere_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tags}{[required] The tags to attach to the resource.}
}
\description{
Attaches tags to a resource.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_tag_resource/} for full documentation.
}
\keyword{internal}
