% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_untag_resource}
\alias{iamrolesanywhere_untag_resource}
\title{Removes tags from the resource}
\usage{
iamrolesanywhere_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}

\item{tagKeys}{[required] A list of keys. Tag keys are the unique identifiers of tags.}
}
\description{
Removes tags from the resource.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_untag_resource/} for full documentation.
}
\keyword{internal}
