% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_update_crl}
\alias{iamrolesanywhere_update_crl}
\title{Updates the certificate revocation list (CRL)}
\usage{
iamrolesanywhere_update_crl(crlData = NULL, crlId, name = NULL)
}
\arguments{
\item{crlData}{The x509 v3 specified certificate revocation list}

\item{crlId}{[required] The unique identifier of the certificate revocation list (CRL).}

\item{name}{The name of the Crl.}
}
\description{
Updates the certificate revocation list (CRL). CRl is a list of certificates that have been revoked by the issuing certificate Authority (CA). IAM Roles Anywhere validates against the crl list before issuing credentials.

See \url{https://paws-r.github.io/docs/iamrolesanywhere/update_crl.html} for full documentation.
}
\keyword{internal}
