% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_describe_group}
\alias{identitystore_describe_group}
\title{Retrieves the group metadata and attributes from GroupId in an identity
store}
\usage{
identitystore_describe_group(IdentityStoreId, GroupId)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store, such as \code{d-1234567890}. In this example, \verb{d-} is a fixed prefix, and \code{1234567890} is a randomly generated string that contains numbers and lower case letters. This value is generated at the time that a new identity store is created.

You can specify the identity store by ID or by Amazon Resource Name (ARN). For example, identity store ID \code{d-1234567890} or identity store ARN \verb{arn:aws:identitystore::111122223333:identitystore/d-1234567890}.}

\item{GroupId}{[required] The identifier for a group in the identity store.

You can specify the group by ID or by Amazon Resource Name (ARN). For example, group ID \verb{a1b2c3d4-5678-90ab-cdef-EXAMPLE22222} or group ARN \verb{arn:aws:identitystore:::group/a1b2c3d4-5678-90ab-cdef-EXAMPLE22222}.}
}
\description{
Retrieves the group metadata and attributes from \code{GroupId} in an identity store.

See \url{https://www.paws-r-sdk.com/docs/identitystore_describe_group/} for full documentation.
}
\keyword{internal}
