% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_describe_group_membership}
\alias{identitystore_describe_group_membership}
\title{Retrieves membership metadata and attributes from MembershipId in an
identity store}
\usage{
identitystore_describe_group_membership(IdentityStoreId, MembershipId)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.

You can specify the identity store by ID or by Amazon Resource Name (ARN). For example, identity store ID \code{d-1234567890} or identity store ARN \verb{arn:aws:identitystore::111122223333:identitystore/d-1234567890}.}

\item{MembershipId}{[required] The identifier for a \code{GroupMembership} in an identity store.

You can specify the group membership by ID or by Amazon Resource Name (ARN). For example, membership ID \verb{a1b2c3d4-5678-90ab-cdef-EXAMPLE33333} or membership ARN \verb{arn:aws:identitystore:::membership/a1b2c3d4-5678-90ab-cdef-EXAMPLE33333}.}
}
\description{
Retrieves membership metadata and attributes from \code{MembershipId} in an identity store.

See \url{https://www.paws-r-sdk.com/docs/identitystore_describe_group_membership/} for full documentation.
}
\keyword{internal}
