% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_describe_identity_store}
\alias{identitystore_describe_identity_store}
\title{Retrieves details about the specified identity store, including its
Amazon Resource Name (ARN) and network configuration}
\usage{
identitystore_describe_identity_store(IdentityStoreId)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.

You can specify the identity store by ID or by Amazon Resource Name (ARN). For example, identity store ID \code{d-1234567890} or identity store ARN \verb{arn:aws:identitystore::111122223333:identitystore/d-1234567890}.}
}
\description{
Retrieves details about the specified identity store, including its Amazon Resource Name (ARN) and network configuration.

See \url{https://www.paws-r-sdk.com/docs/identitystore_describe_identity_store/} for full documentation.
}
\keyword{internal}
