% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_get_group_membership_id}
\alias{identitystore_get_group_membership_id}
\title{Retrieves the MembershipId in an identity store}
\usage{
identitystore_get_group_membership_id(IdentityStoreId, GroupId, MemberId)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.

You can specify the identity store by ID or by Amazon Resource Name (ARN). For example, identity store ID \code{d-1234567890} or identity store ARN \verb{arn:aws:identitystore::111122223333:identitystore/d-1234567890}.}

\item{GroupId}{[required] The identifier for a group in the identity store.

You can specify the group by ID or by Amazon Resource Name (ARN). For example, group ID \verb{a1b2c3d4-5678-90ab-cdef-EXAMPLE22222} or group ARN \verb{arn:aws:identitystore:::group/a1b2c3d4-5678-90ab-cdef-EXAMPLE22222}.}

\item{MemberId}{[required] An object that contains the identifier of a group member. Setting the \code{UserID} field to the specific identifier for a user indicates that the user is a member of the group.}
}
\description{
Retrieves the \code{MembershipId} in an identity store.

See \url{https://www.paws-r-sdk.com/docs/identitystore_get_group_membership_id/} for full documentation.
}
\keyword{internal}
