% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_get_user_id}
\alias{identitystore_get_user_id}
\title{Retrieves the UserId in an identity store}
\usage{
identitystore_get_user_id(IdentityStoreId, AlternateIdentifier)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.

You can specify the identity store by ID or by Amazon Resource Name (ARN). For example, identity store ID \code{d-1234567890} or identity store ARN \verb{arn:aws:identitystore::111122223333:identitystore/d-1234567890}.}

\item{AlternateIdentifier}{[required] A unique identifier for a user or group that is not the primary identifier. This value can be an identifier from an external identity provider (IdP) that is associated with the user, the group, or a unique attribute. For the unique attribute, the only valid paths are \code{ userName} and \code{emails.value}.}
}
\description{
Retrieves the \code{UserId} in an identity store.

See \url{https://www.paws-r-sdk.com/docs/identitystore_get_user_id/} for full documentation.
}
\keyword{internal}
