% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_is_member_in_groups}
\alias{identitystore_is_member_in_groups}
\title{Checks the user's membership in all requested groups and returns if the
member exists in all queried groups}
\usage{
identitystore_is_member_in_groups(IdentityStoreId, MemberId, GroupIds)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.

You can specify the identity store by ID or by Amazon Resource Name (ARN). For example, identity store ID \code{d-1234567890} or identity store ARN \verb{arn:aws:identitystore::111122223333:identitystore/d-1234567890}.}

\item{MemberId}{[required] An object containing the identifier of a group member.}

\item{GroupIds}{[required] A list of identifiers for groups in the identity store.

You can specify each group by ID or by Amazon Resource Name (ARN). For example, group ID \verb{a1b2c3d4-5678-90ab-cdef-EXAMPLE22222} or group ARN \verb{arn:aws:identitystore:::group/a1b2c3d4-5678-90ab-cdef-EXAMPLE22222}.}
}
\description{
Checks the user's membership in all requested groups and returns if the member exists in all queried groups.

See \url{https://www.paws-r-sdk.com/docs/identitystore_is_member_in_groups/} for full documentation.
}
\keyword{internal}
