% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_list_identity_stores}
\alias{identitystore_list_identity_stores}
\title{Lists the identity stores that you have access to}
\usage{
identitystore_list_identity_stores(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to return per request. This parameter is used in all \code{ List} operations to specify how many results to return on one page. If you don't specify a value, the operation uses a default page size.}

\item{NextToken}{The pagination token used for the \code{\link[=identitystore_list_identity_stores]{list_identity_stores}} API operation. This value is generated by the identity store service. It is returned in the API response if the total results are more than the size of one page. This token is also returned when it is used in the API request to retrieve the next page of results.}
}
\description{
Lists the identity stores that you have access to. This operation returns only the identity store ID and Amazon Resource Name (ARN) of each identity store. To obtain additional information about an identity store, call \code{\link[=identitystore_describe_identity_store]{describe_identity_store}}.

See \url{https://www.paws-r-sdk.com/docs/identitystore_list_identity_stores/} for full documentation.
}
\keyword{internal}
