% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_service.R
\name{inspector2}
\alias{inspector2}
\title{Inspector2}
\usage{
inspector2(
  config = list(),
  credentials = list(),
  endpoint = NULL,
  region = NULL
)
}
\arguments{
\item{config}{Optional configuration of credentials, endpoint, and/or region.
\itemize{
\item{\strong{credentials}: \itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}
\item{\strong{endpoint}: The complete URL to use for the constructed client.}
\item{\strong{region}: The AWS Region used in instantiating the client.}
\item{\strong{close_connection}: Immediately close all HTTP connections.}
\item{\strong{timeout}: The time in seconds till a timeout exception is thrown when attempting to make a connection. The default is 60 seconds.}
\item{\strong{s3_force_path_style}: Set this to \code{true} to force the request to use path-style addressing, i.e. \verb{http://s3.amazonaws.com/BUCKET/KEY}.}
\item{\strong{sts_regional_endpoint}: Set sts regional endpoint resolver to regional or legacy \url{https://docs.aws.amazon.com/sdkref/latest/guide/feature-sts-regionalized-endpoints.html}}
\item{\strong{use_dual_stack}: Set this to \code{true} to use the dualstack (IPv4 and IPv6) endpoint for a service, where available, falling back to the regular endpoint when it isn't. Defaults to the \code{AWS_USE_DUALSTACK_ENDPOINT} environment variable when unset.}
}}

\item{credentials}{Optional credentials shorthand for the config parameter
\itemize{
\item{\strong{creds}: \itemize{
\item{\strong{access_key_id}: AWS access key ID}
\item{\strong{secret_access_key}: AWS secret access key}
\item{\strong{session_token}: AWS temporary session token}
}}
\item{\strong{profile}: The name of a profile to use. If not given, then the default profile is used.}
\item{\strong{anonymous}: Set anonymous credentials.}
}}

\item{endpoint}{Optional shorthand for complete URL to use for the constructed client.}

\item{region}{Optional shorthand for AWS Region used in instantiating the client.}
}
\value{
A client for the service. You can call the service's operations using
syntax like \code{svc$operation(...)}, where \code{svc} is the name you've assigned
to the client. The available operations are listed in the
Operations section.
}
\description{
Amazon Inspector is a vulnerability discovery service that automates continuous scanning for security vulnerabilities within your Amazon EC2, Amazon ECR, and Amazon Web Services Lambda environments.
}
\section{Service syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc <- inspector2(
  config = list(
    credentials = list(
      creds = list(
        access_key_id = "string",
        secret_access_key = "string",
        session_token = "string"
      ),
      profile = "string",
      anonymous = "logical"
    ),
    endpoint = "string",
    region = "string",
    close_connection = "logical",
    timeout = "numeric",
    s3_force_path_style = "logical",
    sts_regional_endpoint = "string",
    use_dual_stack = "logical"
  ),
  credentials = list(
    creds = list(
      access_key_id = "string",
      secret_access_key = "string",
      session_token = "string"
    ),
    profile = "string",
    anonymous = "logical"
  ),
  endpoint = "string",
  region = "string"
)
}\if{html}{\out{</div>}}
}

\section{Operations}{

\tabular{ll}{
\link[=inspector2_associate_member]{associate_member} \tab Associates an Amazon Web Services account with an Amazon Inspector delegated administrator\cr
\link[=inspector2_batch_associate_code_security_scan_configuration]{batch_associate_code_security_scan_configuration} \tab Associates multiple code repositories with an Amazon Inspector code security scan configuration\cr
\link[=inspector2_batch_disassociate_code_security_scan_configuration]{batch_disassociate_code_security_scan_configuration} \tab Disassociates multiple code repositories from an Amazon Inspector code security scan configuration\cr
\link[=inspector2_batch_get_account_status]{batch_get_account_status} \tab Retrieves the Amazon Inspector status of multiple Amazon Web Services accounts within your environment\cr
\link[=inspector2_batch_get_code_snippet]{batch_get_code_snippet} \tab Retrieves code snippets from findings that Amazon Inspector detected code vulnerabilities in\cr
\link[=inspector2_batch_get_finding_details]{batch_get_finding_details} \tab Gets vulnerability details for findings\cr
\link[=inspector2_batch_get_free_trial_info]{batch_get_free_trial_info} \tab Gets free trial status for multiple Amazon Web Services accounts\cr
\link[=inspector2_batch_get_member_ec_2_deep_inspection_status]{batch_get_member_ec_2_deep_inspection_status} \tab Retrieves Amazon Inspector deep inspection activation status of multiple member accounts within your organization\cr
\link[=inspector2_batch_update_member_ec_2_deep_inspection_status]{batch_update_member_ec_2_deep_inspection_status} \tab Activates or deactivates Amazon Inspector deep inspection for the provided member accounts in your organization\cr
\link[=inspector2_cancel_findings_report]{cancel_findings_report} \tab Cancels the given findings report\cr
\link[=inspector2_cancel_sbom_export]{cancel_sbom_export} \tab Cancels a software bill of materials (SBOM) report\cr
\link[=inspector2_create_cis_scan_configuration]{create_cis_scan_configuration} \tab Creates a CIS scan configuration\cr
\link[=inspector2_create_code_security_integration]{create_code_security_integration} \tab Creates a code security integration with a source code repository provider\cr
\link[=inspector2_create_code_security_scan_configuration]{create_code_security_scan_configuration} \tab Creates a scan configuration for code security scanning\cr
\link[=inspector2_create_connector]{create_connector} \tab Creates a connector that links an external cloud provider to Amazon Inspector for vulnerability scanning\cr
\link[=inspector2_create_filter]{create_filter} \tab Creates a filter resource using specified filter criteria\cr
\link[=inspector2_create_findings_report]{create_findings_report} \tab Creates a finding report\cr
\link[=inspector2_create_sbom_export]{create_sbom_export} \tab Creates a software bill of materials (SBOM) report\cr
\link[=inspector2_delete_cis_scan_configuration]{delete_cis_scan_configuration} \tab Deletes a CIS scan configuration\cr
\link[=inspector2_delete_code_security_integration]{delete_code_security_integration} \tab Deletes a code security integration\cr
\link[=inspector2_delete_code_security_scan_configuration]{delete_code_security_scan_configuration} \tab Deletes a code security scan configuration\cr
\link[=inspector2_delete_connector]{delete_connector} \tab Deletes a connector from your account\cr
\link[=inspector2_delete_filter]{delete_filter} \tab Deletes a filter resource\cr
\link[=inspector2_describe_organization_configuration]{describe_organization_configuration} \tab Describe Amazon Inspector configuration settings for an Amazon Web Services organization\cr
\link[=inspector2_disable]{disable} \tab Disables Amazon Inspector scans for one or more Amazon Web Services accounts\cr
\link[=inspector2_disable_delegated_admin_account]{disable_delegated_admin_account} \tab Disables the Amazon Inspector delegated administrator for your organization\cr
\link[=inspector2_disassociate_member]{disassociate_member} \tab Disassociates a member account from an Amazon Inspector delegated administrator\cr
\link[=inspector2_enable]{enable} \tab Enables Amazon Inspector scans for one or more Amazon Web Services accounts\cr
\link[=inspector2_enable_delegated_admin_account]{enable_delegated_admin_account} \tab Enables the Amazon Inspector delegated administrator for your Organizations organization\cr
\link[=inspector2_get_cis_scan_report]{get_cis_scan_report} \tab Retrieves a CIS scan report\cr
\link[=inspector2_get_cis_scan_result_details]{get_cis_scan_result_details} \tab Retrieves CIS scan result details\cr
\link[=inspector2_get_clusters_for_image]{get_clusters_for_image} \tab Returns a list of clusters and metadata associated with an image\cr
\link[=inspector2_get_code_security_integration]{get_code_security_integration} \tab Retrieves information about a code security integration\cr
\link[=inspector2_get_code_security_scan]{get_code_security_scan} \tab Retrieves information about a specific code security scan\cr
\link[=inspector2_get_code_security_scan_configuration]{get_code_security_scan_configuration} \tab Retrieves information about a code security scan configuration\cr
\link[=inspector2_get_configuration]{get_configuration} \tab Retrieves setting configurations for Amazon Inspector scans\cr
\link[=inspector2_get_delegated_admin_account]{get_delegated_admin_account} \tab Retrieves information about the Amazon Inspector delegated administrator for your organization\cr
\link[=inspector2_get_ec_2_deep_inspection_configuration]{get_ec_2_deep_inspection_configuration} \tab Retrieves the activation status of Amazon Inspector deep inspection and custom paths associated with your account\cr
\link[=inspector2_get_encryption_key]{get_encryption_key} \tab Gets an encryption key\cr
\link[=inspector2_get_findings_report_status]{get_findings_report_status} \tab Gets the status of a findings report\cr
\link[=inspector2_get_member]{get_member} \tab Gets member information for your organization\cr
\link[=inspector2_get_sbom_export]{get_sbom_export} \tab Gets details of a software bill of materials (SBOM) report\cr
\link[=inspector2_list_account_permissions]{list_account_permissions} \tab Lists the permissions an account has to configure Amazon Inspector\cr
\link[=inspector2_list_cis_scan_configurations]{list_cis_scan_configurations} \tab Lists CIS scan configurations\cr
\link[=inspector2_list_cis_scan_results_aggregated_by_checks]{list_cis_scan_results_aggregated_by_checks} \tab Lists scan results aggregated by checks\cr
\link[=inspector2_list_cis_scan_results_aggregated_by_target_resource]{list_cis_scan_results_aggregated_by_target_resource} \tab Lists scan results aggregated by a target resource\cr
\link[=inspector2_list_cis_scans]{list_cis_scans} \tab Returns a CIS scan list\cr
\link[=inspector2_list_code_security_integrations]{list_code_security_integrations} \tab Lists all code security integrations in your account\cr
\link[=inspector2_list_code_security_scan_configuration_associations]{list_code_security_scan_configuration_associations} \tab Lists the associations between code repositories and Amazon Inspector code security scan configurations\cr
\link[=inspector2_list_code_security_scan_configurations]{list_code_security_scan_configurations} \tab Lists all code security scan configurations in your account\cr
\link[=inspector2_list_connectors]{list_connectors} \tab Lists connectors in your account\cr
\link[=inspector2_list_connector_scan_configurations]{list_connector_scan_configurations} \tab Lists scan configurations for Amazon Web Services Config connectors\cr
\link[=inspector2_list_coverage]{list_coverage} \tab Lists coverage details for your environment\cr
\link[=inspector2_list_coverage_statistics]{list_coverage_statistics} \tab Lists Amazon Inspector coverage statistics for your environment\cr
\link[=inspector2_list_delegated_admin_accounts]{list_delegated_admin_accounts} \tab Lists information about the Amazon Inspector delegated administrator of your organization\cr
\link[=inspector2_list_filters]{list_filters} \tab Lists the filters associated with your account\cr
\link[=inspector2_list_finding_aggregations]{list_finding_aggregations} \tab Lists aggregated finding data for your environment based on specific criteria\cr
\link[=inspector2_list_findings]{list_findings} \tab Lists findings for your environment\cr
\link[=inspector2_list_members]{list_members} \tab List members associated with the Amazon Inspector delegated administrator for your organization\cr
\link[=inspector2_list_tags_for_resource]{list_tags_for_resource} \tab Lists all tags attached to a given resource\cr
\link[=inspector2_list_usage_totals]{list_usage_totals} \tab Lists the Amazon Inspector usage totals over the last 30 days\cr
\link[=inspector2_reset_encryption_key]{reset_encryption_key} \tab Resets an encryption key\cr
\link[=inspector2_search_vulnerabilities]{search_vulnerabilities} \tab Lists Amazon Inspector coverage details for a specific vulnerability\cr
\link[=inspector2_send_cis_session_health]{send_cis_session_health} \tab Sends a CIS session health\cr
\link[=inspector2_send_cis_session_telemetry]{send_cis_session_telemetry} \tab Sends a CIS session telemetry\cr
\link[=inspector2_start_cis_session]{start_cis_session} \tab Starts a CIS session\cr
\link[=inspector2_start_code_security_scan]{start_code_security_scan} \tab Initiates a code security scan on a specified repository\cr
\link[=inspector2_stop_cis_session]{stop_cis_session} \tab Stops a CIS session\cr
\link[=inspector2_tag_resource]{tag_resource} \tab Adds tags to a resource\cr
\link[=inspector2_untag_resource]{untag_resource} \tab Removes tags from a resource\cr
\link[=inspector2_update_cis_scan_configuration]{update_cis_scan_configuration} \tab Updates a CIS scan configuration\cr
\link[=inspector2_update_code_security_integration]{update_code_security_integration} \tab Updates an existing code security integration\cr
\link[=inspector2_update_code_security_scan_configuration]{update_code_security_scan_configuration} \tab Updates an existing code security scan configuration\cr
\link[=inspector2_update_configuration]{update_configuration} \tab Updates the scan configuration for your Amazon Inspector account\cr
\link[=inspector2_update_connector]{update_connector} \tab Updates the description or provider-specific configuration details of an existing connector\cr
\link[=inspector2_update_connector_scan_configuration]{update_connector_scan_configuration} \tab Updates scan configuration settings for resources associated with an Amazon Web Services Config connector\cr
\link[=inspector2_update_ec_2_deep_inspection_configuration]{update_ec_2_deep_inspection_configuration} \tab Activates, deactivates Amazon Inspector deep inspection, or updates custom paths for your account\cr
\link[=inspector2_update_encryption_key]{update_encryption_key} \tab Updates an encryption key\cr
\link[=inspector2_update_filter]{update_filter} \tab Specifies the action that is to be applied to the findings that match the filter\cr
\link[=inspector2_update_organization_configuration]{update_organization_configuration} \tab Updates the configurations for your Amazon Inspector organization\cr
\link[=inspector2_update_org_ec_2_deep_inspection_configuration]{update_org_ec_2_deep_inspection_configuration} \tab Updates the Amazon Inspector deep inspection custom paths for your organization
}
}

\examples{
\dontrun{
svc <- inspector2()
svc$associate_member(
  Foo = 123
)
}

}
