% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_associate_member}
\alias{inspector2_associate_member}
\title{Associates an Amazon Web Services account with an Amazon Inspector
delegated administrator}
\usage{
inspector2_associate_member(accountId)
}
\arguments{
\item{accountId}{[required] The Amazon Web Services account ID of the member account to be
associated.}
}
\description{
Associates an Amazon Web Services account with an Amazon Inspector delegated administrator.

See \url{https://paws-r.github.io/docs/inspector2/associate_member.html} for full documentation.
}
\keyword{internal}
