% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_batch_associate_code_security_scan_configuration}
\alias{inspector2_batch_associate_code_security_scan_configuration}
\title{Associates multiple code repositories with an Amazon Inspector code
security scan configuration}
\usage{
inspector2_batch_associate_code_security_scan_configuration(
  associateConfigurationRequests
)
}
\arguments{
\item{associateConfigurationRequests}{[required] A list of code repositories to associate with the specified scan
configuration.}
}
\description{
Associates multiple code repositories with an Amazon Inspector code security scan configuration.

See \url{https://www.paws-r-sdk.com/docs/inspector2_batch_associate_code_security_scan_configuration/} for full documentation.
}
\keyword{internal}
