% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_batch_disassociate_code_security_scan_configuration}
\alias{inspector2_batch_disassociate_code_security_scan_configuration}
\title{Disassociates multiple code repositories from an Amazon Inspector code
security scan configuration}
\usage{
inspector2_batch_disassociate_code_security_scan_configuration(
  disassociateConfigurationRequests
)
}
\arguments{
\item{disassociateConfigurationRequests}{[required] A list of code repositories to disassociate from the specified scan
configuration.}
}
\description{
Disassociates multiple code repositories from an Amazon Inspector code security scan configuration.

See \url{https://www.paws-r-sdk.com/docs/inspector2_batch_disassociate_code_security_scan_configuration/} for full documentation.
}
\keyword{internal}
