% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_batch_get_account_status}
\alias{inspector2_batch_get_account_status}
\title{Retrieves the Amazon Inspector status of multiple Amazon Web Services
accounts within your environment}
\usage{
inspector2_batch_get_account_status(accountIds = NULL)
}
\arguments{
\item{accountIds}{The 12-digit Amazon Web Services account IDs of the accounts to retrieve
Amazon Inspector status for.}
}
\description{
Retrieves the Amazon Inspector status of multiple Amazon Web Services accounts within your environment.

See \url{https://www.paws-r-sdk.com/docs/inspector2_batch_get_account_status/} for full documentation.
}
\keyword{internal}
