% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_batch_get_code_snippet}
\alias{inspector2_batch_get_code_snippet}
\title{Retrieves code snippets from findings that Amazon Inspector detected
code vulnerabilities in}
\usage{
inspector2_batch_get_code_snippet(findingArns)
}
\arguments{
\item{findingArns}{[required] An array of finding ARNs for the findings you want to retrieve code
snippets from.}
}
\description{
Retrieves code snippets from findings that Amazon Inspector detected code vulnerabilities in.

See \url{https://www.paws-r-sdk.com/docs/inspector2_batch_get_code_snippet/} for full documentation.
}
\keyword{internal}
