% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_batch_get_finding_details}
\alias{inspector2_batch_get_finding_details}
\title{Gets vulnerability details for findings}
\usage{
inspector2_batch_get_finding_details(findingArns)
}
\arguments{
\item{findingArns}{[required] A list of finding ARNs.}
}
\description{
Gets vulnerability details for findings.

See \url{https://www.paws-r-sdk.com/docs/inspector2_batch_get_finding_details/} for full documentation.
}
\keyword{internal}
