% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_batch_get_free_trial_info}
\alias{inspector2_batch_get_free_trial_info}
\title{Gets free trial status for multiple Amazon Web Services accounts}
\usage{
inspector2_batch_get_free_trial_info(accountIds)
}
\arguments{
\item{accountIds}{[required] The account IDs to get free trial status for.}
}
\description{
Gets free trial status for multiple Amazon Web Services accounts.

See \url{https://www.paws-r-sdk.com/docs/inspector2_batch_get_free_trial_info/} for full documentation.
}
\keyword{internal}
