% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_batch_update_member_ec_2_deep_inspection_status}
\alias{inspector2_batch_update_member_ec_2_deep_inspection_status}
\title{Activates or deactivates Amazon Inspector deep inspection for the
provided member accounts in your organization}
\usage{
inspector2_batch_update_member_ec_2_deep_inspection_status(accountIds)
}
\arguments{
\item{accountIds}{[required] The unique identifiers for the Amazon Web Services accounts to change
Amazon Inspector deep inspection status for.}
}
\description{
Activates or deactivates Amazon Inspector deep inspection for the provided member accounts in your organization. You must be the delegated administrator of an organization in Amazon Inspector to use this API.

See \url{https://www.paws-r-sdk.com/docs/inspector2_batch_update_member_ec_2_deep_inspection_status/} for full documentation.
}
\keyword{internal}
