% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_cancel_findings_report}
\alias{inspector2_cancel_findings_report}
\title{Cancels the given findings report}
\usage{
inspector2_cancel_findings_report(reportId)
}
\arguments{
\item{reportId}{[required] The ID of the report to be canceled.}
}
\description{
Cancels the given findings report.

See \url{https://www.paws-r-sdk.com/docs/inspector2_cancel_findings_report/} for full documentation.
}
\keyword{internal}
