% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_create_code_security_integration}
\alias{inspector2_create_code_security_integration}
\title{Creates a code security integration with a source code repository
provider}
\usage{
inspector2_create_code_security_integration(
  name,
  type,
  details = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the code security integration.}

\item{type}{[required] The type of repository provider for the integration.}

\item{details}{The integration details specific to the repository provider type.}

\item{tags}{The tags to apply to the code security integration.}
}
\description{
Creates a code security integration with a source code repository provider.

See \url{https://www.paws-r-sdk.com/docs/inspector2_create_code_security_integration/} for full documentation.
}
\keyword{internal}
