% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_create_code_security_scan_configuration}
\alias{inspector2_create_code_security_scan_configuration}
\title{Creates a scan configuration for code security scanning}
\usage{
inspector2_create_code_security_scan_configuration(
  name,
  level,
  configuration,
  scopeSettings = NULL,
  tags = NULL
)
}
\arguments{
\item{name}{[required] The name of the scan configuration.}

\item{level}{[required] The security level for the scan configuration.}

\item{configuration}{[required] The configuration settings for the code security scan.}

\item{scopeSettings}{The scope settings that define which repositories will be scanned.
Include this parameter to create a default scan configuration. Otherwise
Amazon Inspector creates a general scan configuration.

A default scan configuration automatically applies to all existing and
future projects imported into Amazon Inspector. Use the
\code{\link[=inspector2_batch_associate_code_security_scan_configuration]{batch_associate_code_security_scan_configuration}}
operation to associate a general scan configuration with projects.}

\item{tags}{The tags to apply to the scan configuration.}
}
\description{
Creates a scan configuration for code security scanning.

See \url{https://www.paws-r-sdk.com/docs/inspector2_create_code_security_scan_configuration/} for full documentation.
}
\keyword{internal}
