% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_create_connector}
\alias{inspector2_create_connector}
\title{Creates a connector that links an external cloud provider to Amazon
Inspector for vulnerability scanning}
\usage{
inspector2_create_connector(
  clientToken = NULL,
  name,
  provider,
  description = NULL,
  providerDetail,
  tags = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure that the operation completes no more than one time. If this token matches a previous request, the service ignores the request but does not return an error.}

\item{name}{[required] The name of the connector.}

\item{provider}{[required] The cloud provider for the connector.}

\item{description}{A description of the connector.}

\item{providerDetail}{[required] The provider-specific configuration details for the connector.}

\item{tags}{The tags to apply to the connector.}
}
\description{
Creates a connector that links an external cloud provider to Amazon Inspector for vulnerability scanning.

See \url{https://www.paws-r-sdk.com/docs/inspector2_create_connector/} for full documentation.
}
\keyword{internal}
