% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_create_findings_report}
\alias{inspector2_create_findings_report}
\title{Creates a finding report}
\usage{
inspector2_create_findings_report(
  filterCriteria = NULL,
  reportFormat,
  s3Destination
)
}
\arguments{
\item{filterCriteria}{The filter criteria to apply to the results of the finding report.}

\item{reportFormat}{[required] The format to generate the report in.}

\item{s3Destination}{[required] The Amazon S3 export destination for the report.}
}
\description{
Creates a finding report. By default only \code{ACTIVE} findings are returned in the report. To see \code{SUPRESSED} or \code{CLOSED} findings you must specify a value for the \code{findingStatus} filter criteria.

See \url{https://www.paws-r-sdk.com/docs/inspector2_create_findings_report/} for full documentation.
}
\keyword{internal}
