% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_delete_cis_scan_configuration}
\alias{inspector2_delete_cis_scan_configuration}
\title{Deletes a CIS scan configuration}
\usage{
inspector2_delete_cis_scan_configuration(scanConfigurationArn)
}
\arguments{
\item{scanConfigurationArn}{[required] The ARN of the CIS scan configuration.}
}
\description{
Deletes a CIS scan configuration.

See \url{https://www.paws-r-sdk.com/docs/inspector2_delete_cis_scan_configuration/} for full documentation.
}
\keyword{internal}
