% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_delete_code_security_integration}
\alias{inspector2_delete_code_security_integration}
\title{Deletes a code security integration}
\usage{
inspector2_delete_code_security_integration(integrationArn)
}
\arguments{
\item{integrationArn}{[required] The Amazon Resource Name (ARN) of the code security integration to
delete.}
}
\description{
Deletes a code security integration.

See \url{https://www.paws-r-sdk.com/docs/inspector2_delete_code_security_integration/} for full documentation.
}
\keyword{internal}
