% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_delete_connector}
\alias{inspector2_delete_connector}
\title{Deletes a connector from your account}
\usage{
inspector2_delete_connector(connectorArn)
}
\arguments{
\item{connectorArn}{[required] The Amazon Resource Name (ARN) of the connector to delete.}
}
\description{
Deletes a connector from your account.

See \url{https://www.paws-r-sdk.com/docs/inspector2_delete_connector/} for full documentation.
}
\keyword{internal}
