% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_delete_filter}
\alias{inspector2_delete_filter}
\title{Deletes a filter resource}
\usage{
inspector2_delete_filter(arn)
}
\arguments{
\item{arn}{[required] The Amazon Resource Number (ARN) of the filter to be deleted.}
}
\description{
Deletes a filter resource.

See \url{https://paws-r.github.io/docs/inspector2/delete_filter.html} for full documentation.
}
\keyword{internal}
