% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_describe_organization_configuration}
\alias{inspector2_describe_organization_configuration}
\title{Describe Amazon Inspector configuration settings for an Amazon Web
Services organization}
\usage{
inspector2_describe_organization_configuration()
}
\description{
Describe Amazon Inspector configuration settings for an Amazon Web Services organization.

See \url{https://paws-r.github.io/docs/inspector2/describe_organization_configuration.html} for full documentation.
}
\keyword{internal}
