% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_disable}
\alias{inspector2_disable}
\title{Disables Amazon Inspector scans for one or more Amazon Web Services
accounts}
\usage{
inspector2_disable(accountIds = NULL, resourceTypes = NULL)
}
\arguments{
\item{accountIds}{An array of account IDs you want to disable Amazon Inspector scans for.}

\item{resourceTypes}{The resource scan types you want to disable.}
}
\description{
Disables Amazon Inspector scans for one or more Amazon Web Services accounts. Disabling all scan types in an account disables the Amazon Inspector service.

See \url{https://www.paws-r-sdk.com/docs/inspector2_disable/} for full documentation.
}
\keyword{internal}
