% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_enable}
\alias{inspector2_enable}
\title{Enables Amazon Inspector scans for one or more Amazon Web Services
accounts}
\usage{
inspector2_enable(accountIds = NULL, clientToken = NULL, resourceTypes)
}
\arguments{
\item{accountIds}{A list of account IDs you want to enable Amazon Inspector scans for.}

\item{clientToken}{The idempotency token for the request.}

\item{resourceTypes}{[required] The resource scan types you want to enable.}
}
\description{
Enables Amazon Inspector scans for one or more Amazon Web Services accounts.

See \url{https://paws-r.github.io/docs/inspector2/enable.html} for full documentation.
}
\keyword{internal}
