% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_enable_delegated_admin_account}
\alias{inspector2_enable_delegated_admin_account}
\title{Enables the Amazon Inspector delegated administrator for your
Organizations organization}
\usage{
inspector2_enable_delegated_admin_account(
  delegatedAdminAccountId,
  clientToken = NULL
)
}
\arguments{
\item{delegatedAdminAccountId}{[required] The Amazon Web Services account ID of the Amazon Inspector delegated
administrator.}

\item{clientToken}{The idempotency token for the request.}
}
\description{
Enables the Amazon Inspector delegated administrator for your Organizations organization.

See \url{https://www.paws-r-sdk.com/docs/inspector2_enable_delegated_admin_account/} for full documentation.
}
\keyword{internal}
