% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_cis_scan_report}
\alias{inspector2_get_cis_scan_report}
\title{Retrieves a CIS scan report}
\usage{
inspector2_get_cis_scan_report(
  scanArn,
  targetAccounts = NULL,
  reportFormat = NULL
)
}
\arguments{
\item{scanArn}{[required] The scan ARN.}

\item{targetAccounts}{The target accounts.}

\item{reportFormat}{The format of the report. Valid values are \code{PDF} and \code{CSV}. If no value
is specified, the report format defaults to \code{PDF}.}
}
\description{
Retrieves a CIS scan report.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_cis_scan_report/} for full documentation.
}
\keyword{internal}
