% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_code_security_integration}
\alias{inspector2_get_code_security_integration}
\title{Retrieves information about a code security integration}
\usage{
inspector2_get_code_security_integration(integrationArn, tags = NULL)
}
\arguments{
\item{integrationArn}{[required] The Amazon Resource Name (ARN) of the code security integration to
retrieve.}

\item{tags}{The tags associated with the code security integration.}
}
\description{
Retrieves information about a code security integration.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_code_security_integration/} for full documentation.
}
\keyword{internal}
