% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_code_security_scan}
\alias{inspector2_get_code_security_scan}
\title{Retrieves information about a specific code security scan}
\usage{
inspector2_get_code_security_scan(resource, scanId)
}
\arguments{
\item{resource}{[required] The resource identifier for the code repository that was scanned.}

\item{scanId}{[required] The unique identifier of the scan to retrieve.}
}
\description{
Retrieves information about a specific code security scan.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_code_security_scan/} for full documentation.
}
\keyword{internal}
