% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_code_security_scan_configuration}
\alias{inspector2_get_code_security_scan_configuration}
\title{Retrieves information about a code security scan configuration}
\usage{
inspector2_get_code_security_scan_configuration(scanConfigurationArn)
}
\arguments{
\item{scanConfigurationArn}{[required] The Amazon Resource Name (ARN) of the scan configuration to retrieve.}
}
\description{
Retrieves information about a code security scan configuration.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_code_security_scan_configuration/} for full documentation.
}
\keyword{internal}
