% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_ec_2_deep_inspection_configuration}
\alias{inspector2_get_ec_2_deep_inspection_configuration}
\title{Retrieves the activation status of Amazon Inspector deep inspection and
custom paths associated with your account}
\usage{
inspector2_get_ec_2_deep_inspection_configuration()
}
\description{
Retrieves the activation status of Amazon Inspector deep inspection and custom paths associated with your account.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_ec_2_deep_inspection_configuration/} for full documentation.
}
\keyword{internal}
