% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_encryption_key}
\alias{inspector2_get_encryption_key}
\title{Gets an encryption key}
\usage{
inspector2_get_encryption_key(scanType, resourceType)
}
\arguments{
\item{scanType}{[required] The scan type the key encrypts.}

\item{resourceType}{[required] The resource type the key encrypts.}
}
\description{
Gets an encryption key.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_encryption_key/} for full documentation.
}
\keyword{internal}
