% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_findings_report_status}
\alias{inspector2_get_findings_report_status}
\title{Gets the status of a findings report}
\usage{
inspector2_get_findings_report_status(reportId = NULL)
}
\arguments{
\item{reportId}{The ID of the report to retrieve the status of.}
}
\description{
Gets the status of a findings report.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_findings_report_status/} for full documentation.
}
\keyword{internal}
