% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_member}
\alias{inspector2_get_member}
\title{Gets member information for your organization}
\usage{
inspector2_get_member(accountId)
}
\arguments{
\item{accountId}{[required] The Amazon Web Services account ID of the member account to retrieve
information on.}
}
\description{
Gets member information for your organization.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_member/} for full documentation.
}
\keyword{internal}
