% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_sbom_export}
\alias{inspector2_get_sbom_export}
\title{Gets details of a software bill of materials (SBOM) report}
\usage{
inspector2_get_sbom_export(reportId)
}
\arguments{
\item{reportId}{[required] The report ID of the SBOM export to get details for.}
}
\description{
Gets details of a software bill of materials (SBOM) report.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_sbom_export/} for full documentation.
}
\keyword{internal}
