% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_account_permissions}
\alias{inspector2_list_account_permissions}
\title{Lists the permissions an account has to configure Amazon Inspector}
\usage{
inspector2_list_account_permissions(
  maxResults = NULL,
  nextToken = NULL,
  service = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. For subsequent calls, use the \code{NextToken} value returned from
the previous request to continue listing results after the first page.}

\item{service}{The service scan type to check permissions for.}
}
\description{
Lists the permissions an account has to configure Amazon Inspector.

See \url{https://paws-r.github.io/docs/inspector2/list_account_permissions.html} for full documentation.
}
\keyword{internal}
