% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_cis_scan_configurations}
\alias{inspector2_list_cis_scan_configurations}
\title{Lists CIS scan configurations}
\usage{
inspector2_list_cis_scan_configurations(
  filterCriteria = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sortBy = NULL,
  sortOrder = NULL
)
}
\arguments{
\item{filterCriteria}{The CIS scan configuration filter criteria.}

\item{maxResults}{The maximum number of CIS scan configurations to be returned in a single
page of results.}

\item{nextToken}{The pagination token from a previous request that's used to retrieve the
next page of results.}

\item{sortBy}{The CIS scan configuration sort by order.}

\item{sortOrder}{The CIS scan configuration sort order order.}
}
\description{
Lists CIS scan configurations.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_cis_scan_configurations/} for full documentation.
}
\keyword{internal}
