% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_cis_scan_results_aggregated_by_checks}
\alias{inspector2_list_cis_scan_results_aggregated_by_checks}
\title{Lists scan results aggregated by checks}
\usage{
inspector2_list_cis_scan_results_aggregated_by_checks(
  scanArn,
  filterCriteria = NULL,
  sortBy = NULL,
  sortOrder = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{scanArn}{[required] The scan ARN.}

\item{filterCriteria}{The filter criteria.}

\item{sortBy}{The sort by order.}

\item{sortOrder}{The sort order.}

\item{nextToken}{The pagination token from a previous request that's used to retrieve the
next page of results.}

\item{maxResults}{The maximum number of scan results aggregated by checks to be returned
in a single page of results.}
}
\description{
Lists scan results aggregated by checks.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_cis_scan_results_aggregated_by_checks/} for full documentation.
}
\keyword{internal}
