% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_cis_scans}
\alias{inspector2_list_cis_scans}
\title{Returns a CIS scan list}
\usage{
inspector2_list_cis_scans(
  detailLevel = NULL,
  filterCriteria = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sortBy = NULL,
  sortOrder = NULL
)
}
\arguments{
\item{detailLevel}{The detail applied to the CIS scan.}

\item{filterCriteria}{The CIS scan filter criteria.}

\item{maxResults}{The maximum number of results to be returned.}

\item{nextToken}{The pagination token from a previous request that's used to retrieve the
next page of results.}

\item{sortBy}{The CIS scans sort by order.}

\item{sortOrder}{The CIS scans sort order.}
}
\description{
Returns a CIS scan list.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_cis_scans/} for full documentation.
}
\keyword{internal}
