% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_code_security_scan_configuration_associations}
\alias{inspector2_list_code_security_scan_configuration_associations}
\title{Lists the associations between code repositories and Amazon Inspector
code security scan configurations}
\usage{
inspector2_list_code_security_scan_configuration_associations(
  scanConfigurationArn,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{scanConfigurationArn}{[required] The Amazon Resource Name (ARN) of the scan configuration to list
associations for.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. For subsequent calls, use the \code{NextToken} value returned from
the previous request to continue listing results after the first page.}

\item{maxResults}{The maximum number of results to return in the response. If your request
would return more than the maximum the response will return a
\code{nextToken} value, use this value when you call the action again to get
the remaining results.}
}
\description{
Lists the associations between code repositories and Amazon Inspector code security scan configurations.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_code_security_scan_configuration_associations/} for full documentation.
}
\keyword{internal}
