% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_code_security_integrations}
\alias{inspector2_list_code_security_integrations}
\title{Lists all code security integrations in your account}
\usage{
inspector2_list_code_security_integrations(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request. For
subsequent calls, use the NextToken value returned from the previous
request to continue listing results after the first page.}

\item{maxResults}{The maximum number of results to return in a single call.}
}
\description{
Lists all code security integrations in your account.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_code_security_integrations/} for full documentation.
}
\keyword{internal}
