% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_connector_scan_configurations}
\alias{inspector2_list_connector_scan_configurations}
\title{Lists scan configurations for Amazon Web Services Config connectors}
\usage{
inspector2_list_connector_scan_configurations(
  awsConfigConnectorArns = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{awsConfigConnectorArns}{The list of Amazon Web Services Config connector ARNs to filter results.}

\item{maxResults}{The maximum number of results to return in a single call. Valid range is 1 to 50. To retrieve the remaining results, make another request with the \code{nextToken} value returned from this request.}

\item{nextToken}{A token to use for paginating results. Set this value to null for the first request. For subsequent calls, use the \code{nextToken} value returned from the previous request.}
}
\description{
Lists scan configurations for Amazon Web Services Config connectors. Results are paginated. Use the \code{nextToken} parameter to retrieve the next page of results.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_connector_scan_configurations/} for full documentation.
}
\keyword{internal}
