% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_connectors}
\alias{inspector2_list_connectors}
\title{Lists connectors in your account}
\usage{
inspector2_list_connectors(
  maxResults = NULL,
  nextToken = NULL,
  filterCriteria = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to return in a single call. To retrieve the remaining results, make another request with the \code{nextToken} value returned from this request.}

\item{nextToken}{A token to use for paginating results. Set this value to null for the first request. For subsequent calls, use the \code{nextToken} value returned from the previous request.}

\item{filterCriteria}{The filter criteria to apply to the list of connectors.}
}
\description{
Lists connectors in your account. Results are paginated. Use the \code{nextToken} parameter to retrieve the next page of results.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_connectors/} for full documentation.
}
\keyword{internal}
