% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_coverage_statistics}
\alias{inspector2_list_coverage_statistics}
\title{Lists Amazon Inspector coverage statistics for your environment}
\usage{
inspector2_list_coverage_statistics(
  filterCriteria = NULL,
  groupBy = NULL,
  nextToken = NULL
)
}
\arguments{
\item{filterCriteria}{An object that contains details on the filters to apply to the coverage
data for your environment.}

\item{groupBy}{The value to group the results by.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. For subsequent calls, use the \code{NextToken} value returned from
the previous request to continue listing results after the first page.}
}
\description{
Lists Amazon Inspector coverage statistics for your environment.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_coverage_statistics/} for full documentation.
}
\keyword{internal}
