% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_finding_aggregations}
\alias{inspector2_list_finding_aggregations}
\title{Lists aggregated finding data for your environment based on specific
criteria}
\usage{
inspector2_list_finding_aggregations(
  aggregationType,
  nextToken = NULL,
  maxResults = NULL,
  accountIds = NULL,
  aggregationRequest = NULL
)
}
\arguments{
\item{aggregationType}{[required] The type of the aggregation request.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. If your response returns more than the \code{maxResults} maximum
value it will also return a \code{nextToken} value. For subsequent calls, use
the \code{nextToken} value returned from the previous request to continue
listing results after the first page.}

\item{maxResults}{The maximum number of results the response can return. If your request
would return more than the maximum the response will return a
\code{nextToken} value, use this value when you call the action again to get
the remaining results.}

\item{accountIds}{The Amazon Web Services account IDs to retrieve finding aggregation data
for.}

\item{aggregationRequest}{Details of the aggregation request that is used to filter your
aggregation results.}
}
\description{
Lists aggregated finding data for your environment based on specific criteria.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_finding_aggregations/} for full documentation.
}
\keyword{internal}
