% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_findings}
\alias{inspector2_list_findings}
\title{Lists findings for your environment}
\usage{
inspector2_list_findings(
  filterCriteria = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sortCriteria = NULL
)
}
\arguments{
\item{filterCriteria}{Details on the filters to apply to your finding results.}

\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. For subsequent calls, use the \code{NextToken} value returned from
the previous request to continue listing results after the first page.}

\item{sortCriteria}{Details on the sort criteria to apply to your finding results.}
}
\description{
Lists findings for your environment.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_findings/} for full documentation.
}
\keyword{internal}
