% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_members}
\alias{inspector2_list_members}
\title{List members associated with the Amazon Inspector delegated
administrator for your organization}
\usage{
inspector2_list_members(
  onlyAssociated = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{onlyAssociated}{Specifies whether to list only currently associated members if \code{True} or
to list all members within the organization if \code{False}.}

\item{maxResults}{The maximum number of results the response can return. If your request
would return more than the maximum the response will return a
\code{nextToken} value, use this value when you call the action again to get
the remaining results.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. If your response returns more than the \code{maxResults} maximum
value it will also return a \code{nextToken} value. For subsequent calls, use
the \code{nextToken} value returned from the previous request to continue
listing results after the first page.}
}
\description{
List members associated with the Amazon Inspector delegated administrator for your organization.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_members/} for full documentation.
}
\keyword{internal}
