% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_reset_encryption_key}
\alias{inspector2_reset_encryption_key}
\title{Resets an encryption key}
\usage{
inspector2_reset_encryption_key(scanType, resourceType)
}
\arguments{
\item{scanType}{[required] The scan type the key encrypts.}

\item{resourceType}{[required] The resource type the key encrypts.}
}
\description{
Resets an encryption key. After the key is reset your resources will be encrypted by an Amazon Web Services owned key.

See \url{https://www.paws-r-sdk.com/docs/inspector2_reset_encryption_key/} for full documentation.
}
\keyword{internal}
