% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_search_vulnerabilities}
\alias{inspector2_search_vulnerabilities}
\title{Lists Amazon Inspector coverage details for a specific vulnerability}
\usage{
inspector2_search_vulnerabilities(filterCriteria, nextToken = NULL)
}
\arguments{
\item{filterCriteria}{[required] The criteria used to filter the results of a vulnerability search.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. For subsequent calls, use the \code{NextToken} value returned from
the previous request to continue listing results after the first page.}
}
\description{
Lists Amazon Inspector coverage details for a specific vulnerability.

See \url{https://www.paws-r-sdk.com/docs/inspector2_search_vulnerabilities/} for full documentation.
}
\keyword{internal}
