% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_start_code_security_scan}
\alias{inspector2_start_code_security_scan}
\title{Initiates a code security scan on a specified repository}
\usage{
inspector2_start_code_security_scan(clientToken = NULL, resource)
}
\arguments{
\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{resource}{[required] The resource identifier for the code repository to scan.}
}
\description{
Initiates a code security scan on a specified repository.

See \url{https://www.paws-r-sdk.com/docs/inspector2_start_code_security_scan/} for full documentation.
}
\keyword{internal}
