% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_untag_resource}
\alias{inspector2_untag_resource}
\title{Removes tags from a resource}
\usage{
inspector2_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the resource to remove tags from.}

\item{tagKeys}{[required] The tag keys to remove from the resource.}
}
\description{
Removes tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/inspector2_untag_resource/} for full documentation.
}
\keyword{internal}
