% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_update_cis_scan_configuration}
\alias{inspector2_update_cis_scan_configuration}
\title{Updates a CIS scan configuration}
\usage{
inspector2_update_cis_scan_configuration(
  scanConfigurationArn,
  scanName = NULL,
  securityLevel = NULL,
  schedule = NULL,
  targets = NULL
)
}
\arguments{
\item{scanConfigurationArn}{[required] The CIS scan configuration ARN.}

\item{scanName}{The scan name for the CIS scan configuration.}

\item{securityLevel}{The security level for the CIS scan configuration. Security level refers
to the Benchmark levels that CIS assigns to a profile.}

\item{schedule}{The schedule for the CIS scan configuration.}

\item{targets}{The targets for the CIS scan configuration.}
}
\description{
Updates a CIS scan configuration.

See \url{https://www.paws-r-sdk.com/docs/inspector2_update_cis_scan_configuration/} for full documentation.
}
\keyword{internal}
