% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_update_code_security_integration}
\alias{inspector2_update_code_security_integration}
\title{Updates an existing code security integration}
\usage{
inspector2_update_code_security_integration(integrationArn, details)
}
\arguments{
\item{integrationArn}{[required] The Amazon Resource Name (ARN) of the code security integration to
update.}

\item{details}{[required] The updated integration details specific to the repository provider
type.}
}
\description{
Updates an existing code security integration.

See \url{https://www.paws-r-sdk.com/docs/inspector2_update_code_security_integration/} for full documentation.
}
\keyword{internal}
