% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_update_code_security_scan_configuration}
\alias{inspector2_update_code_security_scan_configuration}
\title{Updates an existing code security scan configuration}
\usage{
inspector2_update_code_security_scan_configuration(
  scanConfigurationArn,
  configuration
)
}
\arguments{
\item{scanConfigurationArn}{[required] The Amazon Resource Name (ARN) of the scan configuration to update.}

\item{configuration}{[required] The updated configuration settings for the code security scan.}
}
\description{
Updates an existing code security scan configuration.

See \url{https://www.paws-r-sdk.com/docs/inspector2_update_code_security_scan_configuration/} for full documentation.
}
\keyword{internal}
