% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_update_connector}
\alias{inspector2_update_connector}
\title{Updates the description or provider-specific configuration details of an
existing connector}
\usage{
inspector2_update_connector(
  connectorArn,
  description = NULL,
  providerDetail = NULL
)
}
\arguments{
\item{connectorArn}{[required] The Amazon Resource Name (ARN) of the connector to update.}

\item{description}{The updated description of the connector.}

\item{providerDetail}{The updated provider-specific configuration details for the connector.}
}
\description{
Updates the description or provider-specific configuration details of an existing connector.

See \url{https://www.paws-r-sdk.com/docs/inspector2_update_connector/} for full documentation.
}
\keyword{internal}
