% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_update_connector_scan_configuration}
\alias{inspector2_update_connector_scan_configuration}
\title{Updates scan configuration settings for resources associated with an
Amazon Web Services Config connector}
\usage{
inspector2_update_connector_scan_configuration(
  awsConfigConnectorArn,
  scanConfiguration
)
}
\arguments{
\item{awsConfigConnectorArn}{[required] The ARN of the Amazon Web Services Config connector.}

\item{scanConfiguration}{[required] The scan configuration settings to apply.}
}
\description{
Updates scan configuration settings for resources associated with an Amazon Web Services Config connector.

See \url{https://www.paws-r-sdk.com/docs/inspector2_update_connector_scan_configuration/} for full documentation.
}
\keyword{internal}
