% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_update_filter}
\alias{inspector2_update_filter}
\title{Specifies the action that is to be applied to the findings that match
the filter}
\usage{
inspector2_update_filter(
  action = NULL,
  description = NULL,
  filterCriteria = NULL,
  name = NULL,
  filterArn,
  reason = NULL
)
}
\arguments{
\item{action}{Specifies the action that is to be applied to the findings that match
the filter.}

\item{description}{A description of the filter.}

\item{filterCriteria}{Defines the criteria to be update in the filter.}

\item{name}{The name of the filter.}

\item{filterArn}{[required] The Amazon Resource Number (ARN) of the filter to update.}

\item{reason}{The reason the filter was updated.}
}
\description{
Specifies the action that is to be applied to the findings that match the filter.

See \url{https://www.paws-r-sdk.com/docs/inspector2_update_filter/} for full documentation.
}
\keyword{internal}
