% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_update_org_ec_2_deep_inspection_configuration}
\alias{inspector2_update_org_ec_2_deep_inspection_configuration}
\title{Updates the Amazon Inspector deep inspection custom paths for your
organization}
\usage{
inspector2_update_org_ec_2_deep_inspection_configuration(orgPackagePaths)
}
\arguments{
\item{orgPackagePaths}{[required] The Amazon Inspector deep inspection custom paths you are adding for
your organization.}
}
\description{
Updates the Amazon Inspector deep inspection custom paths for your organization. You must be an Amazon Inspector delegated administrator to use this API.

See \url{https://www.paws-r-sdk.com/docs/inspector2_update_org_ec_2_deep_inspection_configuration/} for full documentation.
}
\keyword{internal}
