% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_update_organization_configuration}
\alias{inspector2_update_organization_configuration}
\title{Updates the configurations for your Amazon Inspector organization}
\usage{
inspector2_update_organization_configuration(autoEnable)
}
\arguments{
\item{autoEnable}{[required] Defines which scan types are enabled automatically for new members of
your Amazon Inspector organization.}
}
\description{
Updates the configurations for your Amazon Inspector organization.

See \url{https://www.paws-r-sdk.com/docs/inspector2_update_organization_configuration/} for full documentation.
}
\keyword{internal}
