% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_create_assessment_target}
\alias{inspector_create_assessment_target}
\title{Creates a new assessment target using the ARN of the resource group that
is generated by CreateResourceGroup}
\usage{
inspector_create_assessment_target(
  assessmentTargetName,
  resourceGroupArn = NULL
)
}
\arguments{
\item{assessmentTargetName}{[required] The user-defined name that identifies the assessment target that you want to create. The name must be unique within the AWS account.}

\item{resourceGroupArn}{The ARN that specifies the resource group that is used to create the assessment target. If resourceGroupArn is not specified, all EC2 instances in the current AWS account and region are included in the assessment target.}
}
\description{
Creates a new assessment target using the ARN of the resource group that is generated by \code{\link[=inspector_create_resource_group]{create_resource_group}}. If resourceGroupArn is not specified, all EC2 instances in the current AWS account and region are included in the assessment target. If the service-linked role isn’t already registered, this action also creates and registers a service-linked role to grant Amazon Inspector access to AWS Services needed to perform security assessments. You can create up to 50 assessment targets per AWS account. You can run up to 500 concurrent agents per AWS account. For more information, see Amazon Inspector Assessment Targets.

See \url{https://www.paws-r-sdk.com/docs/inspector_create_assessment_target/} for full documentation.
}
\keyword{internal}
