% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_create_exclusions_preview}
\alias{inspector_create_exclusions_preview}
\title{Starts the generation of an exclusions preview for the specified
assessment template}
\usage{
inspector_create_exclusions_preview(assessmentTemplateArn)
}
\arguments{
\item{assessmentTemplateArn}{[required] The ARN that specifies the assessment template for which you want to
create an exclusions preview.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  previewToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Starts the generation of an exclusions preview for the specified
assessment template. The exclusions preview lists the potential
exclusions (ExclusionPreview) that Inspector can detect before it runs
the assessment.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_exclusions_preview(
  assessmentTemplateArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
