% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_delete_assessment_run}
\alias{inspector_delete_assessment_run}
\title{Deletes the assessment run that is specified by the ARN of the
assessment run}
\usage{
inspector_delete_assessment_run(assessmentRunArn)
}
\arguments{
\item{assessmentRunArn}{[required] The ARN that specifies the assessment run that you want to delete.}
}
\description{
Deletes the assessment run that is specified by the ARN of the assessment run.

See \url{https://paws-r.github.io/docs/inspector/delete_assessment_run.html} for full documentation.
}
\keyword{internal}
