% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_delete_assessment_target}
\alias{inspector_delete_assessment_target}
\title{Deletes the assessment target that is specified by the ARN of the
assessment target}
\usage{
inspector_delete_assessment_target(assessmentTargetArn)
}
\arguments{
\item{assessmentTargetArn}{[required] The ARN that specifies the assessment target that you want to delete.}
}
\description{
Deletes the assessment target that is specified by the ARN of the assessment target.

See \url{https://paws-r-sdk.com/docs/delete_assessment_target/} for full documentation.
}
\keyword{internal}
