% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_describe_cross_account_access_role}
\alias{inspector_describe_cross_account_access_role}
\title{Describes the IAM role that enables Amazon Inspector to access your AWS
account}
\usage{
inspector_describe_cross_account_access_role()
}
\description{
Describes the IAM role that enables Amazon Inspector to access your AWS account.

See \url{https://www.paws-r-sdk.com/docs/inspector_describe_cross_account_access_role/} for full documentation.
}
\keyword{internal}
