% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_describe_exclusions}
\alias{inspector_describe_exclusions}
\title{Describes the exclusions that are specified by the exclusions' ARNs}
\usage{
inspector_describe_exclusions(exclusionArns, locale)
}
\arguments{
\item{exclusionArns}{[required] The list of ARNs that specify the exclusions that you want to describe.}

\item{locale}{The locale into which you want to translate the exclusion\'s title,
description, and recommendation.}
}
\description{
Describes the exclusions that are specified by the exclusions\' ARNs.
}
\section{Request syntax}{
\preformatted{svc$describe_exclusions(
  exclusionArns = list(
    "string"
  ),
  locale = "EN_US"
)
}
}

\keyword{internal}
