% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_get_assessment_report}
\alias{inspector_get_assessment_report}
\title{Produces an assessment report that includes detailed and comprehensive
results of a specified assessment run}
\usage{
inspector_get_assessment_report(assessmentRunArn, reportFileFormat, reportType)
}
\arguments{
\item{assessmentRunArn}{[required] The ARN that specifies the assessment run for which you want to generate a report.}

\item{reportFileFormat}{[required] Specifies the file format (html or pdf) of the assessment report that you want to generate.}

\item{reportType}{[required] Specifies the type of the assessment report that you want to generate. There are two types of assessment reports: a finding report and a full report. For more information, see Assessment Reports.}
}
\description{
Produces an assessment report that includes detailed and comprehensive results of a specified assessment run.

See \url{https://www.paws-r-sdk.com/docs/inspector_get_assessment_report/} for full documentation.
}
\keyword{internal}
