% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_list_assessment_runs}
\alias{inspector_list_assessment_runs}
\title{Lists the assessment runs that correspond to the assessment templates
that are specified by the ARNs of the assessment templates}
\usage{
inspector_list_assessment_runs(assessmentTemplateArns, filter,
  nextToken, maxResults)
}
\arguments{
\item{assessmentTemplateArns}{The ARNs that specify the assessment templates whose assessment runs you
want to list.}

\item{filter}{You can use this parameter to specify a subset of data to be included in
the action's response.

For a record to match a filter, all specified filter attributes must
match. When multiple values are specified for a filter attribute, any of
the values can match.}

\item{nextToken}{You can use this parameter when paginating results. Set the value of
this parameter to null on your first call to the \strong{ListAssessmentRuns}
action. Subsequent calls to the action fill \strong{nextToken} in the request
with the value of \strong{NextToken} from the previous response to continue
listing data.}

\item{maxResults}{You can use this parameter to indicate the maximum number of items that
you want in the response. The default value is 10. The maximum value is
500.}
}
\value{
A list with the following syntax:\preformatted{list(
  assessmentRunArns = list(
    "string"
  ),
  nextToken = "string"
)
}
}
\description{
Lists the assessment runs that correspond to the assessment templates
that are specified by the ARNs of the assessment templates.
}
\section{Request syntax}{
\preformatted{svc$list_assessment_runs(
  assessmentTemplateArns = list(
    "string"
  ),
  filter = list(
    namePattern = "string",
    states = list(
      "CREATED"|"START_DATA_COLLECTION_PENDING"|"START_DATA_COLLECTION_IN_PROGRESS"|"COLLECTING_DATA"|"STOP_DATA_COLLECTION_PENDING"|"DATA_COLLECTED"|"START_EVALUATING_RULES_PENDING"|"EVALUATING_RULES"|"FAILED"|"ERROR"|"COMPLETED"|"COMPLETED_WITH_ERRORS"|"CANCELED"
    ),
    durationRange = list(
      minSeconds = 123,
      maxSeconds = 123
    ),
    rulesPackageArns = list(
      "string"
    ),
    startTimeRange = list(
      beginDate = as.POSIXct(
        "2015-01-01"
      ),
      endDate = as.POSIXct(
        "2015-01-01"
      )
    ),
    completionTimeRange = list(
      beginDate = as.POSIXct(
        "2015-01-01"
      ),
      endDate = as.POSIXct(
        "2015-01-01"
      )
    ),
    stateChangeTimeRange = list(
      beginDate = as.POSIXct(
        "2015-01-01"
      ),
      endDate = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\examples{
\dontrun{
# Lists the assessment runs that correspond to the assessment templates
# that are specified by the ARNs of the assessment templates.
svc$list_assessment_runs(
  assessmentTemplateArns = list(
    "arn:aws:inspector:us-west-2:123456789012:target/0-0kFIPusq/template/0-4r1V2mAw"
  ),
  maxResults = 123L
)
}

}
\keyword{internal}
