% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_list_exclusions}
\alias{inspector_list_exclusions}
\title{List exclusions that are generated by the assessment run}
\usage{
inspector_list_exclusions(
  assessmentRunArn,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{assessmentRunArn}{[required] The ARN of the assessment run that generated the exclusions that you
want to list.}

\item{nextToken}{You can use this parameter when paginating results. Set the value of
this parameter to null on your first call to the ListExclusionsRequest
action. Subsequent calls to the action fill nextToken in the request
with the value of nextToken from the previous response to continue
listing data.}

\item{maxResults}{You can use this parameter to indicate the maximum number of items you
want in the response. The default value is 100. The maximum value is
500.}
}
\description{
List exclusions that are generated by the assessment run.

See \url{https://www.paws-r-sdk.com/docs/inspector_list_exclusions/} for full documentation.
}
\keyword{internal}
