% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_subscribe_to_event}
\alias{inspector_subscribe_to_event}
\title{Enables the process of sending Amazon Simple Notification Service (SNS)
notifications about a specified event to a specified SNS topic}
\usage{
inspector_subscribe_to_event(resourceArn, event, topicArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the assessment template that is used during the event for
which you want to receive SNS notifications.}

\item{event}{[required] The event for which you want to receive SNS notifications.}

\item{topicArn}{[required] The ARN of the SNS topic to which the SNS notifications are sent.}
}
\description{
Enables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic.

See \url{https://www.paws-r-sdk.com/docs/inspector_subscribe_to_event/} for full documentation.
}
\keyword{internal}
