% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms_operations.R
\name{kms_connect_custom_key_store}
\alias{kms_connect_custom_key_store}
\title{Connects or reconnects a custom key store to its associated CloudHSM
cluster}
\usage{
kms_connect_custom_key_store(CustomKeyStoreId)
}
\arguments{
\item{CustomKeyStoreId}{[required] Enter the key store ID of the custom key store that you want to connect.
To find the ID of a custom key store, use the
\code{\link[=kms_describe_custom_key_stores]{describe_custom_key_stores}}
operation.}
}
\description{
Connects or reconnects a \href{https://docs.aws.amazon.com/kms/latest/developerguide/custom-key-store-overview.html}{custom key store} to its associated CloudHSM cluster.

See \url{https://paws-r.github.io/docs/kms/connect_custom_key_store.html} for full documentation.
}
\keyword{internal}
