% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms_operations.R
\name{kms_delete_custom_key_store}
\alias{kms_delete_custom_key_store}
\title{Deletes a custom key store}
\usage{
kms_delete_custom_key_store(CustomKeyStoreId)
}
\arguments{
\item{CustomKeyStoreId}{[required] Enter the ID of the custom key store you want to delete. To find the ID
of a custom key store, use the
\code{\link[=kms_describe_custom_key_stores]{describe_custom_key_stores}}
operation.}
}
\description{
Deletes a \href{https://docs.aws.amazon.com/kms/latest/developerguide/key-store-overview.html}{custom key store}. This operation does not affect any backing elements of the custom key store. It does not delete the CloudHSM cluster that is associated with an CloudHSM key store, or affect any users or keys in the cluster. For an external key store, it does not affect the external key store proxy, external key manager, or any external keys.

See \url{https://www.paws-r-sdk.com/docs/kms_delete_custom_key_store/} for full documentation.
}
\keyword{internal}
