% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms_operations.R
\name{kms_get_key_last_usage}
\alias{kms_get_key_last_usage}
\title{Returns usage information about the last successful cryptographic
operation performed with a specified KMS key, including the operation
type, timestamp, and associated CloudTrail event ID}
\usage{
kms_get_key_last_usage(KeyId)
}
\arguments{
\item{KeyId}{[required] Identifies the KMS key to get usage information for. To specify a KMS
key, use its key ID or key ARN. Alias names are not supported.

Specify the key ID or key ARN of the KMS key.

For example:
\itemize{
\item Key ID: \verb{1234abcd-12ab-34cd-56ef-1234567890ab}
\item Key ARN:
\verb{arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab}
}

To get the key ID and key ARN for a KMS key, use
\code{\link[=kms_list_keys]{list_keys}} or \code{\link[=kms_describe_key]{describe_key}}.}
}
\description{
Returns usage information about the last successful cryptographic operation performed with a specified KMS key, including the operation type, timestamp, and associated CloudTrail event ID.

See \url{https://www.paws-r-sdk.com/docs/kms_get_key_last_usage/} for full documentation.
}
\keyword{internal}
