% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_accept_invitation}
\alias{macie2_accept_invitation}
\title{Accepts an Amazon Macie membership invitation that was received from a
specific account}
\usage{
macie2_accept_invitation(
  administratorAccountId = NULL,
  invitationId,
  masterAccount = NULL
)
}
\arguments{
\item{administratorAccountId}{The Amazon Web Services account ID for the account that sent the
invitation.}

\item{invitationId}{[required] The unique identifier for the invitation to accept.}

\item{masterAccount}{(Deprecated) The Amazon Web Services account ID for the account that
sent the invitation. This property has been replaced by the
administratorAccountId property and is retained only for backward
compatibility.}
}
\description{
Accepts an Amazon Macie membership invitation that was received from a specific account.

See \url{https://www.paws-r-sdk.com/docs/macie2_accept_invitation/} for full documentation.
}
\keyword{internal}
