% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_batch_get_custom_data_identifiers}
\alias{macie2_batch_get_custom_data_identifiers}
\title{Retrieves information about one or more custom data identifiers}
\usage{
macie2_batch_get_custom_data_identifiers(ids = NULL)
}
\arguments{
\item{ids}{An array of custom data identifier IDs, one for each custom data
identifier to retrieve information about.}
}
\description{
Retrieves information about one or more custom data identifiers.

See \url{https://www.paws-r-sdk.com/docs/macie2_batch_get_custom_data_identifiers/} for full documentation.
}
\keyword{internal}
