% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_batch_update_automated_discovery_accounts}
\alias{macie2_batch_update_automated_discovery_accounts}
\title{Changes the status of automated sensitive data discovery for one or more
accounts}
\usage{
macie2_batch_update_automated_discovery_accounts(accounts = NULL)
}
\arguments{
\item{accounts}{An array of objects, one for each account to change the status of
automated sensitive data discovery for. Each object specifies the Amazon
Web Services account ID for an account and a new status for that
account.}
}
\description{
Changes the status of automated sensitive data discovery for one or more accounts.

See \url{https://www.paws-r-sdk.com/docs/macie2_batch_update_automated_discovery_accounts/} for full documentation.
}
\keyword{internal}
