% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_create_sample_findings}
\alias{macie2_create_sample_findings}
\title{Creates sample findings}
\usage{
macie2_create_sample_findings(findingTypes = NULL)
}
\arguments{
\item{findingTypes}{An array of finding types, one for each type of sample finding to
create. To create a sample of every type of finding that Amazon Macie
supports, don't include this array in your request.}
}
\description{
Creates sample findings.

See \url{https://www.paws-r-sdk.com/docs/macie2_create_sample_findings/} for full documentation.
}
\keyword{internal}
