% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_delete_custom_data_identifier}
\alias{macie2_delete_custom_data_identifier}
\title{Soft deletes a custom data identifier}
\usage{
macie2_delete_custom_data_identifier(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource or account that the
request applies to.}
}
\description{
Soft deletes a custom data identifier.

See \url{https://paws-r.github.io/docs/macie2/delete_custom_data_identifier.html} for full documentation.
}
\keyword{internal}
