% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_delete_findings_filter}
\alias{macie2_delete_findings_filter}
\title{Deletes a findings filter}
\usage{
macie2_delete_findings_filter(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}
}
\description{
Deletes a findings filter.

See \url{https://paws-r.github.io/docs/macie2/delete_findings_filter.html} for full documentation.
}
\keyword{internal}
