% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_delete_member}
\alias{macie2_delete_member}
\title{Deletes the association between an Amazon Macie administrator account
and an account}
\usage{
macie2_delete_member(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}
}
\description{
Deletes the association between an Amazon Macie administrator account and an account.

See \url{https://paws-r.github.io/docs/macie2/delete_member.html} for full documentation.
}
\keyword{internal}
