% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_describe_classification_job}
\alias{macie2_describe_classification_job}
\title{Retrieves the status and settings for a classification job}
\usage{
macie2_describe_classification_job(jobId)
}
\arguments{
\item{jobId}{[required] The unique identifier for the classification job.}
}
\description{
Retrieves the status and settings for a classification job.

See \url{https://www.paws-r-sdk.com/docs/macie2_describe_classification_job/} for full documentation.
}
\keyword{internal}
