% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_describe_organization_configuration}
\alias{macie2_describe_organization_configuration}
\title{Retrieves the Amazon Macie configuration settings for an organization in
Organizations}
\usage{
macie2_describe_organization_configuration()
}
\description{
Retrieves the Amazon Macie configuration settings for an organization in Organizations.

See \url{https://paws-r.github.io/docs/macie2/describe_organization_configuration.html} for full documentation.
}
\keyword{internal}
