% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_disable_macie}
\alias{macie2_disable_macie}
\title{Disables Amazon Macie and deletes all settings and resources for a Macie
account}
\usage{
macie2_disable_macie()
}
\description{
Disables Amazon Macie and deletes all settings and resources for a Macie account.

See \url{https://www.paws-r-sdk.com/docs/macie2_disable_macie/} for full documentation.
}
\keyword{internal}
