% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_disable_organization_admin_account}
\alias{macie2_disable_organization_admin_account}
\title{Disables an account as the delegated Amazon Macie administrator account
for an organization in Organizations}
\usage{
macie2_disable_organization_admin_account(adminAccountId)
}
\arguments{
\item{adminAccountId}{[required] The Amazon Web Services account ID of the delegated Amazon Macie
administrator account.}
}
\description{
Disables an account as the delegated Amazon Macie administrator account for an organization in Organizations.

See \url{https://www.paws-r-sdk.com/docs/macie2_disable_organization_admin_account/} for full documentation.
}
\keyword{internal}
