% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_disassociate_from_administrator_account}
\alias{macie2_disassociate_from_administrator_account}
\title{Disassociates a member account from its Amazon Macie administrator
account}
\usage{
macie2_disassociate_from_administrator_account()
}
\description{
Disassociates a member account from its Amazon Macie administrator account.

See \url{https://www.paws-r-sdk.com/docs/macie2_disassociate_from_administrator_account/} for full documentation.
}
\keyword{internal}
