% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_disassociate_from_master_account}
\alias{macie2_disassociate_from_master_account}
\title{(Deprecated) Disassociates a member account from its Amazon Macie
administrator account}
\usage{
macie2_disassociate_from_master_account()
}
\description{
(Deprecated) Disassociates a member account from its Amazon Macie administrator account. This operation has been replaced by the DisassociateFromAdministratorAccount operation.

See \url{https://www.paws-r-sdk.com/docs/macie2_disassociate_from_master_account/} for full documentation.
}
\keyword{internal}
