% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_disassociate_member}
\alias{macie2_disassociate_member}
\title{Disassociates an Amazon Macie administrator account from a member
account}
\usage{
macie2_disassociate_member(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}
}
\description{
Disassociates an Amazon Macie administrator account from a member account.

See \url{https://www.paws-r-sdk.com/docs/macie2_disassociate_member/} for full documentation.
}
\keyword{internal}
