% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_enable_macie}
\alias{macie2_enable_macie}
\title{Enables Amazon Macie and specifies the configuration settings for a
Macie account}
\usage{
macie2_enable_macie(
  clientToken = NULL,
  findingPublishingFrequency = NULL,
  status = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive token that you provide to ensure the
idempotency of the request.}

\item{findingPublishingFrequency}{Specifies how often to publish updates to policy findings for the
account. This includes publishing updates to Security Hub and Amazon
EventBridge (formerly Amazon CloudWatch Events).}

\item{status}{Specifies the new status for the account. To enable Amazon Macie and
start all Macie activities for the account, set this value to ENABLED.}
}
\description{
Enables Amazon Macie and specifies the configuration settings for a Macie account.

See \url{https://www.paws-r-sdk.com/docs/macie2_enable_macie/} for full documentation.
}
\keyword{internal}
