% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_enable_organization_admin_account}
\alias{macie2_enable_organization_admin_account}
\title{Designates an account as the delegated Amazon Macie administrator
account for an organization in Organizations}
\usage{
macie2_enable_organization_admin_account(adminAccountId, clientToken = NULL)
}
\arguments{
\item{adminAccountId}{[required] The Amazon Web Services account ID for the account to designate as the
delegated Amazon Macie administrator account for the organization.}

\item{clientToken}{A unique, case-sensitive token that you provide to ensure the
idempotency of the request.}
}
\description{
Designates an account as the delegated Amazon Macie administrator account for an organization in Organizations.

See \url{https://www.paws-r-sdk.com/docs/macie2_enable_organization_admin_account/} for full documentation.
}
\keyword{internal}
