% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_administrator_account}
\alias{macie2_get_administrator_account}
\title{Retrieves information about the Amazon Macie administrator account for
an account}
\usage{
macie2_get_administrator_account()
}
\description{
Retrieves information about the Amazon Macie administrator account for an account.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_administrator_account/} for full documentation.
}
\keyword{internal}
