% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_allow_list}
\alias{macie2_get_allow_list}
\title{Retrieves the settings and status of an allow list}
\usage{
macie2_get_allow_list(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}
}
\description{
Retrieves the settings and status of an allow list.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_allow_list/} for full documentation.
}
\keyword{internal}
