% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_automated_discovery_configuration}
\alias{macie2_get_automated_discovery_configuration}
\title{Retrieves the configuration settings and status of automated sensitive
data discovery for an account}
\usage{
macie2_get_automated_discovery_configuration()
}
\description{
Retrieves the configuration settings and status of automated sensitive data discovery for an account.

See \url{https://paws-r.github.io/docs/macie2/get_automated_discovery_configuration.html} for full documentation.
}
\keyword{internal}
