% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_classification_export_configuration}
\alias{macie2_get_classification_export_configuration}
\title{Retrieves the configuration settings for storing data classification
results}
\usage{
macie2_get_classification_export_configuration()
}
\description{
Retrieves the configuration settings for storing data classification results.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_classification_export_configuration/} for full documentation.
}
\keyword{internal}
