% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_classification_scope}
\alias{macie2_get_classification_scope}
\title{Retrieves the classification scope settings for an account}
\usage{
macie2_get_classification_scope(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}
}
\description{
Retrieves the classification scope settings for an account.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_classification_scope/} for full documentation.
}
\keyword{internal}
