% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_findings}
\alias{macie2_get_findings}
\title{Retrieves the details of one or more findings}
\usage{
macie2_get_findings(findingIds, sortCriteria = NULL)
}
\arguments{
\item{findingIds}{[required] An array of strings that lists the unique identifiers for the findings
to retrieve. You can specify as many as 50 unique identifiers in this
array.}

\item{sortCriteria}{The criteria for sorting the results of the request.}
}
\description{
Retrieves the details of one or more findings.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_findings/} for full documentation.
}
\keyword{internal}
