% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_findings_filter}
\alias{macie2_get_findings_filter}
\title{Retrieves the criteria and other settings for a findings filter}
\usage{
macie2_get_findings_filter(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}
}
\description{
Retrieves the criteria and other settings for a findings filter.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_findings_filter/} for full documentation.
}
\keyword{internal}
