% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_member}
\alias{macie2_get_member}
\title{Retrieves information about an account that's associated with an Amazon
Macie administrator account}
\usage{
macie2_get_member(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource or account that the
request applies to.}
}
\description{
Retrieves information about an account that's associated with an Amazon Macie administrator account.

See \url{https://paws-r.github.io/docs/macie2/get_member.html} for full documentation.
}
\keyword{internal}
