% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_resource_profile}
\alias{macie2_get_resource_profile}
\title{Retrieves (queries) sensitive data discovery statistics and the
sensitivity score for an S3 bucket}
\usage{
macie2_get_resource_profile(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the S3 bucket that the request applies
to.}
}
\description{
Retrieves (queries) sensitive data discovery statistics and the sensitivity score for an S3 bucket.

See \url{https://paws-r.github.io/docs/macie2/get_resource_profile.html} for full documentation.
}
\keyword{internal}
