% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_sensitive_data_occurrences}
\alias{macie2_get_sensitive_data_occurrences}
\title{Retrieves occurrences of sensitive data reported by a finding}
\usage{
macie2_get_sensitive_data_occurrences(findingId)
}
\arguments{
\item{findingId}{[required] The unique identifier for the finding.}
}
\description{
Retrieves occurrences of sensitive data reported by a finding.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_sensitive_data_occurrences/} for full documentation.
}
\keyword{internal}
