% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_sensitive_data_occurrences_availability}
\alias{macie2_get_sensitive_data_occurrences_availability}
\title{Checks whether occurrences of sensitive data can be retrieved for a
finding}
\usage{
macie2_get_sensitive_data_occurrences_availability(findingId)
}
\arguments{
\item{findingId}{[required] The unique identifier for the finding.}
}
\description{
Checks whether occurrences of sensitive data can be retrieved for a finding.

See \url{https://paws-r.github.io/docs/macie2/get_sensitive_data_occurrences_availability.html} for full documentation.
}
\keyword{internal}
