% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_sensitivity_inspection_template}
\alias{macie2_get_sensitivity_inspection_template}
\title{Retrieves the settings for the sensitivity inspection template for an
account}
\usage{
macie2_get_sensitivity_inspection_template(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}
}
\description{
Retrieves the settings for the sensitivity inspection template for an account.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_sensitivity_inspection_template/} for full documentation.
}
\keyword{internal}
