% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_usage_totals}
\alias{macie2_get_usage_totals}
\title{Retrieves (queries) aggregated usage data for an account}
\usage{
macie2_get_usage_totals(timeRange = NULL)
}
\arguments{
\item{timeRange}{The inclusive time period to retrieve the data for. Valid values are:
MONTH_TO_DATE, for the current calendar month to date; and,
PAST_30_DAYS, for the preceding 30 days. If you don't specify a value
for this parameter, Amazon Macie provides aggregated usage data for the
preceding 30 days.}
}
\description{
Retrieves (queries) aggregated usage data for an account.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_usage_totals/} for full documentation.
}
\keyword{internal}
