% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_allow_lists}
\alias{macie2_list_allow_lists}
\title{Retrieves a subset of information about all the allow lists for an
account}
\usage{
macie2_list_allow_lists(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of items to include in each page of a paginated
response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves a subset of information about all the allow lists for an account.

See \url{https://paws-r.github.io/docs/macie2/list_allow_lists.html} for full documentation.
}
\keyword{internal}
