% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_automated_discovery_accounts}
\alias{macie2_list_automated_discovery_accounts}
\title{Retrieves the status of automated sensitive data discovery for one or
more accounts}
\usage{
macie2_list_automated_discovery_accounts(
  accountIds = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{accountIds}{The Amazon Web Services account ID for each account, for as many as 50
accounts. To retrieve the status for multiple accounts, append the
accountIds parameter and argument for each account, separated by an
ampersand (&). To retrieve the status for all the accounts in an
organization, omit this parameter.}

\item{maxResults}{The maximum number of items to include in each page of a paginated
response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves the status of automated sensitive data discovery for one or more accounts.

See \url{https://www.paws-r-sdk.com/docs/macie2_list_automated_discovery_accounts/} for full documentation.
}
\keyword{internal}
