% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_classification_scopes}
\alias{macie2_list_classification_scopes}
\title{Retrieves a subset of information about the classification scope for an
account}
\usage{
macie2_list_classification_scopes(name = NULL, nextToken = NULL)
}
\arguments{
\item{name}{The name of the classification scope to retrieve the unique identifier
for.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves a subset of information about the classification scope for an account.

See \url{https://www.paws-r-sdk.com/docs/macie2_list_classification_scopes/} for full documentation.
}
\keyword{internal}
