% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_custom_data_identifiers}
\alias{macie2_list_custom_data_identifiers}
\title{Retrieves a subset of information about all the custom data identifiers
for an account}
\usage{
macie2_list_custom_data_identifiers(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of items to include in each page of the response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves a subset of information about all the custom data identifiers for an account.

See \url{https://www.paws-r-sdk.com/docs/macie2_list_custom_data_identifiers/} for full documentation.
}
\keyword{internal}
