% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_findings}
\alias{macie2_list_findings}
\title{Retrieves a subset of information about one or more findings}
\usage{
macie2_list_findings(
  findingCriteria = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sortCriteria = NULL
)
}
\arguments{
\item{findingCriteria}{The criteria to use to filter the results.}

\item{maxResults}{The maximum number of items to include in each page of the response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}

\item{sortCriteria}{The criteria to use to sort the results.}
}
\description{
Retrieves a subset of information about one or more findings.

See \url{https://paws-r.github.io/docs/macie2/list_findings.html} for full documentation.
}
\keyword{internal}
