% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_invitations}
\alias{macie2_list_invitations}
\title{Retrieves information about the Amazon Macie membership invitations that
were received by an account}
\usage{
macie2_list_invitations(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of items to include in each page of a paginated
response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves information about the Amazon Macie membership invitations that were received by an account.

See \url{https://www.paws-r-sdk.com/docs/macie2_list_invitations/} for full documentation.
}
\keyword{internal}
