% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_organization_admin_accounts}
\alias{macie2_list_organization_admin_accounts}
\title{Retrieves information about the delegated Amazon Macie administrator
account for an organization in Organizations}
\usage{
macie2_list_organization_admin_accounts(maxResults = NULL, nextToken = NULL)
}
\arguments{
\item{maxResults}{The maximum number of items to include in each page of a paginated
response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves information about the delegated Amazon Macie administrator account for an organization in Organizations.

See \url{https://www.paws-r-sdk.com/docs/macie2_list_organization_admin_accounts/} for full documentation.
}
\keyword{internal}
