% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_resource_profile_artifacts}
\alias{macie2_list_resource_profile_artifacts}
\title{Retrieves information about objects that were selected from an S3 bucket
for automated sensitive data discovery}
\usage{
macie2_list_resource_profile_artifacts(nextToken = NULL, resourceArn)
}
\arguments{
\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the S3 bucket that the request applies
to.}
}
\description{
Retrieves information about objects that were selected from an S3 bucket for automated sensitive data discovery.

See \url{https://www.paws-r-sdk.com/docs/macie2_list_resource_profile_artifacts/} for full documentation.
}
\keyword{internal}
