% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_resource_profile_detections}
\alias{macie2_list_resource_profile_detections}
\title{Retrieves information about the types and amount of sensitive data that
Amazon Macie found in an S3 bucket}
\usage{
macie2_list_resource_profile_detections(
  maxResults = NULL,
  nextToken = NULL,
  resourceArn
)
}
\arguments{
\item{maxResults}{The maximum number of items to include in each page of a paginated
response.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}

\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the S3 bucket that the request applies
to.}
}
\description{
Retrieves information about the types and amount of sensitive data that Amazon Macie found in an S3 bucket.

See \url{https://www.paws-r-sdk.com/docs/macie2_list_resource_profile_detections/} for full documentation.
}
\keyword{internal}
