% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_put_classification_export_configuration}
\alias{macie2_put_classification_export_configuration}
\title{Creates or updates the configuration settings for storing data
classification results}
\usage{
macie2_put_classification_export_configuration(configuration)
}
\arguments{
\item{configuration}{[required] The location to store data classification results in, and the encryption
settings to use when storing results in that location.}
}
\description{
Creates or updates the configuration settings for storing data classification results.

See \url{https://www.paws-r-sdk.com/docs/macie2_put_classification_export_configuration/} for full documentation.
}
\keyword{internal}
