% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_put_findings_publication_configuration}
\alias{macie2_put_findings_publication_configuration}
\title{Updates the configuration settings for publishing findings to Security
Hub}
\usage{
macie2_put_findings_publication_configuration(
  clientToken = NULL,
  securityHubConfiguration = NULL
)
}
\arguments{
\item{clientToken}{A unique, case-sensitive token that you provide to ensure the
idempotency of the request.}

\item{securityHubConfiguration}{The configuration settings that determine which findings to publish to
Security Hub.}
}
\description{
Updates the configuration settings for publishing findings to Security Hub.

See \url{https://www.paws-r-sdk.com/docs/macie2_put_findings_publication_configuration/} for full documentation.
}
\keyword{internal}
