% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_search_resources}
\alias{macie2_search_resources}
\title{Retrieves (queries) statistical data and other information about Amazon
Web Services resources that Amazon Macie monitors and analyzes}
\usage{
macie2_search_resources(
  bucketCriteria = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sortCriteria = NULL
)
}
\arguments{
\item{bucketCriteria}{The filter conditions that determine which S3 buckets to include or
exclude from the query results.}

\item{maxResults}{The maximum number of items to include in each page of the response. The
default value is 50.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}

\item{sortCriteria}{The criteria to use to sort the results.}
}
\description{
Retrieves (queries) statistical data and other information about Amazon Web Services resources that Amazon Macie monitors and analyzes.

See \url{https://www.paws-r-sdk.com/docs/macie2_search_resources/} for full documentation.
}
\keyword{internal}
