% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_allow_list}
\alias{macie2_update_allow_list}
\title{Updates the settings for an allow list}
\usage{
macie2_update_allow_list(criteria, description = NULL, id, name)
}
\arguments{
\item{criteria}{[required] The criteria that specify the text or text pattern to ignore. The
criteria can be the location and name of an S3 object that lists
specific text to ignore (s3WordsList), or a regular expression that
defines a text pattern to ignore (regex).

You can change a list's underlying criteria, such as the name of the S3
object or the regular expression to use. However, you can't change the
type from s3WordsList to regex or the other way around.}

\item{description}{A custom description of the allow list. The description can contain as
many as 512 characters.}

\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}

\item{name}{[required] A custom name for the allow list. The name can contain as many as 128
characters.}
}
\description{
Updates the settings for an allow list.

See \url{https://www.paws-r-sdk.com/docs/macie2_update_allow_list/} for full documentation.
}
\keyword{internal}
