% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_classification_scope}
\alias{macie2_update_classification_scope}
\title{Updates the classification scope settings for an account}
\usage{
macie2_update_classification_scope(id, s3 = NULL)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}

\item{s3}{The S3 buckets to add or remove from the exclusion list defined by the
classification scope.}
}
\description{
Updates the classification scope settings for an account.

See \url{https://www.paws-r-sdk.com/docs/macie2_update_classification_scope/} for full documentation.
}
\keyword{internal}
