% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_macie_session}
\alias{macie2_update_macie_session}
\title{Suspends or re-enables Amazon Macie, or updates the configuration
settings for a Macie account}
\usage{
macie2_update_macie_session(findingPublishingFrequency = NULL, status = NULL)
}
\arguments{
\item{findingPublishingFrequency}{Specifies how often to publish updates to policy findings for the
account. This includes publishing updates to Security Hub and Amazon
EventBridge (formerly Amazon CloudWatch Events).}

\item{status}{Specifies a new status for the account. Valid values are: ENABLED,
resume all Amazon Macie activities for the account; and, PAUSED, suspend
all Macie activities for the account.}
}
\description{
Suspends or re-enables Amazon Macie, or updates the configuration settings for a Macie account.

See \url{https://www.paws-r-sdk.com/docs/macie2_update_macie_session/} for full documentation.
}
\keyword{internal}
