% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_member_session}
\alias{macie2_update_member_session}
\title{Enables an Amazon Macie administrator to suspend or re-enable Macie for
a member account}
\usage{
macie2_update_member_session(id, status)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource or account that the
request applies to.}

\item{status}{[required] Specifies the new status for the account. Valid values are: ENABLED,
resume all Amazon Macie activities for the account; and, PAUSED, suspend
all Macie activities for the account.}
}
\description{
Enables an Amazon Macie administrator to suspend or re-enable Macie for a member account.

See \url{https://paws-r.github.io/docs/macie2/update_member_session.html} for full documentation.
}
\keyword{internal}
