% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_organization_configuration}
\alias{macie2_update_organization_configuration}
\title{Updates the Amazon Macie configuration settings for an organization in
Organizations}
\usage{
macie2_update_organization_configuration(autoEnable)
}
\arguments{
\item{autoEnable}{[required] Specifies whether to enable Amazon Macie automatically for an account
when the account is added to the organization in Organizations.}
}
\description{
Updates the Amazon Macie configuration settings for an organization in Organizations.

See \url{https://paws-r.github.io/docs/macie2/update_organization_configuration.html} for full documentation.
}
\keyword{internal}
