% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_resource_profile}
\alias{macie2_update_resource_profile}
\title{Updates the sensitivity score for an S3 bucket}
\usage{
macie2_update_resource_profile(resourceArn, sensitivityScoreOverride = NULL)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the S3 bucket that the request applies
to.}

\item{sensitivityScoreOverride}{The new sensitivity score for the bucket. Valid values are: 100, assign
the maximum score and apply the \emph{Sensitive} label to the bucket; and,
null (empty), assign a score that Amazon Macie calculates automatically
after you submit the request.}
}
\description{
Updates the sensitivity score for an S3 bucket.

See \url{https://www.paws-r-sdk.com/docs/macie2_update_resource_profile/} for full documentation.
}
\keyword{internal}
