% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_resource_profile_detections}
\alias{macie2_update_resource_profile_detections}
\title{Updates the sensitivity scoring settings for an S3 bucket}
\usage{
macie2_update_resource_profile_detections(
  resourceArn,
  suppressDataIdentifiers = NULL
)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the S3 bucket that the request applies
to.}

\item{suppressDataIdentifiers}{An array of objects, one for each custom data identifier or managed data
identifier that detected the type of sensitive data to start excluding
or including in the bucket's score. To start including all sensitive
data types in the score, don't specify any values for this array.}
}
\description{
Updates the sensitivity scoring settings for an S3 bucket.

See \url{https://paws-r.github.io/docs/macie2/update_resource_profile_detections.html} for full documentation.
}
\keyword{internal}
