% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_reveal_configuration}
\alias{macie2_update_reveal_configuration}
\title{Updates the status and configuration settings for retrieving occurrences
of sensitive data reported by findings}
\usage{
macie2_update_reveal_configuration(configuration)
}
\arguments{
\item{configuration}{[required] The new configuration settings and the status of the configuration for
the account.}
}
\description{
Updates the status and configuration settings for retrieving occurrences of sensitive data reported by findings.

See \url{https://www.paws-r-sdk.com/docs/macie2_update_reveal_configuration/} for full documentation.
}
\keyword{internal}
