% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_sensitivity_inspection_template}
\alias{macie2_update_sensitivity_inspection_template}
\title{Updates the settings for the sensitivity inspection template for an
account}
\usage{
macie2_update_sensitivity_inspection_template(
  description = NULL,
  excludes = NULL,
  id,
  includes = NULL
)
}
\arguments{
\item{description}{A custom description of the template. The description can contain as
many as 200 characters.}

\item{excludes}{The managed data identifiers to explicitly exclude (not use) when
analyzing data.

To exclude an allow list or custom data identifier that's currently
included by the template, update the values for the
SensitivityInspectionTemplateIncludes.allowListIds and
SensitivityInspectionTemplateIncludes.customDataIdentifierIds
properties, respectively.}

\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}

\item{includes}{The allow lists, custom data identifiers, and managed data identifiers
to explicitly include (use) when analyzing data.}
}
\description{
Updates the settings for the sensitivity inspection template for an account.

See \url{https://www.paws-r-sdk.com/docs/macie2_update_sensitivity_inspection_template/} for full documentation.
}
\keyword{internal}
