% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie_operations.R
\name{macie_associate_member_account}
\alias{macie_associate_member_account}
\title{Associates a specified AWS account with Amazon Macie Classic as a member
account}
\usage{
macie_associate_member_account(memberAccountId)
}
\arguments{
\item{memberAccountId}{[required] The ID of the AWS account that you want to associate with Amazon Macie
Classic as a member account.}
}
\value{
An empty list.
}
\description{
Associates a specified AWS account with Amazon Macie Classic as a member
account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_member_account(
  memberAccountId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
