% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie_operations.R
\name{macie_list_member_accounts}
\alias{macie_list_member_accounts}
\title{(Discontinued) Lists all Amazon Macie Classic member accounts for the
current Macie Classic administrator account}
\usage{
macie_list_member_accounts(nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{nextToken}{(Discontinued) Use this parameter when paginating results. Set the value
of this parameter to null on your first call to the
\code{\link[=macie_list_member_accounts]{list_member_accounts}} action. Subsequent
calls to the action fill \code{nextToken} in the request with the value of
\code{nextToken} from the previous response to continue listing data.}

\item{maxResults}{(Discontinued) Use this parameter to indicate the maximum number of
items that you want in the response. The default value is 250.}
}
\description{
(Discontinued) Lists all Amazon Macie Classic member accounts for the current Macie Classic administrator account.

See \url{https://www.paws-r-sdk.com/docs/macie_list_member_accounts/} for full documentation.
}
\keyword{internal}
