% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie_operations.R
\name{macie_update_s3_resources}
\alias{macie_update_s3_resources}
\title{(Discontinued) Updates the classification types for the specified S3
resources}
\usage{
macie_update_s3_resources(memberAccountId = NULL, s3ResourcesUpdate)
}
\arguments{
\item{memberAccountId}{(Discontinued) The Amazon Web Services account ID of the Amazon Macie
Classic member account whose S3 resources' classification types you want
to update.}

\item{s3ResourcesUpdate}{[required] (Discontinued) The S3 resources whose classification types you want to
update.}
}
\description{
(Discontinued) Updates the classification types for the specified S3 resources. If \code{memberAccountId} isn't specified, the action updates the classification types of the S3 resources associated with Amazon Macie Classic for the current Macie Classic administrator account. If \code{memberAccountId} is specified, the action updates the classification types of the S3 resources associated with Macie Classic for the specified member account.

See \url{https://www.paws-r-sdk.com/docs/macie_update_s3_resources/} for full documentation.
}
\keyword{internal}
