% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_create_connector}
\alias{pcaconnectorad_create_connector}
\title{Creates a connector between Amazon Web Services Private CA and an Active
Directory}
\usage{
pcaconnectorad_create_connector(
  CertificateAuthorityArn,
  ClientToken = NULL,
  DirectoryId,
  Tags = NULL,
  VpcInformation
)
}
\arguments{
\item{CertificateAuthorityArn}{[required] The Amazon Resource Name (ARN) of the certificate authority being used.}

\item{ClientToken}{Idempotency token.}

\item{DirectoryId}{[required] The identifier of the Active Directory.}

\item{Tags}{Metadata assigned to a connector consisting of a key-value pair.}

\item{VpcInformation}{[required] Security group IDs that describe the inbound and outbound rules.}
}
\description{
Creates a connector between Amazon Web Services Private CA and an Active Directory. You must specify the private CA, directory ID, and security groups.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_create_connector/} for full documentation.
}
\keyword{internal}
