% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_create_service_principal_name}
\alias{pcaconnectorad_create_service_principal_name}
\title{Creates a service principal name (SPN) for the service account in Active
Directory}
\usage{
pcaconnectorad_create_service_principal_name(
  ClientToken = NULL,
  ConnectorArn,
  DirectoryRegistrationArn
)
}
\arguments{
\item{ClientToken}{Idempotency token.}

\item{ConnectorArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_connector]{create_connector}}.}

\item{DirectoryRegistrationArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_directory_registration]{create_directory_registration}}.}
}
\description{
Creates a service principal name (SPN) for the service account in Active Directory. Kerberos authentication uses SPNs to associate a service instance with a service sign-in account.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_create_service_principal_name/} for full documentation.
}
\keyword{internal}
