% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_delete_directory_registration}
\alias{pcaconnectorad_delete_directory_registration}
\title{Deletes a directory registration}
\usage{
pcaconnectorad_delete_directory_registration(DirectoryRegistrationArn)
}
\arguments{
\item{DirectoryRegistrationArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_directory_registration]{create_directory_registration}}.}
}
\description{
Deletes a directory registration. Deleting a directory registration deauthorizes Amazon Web Services Private CA with the directory.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_delete_directory_registration/} for full documentation.
}
\keyword{internal}
