% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_delete_template}
\alias{pcaconnectorad_delete_template}
\title{Deletes a template}
\usage{
pcaconnectorad_delete_template(TemplateArn)
}
\arguments{
\item{TemplateArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_template]{create_template}}.}
}
\description{
Deletes a template. Certificates issued using the template are still valid until they are revoked or expired.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_delete_template/} for full documentation.
}
\keyword{internal}
