% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_delete_template_group_access_control_entry}
\alias{pcaconnectorad_delete_template_group_access_control_entry}
\title{Deletes a group access control entry}
\usage{
pcaconnectorad_delete_template_group_access_control_entry(
  GroupSecurityIdentifier,
  TemplateArn
)
}
\arguments{
\item{GroupSecurityIdentifier}{[required] Security identifier (SID) of the group object from Active Directory. The
SID starts with "S-".}

\item{TemplateArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_template]{create_template}}.}
}
\description{
Deletes a group access control entry.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_delete_template_group_access_control_entry/} for full documentation.
}
\keyword{internal}
