% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_get_connector}
\alias{pcaconnectorad_get_connector}
\title{Lists information about your connector}
\usage{
pcaconnectorad_get_connector(ConnectorArn)
}
\arguments{
\item{ConnectorArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_connector]{create_connector}}.}
}
\description{
Lists information about your connector. You specify the connector on input by its ARN (Amazon Resource Name).

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_get_connector/} for full documentation.
}
\keyword{internal}
