% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_get_directory_registration}
\alias{pcaconnectorad_get_directory_registration}
\title{A structure that contains information about your directory registration}
\usage{
pcaconnectorad_get_directory_registration(DirectoryRegistrationArn)
}
\arguments{
\item{DirectoryRegistrationArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_directory_registration]{create_directory_registration}}.}
}
\description{
A structure that contains information about your directory registration.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_get_directory_registration/} for full documentation.
}
\keyword{internal}
