% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_get_service_principal_name}
\alias{pcaconnectorad_get_service_principal_name}
\title{Lists the service principal name that the connector uses to authenticate
with Active Directory}
\usage{
pcaconnectorad_get_service_principal_name(
  ConnectorArn,
  DirectoryRegistrationArn
)
}
\arguments{
\item{ConnectorArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_connector]{create_connector}}.}

\item{DirectoryRegistrationArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_directory_registration]{create_directory_registration}}.}
}
\description{
Lists the service principal name that the connector uses to authenticate with Active Directory.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_get_service_principal_name/} for full documentation.
}
\keyword{internal}
