% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_get_template}
\alias{pcaconnectorad_get_template}
\title{Retrieves a certificate template that the connector uses to issue
certificates from a private CA}
\usage{
pcaconnectorad_get_template(TemplateArn)
}
\arguments{
\item{TemplateArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_template]{create_template}}.}
}
\description{
Retrieves a certificate template that the connector uses to issue certificates from a private CA.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_get_template/} for full documentation.
}
\keyword{internal}
