% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_list_directory_registrations}
\alias{pcaconnectorad_list_directory_registrations}
\title{Lists the directory registrations that you created by using the
https://docs}
\usage{
pcaconnectorad_list_directory_registrations(
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{MaxResults}{Use this parameter when paginating results to specify the maximum number
of items to return in the response on each page. If additional items
exist beyond the number you specify, the \code{NextToken} element is sent in
the response. Use this \code{NextToken} value in a subsequent request to
retrieve additional items.}

\item{NextToken}{Use this parameter when paginating results in a subsequent request after
you receive a response with truncated results. Set it to the value of
the \code{NextToken} parameter from the response you just received.}
}
\description{
Lists the directory registrations that you created by using the \href{https://docs.aws.amazon.com/pca-connector-ad/latest/APIReference/API_CreateDirectoryRegistration.html}{https://docs.aws.amazon.com/pca-connector-ad/latest/APIReference/API_CreateDirectoryRegistration} action.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_list_directory_registrations/} for full documentation.
}
\keyword{internal}
