% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_list_tags_for_resource}
\alias{pcaconnectorad_list_tags_for_resource}
\title{Lists the tags, if any, that are associated with your resource}
\usage{
pcaconnectorad_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that was returned when you created the
resource.}
}
\description{
Lists the tags, if any, that are associated with your resource.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
