% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_list_template_group_access_control_entries}
\alias{pcaconnectorad_list_template_group_access_control_entries}
\title{Lists group access control entries you created}
\usage{
pcaconnectorad_list_template_group_access_control_entries(
  MaxResults = NULL,
  NextToken = NULL,
  TemplateArn
)
}
\arguments{
\item{MaxResults}{Use this parameter when paginating results to specify the maximum number
of items to return in the response on each page. If additional items
exist beyond the number you specify, the \code{NextToken} element is sent in
the response. Use this \code{NextToken} value in a subsequent request to
retrieve additional items.}

\item{NextToken}{Use this parameter when paginating results in a subsequent request after
you receive a response with truncated results. Set it to the value of
the \code{NextToken} parameter from the response you just received.}

\item{TemplateArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_template]{create_template}}.}
}
\description{
Lists group access control entries you created.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_list_template_group_access_control_entries/} for full documentation.
}
\keyword{internal}
