% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_tag_resource}
\alias{pcaconnectorad_tag_resource}
\title{Adds one or more tags to your resource}
\usage{
pcaconnectorad_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that was returned when you created the
resource.}

\item{Tags}{[required] Metadata assigned to a directory registration consisting of a key-value
pair.}
}
\description{
Adds one or more tags to your resource.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_tag_resource/} for full documentation.
}
\keyword{internal}
