% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_untag_resource}
\alias{pcaconnectorad_untag_resource}
\title{Removes one or more tags from your resource}
\usage{
pcaconnectorad_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that was returned when you created the
resource.}

\item{TagKeys}{[required] Specifies a list of tag keys that you want to remove from the specified
resources.}
}
\description{
Removes one or more tags from your resource.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_untag_resource/} for full documentation.
}
\keyword{internal}
