% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaconnectorad_operations.R
\name{pcaconnectorad_update_template}
\alias{pcaconnectorad_update_template}
\title{Update template configuration to define the information included in
certificates}
\usage{
pcaconnectorad_update_template(
  Definition = NULL,
  ReenrollAllCertificateHolders = NULL,
  TemplateArn
)
}
\arguments{
\item{Definition}{Template configuration to define the information included in
certificates. Define certificate validity and renewal periods,
certificate request handling and enrollment options, key usage
extensions, application policies, and cryptography settings.}

\item{ReenrollAllCertificateHolders}{This setting allows the major version of a template to be increased
automatically. All members of Active Directory groups that are allowed
to enroll with a template will receive a new certificate issued using
that template.}

\item{TemplateArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=pcaconnectorad_create_template]{create_template}}.}
}
\description{
Update template configuration to define the information included in certificates.

See \url{https://www.paws-r-sdk.com/docs/pcaconnectorad_update_template/} for full documentation.
}
\keyword{internal}
