% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_enable_sharing_with_aws_organization}
\alias{ram_enable_sharing_with_aws_organization}
\title{Enables resource sharing within your AWS Organization}
\usage{
ram_enable_sharing_with_aws_organization()
}
\value{
A list with the following syntax:\preformatted{list(
  returnValue = TRUE|FALSE
)
}
}
\description{
Enables resource sharing within your AWS Organization.

The caller must be the master account for the AWS Organization.
}
\section{Request syntax}{
\preformatted{svc$enable_sharing_with_aws_organization()
}
}

\keyword{internal}
