% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_list_replace_permission_associations_work}
\alias{ram_list_replace_permission_associations_work}
\title{Retrieves the current status of the asynchronous tasks performed by RAM
when you perform the ReplacePermissionAssociationsWork operation}
\usage{
ram_list_replace_permission_associations_work(
  workIds = NULL,
  status = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{workIds}{A list of IDs. These values come from the \code{id}field of the
\code{replacePermissionAssociationsWork}structure returned by the
\code{\link[=ram_replace_permission_associations]{replace_permission_associations}}
operation.}

\item{status}{Specifies that you want to see only the details about requests with a
status that matches this value.}

\item{nextToken}{Specifies that you want to receive the next page of results. Valid only
if you received a \code{NextToken} response in the previous request. If you
did, it indicates that more output is available. Set this parameter to
the value provided by the previous call's \code{NextToken} response to
request the next page of results.}

\item{maxResults}{Specifies the total number of results that you want included on each
page of the response. If you do not include this parameter, it defaults
to a value that is specific to the operation. If additional items exist
beyond the number you specify, the \code{NextToken} response element is
returned with a value (not null). Include the specified value as the
\code{NextToken} request parameter in the next call to the operation to get
the next part of the results. Note that the service might return fewer
results than the maximum even when there are more results available. You
should check \code{NextToken} after every operation to ensure that you
receive all of the results.}
}
\description{
Retrieves the current status of the asynchronous tasks performed by RAM when you perform the ReplacePermissionAssociationsWork operation.

See \url{https://www.paws-r-sdk.com/docs/ram_list_replace_permission_associations_work/} for full documentation.
}
\keyword{internal}
