% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_list_source_associations}
\alias{ram_list_source_associations}
\title{Lists source associations for resource shares}
\usage{
ram_list_source_associations(
  resourceShareArns = NULL,
  sourceId = NULL,
  sourceType = NULL,
  associationStatus = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{resourceShareArns}{The Amazon Resource Names (ARNs) of the resource shares for which you
want to retrieve source associations.}

\item{sourceId}{The identifier of the source for which you want to retrieve
associations. This can be an account ID, Amazon Resource Name (ARN),
organization ID, or organization path.}

\item{sourceType}{The type of source for which you want to retrieve associations.}

\item{associationStatus}{The status of the source associations that you want to retrieve.}

\item{nextToken}{The pagination token that indicates the next set of results to retrieve.}

\item{maxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}
}
\description{
Lists source associations for resource shares. Source associations control which sources can be used with service principals in resource shares. This operation provides visibility into source associations for resource share owners.

See \url{https://www.paws-r-sdk.com/docs/ram_list_source_associations/} for full documentation.
}
\keyword{internal}
