% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsmanager_operations.R
\name{secretsmanager_cancel_rotate_secret}
\alias{secretsmanager_cancel_rotate_secret}
\title{Turns off automatic rotation, and if a rotation is currently in
progress, cancels the rotation}
\usage{
secretsmanager_cancel_rotate_secret(SecretId)
}
\arguments{
\item{SecretId}{[required] The ARN or name of the secret.

For an ARN, we recommend that you specify a complete ARN rather than a
partial ARN. See \href{https://docs.aws.amazon.com/secretsmanager/latest/userguide/troubleshoot.html#ARN_secretnamehyphen}{Finding a secret from a partial ARN}.}
}
\description{
Turns off automatic rotation, and if a rotation is currently in progress, cancels the rotation.

See \url{https://www.paws-r-sdk.com/docs/secretsmanager_cancel_rotate_secret/} for full documentation.
}
\keyword{internal}
