% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsmanager_operations.R
\name{secretsmanager_remove_regions_from_replication}
\alias{secretsmanager_remove_regions_from_replication}
\title{Remove regions from replication}
\usage{
secretsmanager_remove_regions_from_replication(SecretId,
  RemoveReplicaRegions)
}
\arguments{
\item{SecretId}{[required] Remove a secret by \code{SecretId} from replica Regions.}

\item{RemoveReplicaRegions}{[required] Remove replication from specific Regions.}
}
\value{
A list with the following syntax:\preformatted{list(
  ARN = "string",
  ReplicationStatus = list(
    list(
      Region = "string",
      KmsKeyId = "string",
      Status = "InSync"|"Failed"|"InProgress",
      StatusMessage = "string",
      LastAccessedDate = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}
\description{
Remove regions from replication.
}
\section{Request syntax}{
\preformatted{svc$remove_regions_from_replication(
  SecretId = "string",
  RemoveReplicaRegions = list(
    "string"
  )
)
}
}

\keyword{internal}
