% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsmanager_operations.R
\name{secretsmanager_replicate_secret_to_regions}
\alias{secretsmanager_replicate_secret_to_regions}
\title{Converts an existing secret to a multi-Region secret and begins
replication the secret to a list of new regions}
\usage{
secretsmanager_replicate_secret_to_regions(SecretId, AddReplicaRegions,
  ForceOverwriteReplicaSecret)
}
\arguments{
\item{SecretId}{[required] Use the \verb{Secret Id} to replicate a secret to regions.}

\item{AddReplicaRegions}{[required] Add Regions to replicate the secret.}

\item{ForceOverwriteReplicaSecret}{(Optional) If set, Secrets Manager replication overwrites a secret with
the same name in the destination region.}
}
\value{
A list with the following syntax:\preformatted{list(
  ARN = "string",
  ReplicationStatus = list(
    list(
      Region = "string",
      KmsKeyId = "string",
      Status = "InSync"|"Failed"|"InProgress",
      StatusMessage = "string",
      LastAccessedDate = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}
\description{
Converts an existing secret to a multi-Region secret and begins
replication the secret to a list of new regions.
}
\section{Request syntax}{
\preformatted{svc$replicate_secret_to_regions(
  SecretId = "string",
  AddReplicaRegions = list(
    list(
      Region = "string",
      KmsKeyId = "string"
    )
  ),
  ForceOverwriteReplicaSecret = TRUE|FALSE
)
}
}

\keyword{internal}
