% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsmanager_operations.R
\name{secretsmanager_stop_replication_to_replica}
\alias{secretsmanager_stop_replication_to_replica}
\title{Removes the secret from replication and promotes the secret to a
regional secret in the replica Region}
\usage{
secretsmanager_stop_replication_to_replica(SecretId)
}
\arguments{
\item{SecretId}{[required] Response to
\code{\link[=secretsmanager_stop_replication_to_replica]{stop_replication_to_replica}}
of a secret, based on the \code{SecretId}.}
}
\value{
A list with the following syntax:\preformatted{list(
  ARN = "string"
)
}
}
\description{
Removes the secret from replication and promotes the secret to a
regional secret in the replica Region.
}
\section{Request syntax}{
\preformatted{svc$stop_replication_to_replica(
  SecretId = "string"
)
}
}

\keyword{internal}
