% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_accept_invitation}
\alias{securityhub_accept_invitation}
\title{This method is deprecated}
\usage{
securityhub_accept_invitation(MasterId, InvitationId)
}
\arguments{
\item{MasterId}{[required] The account ID of the Security Hub administrator account that sent the
invitation.}

\item{InvitationId}{[required] The identifier of the invitation sent from the Security Hub
administrator account.}
}
\value{
An empty list.
}
\description{
This method is deprecated. Instead, use
\code{\link[=securityhub_accept_administrator_invitation]{accept_administrator_invitation}}.

The Security Hub console continues to use
\code{\link[=securityhub_accept_invitation]{accept_invitation}}. It will eventually
change to use
\code{\link[=securityhub_accept_administrator_invitation]{accept_administrator_invitation}}.
Any IAM policies that specifically control access to this function must
continue to use \code{\link[=securityhub_accept_invitation]{accept_invitation}}.
You should also add
\code{\link[=securityhub_accept_administrator_invitation]{accept_administrator_invitation}}
to your policies to ensure that the correct permissions are in place
after the console begins to use
\code{\link[=securityhub_accept_administrator_invitation]{accept_administrator_invitation}}.

Accepts the invitation to be a member account and be monitored by the
Security Hub administrator account that the invitation was sent from.

This operation is only used by member accounts that are not added
through Organizations.

When the member account accepts the invitation, permission is granted to
the administrator account to view findings generated in the member
account.
}
\section{Request syntax}{
\preformatted{svc$accept_invitation(
  MasterId = "string",
  InvitationId = "string"
)
}
}

\keyword{internal}
