% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_batch_delete_automation_rules}
\alias{securityhub_batch_delete_automation_rules}
\title{Deletes one or more automation rules}
\usage{
securityhub_batch_delete_automation_rules(AutomationRulesArns)
}
\arguments{
\item{AutomationRulesArns}{[required] A list of Amazon Resource Names (ARNs) for the rules that are to be
deleted.}
}
\description{
Deletes one or more automation rules.

See \url{https://www.paws-r-sdk.com/docs/securityhub_batch_delete_automation_rules/} for full documentation.
}
\keyword{internal}
