% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_batch_get_automation_rules}
\alias{securityhub_batch_get_automation_rules}
\title{Retrieves a list of details for automation rules based on rule Amazon
Resource Names (ARNs)}
\usage{
securityhub_batch_get_automation_rules(AutomationRulesArns)
}
\arguments{
\item{AutomationRulesArns}{[required] A list of rule ARNs to get details for.}
}
\description{
Retrieves a list of details for automation rules based on rule Amazon Resource Names (ARNs).

See \url{https://www.paws-r-sdk.com/docs/securityhub_batch_get_automation_rules/} for full documentation.
}
\keyword{internal}
