% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_batch_get_configuration_policy_associations}
\alias{securityhub_batch_get_configuration_policy_associations}
\title{Returns associations between an Security Hub configuration and a batch
of target accounts, organizational units, or the root}
\usage{
securityhub_batch_get_configuration_policy_associations(
  ConfigurationPolicyAssociationIdentifiers
)
}
\arguments{
\item{ConfigurationPolicyAssociationIdentifiers}{[required] Specifies one or more target account IDs, organizational unit (OU) IDs,
or the root ID to retrieve associations for.}
}
\description{
Returns associations between an Security Hub configuration and a batch of target accounts, organizational units, or the root. Only the Security Hub delegated administrator can invoke this operation from the home Region. A configuration can refer to a configuration policy or to a self-managed configuration.

See \url{https://www.paws-r-sdk.com/docs/securityhub_batch_get_configuration_policy_associations/} for full documentation.
}
\keyword{internal}
