% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_batch_get_security_controls}
\alias{securityhub_batch_get_security_controls}
\title{Provides details about a batch of security controls for the current
Amazon Web Services account and Amazon Web Services Region}
\usage{
securityhub_batch_get_security_controls(SecurityControlIds)
}
\arguments{
\item{SecurityControlIds}{[required] A list of security controls (identified with \code{SecurityControlId},
\code{SecurityControlArn}, or a mix of both parameters). The security control
ID or Amazon Resource Name (ARN) is the same across standards.}
}
\description{
Provides details about a batch of security controls for the current Amazon Web Services account and Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/securityhub_batch_get_security_controls/} for full documentation.
}
\keyword{internal}
