% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_batch_get_standards_control_associations}
\alias{securityhub_batch_get_standards_control_associations}
\title{For a batch of security controls and standards, identifies whether each
control is currently enabled or disabled in a standard}
\usage{
securityhub_batch_get_standards_control_associations(
  StandardsControlAssociationIds
)
}
\arguments{
\item{StandardsControlAssociationIds}{[required] An array with one or more objects that includes a security control
(identified with \code{SecurityControlId}, \code{SecurityControlArn}, or a mix of
both parameters) and the Amazon Resource Name (ARN) of a standard. This
field is used to query the enablement status of a control in a specified
standard. The security control ID or ARN is the same across standards.}
}
\description{
For a batch of security controls and standards, identifies whether each control is currently enabled or disabled in a standard.

See \url{https://www.paws-r-sdk.com/docs/securityhub_batch_get_standards_control_associations/} for full documentation.
}
\keyword{internal}
