% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_batch_update_automation_rules}
\alias{securityhub_batch_update_automation_rules}
\title{Updates one or more automation rules based on rule Amazon Resource Names
(ARNs) and input parameters}
\usage{
securityhub_batch_update_automation_rules(UpdateAutomationRulesRequestItems)
}
\arguments{
\item{UpdateAutomationRulesRequestItems}{[required] An array of ARNs for the rules that are to be updated. Optionally, you
can also include \code{RuleStatus} and \code{RuleOrder}.}
}
\description{
Updates one or more automation rules based on rule Amazon Resource Names (ARNs) and input parameters.

See \url{https://www.paws-r-sdk.com/docs/securityhub_batch_update_automation_rules/} for full documentation.
}
\keyword{internal}
