% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_create_action_target}
\alias{securityhub_create_action_target}
\title{Creates a custom action target in Security Hub}
\usage{
securityhub_create_action_target(Name, Description, Id)
}
\arguments{
\item{Name}{[required] The name of the custom action target. Can contain up to 20 characters.}

\item{Description}{[required] The description for the custom action target.}

\item{Id}{[required] The ID for the custom action target. Can contain up to 20 alphanumeric
characters.}
}
\description{
Creates a custom action target in Security Hub.

See \url{https://www.paws-r-sdk.com/docs/securityhub_create_action_target/} for full documentation.
}
\keyword{internal}
