% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_create_aggregator_v2}
\alias{securityhub_create_aggregator_v2}
\title{Enables aggregation across Amazon Web Services Regions}
\usage{
securityhub_create_aggregator_v2(
  RegionLinkingMode,
  LinkedRegions = NULL,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{RegionLinkingMode}{[required] Determines how Regions are linked to an Aggregator V2.}

\item{LinkedRegions}{The list of Regions that are linked to the aggregation Region.}

\item{Tags}{A list of key-value pairs to be applied to the AggregatorV2.}

\item{ClientToken}{A unique identifier used to ensure idempotency.}
}
\description{
Enables aggregation across Amazon Web Services Regions.

See \url{https://www.paws-r-sdk.com/docs/securityhub_create_aggregator_v2/} for full documentation.
}
\keyword{internal}
