% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_create_automation_rule_v2}
\alias{securityhub_create_automation_rule_v2}
\title{Creates a V2 automation rule}
\usage{
securityhub_create_automation_rule_v2(
  RuleName,
  RuleStatus = NULL,
  Description,
  RuleOrder,
  Criteria,
  Actions,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{RuleName}{[required] The name of the V2 automation rule.}

\item{RuleStatus}{The status of the V2 automation rule.}

\item{Description}{[required] A description of the V2 automation rule.}

\item{RuleOrder}{[required] The value for the rule priority.}

\item{Criteria}{[required] The filtering type and configuration of the automation rule.}

\item{Actions}{[required] A list of actions to be performed when the rule criteria is met.}

\item{Tags}{A list of key-value pairs associated with the V2 automation rule.}

\item{ClientToken}{A unique identifier used to ensure idempotency.}
}
\description{
Creates a V2 automation rule.

See \url{https://www.paws-r-sdk.com/docs/securityhub_create_automation_rule_v2/} for full documentation.
}
\keyword{internal}
