% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_create_connector}
\alias{securityhub_create_connector}
\title{Creates a connector to a third-party cloud provider in Security Hub CSPM}
\usage{
securityhub_create_connector(
  Name,
  Description = NULL,
  Provider,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{Name}{[required] The name of the connector. Must be unique within the account.}

\item{Description}{The description of the connector.}

\item{Provider}{[required] The configuration for the cloud provider to connect to. Currently supports Azure.}

\item{Tags}{The tags to add to the connector resource.}

\item{ClientToken}{A unique identifier used to ensure idempotency of the request.}
}
\description{
Creates a connector to a third-party cloud provider in Security Hub CSPM. A connector establishes a connection between Security Hub CSPM and a third-party cloud provider, enabling Security Hub CSPM to ingest security findings and resource data from the connected environment.

See \url{https://www.paws-r-sdk.com/docs/securityhub_create_connector/} for full documentation.
}
\keyword{internal}
