% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_create_connector_v2}
\alias{securityhub_create_connector_v2}
\title{Grants permission to create a connectorV2 based on input parameters}
\usage{
securityhub_create_connector_v2(
  Name,
  Description = NULL,
  Provider,
  KmsKeyArn = NULL,
  Tags = NULL,
  ClientToken = NULL
)
}
\arguments{
\item{Name}{[required] The unique name of the connectorV2.}

\item{Description}{The description of the connectorV2.}

\item{Provider}{[required] The third-party provider’s service configuration.}

\item{KmsKeyArn}{The Amazon Resource Name (ARN) of KMS key used to encrypt secrets for
the connectorV2.}

\item{Tags}{The tags to add to the connectorV2 when you create.}

\item{ClientToken}{A unique identifier used to ensure idempotency.}
}
\description{
Grants permission to create a connectorV2 based on input parameters.

See \url{https://www.paws-r-sdk.com/docs/securityhub_create_connector_v2/} for full documentation.
}
\keyword{internal}
