% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_create_ticket_v2}
\alias{securityhub_create_ticket_v2}
\title{Grants permission to create a ticket in the chosen ITSM based on finding
information for the provided finding metadata UID}
\usage{
securityhub_create_ticket_v2(
  ConnectorId,
  FindingMetadataUid,
  ClientToken = NULL,
  Mode = NULL
)
}
\arguments{
\item{ConnectorId}{[required] The UUID of the connectorV2 to identify connectorV2 resource.}

\item{FindingMetadataUid}{[required] The the unique ID for the finding.}

\item{ClientToken}{The client idempotency token.}

\item{Mode}{The mode for ticket creation. When set to DRYRUN, the ticket is created
using a Security Hub owned template test finding to verify the
integration is working correctly.}
}
\description{
Grants permission to create a ticket in the chosen ITSM based on finding information for the provided finding metadata UID.

See \url{https://www.paws-r-sdk.com/docs/securityhub_create_ticket_v2/} for full documentation.
}
\keyword{internal}
