% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_decline_invitations}
\alias{securityhub_decline_invitations}
\title{We recommend using Organizations instead of Security Hub invitations to
manage your member accounts}
\usage{
securityhub_decline_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{[required] The list of prospective member account IDs for which to decline an
invitation.}
}
\description{
We recommend using Organizations instead of Security Hub invitations to manage your member accounts. For information, see \href{https://docs.aws.amazon.com/securityhub/latest/userguide/securityhub-accounts-orgs.html}{Managing Security Hub administrator and member accounts with Organizations} in the \emph{Security Hub User Guide}.

See \url{https://www.paws-r-sdk.com/docs/securityhub_decline_invitations/} for full documentation.
}
\keyword{internal}
