% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_action_target}
\alias{securityhub_delete_action_target}
\title{Deletes a custom action target from Security Hub}
\usage{
securityhub_delete_action_target(ActionTargetArn)
}
\arguments{
\item{ActionTargetArn}{[required] The ARN of the custom action target to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ActionTargetArn = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Deletes a custom action target from Security Hub.

Deleting a custom action target does not affect any findings or insights
that were already sent to Amazon CloudWatch Events using the custom
action.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_action_target(
  ActionTargetArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
