% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_automation_rule_v2}
\alias{securityhub_delete_automation_rule_v2}
\title{Deletes a V2 automation rule}
\usage{
securityhub_delete_automation_rule_v2(Identifier)
}
\arguments{
\item{Identifier}{[required] The ARN of the V2 automation rule.}
}
\description{
Deletes a V2 automation rule.

See \url{https://www.paws-r-sdk.com/docs/securityhub_delete_automation_rule_v2/} for full documentation.
}
\keyword{internal}
