% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_configuration_policy}
\alias{securityhub_delete_configuration_policy}
\title{Deletes a configuration policy}
\usage{
securityhub_delete_configuration_policy(Identifier)
}
\arguments{
\item{Identifier}{[required] The Amazon Resource Name (ARN) or universally unique identifier (UUID)
of the configuration policy.}
}
\description{
Deletes a configuration policy. Only the Security Hub delegated administrator can invoke this operation from the home Region. For the deletion to succeed, you must first disassociate a configuration policy from target accounts, organizational units, or the root by invoking the \code{\link[=securityhub_start_configuration_policy_disassociation]{start_configuration_policy_disassociation}} operation.

See \url{https://www.paws-r-sdk.com/docs/securityhub_delete_configuration_policy/} for full documentation.
}
\keyword{internal}
