% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_connector}
\alias{securityhub_delete_connector}
\title{Deletes a CSPM connector}
\usage{
securityhub_delete_connector(ConnectorId)
}
\arguments{
\item{ConnectorId}{[required] The unique identifier of the connector to delete.}
}
\description{
Deletes a CSPM connector. When you delete a connector, Security Hub CSPM stops ingesting findings and resource data from the connected cloud provider environment.

See \url{https://www.paws-r-sdk.com/docs/securityhub_delete_connector/} for full documentation.
}
\keyword{internal}
