% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_connector_v2}
\alias{securityhub_delete_connector_v2}
\title{Grants permission to delete a connectorV2}
\usage{
securityhub_delete_connector_v2(ConnectorId)
}
\arguments{
\item{ConnectorId}{[required] The UUID of the connectorV2 to identify connectorV2 resource.}
}
\description{
Grants permission to delete a connectorV2.

See \url{https://www.paws-r-sdk.com/docs/securityhub_delete_connector_v2/} for full documentation.
}
\keyword{internal}
