% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_finding_aggregator}
\alias{securityhub_delete_finding_aggregator}
\title{The aggregation Region is now called the home Region}
\usage{
securityhub_delete_finding_aggregator(FindingAggregatorArn)
}
\arguments{
\item{FindingAggregatorArn}{[required] The ARN of the finding aggregator to delete. To obtain the ARN, use
\code{\link[=securityhub_list_finding_aggregators]{list_finding_aggregators}}.}
}
\description{
The \emph{aggregation Region} is now called the \emph{home Region}.

See \url{https://www.paws-r-sdk.com/docs/securityhub_delete_finding_aggregator/} for full documentation.
}
\keyword{internal}
