% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_invitations}
\alias{securityhub_delete_invitations}
\title{Deletes invitations received by the Amazon Web Services account to
become a member account}
\usage{
securityhub_delete_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{[required] The list of member account IDs that received the invitations you want to
delete.}
}
\description{
Deletes invitations received by the Amazon Web Services account to become a member account.

See \url{https://paws-r.github.io/docs/securityhub/delete_invitations.html} for full documentation.
}
\keyword{internal}
