% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_members}
\alias{securityhub_delete_members}
\title{Deletes the specified member accounts from Security Hub}
\usage{
securityhub_delete_members(AccountIds)
}
\arguments{
\item{AccountIds}{[required] The list of account IDs for the member accounts to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UnprocessedAccounts = list(
    list(
      AccountId = "string",
      ProcessingResult = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes the specified member accounts from Security Hub.

Can be used to delete member accounts that belong to an organization as
well as member accounts that were invited manually.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_members(
  AccountIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
