% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_describe_action_targets}
\alias{securityhub_describe_action_targets}
\title{Returns a list of the custom action targets in Security Hub in your
account}
\usage{
securityhub_describe_action_targets(ActionTargetArns, NextToken,
  MaxResults)
}
\arguments{
\item{ActionTargetArns}{A list of custom action target ARNs for the custom action targets to
retrieve.}

\item{NextToken}{The token that is required for pagination. On your first call to the
\code{\link[=securityhub_describe_action_targets]{describe_action_targets}}
operation, set the value of this parameter to \code{NULL}.

For subsequent calls to the operation, to continue listing data, set the
value of this parameter to the value returned from the previous
response.}

\item{MaxResults}{The maximum number of results to return.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ActionTargets = list(
    list(
      ActionTargetArn = "string",
      Name = "string",
      Description = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of the custom action targets in Security Hub in your
account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_action_targets(
  ActionTargetArns = list(
    "string"
  ),
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
