% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_describe_hub}
\alias{securityhub_describe_hub}
\title{Returns details about the Hub resource in your account, including the
HubArn and the time when you enabled Security Hub}
\usage{
securityhub_describe_hub(HubArn)
}
\arguments{
\item{HubArn}{The ARN of the Hub resource to retrieve.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  HubArn = "string",
  SubscribedAt = "string",
  AutoEnableControls = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
Returns details about the Hub resource in your account, including the
\code{HubArn} and the time when you enabled Security Hub.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_hub(
  HubArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
