% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_describe_products_v2}
\alias{securityhub_describe_products_v2}
\title{Gets information about the product integration}
\usage{
securityhub_describe_products_v2(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The token required for pagination. On your first call, set the value of
this parameter to \code{NULL}. For subsequent calls, to continue listing
data, set the value of this parameter to the value returned in the
previous response.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Gets information about the product integration.

See \url{https://www.paws-r-sdk.com/docs/securityhub_describe_products_v2/} for full documentation.
}
\keyword{internal}
