% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_describe_security_hub_v2}
\alias{securityhub_describe_security_hub_v2}
\title{Returns details about the service resource in your account}
\usage{
securityhub_describe_security_hub_v2()
}
\description{
Returns details about the service resource in your account.

See \url{https://www.paws-r-sdk.com/docs/securityhub_describe_security_hub_v2/} for full documentation.
}
\keyword{internal}
