% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_describe_standards_controls}
\alias{securityhub_describe_standards_controls}
\title{Returns a list of security standards controls}
\usage{
securityhub_describe_standards_controls(StandardsSubscriptionArn,
  NextToken, MaxResults)
}
\arguments{
\item{StandardsSubscriptionArn}{[required] The ARN of a resource that represents your subscription to a supported
standard. To get the subscription ARNs of the standards you have
enabled, use the
\code{\link[=securityhub_get_enabled_standards]{get_enabled_standards}} operation.}

\item{NextToken}{The token that is required for pagination. On your first call to the
\code{\link[=securityhub_describe_standards_controls]{describe_standards_controls}}
operation, set the value of this parameter to \code{NULL}.

For subsequent calls to the operation, to continue listing data, set the
value of this parameter to the value returned from the previous
response.}

\item{MaxResults}{The maximum number of security standard controls to return.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Controls = list(
    list(
      StandardsControlArn = "string",
      ControlStatus = "ENABLED"|"DISABLED",
      DisabledReason = "string",
      ControlStatusUpdatedAt = as.POSIXct(
        "2015-01-01"
      ),
      ControlId = "string",
      Title = "string",
      Description = "string",
      RemediationUrl = "string",
      SeverityRating = "LOW"|"MEDIUM"|"HIGH"|"CRITICAL",
      RelatedRequirements = list(
        "string"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Returns a list of security standards controls.

For each control, the results include information about whether it is
currently enabled, the severity, and a link to remediation information.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_standards_controls(
  StandardsSubscriptionArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
