% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disable_organization_admin_account}
\alias{securityhub_disable_organization_admin_account}
\title{Disables a Security Hub administrator account}
\usage{
securityhub_disable_organization_admin_account(AdminAccountId)
}
\arguments{
\item{AdminAccountId}{[required] The AWS account identifier of the Security Hub administrator account.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Disables a Security Hub administrator account. Can only be called by the
organization management account.
}
\section{Request syntax}{
\preformatted{svc$disable_organization_admin_account(
  AdminAccountId = "string"
)
}
}

\keyword{internal}
