% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disable_security_hub_feature_v2}
\alias{securityhub_disable_security_hub_feature_v2}
\title{Disables an opt-in feature for the calling account in the current Amazon
Web Services Region}
\usage{
securityhub_disable_security_hub_feature_v2(FeatureName)
}
\arguments{
\item{FeatureName}{[required] The name of the feature to disable.}
}
\description{
Disables an opt-in feature for the calling account in the current Amazon Web Services Region. The operation is idempotent. If the feature is already disabled, no changes are made. You cannot disable a feature that is managed by an organization policy.

See \url{https://www.paws-r-sdk.com/docs/securityhub_disable_security_hub_feature_v2/} for full documentation.
}
\keyword{internal}
