% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disassociate_from_administrator_account}
\alias{securityhub_disassociate_from_administrator_account}
\title{Disassociates the current Security Hub member account from the
associated administrator account}
\usage{
securityhub_disassociate_from_administrator_account()
}
\description{
Disassociates the current Security Hub member account from the associated administrator account.

See \url{https://paws-r.github.io/docs/securityhub/disassociate_from_administrator_account.html} for full documentation.
}
\keyword{internal}
