% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disassociate_members}
\alias{securityhub_disassociate_members}
\title{Disassociates the specified member accounts from the associated
administrator account}
\usage{
securityhub_disassociate_members(AccountIds)
}
\arguments{
\item{AccountIds}{[required] The account IDs of the member accounts to disassociate from the
administrator account.}
}
\description{
Disassociates the specified member accounts from the associated administrator account.

See \url{https://www.paws-r-sdk.com/docs/securityhub_disassociate_members/} for full documentation.
}
\keyword{internal}
