% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_enable_organization_admin_account}
\alias{securityhub_enable_organization_admin_account}
\title{Designates the Security Hub administrator account for an organization}
\usage{
securityhub_enable_organization_admin_account(AdminAccountId)
}
\arguments{
\item{AdminAccountId}{[required] The AWS account identifier of the account to designate as the Security
Hub administrator account.}
}
\description{
Designates the Security Hub administrator account for an organization.
Can only be called by the organization management account.
}
\section{Request syntax}{
\preformatted{svc$enable_organization_admin_account(
  AdminAccountId = "string"
)
}
}

\keyword{internal}
