% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_enable_security_hub}
\alias{securityhub_enable_security_hub}
\title{Enables Security Hub for your account in the current Region or the
Region you specify in the request}
\usage{
securityhub_enable_security_hub(Tags = NULL, EnableDefaultStandards = NULL)
}
\arguments{
\item{Tags}{The tags to add to the hub resource when you enable Security Hub.}

\item{EnableDefaultStandards}{Whether to enable the security standards that Security Hub has
designated as automatically enabled. If you do not provide a value for
\code{EnableDefaultStandards}, it is set to \code{true}. To not enable the
automatically enabled standards, set \code{EnableDefaultStandards} to
\code{false}.}
}
\description{
Enables Security Hub for your account in the current Region or the Region you specify in the request.

See \url{https://paws-r.github.io/docs/securityhub/enable_security_hub.html} for full documentation.
}
\keyword{internal}
