% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_enable_security_hub_feature_v2}
\alias{securityhub_enable_security_hub_feature_v2}
\title{Enables an opt-in feature for the calling account in the current Amazon
Web Services Region}
\usage{
securityhub_enable_security_hub_feature_v2(FeatureName)
}
\arguments{
\item{FeatureName}{[required] The name of the feature to enable.}
}
\description{
Enables an opt-in feature for the calling account in the current Amazon Web Services Region. The service must be enabled before you can enable a feature. The operation is idempotent. If the feature is already enabled, no changes are made. You cannot enable a feature that is managed by an organization policy.

See \url{https://www.paws-r-sdk.com/docs/securityhub_enable_security_hub_feature_v2/} for full documentation.
}
\keyword{internal}
