% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_enable_security_hub_v2}
\alias{securityhub_enable_security_hub_v2}
\title{Enables the service in account for the current Amazon Web Services
Region or specified Amazon Web Services Region}
\usage{
securityhub_enable_security_hub_v2(Tags = NULL)
}
\arguments{
\item{Tags}{The tags to add to the hub V2 resource when you enable Security Hub.}
}
\description{
Enables the service in account for the current Amazon Web Services Region or specified Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/securityhub_enable_security_hub_v2/} for full documentation.
}
\keyword{internal}
