% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_generate_recommended_policy_v2}
\alias{securityhub_generate_recommended_policy_v2}
\title{Begins the recommended policy generation to remediate a Security Hub
finding}
\usage{
securityhub_generate_recommended_policy_v2(MetadataUid)
}
\arguments{
\item{MetadataUid}{[required] The unique identifier (ID) of Security Hub OCSF findings found under the
\code{metadata.uid} field of the finding.}
}
\description{
Begins the recommended policy generation to remediate a Security Hub finding. \code{\link[=securityhub_generate_recommended_policy_v2]{generate_recommended_policy_v2}} only supports findings for unused permissions.

See \url{https://www.paws-r-sdk.com/docs/securityhub_generate_recommended_policy_v2/} for full documentation.
}
\keyword{internal}
