% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_administrator_account}
\alias{securityhub_get_administrator_account}
\title{Provides the details for the Security Hub administrator account for the
current member account}
\usage{
securityhub_get_administrator_account()
}
\value{
A list with the following syntax:\preformatted{list(
  Administrator = list(
    AccountId = "string",
    InvitationId = "string",
    InvitedAt = as.POSIXct(
      "2015-01-01"
    ),
    MemberStatus = "string"
  )
)
}
}
\description{
Provides the details for the Security Hub administrator account for the
current member account.

Can be used by both member accounts that are managed using Organizations
and accounts that were invited manually.
}
\section{Request syntax}{
\preformatted{svc$get_administrator_account()
}
}

\keyword{internal}
