% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_aggregator_v2}
\alias{securityhub_get_aggregator_v2}
\title{Returns the configuration of the specified Aggregator V2}
\usage{
securityhub_get_aggregator_v2(AggregatorV2Arn)
}
\arguments{
\item{AggregatorV2Arn}{[required] The ARN of the Aggregator V2.}
}
\description{
Returns the configuration of the specified Aggregator V2.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_aggregator_v2/} for full documentation.
}
\keyword{internal}
