% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_automation_rule_v2}
\alias{securityhub_get_automation_rule_v2}
\title{Returns an automation rule for the V2 service}
\usage{
securityhub_get_automation_rule_v2(Identifier)
}
\arguments{
\item{Identifier}{[required] The ARN of the V2 automation rule.}
}
\description{
Returns an automation rule for the V2 service.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_automation_rule_v2/} for full documentation.
}
\keyword{internal}
