% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_configuration_policy}
\alias{securityhub_get_configuration_policy}
\title{Provides information about a configuration policy}
\usage{
securityhub_get_configuration_policy(Identifier)
}
\arguments{
\item{Identifier}{[required] The Amazon Resource Name (ARN) or universally unique identifier (UUID)
of the configuration policy.}
}
\description{
Provides information about a configuration policy. Only the Security Hub delegated administrator can invoke this operation from the home Region.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_configuration_policy/} for full documentation.
}
\keyword{internal}
