% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_configuration_policy_association}
\alias{securityhub_get_configuration_policy_association}
\title{Returns the association between a configuration and a target account,
organizational unit, or the root}
\usage{
securityhub_get_configuration_policy_association(Target)
}
\arguments{
\item{Target}{[required] The target account ID, organizational unit ID, or the root ID to
retrieve the association for.}
}
\description{
Returns the association between a configuration and a target account, organizational unit, or the root. The configuration can be a configuration policy or self-managed behavior. Only the Security Hub delegated administrator can invoke this operation from the home Region.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_configuration_policy_association/} for full documentation.
}
\keyword{internal}
