% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_connector}
\alias{securityhub_get_connector}
\title{Retrieves details for a CSPM connector based on the connector ID}
\usage{
securityhub_get_connector(ConnectorId)
}
\arguments{
\item{ConnectorId}{[required] The unique identifier of the connector to retrieve.}
}
\description{
Retrieves details for a CSPM connector based on the connector ID.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_connector/} for full documentation.
}
\keyword{internal}
