% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_enabled_standards}
\alias{securityhub_get_enabled_standards}
\title{Returns a list of the standards that are currently enabled}
\usage{
securityhub_get_enabled_standards(
  StandardsSubscriptionArns = NULL,
  NextToken = NULL,
  MaxResults = NULL,
  Providers = NULL
)
}
\arguments{
\item{StandardsSubscriptionArns}{The list of the standards subscription ARNs for the standards to retrieve.}

\item{NextToken}{The token that is required for pagination. On your first call to the \code{\link[=securityhub_get_enabled_standards]{get_enabled_standards}} operation, set the value of this parameter to \code{NULL}.

For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.}

\item{MaxResults}{The maximum number of results to return in the response.}

\item{Providers}{A list of cloud providers to filter the enabled standards by. For example, specify \code{Azure} to return only enabled standards that evaluate Azure resources.}
}
\description{
Returns a list of the standards that are currently enabled.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_enabled_standards/} for full documentation.
}
\keyword{internal}
