% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_finding_aggregator}
\alias{securityhub_get_finding_aggregator}
\title{Returns the current finding aggregation configuration}
\usage{
securityhub_get_finding_aggregator(FindingAggregatorArn)
}
\arguments{
\item{FindingAggregatorArn}{[required] The ARN of the finding aggregator to return details for. To obtain the
ARN, use
\code{\link[=securityhub_list_finding_aggregators]{list_finding_aggregators}}.}
}
\description{
Returns the current finding aggregation configuration.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_finding_aggregator/} for full documentation.
}
\keyword{internal}
