% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_finding_history}
\alias{securityhub_get_finding_history}
\title{Returns history for a Security Hub finding in the last 90 days}
\usage{
securityhub_get_finding_history(
  FindingIdentifier,
  StartTime = NULL,
  EndTime = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{FindingIdentifier}{[required]}

\item{StartTime}{An ISO 8601-formatted timestamp that indicates the start time of the
requested finding history. A correctly formatted example is
\verb{2020-05-21T20:16:34.724Z}. The value cannot contain spaces, and date
and time should be separated by \code{T}. For more information, see \href{https://www.rfc-editor.org/rfc/rfc3339#section-5.6}{RFC 3339 section 5.6, Internet Date/Time Format}.

If you provide values for both \code{StartTime} and \code{EndTime}, Security Hub
returns finding history for the specified time period. If you provide a
value for \code{StartTime} but not for \code{EndTime}, Security Hub returns
finding history from the \code{StartTime} to the time at which the API is
called. If you provide a value for \code{EndTime} but not for \code{StartTime},
Security Hub returns finding history from the
\href{https://docs.aws.amazon.com/securityhub/1.0/APIReference/API_AwsSecurityFindingFilters.html#securityhub-Type-AwsSecurityFindingFilters-CreatedAt}{CreatedAt}
timestamp of the finding to the \code{EndTime}. If you provide neither
\code{StartTime} nor \code{EndTime}, Security Hub returns finding history from the
CreatedAt timestamp of the finding to the time at which the API is
called. In all of these scenarios, the response is limited to 100
results, and the maximum time period is limited to 90 days.}

\item{EndTime}{An ISO 8601-formatted timestamp that indicates the end time of the
requested finding history. A correctly formatted example is
\verb{2020-05-21T20:16:34.724Z}. The value cannot contain spaces, and date
and time should be separated by \code{T}. For more information, see \href{https://www.rfc-editor.org/rfc/rfc3339#section-5.6}{RFC 3339 section 5.6, Internet Date/Time Format}.

If you provide values for both \code{StartTime} and \code{EndTime}, Security Hub
returns finding history for the specified time period. If you provide a
value for \code{StartTime} but not for \code{EndTime}, Security Hub returns
finding history from the \code{StartTime} to the time at which the API is
called. If you provide a value for \code{EndTime} but not for \code{StartTime},
Security Hub returns finding history from the
\href{https://docs.aws.amazon.com/securityhub/1.0/APIReference/API_AwsSecurityFindingFilters.html#securityhub-Type-AwsSecurityFindingFilters-CreatedAt}{CreatedAt}
timestamp of the finding to the \code{EndTime}. If you provide neither
\code{StartTime} nor \code{EndTime}, Security Hub returns finding history from the
CreatedAt timestamp of the finding to the time at which the API is
called. In all of these scenarios, the response is limited to 100
results, and the maximum time period is limited to 90 days.}

\item{NextToken}{A token for pagination purposes. Provide \code{NULL} as the initial value. In
subsequent requests, provide the token included in the response to get
up to an additional 100 results of finding history. If you don’t provide
\code{NextToken}, Security Hub returns up to 100 results of finding history
for each request.}

\item{MaxResults}{The maximum number of results to be returned. If you don’t provide it,
Security Hub returns up to 100 results of finding history.}
}
\description{
Returns history for a Security Hub finding in the last 90 days. The history includes changes made to any fields in the Amazon Web Services Security Finding Format (ASFF).

See \url{https://paws-r.github.io/docs/securityhub/get_finding_history.html} for full documentation.
}
\keyword{internal}
