% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_finding_statistics_v2}
\alias{securityhub_get_finding_statistics_v2}
\title{Returns aggregated statistical data about findings}
\usage{
securityhub_get_finding_statistics_v2(
  GroupByRules,
  Scopes = NULL,
  SortOrder = NULL,
  MaxStatisticResults = NULL
)
}
\arguments{
\item{GroupByRules}{[required] Specifies how security findings should be aggregated and organized in
the statistical analysis. It can accept up to 5 \code{groupBy} fields in a
single call.}

\item{Scopes}{Limits the results to findings from specific organizational units or
from the delegated administrator's organization. Only the delegated
administrator account can use this parameter. Other accounts receive an
\code{AccessDeniedException}.

This parameter is optional. If you omit it, the delegated administrator
sees statistics from all accounts across the entire organization. Other
accounts see only statistics for their own findings.

You can specify up to 10 entries in \code{Scopes.AwsOrganizations}. If
multiple entries are specified, the entries are combined using OR logic.}

\item{SortOrder}{Orders the aggregation count in descending or ascending order.
Descending order is the default.}

\item{MaxStatisticResults}{The maximum number of results to be returned.}
}
\description{
Returns aggregated statistical data about findings.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_finding_statistics_v2/} for full documentation.
}
\keyword{internal}
