% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_insight_results}
\alias{securityhub_get_insight_results}
\title{Lists the results of the Security Hub insight specified by the insight
ARN}
\usage{
securityhub_get_insight_results(InsightArn)
}
\arguments{
\item{InsightArn}{[required] The ARN of the insight for which to return results.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  InsightResults = list(
    InsightArn = "string",
    GroupByAttribute = "string",
    ResultValues = list(
      list(
        GroupByAttributeValue = "string",
        Count = 123
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists the results of the Security Hub insight specified by the insight
ARN.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_insight_results(
  InsightArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
