% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_recommended_policy_v2}
\alias{securityhub_get_recommended_policy_v2}
\title{Retrieves the recommended policy to remediate a Security Hub finding}
\usage{
securityhub_get_recommended_policy_v2(
  MetadataUid,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{MetadataUid}{[required] The unique identifier (ID) of Security Hub OCSF findings found under the
\code{metadata.uid} field of the finding.}

\item{NextToken}{The token used to paginate the \code{RecommendationSteps} list returned. On
your first call to
\code{\link[=securityhub_get_recommended_policy_v2]{get_recommended_policy_v2}},
omit this parameter or set it to \code{NULL}. For subsequent calls, use the
\code{NextToken} value returned in the previous response to retrieve the next
page of results.}

\item{MaxResults}{The maximum number of recommendation steps to return.}
}
\description{
Retrieves the recommended policy to remediate a Security Hub finding. \code{\link[=securityhub_get_recommended_policy_v2]{get_recommended_policy_v2}} only supports findings for unused permissions.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_recommended_policy_v2/} for full documentation.
}
\keyword{internal}
