% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_resources_trends_v2}
\alias{securityhub_get_resources_trends_v2}
\title{Returns resource trend data based on the specified criteria}
\usage{
securityhub_get_resources_trends_v2(
  Filters = NULL,
  StartTime,
  EndTime,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filters}{The filters to apply to the resources trend data.}

\item{StartTime}{[required] The starting timestamp for the time period to analyze resources trends,
in ISO 8601 format.}

\item{EndTime}{[required] The ending timestamp for the time period to analyze resources trends, in
ISO 8601 format.}

\item{NextToken}{The token to use for paginating results. This value is returned in the
response if more results are available.}

\item{MaxResults}{The maximum number of trend data points to return in a single response.}
}
\description{
Returns resource trend data based on the specified criteria. This operation helps you analyze patterns and changes in resource compliance over time.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_resources_trends_v2/} for full documentation.
}
\keyword{internal}
