% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_security_control_definition}
\alias{securityhub_get_security_control_definition}
\title{Retrieves the definition of a security control}
\usage{
securityhub_get_security_control_definition(SecurityControlId)
}
\arguments{
\item{SecurityControlId}{[required] The ID of the security control to retrieve the definition for. This
field doesn’t accept an Amazon Resource Name (ARN).}
}
\description{
Retrieves the definition of a security control. The definition includes the control title, description, Region availability, parameter definitions, and other details.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_security_control_definition/} for full documentation.
}
\keyword{internal}
