% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_automation_rules}
\alias{securityhub_list_automation_rules}
\title{A list of automation rules and their metadata for the calling account}
\usage{
securityhub_list_automation_rules(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A token to specify where to start paginating the response. This is the
\code{NextToken} from a previously truncated response. On your first call to
the \code{\link[=securityhub_list_automation_rules]{list_automation_rules}} API,
set the value of this parameter to \code{NULL}.}

\item{MaxResults}{The maximum number of rules to return in the response. This currently
ranges from 1 to 100.}
}
\description{
A list of automation rules and their metadata for the calling account.

See \url{https://www.paws-r-sdk.com/docs/securityhub_list_automation_rules/} for full documentation.
}
\keyword{internal}
