% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_configuration_policy_associations}
\alias{securityhub_list_configuration_policy_associations}
\title{Provides information about the associations for your configuration
policies and self-managed behavior}
\usage{
securityhub_list_configuration_policy_associations(
  NextToken = NULL,
  MaxResults = NULL,
  Filters = NULL
)
}
\arguments{
\item{NextToken}{The \code{NextToken} value that's returned from a previous paginated
\code{\link[=securityhub_list_configuration_policy_associations]{list_configuration_policy_associations}}
request where \code{MaxResults} was used but the results exceeded the value
of that parameter. Pagination continues from the end of the previous
response that returned the \code{NextToken} value. This value is \code{null} when
there are no more results to return.}

\item{MaxResults}{The maximum number of results that's returned by
\code{\link[=securityhub_list_configuration_policies]{list_configuration_policies}}
in each page of the response. When this parameter is used,
\code{\link[=securityhub_list_configuration_policy_associations]{list_configuration_policy_associations}}
returns the specified number of results in a single page and a
\code{NextToken} response element. You can see the remaining results of the
initial request by sending another
\code{\link[=securityhub_list_configuration_policy_associations]{list_configuration_policy_associations}}
request with the returned \code{NextToken} value. A valid range for
\code{MaxResults} is between 1 and 100.}

\item{Filters}{Options for filtering the
\code{\link[=securityhub_list_configuration_policy_associations]{list_configuration_policy_associations}}
response. You can filter by the Amazon Resource Name (ARN) or
universally unique identifier (UUID) of a configuration,
\code{AssociationType}, or \code{AssociationStatus}.}
}
\description{
Provides information about the associations for your configuration policies and self-managed behavior. Only the Security Hub delegated administrator can invoke this operation from the home Region.

See \url{https://www.paws-r-sdk.com/docs/securityhub_list_configuration_policy_associations/} for full documentation.
}
\keyword{internal}
