% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_connectors}
\alias{securityhub_list_connectors}
\title{Lists the CSPM connectors and their metadata for the calling account}
\usage{
securityhub_list_connectors(
  NextToken = NULL,
  MaxResults = NULL,
  ProviderName = NULL,
  ConnectorStatus = NULL,
  EnablementStatus = NULL
)
}
\arguments{
\item{NextToken}{The pagination token to request the next page of results.}

\item{MaxResults}{The maximum number of results to return.}

\item{ProviderName}{The name of the cloud provider to filter connectors by.}

\item{ConnectorStatus}{The connectivity status to filter connectors by.}

\item{EnablementStatus}{The enablement status to filter connectors by.}
}
\description{
Lists the CSPM connectors and their metadata for the calling account.

See \url{https://www.paws-r-sdk.com/docs/securityhub_list_connectors/} for full documentation.
}
\keyword{internal}
