% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_connectors_v2}
\alias{securityhub_list_connectors_v2}
\title{Grants permission to retrieve a list of connectorsV2 and their metadata
for the calling account}
\usage{
securityhub_list_connectors_v2(
  NextToken = NULL,
  MaxResults = NULL,
  ProviderName = NULL,
  ConnectorStatus = NULL,
  EnablementStatus = NULL
)
}
\arguments{
\item{NextToken}{The pagination token per the Amazon Web Services Pagination standard}

\item{MaxResults}{The maximum number of results to be returned.}

\item{ProviderName}{The name of the third-party provider.}

\item{ConnectorStatus}{The status for the connectorV2.}

\item{EnablementStatus}{The enablement status to filter connectors by.}
}
\description{
Grants permission to retrieve a list of connectorsV2 and their metadata for the calling account.

See \url{https://www.paws-r-sdk.com/docs/securityhub_list_connectors_v2/} for full documentation.
}
\keyword{internal}
