% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_exposures_by_remediation_v2}
\alias{securityhub_list_exposures_by_remediation_v2}
\title{Retrieves the exposure findings tied to a specific remediation target}
\usage{
securityhub_list_exposures_by_remediation_v2(
  TargetUid,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{TargetUid}{[required] The unique identifier (ID) of an existing remediation target to list exposure findings for.}

\item{MaxResults}{The maximum number of results to return. Valid range is 1-100. If you don't specify a value, the operation returns up to 25 results.}

\item{NextToken}{The token used to paginate the exposures list returned. On your first call to \code{\link[=securityhub_list_exposures_by_remediation_v2]{list_exposures_by_remediation_v2}}, omit this parameter or set it to \code{NULL}. For subsequent calls, use the \code{NextToken} value returned in the previous response to retrieve the next page of results.}
}
\description{
Retrieves the exposure findings tied to a specific remediation target. Results are sorted by previous severity, highest first, and are paginated.

See \url{https://www.paws-r-sdk.com/docs/securityhub_list_exposures_by_remediation_v2/} for full documentation.
}
\keyword{internal}
