% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_finding_aggregators}
\alias{securityhub_list_finding_aggregators}
\title{If finding aggregation is enabled, then ListFindingAggregators returns
the ARN of the finding aggregator}
\usage{
securityhub_list_finding_aggregators(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The token returned with the previous set of results. Identifies the next
set of results to return.}

\item{MaxResults}{The maximum number of results to return. This operation currently only
returns a single result.}
}
\description{
If finding aggregation is enabled, then \code{\link[=securityhub_list_finding_aggregators]{list_finding_aggregators}} returns the ARN of the finding aggregator. You can run this operation from any Region.

See \url{https://www.paws-r-sdk.com/docs/securityhub_list_finding_aggregators/} for full documentation.
}
\keyword{internal}
