% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_free_trial_statuses_v2}
\alias{securityhub_list_free_trial_statuses_v2}
\title{Lists the free trial status of Security Hub features}
\usage{
securityhub_list_free_trial_statuses_v2(
  AccountIds = NULL,
  Statuses = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{AccountIds}{The Amazon Web Services account identifiers to list free trial status for. You can specify accounts other than your own only if you are a delegated Security Hub administrator.}

\item{Statuses}{The free trial statuses to filter the results by. Valid values:
\itemize{
\item \code{ACTIVE} returns only features with an ongoing free trial period.
\item \code{INACTIVE} returns only features whose free trial period has ended, or that never started.
}}

\item{MaxResults}{The maximum number of results to return. If you don't specify a value, Security Hub returns up to 100 results.}

\item{NextToken}{The pagination token to request the next page of results.}
}
\description{
Lists the free trial status of Security Hub features. A delegated Security Hub administrator can list the status for accounts in its organization. Any other account can list the status only for itself. Free trial status remains available after a feature is disabled.

See \url{https://www.paws-r-sdk.com/docs/securityhub_list_free_trial_statuses_v2/} for full documentation.
}
\keyword{internal}
