% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_organization_admin_accounts}
\alias{securityhub_list_organization_admin_accounts}
\title{Lists the Security Hub administrator accounts}
\usage{
securityhub_list_organization_admin_accounts(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of items to return in the response.}

\item{NextToken}{The token that is required for pagination. On your first call to the
\code{\link[=securityhub_list_organization_admin_accounts]{list_organization_admin_accounts}}
operation, set the value of this parameter to \code{NULL}. For subsequent
calls to the operation, to continue listing data, set the value of this
parameter to the value returned from the previous response.}
}
\value{
A list with the following syntax:\preformatted{list(
  AdminAccounts = list(
    list(
      AccountId = "string",
      Status = "ENABLED"|"DISABLE_IN_PROGRESS"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists the Security Hub administrator accounts. Can only be called by the
organization management account.
}
\section{Request syntax}{
\preformatted{svc$list_organization_admin_accounts(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
