% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_tags_for_resource}
\alias{securityhub_list_tags_for_resource}
\title{Returns a list of tags associated with a resource}
\usage{
securityhub_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource to retrieve tags for.}
}
\value{
A list with the following syntax:\preformatted{list(
  Tags = list(
    "string"
  )
)
}
}
\description{
Returns a list of tags associated with a resource.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}
}

\keyword{internal}
