% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_register_connector_v2}
\alias{securityhub_register_connector_v2}
\title{Grants permission to complete the authorization based on input
parameters}
\usage{
securityhub_register_connector_v2(AuthCode, AuthState)
}
\arguments{
\item{AuthCode}{[required] The authCode retrieved from authUrl to complete the OAuth 2.0
authorization code flow.}

\item{AuthState}{[required] The authState retrieved from authUrl to complete the OAuth 2.0
authorization code flow.}
}
\description{
Grants permission to complete the authorization based on input parameters.

See \url{https://www.paws-r-sdk.com/docs/securityhub_register_connector_v2/} for full documentation.
}
\keyword{internal}
