% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_tag_resource}
\alias{securityhub_tag_resource}
\title{Adds one or more tags to a resource}
\usage{
securityhub_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource to apply the tags to.}

\item{Tags}{[required] The tags to add to the resource. You can add up to 50 tags at a time.
The tag keys can be no longer than 128 characters. The tag values can be
no longer than 256 characters.}
}
\description{
Adds one or more tags to a resource.

See \url{https://www.paws-r-sdk.com/docs/securityhub_tag_resource/} for full documentation.
}
\keyword{internal}
