% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_untag_resource}
\alias{securityhub_untag_resource}
\title{Removes one or more tags from a resource}
\usage{
securityhub_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource to remove the tags from.}

\item{TagKeys}{[required] The tag keys associated with the tags to remove from the resource.}
}
\value{
An empty list.
}
\description{
Removes one or more tags from a resource.
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
