% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_action_target}
\alias{securityhub_update_action_target}
\title{Updates the name and description of a custom action target in Security
Hub}
\usage{
securityhub_update_action_target(ActionTargetArn, Name, Description)
}
\arguments{
\item{ActionTargetArn}{[required] The ARN of the custom action target to update.}

\item{Name}{The updated name of the custom action target.}

\item{Description}{The updated description for the custom action target.}
}
\description{
Updates the name and description of a custom action target in Security
Hub.
}
\section{Request syntax}{
\preformatted{svc$update_action_target(
  ActionTargetArn = "string",
  Name = "string",
  Description = "string"
)
}
}

\keyword{internal}
