% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_aggregator_v2}
\alias{securityhub_update_aggregator_v2}
\title{Udpates the configuration for the Aggregator V2}
\usage{
securityhub_update_aggregator_v2(
  AggregatorV2Arn,
  RegionLinkingMode,
  LinkedRegions = NULL
)
}
\arguments{
\item{AggregatorV2Arn}{[required] The ARN of the Aggregator V2.}

\item{RegionLinkingMode}{[required] Determines how Amazon Web Services Regions should be linked to the
Aggregator V2.}

\item{LinkedRegions}{A list of Amazon Web Services Regions linked to the aggegation Region.}
}
\description{
Udpates the configuration for the Aggregator V2.

See \url{https://www.paws-r-sdk.com/docs/securityhub_update_aggregator_v2/} for full documentation.
}
\keyword{internal}
