% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_automation_rule_v2}
\alias{securityhub_update_automation_rule_v2}
\title{Updates a V2 automation rule}
\usage{
securityhub_update_automation_rule_v2(
  Identifier,
  RuleStatus = NULL,
  RuleOrder = NULL,
  Description = NULL,
  RuleName = NULL,
  Criteria = NULL,
  Actions = NULL
)
}
\arguments{
\item{Identifier}{[required] The ARN of the automation rule.}

\item{RuleStatus}{The status of the automation rule.}

\item{RuleOrder}{Represents a value for the rule priority.}

\item{Description}{A description of the automation rule.}

\item{RuleName}{The name of the automation rule.}

\item{Criteria}{The filtering type and configuration of the automation rule.}

\item{Actions}{A list of actions to be performed when the rule criteria is met.}
}
\description{
Updates a V2 automation rule.

See \url{https://www.paws-r-sdk.com/docs/securityhub_update_automation_rule_v2/} for full documentation.
}
\keyword{internal}
