% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_connector}
\alias{securityhub_update_connector}
\title{Updates a CSPM connector's configuration, such as the scope or regions
for the connected cloud provider}
\usage{
securityhub_update_connector(ConnectorId, Description = NULL, Provider = NULL)
}
\arguments{
\item{ConnectorId}{[required] The unique identifier of the connector to update.}

\item{Description}{The updated description of the connector.}

\item{Provider}{The updated cloud provider configuration for the connector.}
}
\description{
Updates a CSPM connector's configuration, such as the scope or regions for the connected cloud provider.

See \url{https://www.paws-r-sdk.com/docs/securityhub_update_connector/} for full documentation.
}
\keyword{internal}
