% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_connector_v2}
\alias{securityhub_update_connector_v2}
\title{Grants permission to update a connectorV2 based on its id and input
parameters}
\usage{
securityhub_update_connector_v2(
  ConnectorId,
  Description = NULL,
  Provider = NULL
)
}
\arguments{
\item{ConnectorId}{[required] The UUID of the connectorV2 to identify connectorV2 resource.}

\item{Description}{The description of the connectorV2.}

\item{Provider}{The third-party provider’s service configuration.}
}
\description{
Grants permission to update a connectorV2 based on its id and input parameters.

See \url{https://www.paws-r-sdk.com/docs/securityhub_update_connector_v2/} for full documentation.
}
\keyword{internal}
