% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_insight}
\alias{securityhub_update_insight}
\title{Updates the Security Hub insight identified by the specified insight ARN}
\usage{
securityhub_update_insight(
  InsightArn,
  Name = NULL,
  Filters = NULL,
  GroupByAttribute = NULL
)
}
\arguments{
\item{InsightArn}{[required] The ARN of the insight that you want to update.}

\item{Name}{The updated name for the insight.}

\item{Filters}{The updated filters that define this insight.}

\item{GroupByAttribute}{The updated \code{GroupBy} attribute that defines this insight.}
}
\description{
Updates the Security Hub insight identified by the specified insight ARN.

See \url{https://www.paws-r-sdk.com/docs/securityhub_update_insight/} for full documentation.
}
\keyword{internal}
