% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_organization_configuration}
\alias{securityhub_update_organization_configuration}
\title{Used to update the configuration related to Organizations}
\usage{
securityhub_update_organization_configuration(AutoEnable)
}
\arguments{
\item{AutoEnable}{[required] Whether to automatically enable Security Hub for new accounts in the
organization.

By default, this is \code{false}, and new accounts are not added
automatically.

To automatically enable Security Hub for new accounts, set this to
\code{true}.}
}
\value{
An empty list.
}
\description{
Used to update the configuration related to Organizations. Can only be
called from a Security Hub administrator account.
}
\section{Request syntax}{
\preformatted{svc$update_organization_configuration(
  AutoEnable = TRUE|FALSE
)
}
}

\keyword{internal}
