% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_security_control}
\alias{securityhub_update_security_control}
\title{Updates the properties of a security control}
\usage{
securityhub_update_security_control(
  SecurityControlId,
  Parameters,
  LastUpdateReason = NULL
)
}
\arguments{
\item{SecurityControlId}{[required] The Amazon Resource Name (ARN) or ID of the control to update.}

\item{Parameters}{[required] An object that specifies which security control parameters to update.}

\item{LastUpdateReason}{The most recent reason for updating the properties of the security
control. This field accepts alphanumeric characters in addition to white
spaces, dashes, and underscores.}
}
\description{
Updates the properties of a security control.

See \url{https://www.paws-r-sdk.com/docs/securityhub_update_security_control/} for full documentation.
}
\keyword{internal}
