% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_create_aws_log_source}
\alias{securitylake_create_aws_log_source}
\title{Adds a natively supported Amazon Web Service as an Amazon Security Lake
source}
\usage{
securitylake_create_aws_log_source(sources)
}
\arguments{
\item{sources}{[required] Specify the natively-supported Amazon Web Services service to add as a
source in Security Lake.}
}
\description{
Adds a natively supported Amazon Web Service as an Amazon Security Lake source. Enables source types for member accounts in required Amazon Web Services Regions, based on the parameters you specify. You can choose any source type in any Region for either accounts that are part of a trusted organization or standalone accounts. Once you add an Amazon Web Service as a source, Security Lake starts collecting logs and events from it.

See \url{https://www.paws-r-sdk.com/docs/securitylake_create_aws_log_source/} for full documentation.
}
\keyword{internal}
