% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_create_data_lake_organization_configuration}
\alias{securitylake_create_data_lake_organization_configuration}
\title{Automatically enables Amazon Security Lake for new member accounts in
your organization}
\usage{
securitylake_create_data_lake_organization_configuration(
  autoEnableNewAccount = NULL
)
}
\arguments{
\item{autoEnableNewAccount}{Enable Security Lake with the specified configuration settings, to begin
collecting security data for new accounts in your organization.}
}
\description{
Automatically enables Amazon Security Lake for new member accounts in your organization. Security Lake is not automatically enabled for any existing member accounts in your organization.

See \url{https://www.paws-r-sdk.com/docs/securitylake_create_data_lake_organization_configuration/} for full documentation.
}
\keyword{internal}
