% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_create_subscriber_notification}
\alias{securitylake_create_subscriber_notification}
\title{Notifies the subscriber when new data is written to the data lake for
the sources that the subscriber consumes in Security Lake}
\usage{
securitylake_create_subscriber_notification(configuration, subscriberId)
}
\arguments{
\item{configuration}{[required] Specify the configuration using which you want to create the subscriber
notification.}

\item{subscriberId}{[required] The subscriber ID for the notification subscription.}
}
\description{
Notifies the subscriber when new data is written to the data lake for the sources that the subscriber consumes in Security Lake. You can create only one subscriber notification per subscriber.

See \url{https://www.paws-r-sdk.com/docs/securitylake_create_subscriber_notification/} for full documentation.
}
\keyword{internal}
