% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_delete_aws_log_source}
\alias{securitylake_delete_aws_log_source}
\title{Removes a natively supported Amazon Web Service as an Amazon Security
Lake source}
\usage{
securitylake_delete_aws_log_source(sources)
}
\arguments{
\item{sources}{[required] Specify the natively-supported Amazon Web Services service to remove as
a source in Security Lake.}
}
\description{
Removes a natively supported Amazon Web Service as an Amazon Security Lake source. You can remove a source for one or more Regions. When you remove the source, Security Lake stops collecting data from that source in the specified Regions and accounts, and subscribers can no longer consume new data from the source. However, subscribers can still consume data that Security Lake collected from the source before removal.

See \url{https://www.paws-r-sdk.com/docs/securitylake_delete_aws_log_source/} for full documentation.
}
\keyword{internal}
