% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_delete_data_lake}
\alias{securitylake_delete_data_lake}
\title{When you disable Amazon Security Lake from your account, Security Lake
is disabled in all Amazon Web Services Regions and it stops collecting
data from your sources}
\usage{
securitylake_delete_data_lake(regions)
}
\arguments{
\item{regions}{[required] The list of Regions where Security Lake is enabled.}
}
\description{
When you disable Amazon Security Lake from your account, Security Lake is disabled in all Amazon Web Services Regions and it stops collecting data from your sources. Also, this API automatically takes steps to remove the account from Security Lake. However, Security Lake retains all of your existing settings and the resources that it created in your Amazon Web Services account in the current Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/securitylake_delete_data_lake/} for full documentation.
}
\keyword{internal}
