% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_delete_data_lake_exception_subscription}
\alias{securitylake_delete_data_lake_exception_subscription}
\title{Deletes the specified notification subscription in Amazon Security Lake
for the organization you specify}
\usage{
securitylake_delete_data_lake_exception_subscription()
}
\description{
Deletes the specified notification subscription in Amazon Security Lake for the organization you specify.

See \url{https://www.paws-r-sdk.com/docs/securitylake_delete_data_lake_exception_subscription/} for full documentation.
}
\keyword{internal}
