% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_delete_data_lake_organization_configuration}
\alias{securitylake_delete_data_lake_organization_configuration}
\title{Turns off automatic enablement of Amazon Security Lake for member
accounts that are added to an organization in Organizations}
\usage{
securitylake_delete_data_lake_organization_configuration(
  autoEnableNewAccount = NULL
)
}
\arguments{
\item{autoEnableNewAccount}{Turns off automatic enablement of Security Lake for member accounts that
are added to an organization.}
}
\description{
Turns off automatic enablement of Amazon Security Lake for member accounts that are added to an organization in Organizations. Only the delegated Security Lake administrator for an organization can perform this operation. If the delegated Security Lake administrator performs this operation, new member accounts won't automatically contribute data to the data lake.

See \url{https://www.paws-r-sdk.com/docs/securitylake_delete_data_lake_organization_configuration/} for full documentation.
}
\keyword{internal}
