% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_delete_subscriber}
\alias{securitylake_delete_subscriber}
\title{Deletes the subscription permission and all notification settings for
accounts that are already enabled in Amazon Security Lake}
\usage{
securitylake_delete_subscriber(subscriberId)
}
\arguments{
\item{subscriberId}{[required] A value created by Security Lake that uniquely identifies your
\code{\link[=securitylake_delete_subscriber]{delete_subscriber}} API request.}
}
\description{
Deletes the subscription permission and all notification settings for accounts that are already enabled in Amazon Security Lake. When you run \code{\link[=securitylake_delete_subscriber]{delete_subscriber}}, the subscriber will no longer consume data from Security Lake and the subscriber is removed. This operation deletes the subscriber and removes access to data in the current Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/securitylake_delete_subscriber/} for full documentation.
}
\keyword{internal}
