% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_delete_subscriber_notification}
\alias{securitylake_delete_subscriber_notification}
\title{Deletes the specified notification subscription in Amazon Security Lake
for the organization you specify}
\usage{
securitylake_delete_subscriber_notification(subscriberId)
}
\arguments{
\item{subscriberId}{[required] The ID of the Security Lake subscriber account.}
}
\description{
Deletes the specified notification subscription in Amazon Security Lake for the organization you specify.

See \url{https://www.paws-r-sdk.com/docs/securitylake_delete_subscriber_notification/} for full documentation.
}
\keyword{internal}
