% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_get_data_lake_exception_subscription}
\alias{securitylake_get_data_lake_exception_subscription}
\title{Retrieves the protocol and endpoint that were provided when subscribing
to Amazon SNS topics for exception notifications}
\usage{
securitylake_get_data_lake_exception_subscription()
}
\description{
Retrieves the protocol and endpoint that were provided when subscribing to Amazon SNS topics for exception notifications.

See \url{https://www.paws-r-sdk.com/docs/securitylake_get_data_lake_exception_subscription/} for full documentation.
}
\keyword{internal}
