% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_get_data_lake_organization_configuration}
\alias{securitylake_get_data_lake_organization_configuration}
\title{Retrieves the configuration that will be automatically set up for
accounts added to the organization after the organization has onboarded
to Amazon Security Lake}
\usage{
securitylake_get_data_lake_organization_configuration()
}
\description{
Retrieves the configuration that will be automatically set up for accounts added to the organization after the organization has onboarded to Amazon Security Lake. This API does not take input parameters.

See \url{https://www.paws-r-sdk.com/docs/securitylake_get_data_lake_organization_configuration/} for full documentation.
}
\keyword{internal}
