% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_get_subscriber}
\alias{securitylake_get_subscriber}
\title{Retrieves the subscription information for the specified subscription ID}
\usage{
securitylake_get_subscriber(subscriberId)
}
\arguments{
\item{subscriberId}{[required] A value created by Amazon Security Lake that uniquely identifies your
\code{\link[=securitylake_get_subscriber]{get_subscriber}} API request.}
}
\description{
Retrieves the subscription information for the specified subscription ID. You can get information about a specific subscriber.

See \url{https://www.paws-r-sdk.com/docs/securitylake_get_subscriber/} for full documentation.
}
\keyword{internal}
