% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_list_data_lakes}
\alias{securitylake_list_data_lakes}
\title{Retrieves the Amazon Security Lake configuration object for the
specified Amazon Web Services Regions}
\usage{
securitylake_list_data_lakes(regions = NULL)
}
\arguments{
\item{regions}{The list of Regions where Security Lake is enabled.}
}
\description{
Retrieves the Amazon Security Lake configuration object for the specified Amazon Web Services Regions. You can use this operation to determine whether Security Lake is enabled for a Region.

See \url{https://www.paws-r-sdk.com/docs/securitylake_list_data_lakes/} for full documentation.
}
\keyword{internal}
