% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_list_log_sources}
\alias{securitylake_list_log_sources}
\title{Retrieves the log sources in the current Amazon Web Services Region}
\usage{
securitylake_list_log_sources(
  accounts = NULL,
  maxResults = NULL,
  nextToken = NULL,
  regions = NULL,
  sources = NULL
)
}
\arguments{
\item{accounts}{The list of Amazon Web Services accounts for which log sources are
displayed.}

\item{maxResults}{The maximum number of accounts for which the log sources are displayed.}

\item{nextToken}{If nextToken is returned, there are more results available. You can
repeat the call using the returned token to retrieve the next page.}

\item{regions}{The list of Regions for which log sources are displayed.}

\item{sources}{The list of sources for which log sources are displayed.}
}
\description{
Retrieves the log sources in the current Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/securitylake_list_log_sources/} for full documentation.
}
\keyword{internal}
