% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_list_tags_for_resource}
\alias{securitylake_list_tags_for_resource}
\title{Retrieves the tags (keys and values) that are associated with an Amazon
Security Lake resource: a subscriber, or the data lake configuration for
your Amazon Web Services account in a particular Amazon Web Services
Region}
\usage{
securitylake_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon Security Lake resource for
which you want to retrieve the tags.}
}
\description{
Retrieves the tags (keys and values) that are associated with an Amazon Security Lake resource: a subscriber, or the data lake configuration for your Amazon Web Services account in a particular Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/securitylake_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
