% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_register_data_lake_delegated_administrator}
\alias{securitylake_register_data_lake_delegated_administrator}
\title{Designates the Amazon Security Lake delegated administrator account for
the organization}
\usage{
securitylake_register_data_lake_delegated_administrator(accountId)
}
\arguments{
\item{accountId}{[required] The Amazon Web Services account ID of the Security Lake delegated
administrator.}
}
\description{
Designates the Amazon Security Lake delegated administrator account for the organization. This API can only be called by the organization management account. The organization management account cannot be the delegated administrator account.

See \url{https://www.paws-r-sdk.com/docs/securitylake_register_data_lake_delegated_administrator/} for full documentation.
}
\keyword{internal}
