% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_untag_resource}
\alias{securitylake_untag_resource}
\title{Removes one or more tags (keys and values) from an Amazon Security Lake
resource: a subscriber, or the data lake configuration for your Amazon
Web Services account in a particular Amazon Web Services Region}
\usage{
securitylake_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the Amazon Security Lake resource to
remove one or more tags from.}

\item{tagKeys}{[required] A list of one or more tag keys. For each value in the list, specify the
tag key for a tag to remove from the Amazon Security Lake resource.}
}
\description{
Removes one or more tags (keys and values) from an Amazon Security Lake resource: a subscriber, or the data lake configuration for your Amazon Web Services account in a particular Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/securitylake_untag_resource/} for full documentation.
}
\keyword{internal}
