% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_update_data_lake_exception_subscription}
\alias{securitylake_update_data_lake_exception_subscription}
\title{Updates the specified notification subscription in Amazon Security Lake
for the organization you specify}
\usage{
securitylake_update_data_lake_exception_subscription(
  exceptionTimeToLive = NULL,
  notificationEndpoint,
  subscriptionProtocol
)
}
\arguments{
\item{exceptionTimeToLive}{The time-to-live (TTL) for the exception message to remain.}

\item{notificationEndpoint}{[required] The account that is subscribed to receive exception notifications.}

\item{subscriptionProtocol}{[required] The subscription protocol to which exception messages are posted.}
}
\description{
Updates the specified notification subscription in Amazon Security Lake for the organization you specify.

See \url{https://www.paws-r-sdk.com/docs/securitylake_update_data_lake_exception_subscription/} for full documentation.
}
\keyword{internal}
