% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_update_subscriber_notification}
\alias{securitylake_update_subscriber_notification}
\title{Updates an existing notification method for the subscription (SQS or
HTTPs endpoint) or switches the notification subscription endpoint for a
subscriber}
\usage{
securitylake_update_subscriber_notification(configuration, subscriberId)
}
\arguments{
\item{configuration}{[required] The configuration for subscriber notification.}

\item{subscriberId}{[required] The subscription ID for which the subscription notification is
specified.}
}
\description{
Updates an existing notification method for the subscription (SQS or HTTPs endpoint) or switches the notification subscription endpoint for a subscriber.

See \url{https://www.paws-r-sdk.com/docs/securitylake_update_subscriber_notification/} for full documentation.
}
\keyword{internal}
