% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_associate_proactive_engagement_details}
\alias{shield_associate_proactive_engagement_details}
\title{Initializes proactive engagement and sets the list of contacts for the
Shield Response Team (SRT) to use}
\usage{
shield_associate_proactive_engagement_details(EmergencyContactList)
}
\arguments{
\item{EmergencyContactList}{[required] A list of email addresses and phone numbers that the Shield Response
Team (SRT) can use to contact you for escalations to the SRT and to
initiate proactive customer support.

To enable proactive engagement, the contact list must include at least
one phone number.

The contacts that you provide here replace any contacts that were
already defined. If you already have contacts defined and want to use
them, retrieve the list using
\code{\link[=shield_describe_emergency_contact_settings]{describe_emergency_contact_settings}}
and then provide it here.}
}
\description{
Initializes proactive engagement and sets the list of contacts for the Shield Response Team (SRT) to use. You must provide at least one phone number in the emergency contact list.

See \url{https://www.paws-r-sdk.com/docs/shield_associate_proactive_engagement_details/} for full documentation.
}
\keyword{internal}
