% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_delete_protection_group}
\alias{shield_delete_protection_group}
\title{Removes the specified protection group}
\usage{
shield_delete_protection_group(ProtectionGroupId)
}
\arguments{
\item{ProtectionGroupId}{[required] The name of the protection group. You use this to identify the
protection group in lists and to manage the protection group, for
example to update, delete, or describe it.}
}
\value{
An empty list.
}
\description{
Removes the specified protection group.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_protection_group(
  ProtectionGroupId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
