% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_delete_subscription}
\alias{shield_delete_subscription}
\title{Removes AWS Shield Advanced from an account}
\usage{
shield_delete_subscription()
}
\value{
An empty list.
}
\description{
Removes AWS Shield Advanced from an account. AWS Shield Advanced
requires a 1-year subscription commitment. You cannot delete a
subscription prior to the completion of that commitment.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_subscription()
}\if{html}{\out{</div>}}
}

\keyword{internal}
