% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_attack_statistics}
\alias{shield_describe_attack_statistics}
\title{Provides information about the number and type of attacks AWS Shield has
detected in the last year for all resources that belong to your account,
regardless of whether you've defined Shield protections for them}
\usage{
shield_describe_attack_statistics()
}
\value{
A list with the following syntax:\preformatted{list(
  TimeRange = list(
    FromInclusive = as.POSIXct(
      "2015-01-01"
    ),
    ToExclusive = as.POSIXct(
      "2015-01-01"
    )
  ),
  DataItems = list(
    list(
      AttackVolume = list(
        BitsPerSecond = list(
          Max = 123.0
        ),
        PacketsPerSecond = list(
          Max = 123.0
        ),
        RequestsPerSecond = list(
          Max = 123.0
        )
      ),
      AttackCount = 123
    )
  )
)
}
}
\description{
Provides information about the number and type of attacks AWS Shield has
detected in the last year for all resources that belong to your account,
regardless of whether you've defined Shield protections for them. This
operation is available to Shield customers as well as to Shield Advanced
customers.

The operation returns data for the time range of midnight UTC, one year
ago, to midnight UTC, today. For example, if the current time is
\verb{2020-10-26 15:39:32 PDT}, equal to \verb{2020-10-26 22:39:32 UTC}, then the
time range for the attack data returned is from
\verb{2019-10-26 00:00:00 UTC} to \verb{2020-10-26 00:00:00 UTC}.

The time range indicates the period covered by the attack statistics
data items.
}
\section{Request syntax}{
\preformatted{svc$describe_attack_statistics()
}
}

\keyword{internal}
