% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_subscription}
\alias{shield_describe_subscription}
\title{Provides details about the AWS Shield Advanced subscription for an
account}
\usage{
shield_describe_subscription()
}
\description{
Provides details about the AWS Shield Advanced subscription for an
account.
}
\section{Request syntax}{
\preformatted{svc$describe_subscription()
}
}

\keyword{internal}
