% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_disable_proactive_engagement}
\alias{shield_disable_proactive_engagement}
\title{Removes authorization from the DDoS Response Team (DRT) to notify
contacts about escalations to the DRT and to initiate proactive customer
support}
\usage{
shield_disable_proactive_engagement()
}
\value{
An empty list.
}
\description{
Removes authorization from the DDoS Response Team (DRT) to notify
contacts about escalations to the DRT and to initiate proactive customer
support.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disable_proactive_engagement()
}\if{html}{\out{</div>}}
}

\keyword{internal}
