% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_disassociate_drt_log_bucket}
\alias{shield_disassociate_drt_log_bucket}
\title{Removes the Shield Response Team's (SRT) access to the specified Amazon
S3 bucket containing the logs that you shared previously}
\usage{
shield_disassociate_drt_log_bucket(LogBucket)
}
\arguments{
\item{LogBucket}{[required] The Amazon S3 bucket that contains the logs that you want to share.}
}
\description{
Removes the Shield Response Team's (SRT) access to the specified Amazon S3 bucket containing the logs that you shared previously.

See \url{https://www.paws-r-sdk.com/docs/shield_disassociate_drt_log_bucket/} for full documentation.
}
\keyword{internal}
