% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_enable_proactive_engagement}
\alias{shield_enable_proactive_engagement}
\title{Authorizes the DDoS Response Team (DRT) to use email and phone to notify
contacts about escalations to the DRT and to initiate proactive customer
support}
\usage{
shield_enable_proactive_engagement()
}
\value{
An empty list.
}
\description{
Authorizes the DDoS Response Team (DRT) to use email and phone to notify
contacts about escalations to the DRT and to initiate proactive customer
support.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$enable_proactive_engagement()
}\if{html}{\out{</div>}}
}

\keyword{internal}
