% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_get_subscription_state}
\alias{shield_get_subscription_state}
\title{Returns the SubscriptionState, either Active or Inactive}
\usage{
shield_get_subscription_state()
}
\value{
A list with the following syntax:\preformatted{list(
  SubscriptionState = "ACTIVE"|"INACTIVE"
)
}
}
\description{
Returns the \code{SubscriptionState}, either \code{Active} or \code{Inactive}.
}
\section{Request syntax}{
\preformatted{svc$get_subscription_state()
}
}

\keyword{internal}
