% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_tag_resource}
\alias{shield_tag_resource}
\title{Adds or updates tags for a resource in AWS Shield}
\usage{
shield_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource that you want to add or
update tags for.}

\item{Tags}{[required] The tags that you want to modify or add to the resource.}
}
\value{
An empty list.
}
\description{
Adds or updates tags for a resource in AWS Shield.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceARN = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
