% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_update_emergency_contact_settings}
\alias{shield_update_emergency_contact_settings}
\title{Updates the details of the list of email addresses that the DRT can use
to contact you during a suspected attack}
\usage{
shield_update_emergency_contact_settings(EmergencyContactList)
}
\arguments{
\item{EmergencyContactList}{A list of email addresses that the DRT can use to contact you during a
suspected attack.}
}
\description{
Updates the details of the list of email addresses that the DRT can use
to contact you during a suspected attack.
}
\section{Request syntax}{
\preformatted{svc$update_emergency_contact_settings(
  EmergencyContactList = list(
    list(
      EmailAddress = "string"
    )
  )
)
}
}

\keyword{internal}
