% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sso_operations.R
\name{sso_logout}
\alias{sso_logout}
\title{Removes the client- and server-side session that is associated with the
user}
\usage{
sso_logout(accessToken)
}
\arguments{
\item{accessToken}{[required] The token issued by the \code{CreateToken} API call. For more information,
see
\href{https://docs.aws.amazon.com/singlesignon/latest/OIDCAPIReference/API_CreateToken.html}{CreateToken}
in the \emph{AWS SSO OIDC API Reference Guide}.}
}
\value{
An empty list.
}
\description{
Removes the client- and server-side session that is associated with the
user.
}
\section{Request syntax}{
\preformatted{svc$logout(
  accessToken = "string"
)
}
}

\keyword{internal}
