% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_add_region}
\alias{ssoadmin_add_region}
\title{Adds a Region to an IAM Identity Center instance}
\usage{
ssoadmin_add_region(InstanceArn, RegionName)
}
\arguments{
\item{InstanceArn}{[required] The Amazon Resource Name (ARN) of the IAM Identity Center instance to
replicate to the target Region.}

\item{RegionName}{[required] The name of the Amazon Web Services Region to add to the IAM Identity
Center instance. The Region name must be 1-32 characters long and follow
the pattern of Amazon Web Services Region names (for example,
us-east-1).}
}
\description{
Adds a Region to an IAM Identity Center instance. This operation initiates an asynchronous workflow to replicate the IAM Identity Center instance to the target Region. The Region status is set to ADDING at first and changes to ACTIVE when the workflow completes.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_add_region/} for full documentation.
}
\keyword{internal}
