% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_delete_application}
\alias{ssoadmin_delete_application}
\title{Deletes the association with the application}
\usage{
ssoadmin_delete_application(ApplicationArn)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application. For more information about ARNs,
see Amazon Resource Names (ARNs) and Amazon Web Services Service
Namespaces in the \emph{Amazon Web Services General Reference}.}
}
\description{
Deletes the association with the application. The connected service resource still exists.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_delete_application/} for full documentation.
}
\keyword{internal}
