% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_delete_application_access_scope}
\alias{ssoadmin_delete_application_access_scope}
\title{Deletes an IAM Identity Center access scope from an application}
\usage{
ssoadmin_delete_application_access_scope(ApplicationArn, Scope)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application with the access scope to delete.}

\item{Scope}{[required] Specifies the name of the access scope to remove from the application.}
}
\description{
Deletes an IAM Identity Center access scope from an application.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_delete_application_access_scope/} for full documentation.
}
\keyword{internal}
