% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_delete_application_authentication_method}
\alias{ssoadmin_delete_application_authentication_method}
\title{Deletes an authentication method from an application}
\usage{
ssoadmin_delete_application_authentication_method(
  ApplicationArn,
  AuthenticationMethodType
)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application with the authentication method to
delete.}

\item{AuthenticationMethodType}{[required] Specifies the authentication method type to delete from the application.}
}
\description{
Deletes an authentication method from an application.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_delete_application_authentication_method/} for full documentation.
}
\keyword{internal}
