% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_delete_application_grant}
\alias{ssoadmin_delete_application_grant}
\title{Deletes a grant from an application}
\usage{
ssoadmin_delete_application_grant(ApplicationArn, GrantType)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application with the grant to delete.}

\item{GrantType}{[required] Specifies the type of grant to delete from the application.}
}
\description{
Deletes a grant from an application.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_delete_application_grant/} for full documentation.
}
\keyword{internal}
