% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_delete_instance}
\alias{ssoadmin_delete_instance}
\title{Deletes the instance of IAM Identity Center}
\usage{
ssoadmin_delete_instance(InstanceArn)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the instance of IAM Identity Center under which the operation
will run.}
}
\description{
Deletes the instance of IAM Identity Center. Only the account that owns the instance can call this API. Neither the delegated administrator nor member account can delete the organization instance, but those roles can delete their own instance.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_delete_instance/} for full documentation.
}
\keyword{internal}
