% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_delete_permission_set}
\alias{ssoadmin_delete_permission_set}
\title{Deletes the specified permission set}
\usage{
ssoadmin_delete_permission_set(InstanceArn, PermissionSetArn)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{PermissionSetArn}{[required] The ARN of the permission set that should be deleted.}
}
\description{
Deletes the specified permission set.

See \url{https://paws-r.github.io/docs/ssoadmin/delete_permission_set.html} for full documentation.
}
\keyword{internal}
