% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_delete_trusted_token_issuer}
\alias{ssoadmin_delete_trusted_token_issuer}
\title{Deletes a trusted token issuer configuration from an instance of IAM
Identity Center}
\usage{
ssoadmin_delete_trusted_token_issuer(TrustedTokenIssuerArn)
}
\arguments{
\item{TrustedTokenIssuerArn}{[required] Specifies the ARN of the trusted token issuer configuration to delete.}
}
\description{
Deletes a trusted token issuer configuration from an instance of IAM Identity Center.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_delete_trusted_token_issuer/} for full documentation.
}
\keyword{internal}
