% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_describe_application}
\alias{ssoadmin_describe_application}
\title{Retrieves the details of an application associated with an instance of
IAM Identity Center}
\usage{
ssoadmin_describe_application(ApplicationArn)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application. For more information about ARNs,
see Amazon Resource Names (ARNs) and Amazon Web Services Service
Namespaces in the \emph{Amazon Web Services General Reference}.}
}
\description{
Retrieves the details of an application associated with an instance of IAM Identity Center.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_describe_application/} for full documentation.
}
\keyword{internal}
