% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_describe_application_provider}
\alias{ssoadmin_describe_application_provider}
\title{Retrieves details about a provider that can be used to connect an Amazon
Web Services managed application or customer managed application to IAM
Identity Center}
\usage{
ssoadmin_describe_application_provider(ApplicationProviderArn)
}
\arguments{
\item{ApplicationProviderArn}{[required] Specifies the ARN of the application provider for which you want
details.}
}
\description{
Retrieves details about a provider that can be used to connect an Amazon Web Services managed application or customer managed application to IAM Identity Center.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_describe_application_provider/} for full documentation.
}
\keyword{internal}
