% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_describe_instance}
\alias{ssoadmin_describe_instance}
\title{Returns the details of an instance of IAM Identity Center}
\usage{
ssoadmin_describe_instance(InstanceArn)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the instance of IAM Identity Center under which the operation
will run.}
}
\description{
Returns the details of an instance of IAM Identity Center. The status can be one of the following:

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_describe_instance/} for full documentation.
}
\keyword{internal}
