% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_describe_permission_set_provisioning_status}
\alias{ssoadmin_describe_permission_set_provisioning_status}
\title{Describes the status for the given permission set provisioning request}
\usage{
ssoadmin_describe_permission_set_provisioning_status(
  InstanceArn,
  ProvisionPermissionSetRequestId
)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and Amazon Web Services Service Namespaces in the \emph{Amazon
Web Services General Reference}.}

\item{ProvisionPermissionSetRequestId}{[required] The identifier that is provided by the
\code{\link[=ssoadmin_provision_permission_set]{provision_permission_set}} call to
retrieve the current status of the provisioning workflow.}
}
\description{
Describes the status for the given permission set provisioning request.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_describe_permission_set_provisioning_status/} for full documentation.
}
\keyword{internal}
