% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_describe_trusted_token_issuer}
\alias{ssoadmin_describe_trusted_token_issuer}
\title{Retrieves details about a trusted token issuer configuration stored in
an instance of IAM Identity Center}
\usage{
ssoadmin_describe_trusted_token_issuer(TrustedTokenIssuerArn)
}
\arguments{
\item{TrustedTokenIssuerArn}{[required] Specifies the ARN of the trusted token issuer configuration that you
want details about.}
}
\description{
Retrieves details about a trusted token issuer configuration stored in an instance of IAM Identity Center. Details include the name of the trusted token issuer, the issuer URL, and the path of the source attribute and the destination attribute for a trusted token issuer configuration.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_describe_trusted_token_issuer/} for full documentation.
}
\keyword{internal}
