% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_detach_customer_managed_policy_reference_from_permission_set}
\alias{ssoadmin_detach_customer_managed_policy_reference_from_permission_set}
\title{Detaches the specified customer managed policy from the specified
PermissionSet}
\usage{
ssoadmin_detach_customer_managed_policy_reference_from_permission_set(
  CustomerManagedPolicyReference,
  InstanceArn,
  PermissionSetArn
)
}
\arguments{
\item{CustomerManagedPolicyReference}{[required] Specifies the name and path of a customer managed policy. You must have
an IAM policy that matches the name and path in each Amazon Web Services
account where you want to deploy your permission set.}

\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed.}

\item{PermissionSetArn}{[required] The ARN of the \code{PermissionSet}.}
}
\description{
Detaches the specified customer managed policy from the specified PermissionSet.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_detach_customer_managed_policy_reference_from_permission_set/} for full documentation.
}
\keyword{internal}
