% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_get_application_access_scope}
\alias{ssoadmin_get_application_access_scope}
\title{Retrieves the authorized targets for an IAM Identity Center access scope
for an application}
\usage{
ssoadmin_get_application_access_scope(ApplicationArn, Scope)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application with the access scope that you want
to retrieve.}

\item{Scope}{[required] Specifies the name of the access scope for which you want the authorized
targets.}
}
\description{
Retrieves the authorized targets for an IAM Identity Center access scope for an application.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_get_application_access_scope/} for full documentation.
}
\keyword{internal}
