% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_get_application_assignment_configuration}
\alias{ssoadmin_get_application_assignment_configuration}
\title{Retrieves the configuration of PutApplicationAssignmentConfiguration}
\usage{
ssoadmin_get_application_assignment_configuration(ApplicationArn)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application. For more information about ARNs,
see Amazon Resource Names (ARNs) and Amazon Web Services Service
Namespaces in the \emph{Amazon Web Services General Reference}.}
}
\description{
Retrieves the configuration of \code{\link[=ssoadmin_put_application_assignment_configuration]{put_application_assignment_configuration}}.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_get_application_assignment_configuration/} for full documentation.
}
\keyword{internal}
