% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_get_application_authentication_method}
\alias{ssoadmin_get_application_authentication_method}
\title{Retrieves details about an authentication method used by an application}
\usage{
ssoadmin_get_application_authentication_method(
  ApplicationArn,
  AuthenticationMethodType
)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application.}

\item{AuthenticationMethodType}{[required] Specifies the type of authentication method for which you want details.}
}
\description{
Retrieves details about an authentication method used by an application.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_get_application_authentication_method/} for full documentation.
}
\keyword{internal}
