% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_get_application_grant}
\alias{ssoadmin_get_application_grant}
\title{Retrieves details about an application grant}
\usage{
ssoadmin_get_application_grant(ApplicationArn, GrantType)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application that contains the grant.}

\item{GrantType}{[required] Specifies the type of grant.}
}
\description{
Retrieves details about an application grant.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_get_application_grant/} for full documentation.
}
\keyword{internal}
