% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_get_application_session_configuration}
\alias{ssoadmin_get_application_session_configuration}
\title{Retrieves the session configuration for an application in IAM Identity
Center}
\usage{
ssoadmin_get_application_session_configuration(ApplicationArn)
}
\arguments{
\item{ApplicationArn}{[required] The Amazon Resource Name (ARN) of the application for which to retrieve
the session configuration.}
}
\description{
Retrieves the session configuration for an application in IAM Identity Center.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_get_application_session_configuration/} for full documentation.
}
\keyword{internal}
