% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_get_inline_policy_for_permission_set}
\alias{ssoadmin_get_inline_policy_for_permission_set}
\title{Obtains the inline policy assigned to the permission set}
\usage{
ssoadmin_get_inline_policy_for_permission_set(InstanceArn, PermissionSetArn)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and Amazon Web Services Service Namespaces in the \emph{Amazon
Web Services General Reference}.}

\item{PermissionSetArn}{[required] The ARN of the permission set.}
}
\description{
Obtains the inline policy assigned to the permission set.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_get_inline_policy_for_permission_set/} for full documentation.
}
\keyword{internal}
