% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_get_permissions_boundary_for_permission_set}
\alias{ssoadmin_get_permissions_boundary_for_permission_set}
\title{Obtains the permissions boundary for a specified PermissionSet}
\usage{
ssoadmin_get_permissions_boundary_for_permission_set(
  InstanceArn,
  PermissionSetArn
)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed.}

\item{PermissionSetArn}{[required] The ARN of the \code{PermissionSet}.}
}
\description{
Obtains the permissions boundary for a specified PermissionSet.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_get_permissions_boundary_for_permission_set/} for full documentation.
}
\keyword{internal}
