% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_account_assignment_creation_status}
\alias{ssoadmin_list_account_assignment_creation_status}
\title{Lists the status of the AWS account assignment creation requests for a
specified IAM Identity Center instance}
\usage{
ssoadmin_list_account_assignment_creation_status(
  InstanceArn,
  MaxResults = NULL,
  NextToken = NULL,
  Filter = NULL
)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{MaxResults}{The maximum number of results to display for the assignment.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}

\item{Filter}{Filters results based on the passed attribute value.}
}
\description{
Lists the status of the AWS account assignment creation requests for a specified IAM Identity Center instance.

See \url{https://paws-r.github.io/docs/ssoadmin/list_account_assignment_creation_status.html} for full documentation.
}
\keyword{internal}
