% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_account_assignments}
\alias{ssoadmin_list_account_assignments}
\title{Lists the assignee of the specified Amazon Web Services account with the
specified permission set}
\usage{
ssoadmin_list_account_assignments(
  InstanceArn,
  AccountId,
  PermissionSetArn,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the Amazon Web Services SSO instance under which the
operation will be executed. For more information about ARNs, see Amazon
Resource Names (ARNs) and Amazon Web Services Service Namespaces in the
\emph{Amazon Web Services General Reference}.}

\item{AccountId}{[required] The identifier of the Amazon Web Services account from which to list the
assignments.}

\item{PermissionSetArn}{[required] The ARN of the permission set from which to list assignments.}

\item{MaxResults}{The maximum number of results to display for the assignment.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}
}
\description{
Lists the assignee of the specified Amazon Web Services account with the specified permission set.

See \url{https://paws-r.github.io/docs/ssoadmin/list_account_assignments.html} for full documentation.
}
\keyword{internal}
