% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_application_authentication_methods}
\alias{ssoadmin_list_application_authentication_methods}
\title{Lists all of the authentication methods supported by the specified
application}
\usage{
ssoadmin_list_application_authentication_methods(
  ApplicationArn,
  NextToken = NULL
)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application with the authentication methods you
want to list.}

\item{NextToken}{Specifies that you want to receive the next page of results. Valid only
if you received a \code{NextToken} response in the previous request. If you
did, it indicates that more output is available. Set this parameter to
the value provided by the previous call's \code{NextToken} response to
request the next page of results.}
}
\description{
Lists all of the authentication methods supported by the specified application.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_list_application_authentication_methods/} for full documentation.
}
\keyword{internal}
