% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_application_grants}
\alias{ssoadmin_list_application_grants}
\title{List the grants associated with an application}
\usage{
ssoadmin_list_application_grants(ApplicationArn, NextToken = NULL)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application whose grants you want to list.}

\item{NextToken}{Specifies that you want to receive the next page of results. Valid only
if you received a \code{NextToken} response in the previous request. If you
did, it indicates that more output is available. Set this parameter to
the value provided by the previous call's \code{NextToken} response to
request the next page of results.}
}
\description{
List the grants associated with an application.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_list_application_grants/} for full documentation.
}
\keyword{internal}
