% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_customer_managed_policy_references_in_permission_set}
\alias{ssoadmin_list_customer_managed_policy_references_in_permission_set}
\title{Lists all customer managed policies attached to a specified
PermissionSet}
\usage{
ssoadmin_list_customer_managed_policy_references_in_permission_set(
  InstanceArn,
  PermissionSetArn,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed.}

\item{PermissionSetArn}{[required] The ARN of the \code{PermissionSet}.}

\item{MaxResults}{The maximum number of results to display for the list call.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}
}
\description{
Lists all customer managed policies attached to a specified PermissionSet.

See \url{https://paws-r.github.io/docs/ssoadmin/list_customer_managed_policy_references_in_permission_set.html} for full documentation.
}
\keyword{internal}
