% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_instances}
\alias{ssoadmin_list_instances}
\title{Lists the details of the organization and account instances of IAM
Identity Center that were created in or visible to the account calling
this API}
\usage{
ssoadmin_list_instances(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to display for the instance.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}
}
\description{
Lists the details of the organization and account instances of IAM Identity Center that were created in or visible to the account calling this API.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_list_instances/} for full documentation.
}
\keyword{internal}
