% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_managed_policies_in_permission_set}
\alias{ssoadmin_list_managed_policies_in_permission_set}
\title{Lists the AWS managed policy that is attached to a specified permission
set}
\usage{
ssoadmin_list_managed_policies_in_permission_set(
  InstanceArn,
  PermissionSetArn,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{PermissionSetArn}{[required] The ARN of the PermissionSet whose managed policies will be listed.}

\item{MaxResults}{The maximum number of results to display for the PermissionSet.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}
}
\description{
Lists the AWS managed policy that is attached to a specified permission set.

See \url{https://paws-r.github.io/docs/ssoadmin/list_managed_policies_in_permission_set.html} for full documentation.
}
\keyword{internal}
