% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_permission_sets}
\alias{ssoadmin_list_permission_sets}
\title{Lists the PermissionSets in an IAM Identity Center instance}
\usage{
ssoadmin_list_permission_sets(InstanceArn, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and Amazon Web Services Service Namespaces in the \emph{Amazon
Web Services General Reference}.}

\item{MaxResults}{The maximum number of results to display for the assignment.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}
}
\description{
Lists the PermissionSets in an IAM Identity Center instance.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_list_permission_sets/} for full documentation.
}
\keyword{internal}
