% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_permission_sets_provisioned_to_account}
\alias{ssoadmin_list_permission_sets_provisioned_to_account}
\title{Lists all the permission sets that are provisioned to a specified AWS
account}
\usage{
ssoadmin_list_permission_sets_provisioned_to_account(
  InstanceArn,
  AccountId,
  ProvisioningStatus = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{AccountId}{[required] The identifier of the AWS account from which to list the assignments.}

\item{ProvisioningStatus}{The status object for the permission set provisioning operation.}

\item{MaxResults}{The maximum number of results to display for the assignment.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}
}
\description{
Lists all the permission sets that are provisioned to a specified AWS account.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_list_permission_sets_provisioned_to_account/} for full documentation.
}
\keyword{internal}
