% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_tags_for_resource}
\alias{ssoadmin_list_tags_for_resource}
\title{Lists the tags that are attached to a specified resource}
\usage{
ssoadmin_list_tags_for_resource(
  InstanceArn = NULL,
  NextToken = NULL,
  ResourceArn
)
}
\arguments{
\item{InstanceArn}{The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and Amazon Web Services Service Namespaces in the \emph{Amazon
Web Services General Reference}.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}

\item{ResourceArn}{[required] The ARN of the resource with the tags to be listed.}
}
\description{
Lists the tags that are attached to a specified resource.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
