% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_trusted_token_issuers}
\alias{ssoadmin_list_trusted_token_issuers}
\title{Lists all the trusted token issuers configured in an instance of IAM
Identity Center}
\usage{
ssoadmin_list_trusted_token_issuers(
  InstanceArn,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{InstanceArn}{[required] Specifies the ARN of the instance of IAM Identity Center with the
trusted token issuer configurations that you want to list.}

\item{MaxResults}{Specifies the total number of results that you want included in each
response. If additional items exist beyond the number you specify, the
\code{NextToken} response element is returned with a value (not null).
Include the specified value as the \code{NextToken} request parameter in the
next call to the operation to get the next set of results. Note that the
service might return fewer results than the maximum even when there are
more results available. You should check \code{NextToken} after every
operation to ensure that you receive all of the results.}

\item{NextToken}{Specifies that you want to receive the next page of results. Valid only
if you received a \code{NextToken} response in the previous request. If you
did, it indicates that more output is available. Set this parameter to
the value provided by the previous call's \code{NextToken} response to
request the next page of results.}
}
\description{
Lists all the trusted token issuers configured in an instance of IAM Identity Center.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_list_trusted_token_issuers/} for full documentation.
}
\keyword{internal}
