% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_provision_permission_set}
\alias{ssoadmin_provision_permission_set}
\title{The process by which a specified permission set is provisioned to the
specified target}
\usage{
ssoadmin_provision_permission_set(
  InstanceArn,
  PermissionSetArn,
  TargetId = NULL,
  TargetType
)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the Amazon Web Services SSO instance under which the
operation will be executed. For more information about ARNs, see Amazon
Resource Names (ARNs) and Amazon Web Services Service Namespaces in the
\emph{Amazon Web Services General Reference}.}

\item{PermissionSetArn}{[required] The ARN of the permission set.}

\item{TargetId}{TargetID is an Amazon Web Services account identifier, typically a 10-12
digit string (For example, 123456789012).}

\item{TargetType}{[required] The entity type for which the assignment will be created.}
}
\description{
The process by which a specified permission set is provisioned to the specified target.

See \url{https://paws-r.github.io/docs/ssoadmin/provision_permission_set.html} for full documentation.
}
\keyword{internal}
