% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_put_application_access_scope}
\alias{ssoadmin_put_application_access_scope}
\title{Adds or updates the list of authorized targets for an IAM Identity
Center access scope for an application}
\usage{
ssoadmin_put_application_access_scope(
  ApplicationArn,
  AuthorizedTargets = NULL,
  Scope
)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application with the access scope with the
targets to add or update.}

\item{AuthorizedTargets}{Specifies an array list of ARNs that represent the authorized targets
for this access scope.}

\item{Scope}{[required] Specifies the name of the access scope to be associated with the
specified targets.}
}
\description{
Adds or updates the list of authorized targets for an IAM Identity Center access scope for an application.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_put_application_access_scope/} for full documentation.
}
\keyword{internal}
