% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_put_application_authentication_method}
\alias{ssoadmin_put_application_authentication_method}
\title{Adds or updates an authentication method for an application}
\usage{
ssoadmin_put_application_authentication_method(
  ApplicationArn,
  AuthenticationMethod,
  AuthenticationMethodType
)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application with the authentication method to
add or update.}

\item{AuthenticationMethod}{[required] Specifies a structure that describes the authentication method to add or
update. The structure type you provide is determined by the
\code{AuthenticationMethodType} parameter.}

\item{AuthenticationMethodType}{[required] Specifies the type of the authentication method that you want to add or
update.}
}
\description{
Adds or updates an authentication method for an application.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_put_application_authentication_method/} for full documentation.
}
\keyword{internal}
