% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_put_application_session_configuration}
\alias{ssoadmin_put_application_session_configuration}
\title{Updates the session configuration for an application in IAM Identity
Center}
\usage{
ssoadmin_put_application_session_configuration(
  ApplicationArn,
  UserBackgroundSessionApplicationStatus = NULL
)
}
\arguments{
\item{ApplicationArn}{[required] The Amazon Resource Name (ARN) of the application for which to update
the session configuration.}

\item{UserBackgroundSessionApplicationStatus}{The status of user background sessions for the application.}
}
\description{
Updates the session configuration for an application in IAM Identity Center.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_put_application_session_configuration/} for full documentation.
}
\keyword{internal}
