% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_tag_resource}
\alias{ssoadmin_tag_resource}
\title{Associates a set of tags with a specified resource}
\usage{
ssoadmin_tag_resource(InstanceArn = NULL, ResourceArn, Tags)
}
\arguments{
\item{InstanceArn}{The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and Amazon Web Services Service Namespaces in the \emph{Amazon
Web Services General Reference}.}

\item{ResourceArn}{[required] The ARN of the resource with the tags to be listed.}

\item{Tags}{[required] A set of key-value pairs that are used to manage the resource.}
}
\description{
Associates a set of tags with a specified resource.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_tag_resource/} for full documentation.
}
\keyword{internal}
