% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_update_application}
\alias{ssoadmin_update_application}
\title{Updates application properties}
\usage{
ssoadmin_update_application(
  ApplicationArn,
  Name = NULL,
  Description = NULL,
  Status = NULL,
  PortalOptions = NULL
)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application. For more information about ARNs,
see Amazon Resource Names (ARNs) and Amazon Web Services Service
Namespaces in the \emph{Amazon Web Services General Reference}.}

\item{Name}{Specifies the updated name for the application.}

\item{Description}{The description of the .}

\item{Status}{Specifies whether the application is enabled or disabled.}

\item{PortalOptions}{A structure that describes the options for the portal associated with an
application.}
}
\description{
Updates application properties.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_update_application/} for full documentation.
}
\keyword{internal}
