% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_update_permission_set}
\alias{ssoadmin_update_permission_set}
\title{Updates an existing permission set}
\usage{
ssoadmin_update_permission_set(
  Description = NULL,
  InstanceArn,
  PermissionSetArn,
  RelayState = NULL,
  SessionDuration = NULL
)
}
\arguments{
\item{Description}{The description of the PermissionSet.}

\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and Amazon Web Services Service Namespaces in the \emph{Amazon
Web Services General Reference}.}

\item{PermissionSetArn}{[required] The ARN of the permission set.}

\item{RelayState}{Used to redirect users within the application during the federation
authentication process.}

\item{SessionDuration}{The length of time that the application user sessions are valid for in
the ISO-8601 standard.}
}
\description{
Updates an existing permission set.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_update_permission_set/} for full documentation.
}
\keyword{internal}
