% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_update_trusted_token_issuer}
\alias{ssoadmin_update_trusted_token_issuer}
\title{Updates the name of the trusted token issuer, or the path of a source
attribute or destination attribute for a trusted token issuer
configuration}
\usage{
ssoadmin_update_trusted_token_issuer(
  Name = NULL,
  TrustedTokenIssuerArn,
  TrustedTokenIssuerConfiguration = NULL
)
}
\arguments{
\item{Name}{Specifies the updated name to be applied to the trusted token issuer
configuration.}

\item{TrustedTokenIssuerArn}{[required] Specifies the ARN of the trusted token issuer configuration that you
want to update.}

\item{TrustedTokenIssuerConfiguration}{Specifies a structure with settings to apply to the specified trusted
token issuer. The settings that you can provide are determined by the
type of the trusted token issuer that you are updating.}
}
\description{
Updates the name of the trusted token issuer, or the path of a source attribute or destination attribute for a trusted token issuer configuration.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_update_trusted_token_issuer/} for full documentation.
}
\keyword{internal}
