% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssooidc_operations.R
\name{ssooidc_start_device_authorization}
\alias{ssooidc_start_device_authorization}
\title{Initiates device authorization by requesting a pair of verification
codes from the authorization service}
\usage{
ssooidc_start_device_authorization(clientId, clientSecret, startUrl)
}
\arguments{
\item{clientId}{[required] The unique identifier string for the client that is registered with IAM
Identity Center. This value should come from the persisted result of the
\code{\link[=ssooidc_register_client]{register_client}} API operation.}

\item{clientSecret}{[required] A secret string that is generated for the client. This value should come
from the persisted result of the
\code{\link[=ssooidc_register_client]{register_client}} API operation.}

\item{startUrl}{[required] The URL for the Amazon Web Services access portal. For more information,
see \href{https://docs.aws.amazon.com/singlesignon/latest/userguide/using-the-portal.html}{Using the Amazon Web Services access portal}
in the \emph{IAM Identity Center User Guide}.}
}
\description{
Initiates device authorization by requesting a pair of verification codes from the authorization service.

See \url{https://www.paws-r-sdk.com/docs/ssooidc_start_device_authorization/} for full documentation.
}
\keyword{internal}
