% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts_operations.R
\name{sts_get_caller_identity}
\alias{sts_get_caller_identity}
\title{Returns details about the IAM user or role whose credentials are used to
call the operation}
\usage{
sts_get_caller_identity()
}
\description{
Returns details about the IAM user or role whose credentials are used to call the operation.

See \url{https://www.paws-r-sdk.com/docs/sts_get_caller_identity/} for full documentation.
}
\keyword{internal}
