% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts_operations.R
\name{sts_get_delegated_access_token}
\alias{sts_get_delegated_access_token}
\title{Exchanges a trade-in token for temporary Amazon Web Services credentials
with the permissions associated with the assumed principal}
\usage{
sts_get_delegated_access_token(TradeInToken)
}
\arguments{
\item{TradeInToken}{[required] The token to exchange for temporary Amazon Web Services credentials.
This token must be valid and unexpired at the time of the request.}
}
\description{
Exchanges a trade-in token for temporary Amazon Web Services credentials with the permissions associated with the assumed principal. This operation allows you to obtain credentials for a specific principal based on a trade-in token, enabling delegation of access to Amazon Web Services resources.

See \url{https://www.paws-r-sdk.com/docs/sts_get_delegated_access_token/} for full documentation.
}
\keyword{internal}
