% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_batch_get_policy}
\alias{verifiedpermissions_batch_get_policy}
\title{Retrieves information about a group (batch) of policies}
\usage{
verifiedpermissions_batch_get_policy(requests)
}
\arguments{
\item{requests}{[required] An array of up to 100 policies you want information about.}
}
\description{
Retrieves information about a group (batch) of policies.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_batch_get_policy/} for full documentation.
}
\keyword{internal}
