% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_delete_policy_store}
\alias{verifiedpermissions_delete_policy_store}
\title{Deletes the specified policy store}
\usage{
verifiedpermissions_delete_policy_store(policyStoreId)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that you want to delete.}
}
\description{
Deletes the specified policy store.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_delete_policy_store/} for full documentation.
}
\keyword{internal}
