% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_delete_policy_store_alias}
\alias{verifiedpermissions_delete_policy_store_alias}
\title{Deletes the specified policy store alias}
\usage{
verifiedpermissions_delete_policy_store_alias(aliasName, deletionMode = NULL)
}
\arguments{
\item{aliasName}{[required] Specifies the name of the policy store alias that you want to delete.

The alias name must always be prefixed with \verb{policy-store-alias/}.}

\item{deletionMode}{Specifies the deletion mode for the policy store alias. The valid values are:
\itemize{
\item \strong{SoftDelete} – The policy store alias enters the \code{PendingDeletion} state. This is the default behavior when no \code{deletionMode} is specified.
\item \strong{HardDelete} – The policy store alias is immediately deleted, bypassing the \code{PendingDeletion} state.
}}
}
\description{
Deletes the specified policy store alias.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_delete_policy_store_alias/} for full documentation.
}
\keyword{internal}
