% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_get_policy_store}
\alias{verifiedpermissions_get_policy_store}
\title{Retrieves details about a policy store}
\usage{
verifiedpermissions_get_policy_store(policyStoreId, tags = NULL)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that you want information about.}

\item{tags}{Specifies whether to return the tags that are attached to the policy
store. If this parameter is included in the API call, the tags are
returned, otherwise they are not returned.

If this parameter is included in the API call but there are no tags
attached to the policy store, the \code{tags} response parameter is omitted
from the response.}
}
\description{
Retrieves details about a policy store.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_get_policy_store/} for full documentation.
}
\keyword{internal}
