% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_get_policy_store_alias}
\alias{verifiedpermissions_get_policy_store_alias}
\title{Retrieves details about the specified policy store alias}
\usage{
verifiedpermissions_get_policy_store_alias(aliasName)
}
\arguments{
\item{aliasName}{[required] Specifies the name of the policy store alias that you want information
about.

The alias name must always be prefixed with \verb{policy-store-alias/}.}
}
\description{
Retrieves details about the specified policy store alias.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_get_policy_store_alias/} for full documentation.
}
\keyword{internal}
