% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_get_policy_template}
\alias{verifiedpermissions_get_policy_template}
\title{Retrieve the details for the specified policy template in the specified
policy store}
\usage{
verifiedpermissions_get_policy_template(policyStoreId, policyTemplateId)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that contains the policy template
that you want information about.

To specify a policy store, use its ID or alias name. When using an alias
name, prefix it with \verb{policy-store-alias/}. For example:
\itemize{
\item ID: \code{PSEXAMPLEabcdefg111111}
\item Alias name: \code{policy-store-alias/example-policy-store}
}

To view aliases, use
\code{\link[=verifiedpermissions_list_policy_store_aliases]{list_policy_store_aliases}}.}

\item{policyTemplateId}{[required] Specifies the ID of the policy template that you want information about.

You can use the policy template name in place of the policy template ID.
When using a name, prefix it with \verb{name/}. For example:
\itemize{
\item ID: \code{PTEXAMPLEabcdefg111111}
\item Name: \code{name/example-policy-template}
}}
}
\description{
Retrieve the details for the specified policy template in the specified policy store.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_get_policy_template/} for full documentation.
}
\keyword{internal}
