% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_get_schema}
\alias{verifiedpermissions_get_schema}
\title{Retrieve the details for the specified schema in the specified policy
store}
\usage{
verifiedpermissions_get_schema(policyStoreId)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that contains the schema.}
}
\description{
Retrieve the details for the specified schema in the specified policy store.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_get_schema/} for full documentation.
}
\keyword{internal}
