% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_is_authorized_with_token}
\alias{verifiedpermissions_is_authorized_with_token}
\title{Makes an authorization decision about a service request described in the
parameters}
\usage{
verifiedpermissions_is_authorized_with_token(
  policyStoreId,
  identityToken = NULL,
  accessToken = NULL,
  action = NULL,
  resource = NULL,
  context = NULL,
  entities = NULL
)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store. Policies in this policy store will
be used to make an authorization decision for the input.}

\item{identityToken}{Specifies an identity token for the principal to be authorized. This
token is provided to you by the identity provider (IdP) associated with
the specified identity source. You must specify either an \code{AccessToken}
or an \code{IdentityToken}, or both.}

\item{accessToken}{Specifies an access token for the principal to be authorized. This token
is provided to you by the identity provider (IdP) associated with the
specified identity source. You must specify either an \code{AccessToken}, or
an \code{IdentityToken}, or both.}

\item{action}{Specifies the requested action to be authorized. Is the specified
principal authorized to perform this action on the specified resource.}

\item{resource}{Specifies the resource for which the authorization decision is made. For
example, is the principal allowed to perform the action on the resource?}

\item{context}{Specifies additional context that can be used to make more granular
authorization decisions.}

\item{entities}{Specifies the list of resources and their associated attributes that
Verified Permissions can examine when evaluating the policies.

You can include only resource and action entities in this parameter; you
can't include principals.
\itemize{
\item The
\code{\link[=verifiedpermissions_is_authorized_with_token]{is_authorized_with_token}}
operation takes principal attributes from \emph{\strong{only}} the
\code{identityToken} or \code{accessToken} passed to the operation.
\item For action entities, you can include only their \code{Identifier} and
\code{EntityType}.
}}
}
\description{
Makes an authorization decision about a service request described in the parameters. The principal in this request comes from an external identity source in the form of an identity token formatted as a \href{https://en.wikipedia.org/wiki/JSON_Web_Token}{JSON web token (JWT)}. The information in the parameters can also define additional context that Verified Permissions can include in the evaluation. The request is evaluated against all matching policies in the specified policy store. The result of the decision is either \code{Allow} or \code{Deny}, along with a list of the policies that resulted in the decision.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_is_authorized_with_token/} for full documentation.
}
\keyword{internal}
