% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_list_tags_for_resource}
\alias{verifiedpermissions_list_tags_for_resource}
\title{Returns the tags associated with the specified Amazon Verified
Permissions resource}
\usage{
verifiedpermissions_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource for which you want to view tags.}
}
\description{
Returns the tags associated with the specified Amazon Verified Permissions resource. In Verified Permissions, policy stores can be tagged.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
