% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_put_schema}
\alias{verifiedpermissions_put_schema}
\title{Creates or updates the policy schema in the specified policy store}
\usage{
verifiedpermissions_put_schema(policyStoreId, definition)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store in which to place the schema.}

\item{definition}{[required] Specifies the definition of the schema to be stored. The schema
definition must be written in Cedar schema JSON.}
}
\description{
Creates or updates the policy schema in the specified policy store. The schema is used to validate any Cedar policies and policy templates submitted to the policy store. Any changes to the schema validate only policies and templates submitted after the schema change. Existing policies and templates are not re-evaluated against the changed schema. If you later update a policy, then it is evaluated against the new schema at that time.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_put_schema/} for full documentation.
}
\keyword{internal}
