% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_tag_resource}
\alias{verifiedpermissions_tag_resource}
\title{Assigns one or more tags (key-value pairs) to the specified Amazon
Verified Permissions resource}
\usage{
verifiedpermissions_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource that you're adding tags to.}

\item{tags}{[required] The list of key-value pairs to associate with the resource.}
}
\description{
Assigns one or more tags (key-value pairs) to the specified Amazon Verified Permissions resource. Tags can help you organize and categorize your resources. You can also use them to scope user permissions by granting a user permission to access or change only resources with certain tag values. In Verified Permissions, policy stores can be tagged.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_tag_resource/} for full documentation.
}
\keyword{internal}
