% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_untag_resource}
\alias{verifiedpermissions_untag_resource}
\title{Removes one or more tags from the specified Amazon Verified Permissions
resource}
\usage{
verifiedpermissions_untag_resource(resourceArn, tagKeys)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource from which you are removing tags.}

\item{tagKeys}{[required] The list of tag keys to remove from the resource.}
}
\description{
Removes one or more tags from the specified Amazon Verified Permissions resource. In Verified Permissions, policy stores can be tagged.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_untag_resource/} for full documentation.
}
\keyword{internal}
