% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_update_policy_store}
\alias{verifiedpermissions_update_policy_store}
\title{Modifies the validation setting for a policy store}
\usage{
verifiedpermissions_update_policy_store(
  policyStoreId,
  validationSettings,
  deletionProtection = NULL,
  description = NULL
)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that you want to update}

\item{validationSettings}{[required] A structure that defines the validation settings that want to enable for
the policy store.}

\item{deletionProtection}{Specifies whether the policy store can be deleted. If enabled, the
policy store can't be deleted.

When you call
\code{\link[=verifiedpermissions_update_policy_store]{update_policy_store}}, this
parameter is unchanged unless explicitly included in the call.}

\item{description}{Descriptive text that you can provide to help with identification of the
current policy store.}
}
\description{
Modifies the validation setting for a policy store.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_update_policy_store/} for full documentation.
}
\keyword{internal}
