% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_update_policy_template}
\alias{verifiedpermissions_update_policy_template}
\title{Updates the specified policy template}
\usage{
verifiedpermissions_update_policy_template(
  policyStoreId,
  policyTemplateId,
  description = NULL,
  statement,
  name = NULL
)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that contains the policy template
that you want to update.

To specify a policy store, use its ID or alias name. When using an alias
name, prefix it with \verb{policy-store-alias/}. For example:
\itemize{
\item ID: \code{PSEXAMPLEabcdefg111111}
\item Alias name: \code{policy-store-alias/example-policy-store}
}

To view aliases, use
\code{\link[=verifiedpermissions_list_policy_store_aliases]{list_policy_store_aliases}}.}

\item{policyTemplateId}{[required] Specifies the ID of the policy template that you want to update.

You can use the policy template name in place of the policy template ID.
When using a name, prefix it with \verb{name/}. For example:
\itemize{
\item ID: \code{PTEXAMPLEabcdefg111111}
\item Name: \code{name/example-policy-template}
}}

\item{description}{Specifies a new description to apply to the policy template.}

\item{statement}{[required] Specifies new statement content written in Cedar policy language to
replace the current body of the policy template.

You can change only the following elements of the policy body:
\itemize{
\item The \code{action} referenced by the policy template.
\item Any conditional clauses, such as \code{when} or \code{unless} clauses.
}

You \strong{can't} change the following elements:
\itemize{
\item The effect (\code{permit} or \code{forbid}) of the policy template.
\item The \code{principal} referenced by the policy template.
\item The \code{resource} referenced by the policy template.
}}

\item{name}{Specifies a name for the policy template that is unique among all policy
templates within the policy store. You can use the name in place of the
policy template ID in API operations that reference the policy template.
The name must be prefixed with \verb{name/}.

If you don't include the name in an update request, the existing name is
unchanged. To remove a name, set it to an empty string (\code{""}).

If you specify a name that is already associated with another policy
template in the policy store, you receive a \code{ConflictException} error.}
}
\description{
Updates the specified policy template. You can update only the description and the some elements of the \href{https://docs.aws.amazon.com/verifiedpermissions/latest/apireference/API_UpdatePolicyTemplate.html#amazonverifiedpermissions-UpdatePolicyTemplate-request-policyBody}{policyBody}.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_update_policy_template/} for full documentation.
}
\keyword{internal}
