% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_create_rule}
\alias{waf_create_rule}
\title{This is AWS WAF Classic documentation}
\usage{
waf_create_rule(Name, MetricName, ChangeToken, Tags = NULL)
}
\arguments{
\item{Name}{[required] A friendly name or description of the Rule. You can't change the name of
a \code{Rule} after you create it.}

\item{MetricName}{[required] A friendly name or description for the metrics for this \code{Rule}. The name
can contain only alphanumeric characters (A-Z, a-z, 0-9), with maximum
length 128 and minimum length one. It can't contain whitespace or metric
names reserved for AWS WAF, including "All" and "Default_Action." You
can't change the name of the metric after you create the \code{Rule}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=waf_get_change_token]{get_change_token}}.}

\item{Tags}{}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_create_rule/} for full documentation.
}
\keyword{internal}
