% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_delete_geo_match_set}
\alias{waf_delete_geo_match_set}
\title{This is AWS WAF Classic documentation}
\usage{
waf_delete_geo_match_set(GeoMatchSetId, ChangeToken)
}
\arguments{
\item{GeoMatchSetId}{[required] The \code{GeoMatchSetID} of the GeoMatchSet that you want to delete.
\code{GeoMatchSetId} is returned by
\code{\link[=waf_create_geo_match_set]{create_geo_match_set}} and by
\code{\link[=waf_list_geo_match_sets]{list_geo_match_sets}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=waf_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_delete_geo_match_set/} for full documentation.
}
\keyword{internal}
