% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_delete_regex_pattern_set}
\alias{waf_delete_regex_pattern_set}
\title{This is AWS WAF CLASSIC documentation}
\usage{
waf_delete_regex_pattern_set(RegexPatternSetId, ChangeToken)
}
\arguments{
\item{RegexPatternSetId}{[required] The \code{RegexPatternSetId} of the RegexPatternSet that you want to delete.
\code{RegexPatternSetId} is returned by CreateRegexPatternSet and by
ListRegexPatternSets.}

\item{ChangeToken}{[required] The value returned by the most recent call to GetChangeToken.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.
}
\details{
\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Permanently deletes a RegexPatternSet. You can\'t delete a
\code{RegexPatternSet} if it\'s still used in any \code{RegexMatchSet} or if the
\code{RegexPatternSet} is not empty.
}
\section{Request syntax}{
\preformatted{svc$delete_regex_pattern_set(
  RegexPatternSetId = "string",
  ChangeToken = "string"
)
}
}

\keyword{internal}
