% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_delete_rule}
\alias{waf_delete_rule}
\title{This is AWS WAF Classic documentation}
\usage{
waf_delete_rule(RuleId, ChangeToken)
}
\arguments{
\item{RuleId}{[required] The \code{RuleId} of the Rule that you want to delete. \code{RuleId} is returned
by \code{\link[=waf_create_rule]{create_rule}} and by
\code{\link[=waf_list_rules]{list_rules}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=waf_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_delete_rule/} for full documentation.
}
\keyword{internal}
