% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_delete_web_acl}
\alias{waf_delete_web_acl}
\title{This is AWS WAF Classic documentation}
\usage{
waf_delete_web_acl(WebACLId, ChangeToken)
}
\arguments{
\item{WebACLId}{[required] The \code{WebACLId} of the WebACL that you want to delete. \code{WebACLId} is
returned by \code{\link[=waf_create_web_acl]{create_web_acl}} and by
\code{\link[=waf_list_web_ac_ls]{list_web_ac_ls}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=waf_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_delete_web_acl/} for full documentation.
}
\keyword{internal}
