% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_get_change_token}
\alias{waf_get_change_token}
\title{This is AWS WAF Classic documentation}
\usage{
waf_get_change_token()
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_get_change_token/} for full documentation.
}
\keyword{internal}
