% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_get_ip_set}
\alias{waf_get_ip_set}
\title{This is AWS WAF Classic documentation}
\usage{
waf_get_ip_set(IPSetId)
}
\arguments{
\item{IPSetId}{[required] The \code{IPSetId} of the IPSet that you want to get. \code{IPSetId} is returned
by \code{\link[=waf_create_ip_set]{create_ip_set}} and by
\code{\link[=waf_list_ip_sets]{list_ip_sets}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_get_ip_set/} for full documentation.
}
\keyword{internal}
