% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_get_xss_match_set}
\alias{waf_get_xss_match_set}
\title{This is AWS WAF Classic documentation}
\usage{
waf_get_xss_match_set(XssMatchSetId)
}
\arguments{
\item{XssMatchSetId}{[required] The \code{XssMatchSetId} of the XssMatchSet that you want to get.
\code{XssMatchSetId} is returned by
\code{\link[=waf_create_xss_match_set]{create_xss_match_set}} and by
\code{\link[=waf_list_xss_match_sets]{list_xss_match_sets}}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  XssMatchSet = list(
    XssMatchSetId = "string",
    Name = "string",
    XssMatchTuples = list(
      list(
        FieldToMatch = list(
          Type = "URI"|"QUERY_STRING"|"HEADER"|"METHOD"|"BODY"|"SINGLE_QUERY_ARG"|"ALL_QUERY_ARGS",
          Data = "string"
        ),
        TextTransformation = "NONE"|"COMPRESS_WHITE_SPACE"|"HTML_ENTITY_DECODE"|"LOWERCASE"|"CMD_LINE"|"URL_DECODE"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Returns the XssMatchSet that is specified by \code{XssMatchSetId}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_xss_match_set(
  XssMatchSetId = "string"
)
}\if{html}{\out{</div>}}
}

\examples{
\dontrun{
# The following example returns the details of an XSS match set with the
# ID example1ds3t-46da-4fdb-b8d5-abc321j569j5.
svc$get_xss_match_set(
  XssMatchSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)
}

}
\keyword{internal}
