% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_list_tags_for_resource}
\alias{waf_list_tags_for_resource}
\title{List tags for resource}
\usage{
waf_list_tags_for_resource(NextMarker, Limit, ResourceARN)
}
\arguments{
\item{NextMarker}{}

\item{Limit}{}

\item{ResourceARN}{[required]}
}
\description{
List tags for resource
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  NextMarker = "string",
  Limit = 123,
  ResourceARN = "string"
)
}
}

\keyword{internal}
