% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_put_permission_policy}
\alias{waf_put_permission_policy}
\title{This is AWS WAF Classic documentation}
\usage{
waf_put_permission_policy(ResourceArn, Policy)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the RuleGroup to which you want to
attach the policy.}

\item{Policy}{[required] The policy to attach to the specified RuleGroup.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_put_permission_policy/} for full documentation.
}
\keyword{internal}
