% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_untag_resource}
\alias{waf_untag_resource}
\title{Untag resource}
\usage{
waf_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required]}

\item{TagKeys}{[required]}
}
\description{
Untag resource
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
