% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_update_rule_group}
\alias{waf_update_rule_group}
\title{This is AWS WAF Classic documentation}
\usage{
waf_update_rule_group(RuleGroupId, Updates, ChangeToken)
}
\arguments{
\item{RuleGroupId}{[required] The \code{RuleGroupId} of the RuleGroup that you want to update.
\code{RuleGroupId} is returned by
\code{\link[=waf_create_rule_group]{create_rule_group}} and by
\code{\link[=waf_list_rule_groups]{list_rule_groups}}.}

\item{Updates}{[required] An array of \code{RuleGroupUpdate} objects that you want to insert into or
delete from a RuleGroup.

You can only insert \code{REGULAR} rules into a rule group.

\code{ActivatedRule|OverrideAction} applies only when updating or adding a
\code{RuleGroup} to a \code{WebACL}. In this case you do not use
\code{ActivatedRule|Action}. For all other update requests,
\code{ActivatedRule|Action} is used instead of
\code{ActivatedRule|OverrideAction}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=waf_get_change_token]{get_change_token}}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ChangeToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Inserts or deletes ActivatedRule objects in a \code{RuleGroup}.

You can only insert \code{REGULAR} rules into a rule group.

You can have a maximum of ten rules per rule group.

To create and configure a \code{RuleGroup}, perform the following steps:
\enumerate{
\item Create and update the \code{Rules} that you want to include in the
\code{RuleGroup}. See \code{\link[=waf_create_rule]{create_rule}}.
\item Use \code{\link[=waf_get_change_token]{get_change_token}} to get the change
token that you provide in the \code{ChangeToken} parameter of an
\code{\link[=waf_update_rule_group]{update_rule_group}} request.
\item Submit an \code{\link[=waf_update_rule_group]{update_rule_group}} request to
add \code{Rules} to the \code{RuleGroup}.
\item Create and update a \code{WebACL} that contains the \code{RuleGroup}. See
\code{\link[=waf_create_web_acl]{create_web_acl}}.
}

If you want to replace one \code{Rule} with another, you delete the existing
one and add the new one.

For more information about how to use the AWS WAF API to allow or block
HTTP requests, see the \href{https://docs.aws.amazon.com/waf/latest/developerguide/}{AWS WAF Developer Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_rule_group(
  RuleGroupId = "string",
  Updates = list(
    list(
      Action = "INSERT"|"DELETE",
      ActivatedRule = list(
        Priority = 123,
        RuleId = "string",
        Action = list(
          Type = "BLOCK"|"ALLOW"|"COUNT"
        ),
        OverrideAction = list(
          Type = "NONE"|"COUNT"
        ),
        Type = "REGULAR"|"RATE_BASED"|"GROUP",
        ExcludedRules = list(
          list(
            RuleId = "string"
          )
        )
      )
    )
  ),
  ChangeToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
