% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_update_sql_injection_match_set}
\alias{waf_update_sql_injection_match_set}
\title{This is AWS WAF Classic documentation}
\usage{
waf_update_sql_injection_match_set(
  SqlInjectionMatchSetId,
  ChangeToken,
  Updates
)
}
\arguments{
\item{SqlInjectionMatchSetId}{[required] The \code{SqlInjectionMatchSetId} of the \code{SqlInjectionMatchSet} that you want
to update. \code{SqlInjectionMatchSetId} is returned by
\code{\link[=waf_create_sql_injection_match_set]{create_sql_injection_match_set}}
and by
\code{\link[=waf_list_sql_injection_match_sets]{list_sql_injection_match_sets}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=waf_get_change_token]{get_change_token}}.}

\item{Updates}{[required] An array of \code{SqlInjectionMatchSetUpdate} objects that you want to insert
into or delete from a SqlInjectionMatchSet. For more information, see
the applicable data types:
\itemize{
\item SqlInjectionMatchSetUpdate: Contains \code{Action} and
\code{SqlInjectionMatchTuple}
\item SqlInjectionMatchTuple: Contains \code{FieldToMatch} and
\code{TextTransformation}
\item FieldToMatch: Contains \code{Data} and \code{Type}
}}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_update_sql_injection_match_set/} for full documentation.
}
\keyword{internal}
