% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_create_regex_match_set}
\alias{wafregional_create_regex_match_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_create_regex_match_set(Name, ChangeToken)
}
\arguments{
\item{Name}{[required] A friendly name or description of the RegexMatchSet. You can't change
\code{Name} after you create a \code{RegexMatchSet}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=wafregional_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_create_regex_match_set/} for full documentation.
}
\keyword{internal}
