% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_create_rule_group}
\alias{wafregional_create_rule_group}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_create_rule_group(Name, MetricName, ChangeToken, Tags = NULL)
}
\arguments{
\item{Name}{[required] A friendly name or description of the RuleGroup. You can't change \code{Name}
after you create a \code{RuleGroup}.}

\item{MetricName}{[required] A friendly name or description for the metrics for this \code{RuleGroup}. The
name can contain only alphanumeric characters (A-Z, a-z, 0-9), with
maximum length 128 and minimum length one. It can't contain whitespace
or metric names reserved for AWS WAF, including "All" and
"Default_Action." You can't change the name of the metric after you
create the \code{RuleGroup}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=wafregional_get_change_token]{get_change_token}}.}

\item{Tags}{}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_create_rule_group/} for full documentation.
}
\keyword{internal}
