% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_delete_ip_set}
\alias{wafregional_delete_ip_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_delete_ip_set(IPSetId, ChangeToken)
}
\arguments{
\item{IPSetId}{[required] The \code{IPSetId} of the IPSet that you want to delete. \code{IPSetId} is
returned by \code{\link[=wafregional_create_ip_set]{create_ip_set}} and by
\code{\link[=wafregional_list_ip_sets]{list_ip_sets}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=wafregional_get_change_token]{get_change_token}}.}
}
\value{
A list with the following syntax:\preformatted{list(
  ChangeToken = "string"
)
}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Permanently deletes an IPSet. You can't delete an \code{IPSet} if it's still
used in any \code{Rules} or if it still includes any IP addresses.

If you just want to remove an \code{IPSet} from a \code{Rule}, use
\code{\link[=wafregional_update_rule]{update_rule}}.

To permanently delete an \code{IPSet} from AWS WAF, perform the following
steps:
\enumerate{
\item Update the \code{IPSet} to remove IP address ranges, if any. For more
information, see \code{\link[=wafregional_update_ip_set]{update_ip_set}}.
\item Use \code{\link[=wafregional_get_change_token]{get_change_token}} to get the
change token that you provide in the \code{ChangeToken} parameter of a
\code{\link[=wafregional_delete_ip_set]{delete_ip_set}} request.
\item Submit a \code{\link[=wafregional_delete_ip_set]{delete_ip_set}} request.
}
}
\section{Request syntax}{
\preformatted{svc$delete_ip_set(
  IPSetId = "string",
  ChangeToken = "string"
)
}
}

\examples{
\dontrun{
# The following example deletes an IP match set  with the ID
# example1ds3t-46da-4fdb-b8d5-abc321j569j5.
svc$delete_ip_set(
  ChangeToken = "abcd12f2-46da-4fdb-b8d5-fbd4c466928f",
  IPSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)
}

}
\keyword{internal}
