% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_delete_xss_match_set}
\alias{wafregional_delete_xss_match_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_delete_xss_match_set(XssMatchSetId, ChangeToken)
}
\arguments{
\item{XssMatchSetId}{[required] The \code{XssMatchSetId} of the XssMatchSet that you want to delete.
\code{XssMatchSetId} is returned by
\code{\link[=wafregional_create_xss_match_set]{create_xss_match_set}} and by
\code{\link[=wafregional_list_xss_match_sets]{list_xss_match_sets}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=wafregional_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_delete_xss_match_set/} for full documentation.
}
\keyword{internal}
