% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_disassociate_web_acl}
\alias{wafregional_disassociate_web_acl}
\title{This is AWS WAF Classic Regional documentation}
\usage{
wafregional_disassociate_web_acl(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN (Amazon Resource Name) of the resource from which the web ACL is
being removed, either an application load balancer or Amazon API Gateway
stage.

The ARN should be in one of the following formats:
\itemize{
\item For an Application Load Balancer:
\code{arn:aws:elasticloadbalancing:region:account-id:loadbalancer/app/load-balancer-name/load-balancer-id }
\item For an Amazon API Gateway stage:
\verb{arn:aws:apigateway:region::/restapis/api-id/stages/stage-name }
}}
}
\description{
This is \strong{AWS WAF Classic Regional} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_disassociate_web_acl/} for full documentation.
}
\keyword{internal}
