% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_ip_set}
\alias{wafregional_get_ip_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_get_ip_set(IPSetId)
}
\arguments{
\item{IPSetId}{[required] The \code{IPSetId} of the IPSet that you want to get. \code{IPSetId} is returned
by CreateIPSet and by ListIPSets.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Returns the IPSet that is specified by \code{IPSetId}.
}
\section{Request syntax}{
\preformatted{svc$get_ip_set(
  IPSetId = "string"
)
}
}

\examples{
\dontrun{
# The following example returns the details of an IP match set with the ID
# example1ds3t-46da-4fdb-b8d5-abc321j569j5.
svc$get_ip_set(
  IPSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)
}

}
\keyword{internal}
