% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_permission_policy}
\alias{wafregional_get_permission_policy}
\title{Returns the IAM policy attached to the RuleGroup}
\usage{
wafregional_get_permission_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the RuleGroup for which you want to
get the policy.}
}
\description{
Returns the IAM policy attached to the RuleGroup.
}
\section{Request syntax}{
\preformatted{svc$get_permission_policy(
  ResourceArn = "string"
)
}
}

\keyword{internal}
