% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_xss_match_set}
\alias{wafregional_get_xss_match_set}
\title{Returns the XssMatchSet that is specified by XssMatchSetId}
\usage{
wafregional_get_xss_match_set(XssMatchSetId)
}
\arguments{
\item{XssMatchSetId}{[required] The \code{XssMatchSetId} of the XssMatchSet that you want to get.
\code{XssMatchSetId} is returned by CreateXssMatchSet and by
ListXssMatchSets.}
}
\description{
Returns the XssMatchSet that is specified by \code{XssMatchSetId}.
}
\section{Request syntax}{
\preformatted{svc$get_xss_match_set(
  XssMatchSetId = "string"
)
}
}

\examples{
\dontrun{
# The following example returns the details of an XSS match set with the
# ID example1ds3t-46da-4fdb-b8d5-abc321j569j5.
svc$get_xss_match_set(
  XssMatchSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)
}

}
\keyword{internal}
