% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_list_rate_based_rules}
\alias{wafregional_list_rate_based_rules}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_list_rate_based_rules(NextMarker, Limit)
}
\arguments{
\item{NextMarker}{If you specify a value for \code{Limit} and you have more \code{Rules} than the
value of \code{Limit}, AWS WAF returns a \code{NextMarker} value in the response
that allows you to list another group of \code{Rules}. For the second and
subsequent \code{\link[=wafregional_list_rate_based_rules]{list_rate_based_rules}}
requests, specify the value of \code{NextMarker} from the previous response
to get information about another batch of \code{Rules}.}

\item{Limit}{Specifies the number of \code{Rules} that you want AWS WAF to return for this
request. If you have more \code{Rules} than the number that you specify for
\code{Limit}, the response includes a \code{NextMarker} value that you can use to
get another batch of \code{Rules}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NextMarker = "string",
  Rules = list(
    list(
      RuleId = "string",
      Name = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Returns an array of RuleSummary objects.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_rate_based_rules(
  NextMarker = "string",
  Limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
