% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_list_subscribed_rule_groups}
\alias{wafregional_list_subscribed_rule_groups}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_list_subscribed_rule_groups(NextMarker, Limit)
}
\arguments{
\item{NextMarker}{If you specify a value for \code{Limit} and you have more
\code{ByteMatchSets}subscribed rule groups than the value of \code{Limit}, AWS WAF
returns a \code{NextMarker} value in the response that allows you to list
another group of subscribed rule groups. For the second and subsequent
\code{ListSubscribedRuleGroupsRequest} requests, specify the value of
\code{NextMarker} from the previous response to get information about another
batch of subscribed rule groups.}

\item{Limit}{Specifies the number of subscribed rule groups that you want AWS WAF to
return for this request. If you have more objects than the number you
specify for \code{Limit}, the response includes a \code{NextMarker} value that you
can use to get another batch of objects.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  NextMarker = "string",
  RuleGroups = list(
    list(
      RuleGroupId = "string",
      Name = "string",
      MetricName = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Returns an array of RuleGroup objects that you are subscribed to.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_subscribed_rule_groups(
  NextMarker = "string",
  Limit = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
