% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_list_xss_match_sets}
\alias{wafregional_list_xss_match_sets}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_list_xss_match_sets(NextMarker, Limit)
}
\arguments{
\item{NextMarker}{If you specify a value for \code{Limit} and you have more XssMatchSet objects
than the value of \code{Limit}, AWS WAF returns a \code{NextMarker} value in the
response that allows you to list another group of \code{XssMatchSets}. For
the second and subsequent
\code{\link[=wafregional_list_xss_match_sets]{list_xss_match_sets}} requests,
specify the value of \code{NextMarker} from the previous response to get
information about another batch of \code{XssMatchSets}.}

\item{Limit}{Specifies the number of XssMatchSet objects that you want AWS WAF to
return for this request. If you have more \code{XssMatchSet} objects than the
number you specify for \code{Limit}, the response includes a \code{NextMarker}
value that you can use to get another batch of \code{Rules}.}
}
\value{
A list with the following syntax:\preformatted{list(
  NextMarker = "string",
  XssMatchSets = list(
    list(
      XssMatchSetId = "string",
      Name = "string"
    )
  )
)
}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Returns an array of XssMatchSet objects.
}
\section{Request syntax}{
\preformatted{svc$list_xss_match_sets(
  NextMarker = "string",
  Limit = 123
)
}
}

\examples{
\dontrun{
# The following example returns an array of up to 100 XSS match sets.
svc$list_xss_match_sets(
  Limit = 100L
)
}

}
\keyword{internal}
