% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_update_geo_match_set}
\alias{wafregional_update_geo_match_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_update_geo_match_set(GeoMatchSetId, ChangeToken, Updates)
}
\arguments{
\item{GeoMatchSetId}{[required] The \code{GeoMatchSetId} of the GeoMatchSet that you want to update.
\code{GeoMatchSetId} is returned by
\code{\link[=wafregional_create_geo_match_set]{create_geo_match_set}} and by
\code{\link[=wafregional_list_geo_match_sets]{list_geo_match_sets}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=wafregional_get_change_token]{get_change_token}}.}

\item{Updates}{[required] An array of \code{GeoMatchSetUpdate} objects that you want to insert into or
delete from an GeoMatchSet. For more information, see the applicable
data types:
\itemize{
\item GeoMatchSetUpdate: Contains \code{Action} and \code{GeoMatchConstraint}
\item GeoMatchConstraint: Contains \code{Type} and \code{Value}

You can have only one \code{Type} and \code{Value} per \code{GeoMatchConstraint}.
To add multiple countries, include multiple \code{GeoMatchSetUpdate}
objects in your request.
}}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_update_geo_match_set/} for full documentation.
}
\keyword{internal}
