% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_update_regex_match_set}
\alias{wafregional_update_regex_match_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_update_regex_match_set(RegexMatchSetId, Updates, ChangeToken)
}
\arguments{
\item{RegexMatchSetId}{[required] The \code{RegexMatchSetId} of the RegexMatchSet that you want to update.
\code{RegexMatchSetId} is returned by
\code{\link[=wafregional_create_regex_match_set]{create_regex_match_set}} and by
\code{\link[=wafregional_list_regex_match_sets]{list_regex_match_sets}}.}

\item{Updates}{[required] An array of \code{RegexMatchSetUpdate} objects that you want to insert into
or delete from a RegexMatchSet. For more information, see
RegexMatchTuple.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=wafregional_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_update_regex_match_set/} for full documentation.
}
\keyword{internal}
