% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_update_rule}
\alias{wafregional_update_rule}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_update_rule(RuleId, ChangeToken, Updates)
}
\arguments{
\item{RuleId}{[required] The \code{RuleId} of the \code{Rule} that you want to update. \code{RuleId} is returned
by \code{\link[=wafregional_create_rule]{create_rule}} and by
\code{\link[=wafregional_list_rules]{list_rules}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=wafregional_get_change_token]{get_change_token}}.}

\item{Updates}{[required] An array of \code{RuleUpdate} objects that you want to insert into or delete
from a Rule. For more information, see the applicable data types:
\itemize{
\item RuleUpdate: Contains \code{Action} and \code{Predicate}
\item Predicate: Contains \code{DataId}, \code{Negated}, and \code{Type}
\item FieldToMatch: Contains \code{Data} and \code{Type}
}}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_update_rule/} for full documentation.
}
\keyword{internal}
