% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_delete_api_key}
\alias{wafv2_delete_api_key}
\title{Deletes the specified API key}
\usage{
wafv2_delete_api_key(Scope, APIKey)
}
\arguments{
\item{Scope}{[required] Specifies whether this is for a global resource type, such as a Amazon
CloudFront distribution. For an Amplify application, use \code{CLOUDFRONT}.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}

\item{APIKey}{[required] The encrypted API key that you want to delete.}
}
\description{
Deletes the specified API key.

See \url{https://www.paws-r-sdk.com/docs/wafv2_delete_api_key/} for full documentation.
}
\keyword{internal}
