% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_delete_permission_policy}
\alias{wafv2_delete_permission_policy}
\title{Permanently deletes an IAM policy from the specified rule group}
\usage{
wafv2_delete_permission_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the rule group from which you want to
delete the policy.

You must be the owner of the rule group to perform this operation.}
}
\description{
Permanently deletes an IAM policy from the specified rule group.

See \url{https://www.paws-r-sdk.com/docs/wafv2_delete_permission_policy/} for full documentation.
}
\keyword{internal}
