% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_describe_managed_products_by_vendor}
\alias{wafv2_describe_managed_products_by_vendor}
\title{Provides high-level information for the managed rule groups owned by a
specific vendor}
\usage{
wafv2_describe_managed_products_by_vendor(VendorName, Scope)
}
\arguments{
\item{VendorName}{[required] The name of the managed rule group vendor. You use this, along with the
rule group name, to identify a rule group.}

\item{Scope}{[required] Specifies whether this is for an Amazon CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB), an Amazon API Gateway REST API, an AppSync GraphQL API,
an Amazon Cognito user pool, an App Runner service, or an Amazon Web
Services Verified Access instance.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}
}
\description{
Provides high-level information for the managed rule groups owned by a specific vendor.

See \url{https://www.paws-r-sdk.com/docs/wafv2_describe_managed_products_by_vendor/} for full documentation.
}
\keyword{internal}
