% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_generate_mobile_sdk_release_url}
\alias{wafv2_generate_mobile_sdk_release_url}
\title{Generates a presigned download URL for the specified release of the
mobile SDK}
\usage{
wafv2_generate_mobile_sdk_release_url(Platform, ReleaseVersion)
}
\arguments{
\item{Platform}{[required] The device platform.}

\item{ReleaseVersion}{[required] The release version. For the latest available version, specify \code{LATEST}.}
}
\description{
Generates a presigned download URL for the specified release of the mobile SDK.

See \url{https://www.paws-r-sdk.com/docs/wafv2_generate_mobile_sdk_release_url/} for full documentation.
}
\keyword{internal}
