% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_get_logging_configuration}
\alias{wafv2_get_logging_configuration}
\title{Returns the LoggingConfiguration for the specified web ACL}
\usage{
wafv2_get_logging_configuration(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the web ACL for which you want to get
the LoggingConfiguration.}
}
\description{
Returns the LoggingConfiguration for the specified web ACL.

See \url{https://www.paws-r-sdk.com/docs/wafv2_get_logging_configuration/} for full documentation.
}
\keyword{internal}
