% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_get_managed_rule_set}
\alias{wafv2_get_managed_rule_set}
\title{Retrieves the specified managed rule set}
\usage{
wafv2_get_managed_rule_set(Name, Scope, Id)
}
\arguments{
\item{Name}{[required] The name of the managed rule set. You use this, along with the rule set
ID, to identify the rule set.

This name is assigned to the corresponding managed rule group, which
your customers can access and use.}

\item{Scope}{[required] Specifies whether this is for an Amazon CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB), an Amazon API Gateway REST API, an AppSync GraphQL API,
an Amazon Cognito user pool, an App Runner service, or an Amazon Web
Services Verified Access instance.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}

\item{Id}{[required] A unique identifier for the managed rule set. The ID is returned in the
responses to commands like \code{list}. You provide it to operations like
\code{get} and \code{update}.}
}
\description{
Retrieves the specified managed rule set.

See \url{https://www.paws-r-sdk.com/docs/wafv2_get_managed_rule_set/} for full documentation.
}
\keyword{internal}
