% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_get_mobile_sdk_release}
\alias{wafv2_get_mobile_sdk_release}
\title{Retrieves information for the specified mobile SDK release, including
release notes and tags}
\usage{
wafv2_get_mobile_sdk_release(Platform, ReleaseVersion)
}
\arguments{
\item{Platform}{[required] The device platform.}

\item{ReleaseVersion}{[required] The release version. For the latest available version, specify \code{LATEST}.}
}
\description{
Retrieves information for the specified mobile SDK release, including release notes and tags.

See \url{https://www.paws-r-sdk.com/docs/wafv2_get_mobile_sdk_release/} for full documentation.
}
\keyword{internal}
