% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_get_rule_group}
\alias{wafv2_get_rule_group}
\title{Retrieves the specified RuleGroup}
\usage{
wafv2_get_rule_group(Name = NULL, Scope = NULL, Id = NULL, ARN = NULL)
}
\arguments{
\item{Name}{The name of the rule group. You cannot change the name of a rule group
after you create it.}

\item{Scope}{Specifies whether this is for an Amazon CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB), an Amazon API Gateway REST API, an AppSync GraphQL API,
an Amazon Cognito user pool, an App Runner service, or an Amazon Web
Services Verified Access instance.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}

\item{Id}{A unique identifier for the rule group. This ID is returned in the
responses to create and list commands. You provide it to operations like
update and delete.}

\item{ARN}{The Amazon Resource Name (ARN) of the entity.}
}
\description{
Retrieves the specified RuleGroup.

See \url{https://www.paws-r-sdk.com/docs/wafv2_get_rule_group/} for full documentation.
}
\keyword{internal}
