% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_get_web_acl}
\alias{wafv2_get_web_acl}
\title{Retrieves the specified WebACL}
\usage{
wafv2_get_web_acl(Name = NULL, Scope = NULL, Id = NULL, ARN = NULL)
}
\arguments{
\item{Name}{The name of the web ACL. You cannot change the name of a web ACL after
you create it.}

\item{Scope}{Specifies whether this is for a global resource type, such as a Amazon
CloudFront distribution. For an Amplify application, use \code{CLOUDFRONT}.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}

\item{Id}{The unique identifier for the web ACL. This ID is returned in the
responses to create and list commands. You provide it to operations like
update and delete.}

\item{ARN}{The Amazon Resource Name (ARN) of the web ACL that you want to retrieve.}
}
\description{
Retrieves the specified WebACL.

See \url{https://www.paws-r-sdk.com/docs/wafv2_get_web_acl/} for full documentation.
}
\keyword{internal}
