% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_list_available_managed_rule_group_versions}
\alias{wafv2_list_available_managed_rule_group_versions}
\title{Returns a list of the available versions for the specified managed rule
group}
\usage{
wafv2_list_available_managed_rule_group_versions(
  VendorName,
  Name,
  Scope,
  NextMarker = NULL,
  Limit = NULL
)
}
\arguments{
\item{VendorName}{[required] The name of the managed rule group vendor. You use this, along with the
rule group name, to identify a rule group.}

\item{Name}{[required] The name of the managed rule group. You use this, along with the vendor
name, to identify the rule group.}

\item{Scope}{[required] Specifies whether this is for an Amazon CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB), an Amazon API Gateway REST API, an AppSync GraphQL API,
an Amazon Cognito user pool, an App Runner service, or an Amazon Web
Services Verified Access instance.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}

\item{NextMarker}{When you request a list of objects with a \code{Limit} setting, if the number
of objects that are still available for retrieval exceeds the limit, WAF
returns a \code{NextMarker} value in the response. To retrieve the next batch
of objects, provide the marker from the prior call in your next request.}

\item{Limit}{The maximum number of objects that you want WAF to return for this
request. If more objects are available, in the response, WAF provides a
\code{NextMarker} value that you can use in a subsequent call to get the next
batch of objects.}
}
\description{
Returns a list of the available versions for the specified managed rule group.

See \url{https://www.paws-r-sdk.com/docs/wafv2_list_available_managed_rule_group_versions/} for full documentation.
}
\keyword{internal}
