% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_put_logging_configuration}
\alias{wafv2_put_logging_configuration}
\title{Enables the specified LoggingConfiguration, to start logging from a web
ACL, according to the configuration provided}
\usage{
wafv2_put_logging_configuration(LoggingConfiguration)
}
\arguments{
\item{LoggingConfiguration}{[required]}
}
\description{
Enables the specified LoggingConfiguration, to start logging from a web ACL, according to the configuration provided.

See \url{https://www.paws-r-sdk.com/docs/wafv2_put_logging_configuration/} for full documentation.
}
\keyword{internal}
