% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_tag_resource}
\alias{wafv2_tag_resource}
\title{Associates tags with the specified Amazon Web Services resource}
\usage{
wafv2_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{Tags}{[required] An array of key:value pairs to associate with the resource.}
}
\description{
Associates tags with the specified Amazon Web Services resource. Tags are key:value pairs that you can use to categorize and manage your resources, for purposes like billing. For example, you might set the tag key to "customer" and the value to the customer name or ID. You can specify one or more tags to add to each Amazon Web Services resource, up to 50 tags for a resource.

See \url{https://www.paws-r-sdk.com/docs/wafv2_tag_resource/} for full documentation.
}
\keyword{internal}
