% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_untag_resource}
\alias{wafv2_untag_resource}
\title{Disassociates tags from an Amazon Web Services resource}
\usage{
wafv2_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{TagKeys}{[required] An array of keys identifying the tags to disassociate from the resource.}
}
\description{
Disassociates tags from an Amazon Web Services resource. Tags are key:value pairs that you can associate with Amazon Web Services resources. For example, the tag key might be "customer" and the tag value might be "companyA." You can specify one or more tags to add to each container. You can add up to 50 tags to each Amazon Web Services resource.

See \url{https://www.paws-r-sdk.com/docs/wafv2_untag_resource/} for full documentation.
}
\keyword{internal}
