% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_create_analyzer}
\alias{accessanalyzer_create_analyzer}
\title{Creates an analyzer for your account}
\usage{
accessanalyzer_create_analyzer(
  analyzerName,
  type,
  archiveRules = NULL,
  tags = NULL,
  clientToken = NULL,
  configuration = NULL
)
}
\arguments{
\item{analyzerName}{[required] The name of the analyzer to create.}

\item{type}{[required] The type of analyzer to create. You can create only one analyzer per
account per Region. You can create up to 5 analyzers per organization
per Region.}

\item{archiveRules}{Specifies the archive rules to add for the analyzer. Archive rules
automatically archive findings that meet the criteria you define for the
rule.}

\item{tags}{An array of key-value pairs to apply to the analyzer. You can use the
set of Unicode letters, digits, whitespace, \verb{_}, \code{.}, \code{/}, \code{=}, \code{+}, and
\code{-}.

For the tag key, you can specify a value that is 1 to 128 characters in
length and cannot be prefixed with \verb{aws:}.

For the tag value, you can specify a value that is 0 to 256 characters
in length.}

\item{clientToken}{A client token.}

\item{configuration}{Specifies the configuration of the analyzer. If the analyzer is an
unused access analyzer, the specified scope of unused access is used for
the configuration. If the analyzer is an internal access analyzer, the
specified internal access analysis rules are used for the configuration.}
}
\description{
Creates an analyzer for your account.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_create_analyzer/} for full documentation.
}
\keyword{internal}
