% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_list_access_preview_findings}
\alias{accessanalyzer_list_access_preview_findings}
\title{Retrieves a list of access preview findings generated by the specified
access preview}
\usage{
accessanalyzer_list_access_preview_findings(
  accessPreviewId,
  analyzerArn,
  filter = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{accessPreviewId}{[required] The unique ID for the access preview.}

\item{analyzerArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access-analyzer-getting-started.html#permission-resources}{ARN of the analyzer}
used to generate the access.}

\item{filter}{Criteria to filter the returned findings.}

\item{nextToken}{A token used for pagination of results returned.}

\item{maxResults}{The maximum number of results to return in the response.}
}
\description{
Retrieves a list of access preview findings generated by the specified access preview.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_list_access_preview_findings/} for full documentation.
}
\keyword{internal}
