% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_update_archive_rule}
\alias{accessanalyzer_update_archive_rule}
\title{Updates the criteria and values for the specified archive rule}
\usage{
accessanalyzer_update_archive_rule(
  analyzerName,
  ruleName,
  filter,
  clientToken = NULL
)
}
\arguments{
\item{analyzerName}{[required] The name of the analyzer to update the archive rules for.}

\item{ruleName}{[required] The name of the rule to update.}

\item{filter}{[required] A filter to match for the rules to update. Only rules that match the
filter are updated.}

\item{clientToken}{A client token.}
}
\description{
Updates the criteria and values for the specified archive rule.

See \url{https://www.paws-r-sdk.com/docs/accessanalyzer_update_archive_rule/} for full documentation.
}
\keyword{internal}
