% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_get_account_configuration}
\alias{acm_get_account_configuration}
\title{Returns the account configuration options associated with an Amazon Web
Services account}
\usage{
acm_get_account_configuration()
}
\description{
Returns the account configuration options associated with an Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/acm_get_account_configuration/} for full documentation.
}
\keyword{internal}
