% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_renew_certificate}
\alias{acm_renew_certificate}
\title{Renews an eligible ACM certificate}
\usage{
acm_renew_certificate(CertificateArn)
}
\arguments{
\item{CertificateArn}{[required] String that contains the ARN of the ACM certificate to be renewed. This
must be of the form:

\code{arn:aws:acm:region:123456789012:certificate/12345678-1234-1234-1234-123456789012}

For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}.}
}
\description{
Renews an \href{https://docs.aws.amazon.com/acm/latest/userguide/managed-renewal.html}{eligible ACM certificate}. In order to renew your Amazon Web Services Private CA certificates with ACM, you must first \href{https://docs.aws.amazon.com/privateca/latest/userguide/}{grant the ACM service principal permission to do so}. For more information, see \href{https://docs.aws.amazon.com/acm/latest/userguide/}{Testing Managed Renewal} in the ACM User Guide.

See \url{https://www.paws-r-sdk.com/docs/acm_renew_certificate/} for full documentation.
}
\keyword{internal}
