% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm_operations.R
\name{acm_update_certificate_options}
\alias{acm_update_certificate_options}
\title{Updates a certificate}
\usage{
acm_update_certificate_options(CertificateArn, Options)
}
\arguments{
\item{CertificateArn}{[required] ARN of the requested certificate to update. This must be of the form:

\code{arn:aws:acm:us-east-1:account:certificate/12345678-1234-1234-1234-123456789012 }}

\item{Options}{[required] Use to update the options for your certificate. Currently, you can
specify whether to add your certificate to a transparency log or export
your certificate. Certificate transparency makes it possible to detect
SSL/TLS certificates that have been mistakenly or maliciously issued.
Certificates that have not been logged typically produce an error
message in a browser.}
}
\description{
Updates a certificate. You can use this function to specify whether to opt in to or out of recording your certificate in a certificate transparency log and exporting. For more information, see \href{https://docs.aws.amazon.com/acm/latest/userguide/acm-bestpractices.html#best-practices-transparency}{Opting Out of Certificate Transparency Logging} and \href{https://docs.aws.amazon.com/acm/latest/userguide/acm-exportable-certificates.html}{Certificate Manager Exportable Managed Certificates}.

See \url{https://www.paws-r-sdk.com/docs/acm_update_certificate_options/} for full documentation.
}
\keyword{internal}
