% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acmpca_operations.R
\name{acmpca_describe_certificate_authority}
\alias{acmpca_describe_certificate_authority}
\title{Lists information about your private certificate authority (CA) or one
that has been shared with you}
\usage{
acmpca_describe_certificate_authority(CertificateAuthorityArn)
}
\arguments{
\item{CertificateAuthorityArn}{[required] The Amazon Resource Name (ARN) that was returned when you called
\code{\link[=acmpca_create_certificate_authority]{create_certificate_authority}}.
This must be of the form:

\code{arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012 }.}
}
\description{
Lists information about your private certificate authority (CA) or one that has been shared with you. You specify the private CA on input by its ARN (Amazon Resource Name). The output contains the status of your CA. This can be any of the following:

See \url{https://www.paws-r-sdk.com/docs/acmpca_describe_certificate_authority/} for full documentation.
}
\keyword{internal}
