% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acmpca_operations.R
\name{acmpca_get_policy}
\alias{acmpca_get_policy}
\title{Retrieves the resource-based policy attached to a private CA}
\usage{
acmpca_get_policy(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Number (ARN) of the private CA that will have its
policy retrieved. You can find the CA's ARN by calling the
ListCertificateAuthorities action.}
}
\description{
Retrieves the resource-based policy attached to a private CA. If either the private CA resource or the policy cannot be found, this action returns a \code{ResourceNotFoundException}.

See \url{https://www.paws-r-sdk.com/docs/acmpca_get_policy/} for full documentation.
}
\keyword{internal}
