% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acmpca_operations.R
\name{acmpca_list_certificate_authorities}
\alias{acmpca_list_certificate_authorities}
\title{Lists the private certificate authorities that you created by using the
CreateCertificateAuthority action}
\usage{
acmpca_list_certificate_authorities(
  NextToken = NULL,
  MaxResults = NULL,
  ResourceOwner = NULL
)
}
\arguments{
\item{NextToken}{Use this parameter when paginating results in a subsequent request after
you receive a response with truncated results. Set it to the value of
the \code{NextToken} parameter from the response you just received.}

\item{MaxResults}{Use this parameter when paginating results to specify the maximum number
of items to return in the response on each page. If additional items
exist beyond the number you specify, the \code{NextToken} element is sent in
the response. Use this \code{NextToken} value in a subsequent request to
retrieve additional items.

Although the maximum value is 1000, the action only returns a maximum of
100 items.}

\item{ResourceOwner}{Use this parameter to filter the returned set of certificate authorities
based on their owner. The default is SELF.}
}
\description{
Lists the private certificate authorities that you created by using the \code{\link[=acmpca_create_certificate_authority]{create_certificate_authority}} action.

See \url{https://www.paws-r-sdk.com/docs/acmpca_list_certificate_authorities/} for full documentation.
}
\keyword{internal}
