% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acmpca_operations.R
\name{acmpca_list_permissions}
\alias{acmpca_list_permissions}
\title{List all permissions on a private CA, if any, granted to the AWS
Certificate Manager (ACM) service principal (acm}
\usage{
acmpca_list_permissions(CertificateAuthorityArn, NextToken, MaxResults)
}
\arguments{
\item{CertificateAuthorityArn}{[required] The Amazon Resource Number (ARN) of the private CA to inspect. You can
find the ARN by calling the
\code{\link[=acmpca_list_certificate_authorities]{list_certificate_authorities}}
action. This must be of the form:
\code{arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012}
You can get a private CA's ARN by running the
\code{\link[=acmpca_list_certificate_authorities]{list_certificate_authorities}}
action.}

\item{NextToken}{When paginating results, use this parameter in a subsequent request
after you receive a response with truncated results. Set it to the value
of \strong{NextToken} from the response you just received.}

\item{MaxResults}{When paginating results, use this parameter to specify the maximum
number of items to return in the response. If additional items exist
beyond the number you specify, the \strong{NextToken} element is sent in the
response. Use this \strong{NextToken} value in a subsequent request to
retrieve additional items.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Permissions = list(
    list(
      CertificateAuthorityArn = "string",
      CreatedAt = as.POSIXct(
        "2015-01-01"
      ),
      Principal = "string",
      SourceAccount = "string",
      Actions = list(
        "IssueCertificate"|"GetCertificate"|"ListPermissions"
      ),
      Policy = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
List all permissions on a private CA, if any, granted to the AWS
Certificate Manager (ACM) service principal (acm.amazonaws.com).

These permissions allow ACM to issue and renew ACM certificates that
reside in the same AWS account as the CA.

Permissions can be granted with the
\code{\link[=acmpca_create_permission]{create_permission}} action and revoked with
the \code{\link[=acmpca_delete_permission]{delete_permission}} action.

\strong{About Permissions}
\itemize{
\item If the private CA and the certificates it issues reside in the same
account, you can use \code{\link[=acmpca_create_permission]{create_permission}}
to grant permissions for ACM to carry out automatic certificate
renewals.
\item For automatic certificate renewal to succeed, the ACM service
principal needs permissions to create, retrieve, and list
certificates.
\item If the private CA and the ACM certificates reside in different
accounts, then permissions cannot be used to enable automatic
renewals. Instead, the ACM certificate owner must set up a
resource-based policy to enable cross-account issuance and renewals.
For more information, see \href{https://docs.aws.amazon.com/acm-pca/latest/userguide/pca-rbp.html}{Using a Resource Based Policy with ACM Private CA}.
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_permissions(
  CertificateAuthorityArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
