% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_attach_object}
\alias{clouddirectory_attach_object}
\title{Attaches an existing object to another object}
\usage{
clouddirectory_attach_object(
  DirectoryArn,
  ParentReference,
  ChildReference,
  LinkName
)
}
\arguments{
\item{DirectoryArn}{[required] Amazon Resource Name (ARN) that is associated with the Directory where
both objects reside. For more information, see arns.}

\item{ParentReference}{[required] The parent object reference.}

\item{ChildReference}{[required] The child object reference to be attached to the object.}

\item{LinkName}{[required] The link name with which the child object is attached to the parent.}
}
\description{
Attaches an existing object to another object. An object can be accessed in two ways:

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_attach_object/} for full documentation.
}
\keyword{internal}
