% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_create_directory}
\alias{clouddirectory_create_directory}
\title{Creates a Directory by copying the published schema into the directory}
\usage{
clouddirectory_create_directory(Name, SchemaArn)
}
\arguments{
\item{Name}{[required] The name of the Directory. Should be unique per account, per region.}

\item{SchemaArn}{[required] The Amazon Resource Name (ARN) of the published schema that will be
copied into the data Directory. For more information, see arns.}
}
\description{
Creates a Directory by copying the published schema into the directory. A directory cannot be created without a schema.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_create_directory/} for full documentation.
}
\keyword{internal}
