% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_create_typed_link_facet}
\alias{clouddirectory_create_typed_link_facet}
\title{Creates a TypedLinkFacet}
\usage{
clouddirectory_create_typed_link_facet(SchemaArn, Facet)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) that is associated with the schema. For
more information, see arns.}

\item{Facet}{[required] Facet structure that is associated with the typed link facet.}
}
\description{
Creates a TypedLinkFacet. For more information, see \href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/directory_objects_links.html#directory_objects_links_typedlink}{Typed Links}.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_create_typed_link_facet/} for full documentation.
}
\keyword{internal}
