% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_delete_directory}
\alias{clouddirectory_delete_directory}
\title{Deletes a directory}
\usage{
clouddirectory_delete_directory(DirectoryArn)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory to delete.}
}
\description{
Deletes a directory. Only disabled directories can be deleted. A deleted directory cannot be undone. Exercise extreme caution when deleting directories.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_delete_directory/} for full documentation.
}
\keyword{internal}
