% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_delete_facet}
\alias{clouddirectory_delete_facet}
\title{Deletes a given Facet}
\usage{
clouddirectory_delete_facet(SchemaArn, Name)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) that is associated with the Facet. For
more information, see arns.}

\item{Name}{[required] The name of the facet to delete.}
}
\description{
Deletes a given Facet. All attributes and Rules that are associated with the facet will be deleted. Only development schema facets are allowed deletion.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_delete_facet/} for full documentation.
}
\keyword{internal}
