% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_delete_schema}
\alias{clouddirectory_delete_schema}
\title{Deletes a given schema}
\usage{
clouddirectory_delete_schema(SchemaArn)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) of the development schema. For more
information, see arns.}
}
\description{
Deletes a given schema. Schemas in a development and published state can only be deleted.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_delete_schema/} for full documentation.
}
\keyword{internal}
