% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_disable_directory}
\alias{clouddirectory_disable_directory}
\title{Disables the specified directory}
\usage{
clouddirectory_disable_directory(DirectoryArn)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory to disable.}
}
\description{
Disables the specified directory. Disabled directories cannot be read or written to. Only enabled directories can be disabled. Disabled directories may be reenabled.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_disable_directory/} for full documentation.
}
\keyword{internal}
