% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_enable_directory}
\alias{clouddirectory_enable_directory}
\title{Enables the specified directory}
\usage{
clouddirectory_enable_directory(DirectoryArn)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory to enable.}
}
\value{
A list with the following syntax:\preformatted{list(
  DirectoryArn = "string"
)
}
}
\description{
Enables the specified directory. Only disabled directories can be
enabled. Once enabled, the directory can then be read and written to.
}
\section{Request syntax}{
\preformatted{svc$enable_directory(
  DirectoryArn = "string"
)
}
}

\keyword{internal}
