% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_get_applied_schema_version}
\alias{clouddirectory_get_applied_schema_version}
\title{Returns current applied schema version ARN, including the minor version
in use}
\usage{
clouddirectory_get_applied_schema_version(SchemaArn)
}
\arguments{
\item{SchemaArn}{[required] The ARN of the applied schema.}
}
\description{
Returns current applied schema version ARN, including the minor version in use.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_get_applied_schema_version/} for full documentation.
}
\keyword{internal}
