% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_get_directory}
\alias{clouddirectory_get_directory}
\title{Retrieves metadata about a directory}
\usage{
clouddirectory_get_directory(DirectoryArn)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory.}
}
\value{
A list with the following syntax:\preformatted{list(
  Directory = list(
    Name = "string",
    DirectoryArn = "string",
    State = "ENABLED"|"DISABLED"|"DELETED",
    CreationDateTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
Retrieves metadata about a directory.
}
\section{Request syntax}{
\preformatted{svc$get_directory(
  DirectoryArn = "string"
)
}
}

\keyword{internal}
