% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_get_directory}
\alias{clouddirectory_get_directory}
\title{Retrieves metadata about a directory}
\usage{
clouddirectory_get_directory(DirectoryArn)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory.}
}
\description{
Retrieves metadata about a directory.

See \url{https://paws-r.github.io/docs/clouddirectory/get_directory.html} for full documentation.
}
\keyword{internal}
