% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_get_object_attributes}
\alias{clouddirectory_get_object_attributes}
\title{Retrieves attributes within a facet that are associated with an object}
\usage{
clouddirectory_get_object_attributes(
  DirectoryArn,
  ObjectReference,
  ConsistencyLevel = NULL,
  SchemaFacet,
  AttributeNames
)
}
\arguments{
\item{DirectoryArn}{[required] The Amazon Resource Name (ARN) that is associated with the Directory
where the object resides.}

\item{ObjectReference}{[required] Reference that identifies the object whose attributes will be retrieved.}

\item{ConsistencyLevel}{The consistency level at which to retrieve the attributes on an object.}

\item{SchemaFacet}{[required] Identifier for the facet whose attributes will be retrieved. See
SchemaFacet for details.}

\item{AttributeNames}{[required] List of attribute names whose values will be retrieved.}
}
\description{
Retrieves attributes within a facet that are associated with an object.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_get_object_attributes/} for full documentation.
}
\keyword{internal}
