% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_get_object_information}
\alias{clouddirectory_get_object_information}
\title{Retrieves metadata about an object}
\usage{
clouddirectory_get_object_information(
  DirectoryArn,
  ObjectReference,
  ConsistencyLevel = NULL
)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory being retrieved.}

\item{ObjectReference}{[required] A reference to the object.}

\item{ConsistencyLevel}{The consistency level at which to retrieve the object information.}
}
\description{
Retrieves metadata about an object.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_get_object_information/} for full documentation.
}
\keyword{internal}
