% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_get_schema_as_json}
\alias{clouddirectory_get_schema_as_json}
\title{Retrieves a JSON representation of the schema}
\usage{
clouddirectory_get_schema_as_json(SchemaArn)
}
\arguments{
\item{SchemaArn}{[required] The ARN of the schema to retrieve.}
}
\description{
Retrieves a JSON representation of the schema. See \href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/schemas_jsonformat.html#schemas_json}{JSON Schema Format} for more information.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_get_schema_as_json/} for full documentation.
}
\keyword{internal}
