% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_get_typed_link_facet_information}
\alias{clouddirectory_get_typed_link_facet_information}
\title{Returns the identity attribute order for a specific TypedLinkFacet}
\usage{
clouddirectory_get_typed_link_facet_information(SchemaArn, Name)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) that is associated with the schema. For
more information, see arns.}

\item{Name}{[required] The unique name of the typed link facet.}
}
\description{
Returns the identity attribute order for a specific TypedLinkFacet. For more information, see \href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/directory_objects_links.html#directory_objects_links_typedlink}{Typed Links}.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_get_typed_link_facet_information/} for full documentation.
}
\keyword{internal}
