% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_list_applied_schema_arns}
\alias{clouddirectory_list_applied_schema_arns}
\title{Lists schema major versions applied to a directory}
\usage{
clouddirectory_list_applied_schema_arns(
  DirectoryArn,
  SchemaArn = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{DirectoryArn}{[required] The ARN of the directory you are listing.}

\item{SchemaArn}{The response for
\code{\link[=clouddirectory_list_applied_schema_arns]{list_applied_schema_arns}}
when this parameter is used will list all minor version ARNs for a major
version.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\description{
Lists schema major versions applied to a directory. If \code{SchemaArn} is provided, lists the minor version.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_list_applied_schema_arns/} for full documentation.
}
\keyword{internal}
