% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_list_development_schema_arns}
\alias{clouddirectory_list_development_schema_arns}
\title{Retrieves each Amazon Resource Name (ARN) of schemas in the development
state}
\usage{
clouddirectory_list_development_schema_arns(
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\description{
Retrieves each Amazon Resource Name (ARN) of schemas in the development state.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_list_development_schema_arns/} for full documentation.
}
\keyword{internal}
