% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_list_facet_attributes}
\alias{clouddirectory_list_facet_attributes}
\title{Retrieves attributes attached to the facet}
\usage{
clouddirectory_list_facet_attributes(
  SchemaArn,
  Name,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{SchemaArn}{[required] The ARN of the schema where the facet resides.}

\item{Name}{[required] The name of the facet whose attributes will be retrieved.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\description{
Retrieves attributes attached to the facet.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_list_facet_attributes/} for full documentation.
}
\keyword{internal}
