% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_list_facet_names}
\alias{clouddirectory_list_facet_names}
\title{Retrieves the names of facets that exist in a schema}
\usage{
clouddirectory_list_facet_names(SchemaArn, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) to retrieve facet names from.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\description{
Retrieves the names of facets that exist in a schema.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_list_facet_names/} for full documentation.
}
\keyword{internal}
