% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_list_typed_link_facet_attributes}
\alias{clouddirectory_list_typed_link_facet_attributes}
\title{Returns a paginated list of all attribute definitions for a particular
TypedLinkFacet}
\usage{
clouddirectory_list_typed_link_facet_attributes(
  SchemaArn,
  Name,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) that is associated with the schema. For
more information, see arns.}

\item{Name}{[required] The unique name of the typed link facet.}

\item{NextToken}{The pagination token.}

\item{MaxResults}{The maximum number of results to retrieve.}
}
\description{
Returns a paginated list of all attribute definitions for a particular TypedLinkFacet. For more information, see \href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/directory_objects_links.html#directory_objects_links_typedlink}{Typed Links}.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_list_typed_link_facet_attributes/} for full documentation.
}
\keyword{internal}
