% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_update_facet}
\alias{clouddirectory_update_facet}
\title{Does the following:}
\usage{
clouddirectory_update_facet(
  SchemaArn,
  Name,
  AttributeUpdates = NULL,
  ObjectType = NULL
)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) that is associated with the Facet. For
more information, see arns.}

\item{Name}{[required] The name of the facet.}

\item{AttributeUpdates}{List of attributes that need to be updated in a given schema Facet. Each
attribute is followed by \code{AttributeAction}, which specifies the type of
update operation to perform.}

\item{ObjectType}{The object type that is associated with the facet. See
CreateFacetRequest$ObjectType for more details.}
}
\description{
Does the following:

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_update_facet/} for full documentation.
}
\keyword{internal}
