% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddirectory_operations.R
\name{clouddirectory_update_typed_link_facet}
\alias{clouddirectory_update_typed_link_facet}
\title{Updates a TypedLinkFacet}
\usage{
clouddirectory_update_typed_link_facet(
  SchemaArn,
  Name,
  AttributeUpdates,
  IdentityAttributeOrder
)
}
\arguments{
\item{SchemaArn}{[required] The Amazon Resource Name (ARN) that is associated with the schema. For
more information, see arns.}

\item{Name}{[required] The unique name of the typed link facet.}

\item{AttributeUpdates}{[required] Attributes update structure.}

\item{IdentityAttributeOrder}{[required] The order of identity attributes for the facet, from most significant to
least significant. The ability to filter typed links considers the order
that the attributes are defined on the typed link facet. When providing
ranges to a typed link selection, any inexact ranges must be specified
at the end. Any attributes that do not have a range specified are
presumed to match the entire range. Filters are interpreted in the order
of the attributes on the typed link facet, not the order in which they
are supplied to any API calls. For more information about identity
attributes, see \href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/directory_objects_links.html#directory_objects_links_typedlink}{Typed Links}.}
}
\description{
Updates a TypedLinkFacet. For more information, see \href{https://docs.aws.amazon.com/clouddirectory/latest/developerguide/directory_objects_links.html#directory_objects_links_typedlink}{Typed Links}.

See \url{https://www.paws-r-sdk.com/docs/clouddirectory_update_typed_link_facet/} for full documentation.
}
\keyword{internal}
