% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_create_hapg}
\alias{cloudhsm_create_hapg}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_create_hapg(Label)
}
\arguments{
\item{Label}{[required] The label of the new high-availability partition group.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

See \url{https://www.paws-r-sdk.com/docs/cloudhsm_create_hapg/} for full documentation.
}
\keyword{internal}
