% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_delete_hsm}
\alias{cloudhsm_delete_hsm}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_delete_hsm(HsmArn)
}
\arguments{
\item{HsmArn}{[required] The ARN of the HSM to delete.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more
information, see \href{http://aws.amazon.com/cloudhsm/faqs-classic/}{AWS CloudHSM Classic FAQs}, the \href{http://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide}, and the
\href{http://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.
}
\details{
\strong{For information about the current version of AWS CloudHSM}, see \href{http://aws.amazon.com/cloudhsm/}{AWS CloudHSM}, the \href{http://docs.aws.amazon.com/cloudhsm/latest/userguide/}{AWS CloudHSM User Guide}, and the
\href{http://docs.aws.amazon.com/cloudhsm/latest/APIReference/}{AWS CloudHSM API Reference}.

Deletes an HSM. After completion, this operation cannot be undone and
your key material cannot be recovered.
}
\section{Request syntax}{
\preformatted{svc$delete_hsm(
  HsmArn = "string"
)
}
}

\keyword{internal}
