% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_describe_luna_client}
\alias{cloudhsm_describe_luna_client}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_describe_luna_client(ClientArn = NULL, CertificateFingerprint = NULL)
}
\arguments{
\item{ClientArn}{The ARN of the client.}

\item{CertificateFingerprint}{The certificate fingerprint.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

See \url{https://www.paws-r-sdk.com/docs/cloudhsm_describe_luna_client/} for full documentation.
}
\keyword{internal}
