% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_get_config}
\alias{cloudhsm_get_config}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_get_config(ClientArn, ClientVersion, HapgList)
}
\arguments{
\item{ClientArn}{[required] The ARN of the client.}

\item{ClientVersion}{[required] The client version.}

\item{HapgList}{[required] A list of ARNs that identify the high-availability partition groups that
are associated with the client.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

See \url{https://www.paws-r-sdk.com/docs/cloudhsm_get_config/} for full documentation.
}
\keyword{internal}
