% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_modify_hapg}
\alias{cloudhsm_modify_hapg}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_modify_hapg(HapgArn, Label = NULL, PartitionSerialList = NULL)
}
\arguments{
\item{HapgArn}{[required] The ARN of the high-availability partition group to modify.}

\item{Label}{The new label for the high-availability partition group.}

\item{PartitionSerialList}{The list of partition serial numbers to make members of the
high-availability partition group.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the AWS CloudHSM Classic User Guide, and the AWS CloudHSM Classic API Reference.

See \url{https://www.paws-r-sdk.com/docs/cloudhsm_modify_hapg/} for full documentation.
}
\keyword{internal}
