% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_copy_backup_to_region}
\alias{cloudhsmv2_copy_backup_to_region}
\title{Copy an AWS CloudHSM cluster backup to a different region}
\usage{
cloudhsmv2_copy_backup_to_region(DestinationRegion, BackupId, TagList)
}
\arguments{
\item{DestinationRegion}{[required] The AWS region that will contain your copied CloudHSM cluster backup.}

\item{BackupId}{[required] The ID of the backup that will be copied to the destination region.}

\item{TagList}{Tags to apply to the destination backup during creation. If you specify
tags, only these tags will be applied to the destination backup. If you
do not specify tags, the service copies tags from the source backup to
the destination backup.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DestinationBackup = list(
    CreateTimestamp = as.POSIXct(
      "2015-01-01"
    ),
    SourceRegion = "string",
    SourceBackup = "string",
    SourceCluster = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Copy an AWS CloudHSM cluster backup to a different region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$copy_backup_to_region(
  DestinationRegion = "string",
  BackupId = "string",
  TagList = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
