% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_describe_clusters}
\alias{cloudhsmv2_describe_clusters}
\title{Gets information about AWS CloudHSM clusters}
\usage{
cloudhsmv2_describe_clusters(Filters, NextToken, MaxResults)
}
\arguments{
\item{Filters}{One or more filters to limit the items returned in the response.

Use the \code{clusterIds} filter to return only the specified clusters.
Specify clusters by their cluster identifier (ID).

Use the \code{vpcIds} filter to return only the clusters in the specified
virtual private clouds (VPCs). Specify VPCs by their VPC identifier
(ID).

Use the \code{states} filter to return only clusters that match the specified
state.}

\item{NextToken}{The \code{NextToken} value that you received in the previous response. Use
this value to get more clusters.}

\item{MaxResults}{The maximum number of clusters to return in the response. When there are
more clusters than the number you specify, the response contains a
\code{NextToken} value.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Clusters = list(
    list(
      BackupPolicy = "DEFAULT",
      BackupRetentionPolicy = list(
        Type = "DAYS",
        Value = "string"
      ),
      ClusterId = "string",
      CreateTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      Hsms = list(
        list(
          AvailabilityZone = "string",
          ClusterId = "string",
          SubnetId = "string",
          EniId = "string",
          EniIp = "string",
          HsmId = "string",
          State = "CREATE_IN_PROGRESS"|"ACTIVE"|"DEGRADED"|"DELETE_IN_PROGRESS"|"DELETED",
          StateMessage = "string"
        )
      ),
      HsmType = "string",
      PreCoPassword = "string",
      SecurityGroup = "string",
      SourceBackupId = "string",
      State = "CREATE_IN_PROGRESS"|"UNINITIALIZED"|"INITIALIZE_IN_PROGRESS"|"INITIALIZED"|"ACTIVE"|"UPDATE_IN_PROGRESS"|"DELETE_IN_PROGRESS"|"DELETED"|"DEGRADED",
      StateMessage = "string",
      SubnetMapping = list(
        "string"
      ),
      VpcId = "string",
      Certificates = list(
        ClusterCsr = "string",
        HsmCertificate = "string",
        AwsHardwareCertificate = "string",
        ManufacturerHardwareCertificate = "string",
        ClusterCertificate = "string"
      ),
      TagList = list(
        list(
          Key = "string",
          Value = "string"
        )
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about AWS CloudHSM clusters.

This is a paginated operation, which means that each response might
contain only a subset of all the clusters. When the response contains
only a subset of clusters, it includes a \code{NextToken} value. Use this
value in a subsequent
\code{\link[=cloudhsmv2_describe_clusters]{describe_clusters}} request to get more
clusters. When you receive a response with no \code{NextToken} (or an empty
or null value), that means there are no more clusters to get.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_clusters(
  Filters = list(
    list(
      "string"
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
