% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsmv2_operations.R
\name{cloudhsmv2_untag_resource}
\alias{cloudhsmv2_untag_resource}
\title{Removes the specified tag or tags from the specified AWS CloudHSM
cluster}
\usage{
cloudhsmv2_untag_resource(ResourceId, TagKeyList)
}
\arguments{
\item{ResourceId}{[required] The cluster identifier (ID) for the cluster whose tags you are removing.
To find the cluster ID, use
\code{\link[=cloudhsmv2_describe_clusters]{describe_clusters}}.}

\item{TagKeyList}{[required] A list of one or more tag keys for the tags that you are removing.
Specify only the tag keys, not the tag values.}
}
\description{
Removes the specified tag or tags from the specified AWS CloudHSM cluster.

See \url{https://www.paws-r-sdk.com/docs/cloudhsmv2_untag_resource/} for full documentation.
}
\keyword{internal}
