% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_delete_identities}
\alias{cognitoidentity_delete_identities}
\title{Deletes identities from an identity pool}
\usage{
cognitoidentity_delete_identities(IdentityIdsToDelete)
}
\arguments{
\item{IdentityIdsToDelete}{[required] A list of 1-60 identities that you want to delete.}
}
\description{
Deletes identities from an identity pool. You can specify a list of 1-60 identities that you want to delete.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentity_delete_identities/} for full documentation.
}
\keyword{internal}
