% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_delete_identity_pool}
\alias{cognitoidentity_delete_identity_pool}
\title{Deletes an identity pool}
\usage{
cognitoidentity_delete_identity_pool(IdentityPoolId)
}
\arguments{
\item{IdentityPoolId}{[required] An identity pool ID in the format REGION:GUID.}
}
\description{
Deletes an identity pool. Once a pool is deleted, users will not be able to authenticate with the pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentity_delete_identity_pool/} for full documentation.
}
\keyword{internal}
