% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_get_identity_pool_roles}
\alias{cognitoidentity_get_identity_pool_roles}
\title{Gets the roles for an identity pool}
\usage{
cognitoidentity_get_identity_pool_roles(IdentityPoolId)
}
\arguments{
\item{IdentityPoolId}{[required] An identity pool ID in the format REGION:GUID.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  IdentityPoolId = "string",
  Roles = list(
    "string"
  ),
  RoleMappings = list(
    list(
      Type = "Token"|"Rules",
      AmbiguousRoleResolution = "AuthenticatedRole"|"Deny",
      RulesConfiguration = list(
        Rules = list(
          list(
            Claim = "string",
            MatchType = "Equals"|"Contains"|"StartsWith"|"NotEqual",
            Value = "string",
            RoleARN = "string"
          )
        )
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the roles for an identity pool.

You must use AWS Developer credentials to call this API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_identity_pool_roles(
  IdentityPoolId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
