% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_get_principal_tag_attribute_map}
\alias{cognitoidentity_get_principal_tag_attribute_map}
\title{Use GetPrincipalTagAttributeMap to list all mappings between
PrincipalTags and user attributes}
\usage{
cognitoidentity_get_principal_tag_attribute_map(
  IdentityPoolId,
  IdentityProviderName
)
}
\arguments{
\item{IdentityPoolId}{[required] You can use this operation to get the ID of the Identity Pool you setup
attribute mappings for.}

\item{IdentityProviderName}{[required] You can use this operation to get the provider name.}
}
\description{
Use \code{\link[=cognitoidentity_get_principal_tag_attribute_map]{get_principal_tag_attribute_map}} to list all mappings between \code{PrincipalTags} and user attributes.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentity_get_principal_tag_attribute_map/} for full documentation.
}
\keyword{internal}
