% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_list_identities}
\alias{cognitoidentity_list_identities}
\title{Lists the identities in an identity pool}
\usage{
cognitoidentity_list_identities(
  IdentityPoolId,
  MaxResults,
  NextToken = NULL,
  HideDisabled = NULL
)
}
\arguments{
\item{IdentityPoolId}{[required] An identity pool ID in the format REGION:GUID.}

\item{MaxResults}{[required] The maximum number of identities to return.}

\item{NextToken}{A pagination token.}

\item{HideDisabled}{An optional boolean parameter that allows you to hide disabled
identities. If omitted, the ListIdentities API will include disabled
identities in the response.}
}
\description{
Lists the identities in an identity pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentity_list_identities/} for full documentation.
}
\keyword{internal}
