% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_list_identity_pools}
\alias{cognitoidentity_list_identity_pools}
\title{Lists all of the Cognito identity pools registered for your account}
\usage{
cognitoidentity_list_identity_pools(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{[required] The maximum number of identities to return.}

\item{NextToken}{A pagination token.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  IdentityPools = list(
    list(
      IdentityPoolId = "string",
      IdentityPoolName = "string"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all of the Cognito identity pools registered for your account.

You must use AWS Developer credentials to call this API.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_identity_pools(
  MaxResults = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
