% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_list_identity_pools}
\alias{cognitoidentity_list_identity_pools}
\title{Lists all of the Cognito identity pools registered for your account}
\usage{
cognitoidentity_list_identity_pools(MaxResults, NextToken = NULL)
}
\arguments{
\item{MaxResults}{[required] The maximum number of identities to return.}

\item{NextToken}{A pagination token.}
}
\description{
Lists all of the Cognito identity pools registered for your account.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentity_list_identity_pools/} for full documentation.
}
\keyword{internal}
