% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_list_tags_for_resource}
\alias{cognitoidentity_list_tags_for_resource}
\title{Lists the tags that are assigned to an Amazon Cognito identity pool}
\usage{
cognitoidentity_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the identity pool that the tags are
assigned to.}
}
\description{
Lists the tags that are assigned to an Amazon Cognito identity pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentity_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
