% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_tag_resource}
\alias{cognitoidentity_tag_resource}
\title{Assigns a set of tags to the specified Amazon Cognito identity pool}
\usage{
cognitoidentity_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the identity pool.}

\item{Tags}{[required] The tags to assign to the identity pool.}
}
\description{
Assigns a set of tags to the specified Amazon Cognito identity pool. A tag is a label that you can use to categorize and manage identity pools in different ways, such as by purpose, owner, environment, or other criteria.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentity_tag_resource/} for full documentation.
}
\keyword{internal}
