% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentity_operations.R
\name{cognitoidentity_untag_resource}
\alias{cognitoidentity_untag_resource}
\title{Removes the specified tags from the specified Amazon Cognito identity
pool}
\usage{
cognitoidentity_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the identity pool.}

\item{TagKeys}{[required] The keys of the tags to remove from the user pool.}
}
\description{
Removes the specified tags from the specified Amazon Cognito identity pool. You can use this action up to 5 times per second, per account

See \url{https://www.paws-r-sdk.com/docs/cognitoidentity_untag_resource/} for full documentation.
}
\keyword{internal}
