% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_delete_user}
\alias{cognitoidentityprovider_admin_delete_user}
\title{Deletes a user as an administrator}
\usage{
cognitoidentityprovider_admin_delete_user(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool where you want to delete the user.}

\item{Username}{[required] The user name of the user you want to delete.}
}
\description{
Deletes a user as an administrator. Works on any user.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/admin_delete_user.html} for full documentation.
}
\keyword{internal}
