% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_delete_user}
\alias{cognitoidentityprovider_admin_delete_user}
\title{Deletes a user profile in your user pool}
\usage{
cognitoidentityprovider_admin_delete_user(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to delete the user.}

\item{Username}{[required] The name of the user that you want to query or modify. The value of this
parameter is typically your user's username, but it can be any of their
alias attributes. If \code{username} isn't an alias attribute in your user
pool, this value must be the \code{sub} of a local user or the username of a
user from a third-party IdP.}
}
\description{
Deletes a user profile in your user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_delete_user/} for full documentation.
}
\keyword{internal}
