% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_disable_user}
\alias{cognitoidentityprovider_admin_disable_user}
\title{Disables the specified user}
\usage{
cognitoidentityprovider_admin_disable_user(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool where you want to disable the user.}

\item{Username}{[required] The user name of the user you want to disable.}
}
\description{
Disables the specified user.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/admin_disable_user.html} for full documentation.
}
\keyword{internal}
