% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_disable_user}
\alias{cognitoidentityprovider_admin_disable_user}
\title{Deactivates a user and revokes all access tokens for the user}
\usage{
cognitoidentityprovider_admin_disable_user(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool where you want to disable the user.}

\item{Username}{[required] The user name of the user you want to disable.}
}
\description{
Deactivates a user and revokes all access tokens for the user. A deactivated user can't sign in, but still appears in the responses to \code{\link[=cognitoidentityprovider_get_user]{get_user}} and \code{\link[=cognitoidentityprovider_list_users]{list_users}} API requests.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_disable_user/} for full documentation.
}
\keyword{internal}
