% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_disable_user}
\alias{cognitoidentityprovider_admin_disable_user}
\title{Disables the specified user}
\usage{
cognitoidentityprovider_admin_disable_user(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool where you want to disable the user.}

\item{Username}{[required] The user name of the user you wish to disable.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Disables the specified user.

Calling this action requires developer credentials.
}
\section{Request syntax}{
\preformatted{svc$admin_disable_user(
  UserPoolId = "string",
  Username = "string"
)
}
}

\keyword{internal}
