% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_disable_user}
\alias{cognitoidentityprovider_admin_disable_user}
\title{Deactivates a user and revokes all access tokens for the user}
\usage{
cognitoidentityprovider_admin_disable_user(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool where you want to disable the user.}

\item{Username}{[required] The username of the user that you want to query or modify. The value of
this parameter is typically your user's username, but it can be any of
their alias attributes. If \code{username} isn't an alias attribute in your
user pool, this value must be the \code{sub} of a local user or the username
of a user from a third-party IdP.}
}
\description{
Deactivates a user and revokes all access tokens for the user. A deactivated user can't sign in, but still appears in the responses to \code{\link[=cognitoidentityprovider_get_user]{get_user}} and \code{\link[=cognitoidentityprovider_list_users]{list_users}} API requests.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_disable_user/} for full documentation.
}
\keyword{internal}
