% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_get_device}
\alias{cognitoidentityprovider_admin_get_device}
\title{Gets the device, as an administrator}
\usage{
cognitoidentityprovider_admin_get_device(DeviceKey, UserPoolId, Username)
}
\arguments{
\item{DeviceKey}{[required] The device key.}

\item{UserPoolId}{[required] The user pool ID.}

\item{Username}{[required] The username of the user that you want to query or modify. The value of
this parameter is typically your user's username, but it can be any of
their alias attributes. If \code{username} isn't an alias attribute in your
user pool, this value must be the \code{sub} of a local user or the username
of a user from a third-party IdP.}
}
\description{
Gets the device, as an administrator.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_get_device/} for full documentation.
}
\keyword{internal}
