% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_get_user}
\alias{cognitoidentityprovider_admin_get_user}
\title{Given a username, returns details about a user profile in a user pool}
\usage{
cognitoidentityprovider_admin_get_user(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to get information about the
user.}

\item{Username}{[required] The name of the user that you want to query or modify. The value of this
parameter is typically your user's username, but it can be any of their
alias attributes. If \code{username} isn't an alias attribute in your user
pool, this value must be the \code{sub} of a local user or the username of a
user from a third-party IdP.}
}
\description{
Given a username, returns details about a user profile in a user pool. You can specify alias attributes in the \code{Username} request parameter.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_get_user/} for full documentation.
}
\keyword{internal}
