% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_list_devices}
\alias{cognitoidentityprovider_admin_list_devices}
\title{Lists devices, as an administrator}
\usage{
cognitoidentityprovider_admin_list_devices(
  UserPoolId,
  Username,
  Limit = NULL,
  PaginationToken = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{Username}{[required] The user name.}

\item{Limit}{The limit of the devices request.}

\item{PaginationToken}{The pagination token.}
}
\description{
Lists devices, as an administrator.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/admin_list_devices.html} for full documentation.
}
\keyword{internal}
