% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_list_devices}
\alias{cognitoidentityprovider_admin_list_devices}
\title{Lists devices, as an administrator}
\usage{
cognitoidentityprovider_admin_list_devices(
  UserPoolId,
  Username,
  Limit = NULL,
  PaginationToken = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{Username}{[required] The username of the user that you want to query or modify. The value of
this parameter is typically your user's username, but it can be any of
their alias attributes. If \code{username} isn't an alias attribute in your
user pool, this value must be the \code{sub} of a local user or the username
of a user from a third-party IdP.}

\item{Limit}{The limit of the devices request.}

\item{PaginationToken}{This API operation returns a limited number of results. The pagination
token is an identifier that you can present in an additional API request
with the same parameters. When you include the pagination token, Amazon
Cognito returns the next set of items after the current list. Subsequent
requests return a new pagination token. By use of this token, you can
paginate through the full list of items.}
}
\description{
Lists devices, as an administrator.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_list_devices/} for full documentation.
}
\keyword{internal}
