% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_list_groups_for_user}
\alias{cognitoidentityprovider_admin_list_groups_for_user}
\title{Lists the groups that a user belongs to}
\usage{
cognitoidentityprovider_admin_list_groups_for_user(
  Username,
  UserPoolId,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Username}{[required] The username of the user that you want to query or modify. The value of
this parameter is typically your user's username, but it can be any of
their alias attributes. If \code{username} isn't an alias attribute in your
user pool, this value must be the \code{sub} of a local user or the username
of a user from a third-party IdP.}

\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{Limit}{The limit of the request to list groups.}

\item{NextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Lists the groups that a user belongs to.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_list_groups_for_user/} for full documentation.
}
\keyword{internal}
