% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_remove_user_from_group}
\alias{cognitoidentityprovider_admin_remove_user_from_group}
\title{Removes the specified user from the specified group}
\usage{
cognitoidentityprovider_admin_remove_user_from_group(UserPoolId,
  Username, GroupName)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{Username}{[required] The username for the user.}

\item{GroupName}{[required] The group name.}
}
\value{
An empty list.
}
\description{
Removes the specified user from the specified group.

Calling this action requires developer credentials.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$admin_remove_user_from_group(
  UserPoolId = "string",
  Username = "string",
  GroupName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
