% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_remove_user_from_group}
\alias{cognitoidentityprovider_admin_remove_user_from_group}
\title{Removes the specified user from the specified group}
\usage{
cognitoidentityprovider_admin_remove_user_from_group(
  UserPoolId,
  Username,
  GroupName
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{Username}{[required] The username of the user that you want to query or modify. The value of
this parameter is typically your user's username, but it can be any of
their alias attributes. If \code{username} isn't an alias attribute in your
user pool, this value must be the \code{sub} of a local user or the username
of a user from a third-party IdP.}

\item{GroupName}{[required] The group name.}
}
\description{
Removes the specified user from the specified group.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_remove_user_from_group/} for full documentation.
}
\keyword{internal}
