% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_update_auth_event_feedback}
\alias{cognitoidentityprovider_admin_update_auth_event_feedback}
\title{Provides the feedback for an authentication event generated by threat
protection features}
\usage{
cognitoidentityprovider_admin_update_auth_event_feedback(
  UserPoolId,
  Username,
  EventId,
  FeedbackValue
)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to submit authentication-event
feedback.}

\item{Username}{[required] The name of the user that you want to query or modify. The value of this
parameter is typically your user's username, but it can be any of their
alias attributes. If \code{username} isn't an alias attribute in your user
pool, this value must be the \code{sub} of a local user or the username of a
user from a third-party IdP.}

\item{EventId}{[required] The ID of the threat protection authentication event that you want to
update.}

\item{FeedbackValue}{[required] Your feedback to the authentication event. When you provide a
\code{FeedbackValue} value of \code{valid}, you tell Amazon Cognito that you trust
a user session where Amazon Cognito has evaluated some level of risk.
When you provide a \code{FeedbackValue} value of \code{invalid}, you tell Amazon
Cognito that you don't trust a user session, or you don't believe that
Amazon Cognito evaluated a high-enough risk level.}
}
\description{
Provides the feedback for an authentication event generated by threat protection features. Your response indicates that you think that the event either was from a valid user or was an unwanted authentication attempt. This feedback improves the risk evaluation decision for the user pool as part of Amazon Cognito threat protection. To activate this setting, your user pool must be on the \href{https://docs.aws.amazon.com/cognito/latest/developerguide/feature-plans-features-plus.html}{Plus tier}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_update_auth_event_feedback/} for full documentation.
}
\keyword{internal}
