% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_update_device_status}
\alias{cognitoidentityprovider_admin_update_device_status}
\title{Updates the device status as an administrator}
\usage{
cognitoidentityprovider_admin_update_device_status(
  UserPoolId,
  Username,
  DeviceKey,
  DeviceRememberedStatus = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{Username}{[required] The user name.}

\item{DeviceKey}{[required] The device key.}

\item{DeviceRememberedStatus}{The status indicating whether a device has been remembered or not.}
}
\description{
Updates the device status as an administrator.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_admin_update_device_status/} for full documentation.
}
\keyword{internal}
