% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_admin_user_global_sign_out}
\alias{cognitoidentityprovider_admin_user_global_sign_out}
\title{Signs out users from all devices, as an administrator}
\usage{
cognitoidentityprovider_admin_user_global_sign_out(UserPoolId, Username)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{Username}{[required] The user name.}
}
\value{
An empty list.
}
\description{
Signs out users from all devices, as an administrator. It also
invalidates all refresh tokens issued to a user. The user's current
access and Id tokens remain valid until their expiry. Access and Id
tokens expire one hour after they are issued.

Calling this action requires developer credentials.
}
\section{Request syntax}{
\preformatted{svc$admin_user_global_sign_out(
  UserPoolId = "string",
  Username = "string"
)
}
}

\keyword{internal}
