% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_associate_software_token}
\alias{cognitoidentityprovider_associate_software_token}
\title{Begins setup of time-based one-time password (TOTP) multi-factor
authentication (MFA) for a user, with a unique private key that Amazon
Cognito generates and returns in the API response}
\usage{
cognitoidentityprovider_associate_software_token(
  AccessToken = NULL,
  Session = NULL
)
}
\arguments{
\item{AccessToken}{A valid access token that Amazon Cognito issued to the user whose
software token you want to generate.}

\item{Session}{The session that should be passed both ways in challenge-response calls
to the service. This allows authentication of the user as part of the
MFA setup process.}
}
\description{
Begins setup of time-based one-time password (TOTP) multi-factor authentication (MFA) for a user, with a unique private key that Amazon Cognito generates and returns in the API response. You can authorize an \code{\link[=cognitoidentityprovider_associate_software_token]{associate_software_token}} request with either the user's access token, or a session string from a challenge response that you received from Amazon Cognito.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_associate_software_token/} for full documentation.
}
\keyword{internal}
