% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_change_password}
\alias{cognitoidentityprovider_change_password}
\title{Changes the password for a specified user in a user pool}
\usage{
cognitoidentityprovider_change_password(PreviousPassword,
  ProposedPassword, AccessToken)
}
\arguments{
\item{PreviousPassword}{[required] The old password.}

\item{ProposedPassword}{[required] The new password.}

\item{AccessToken}{[required] The access token.}
}
\value{
An empty list.
}
\description{
Changes the password for a specified user in a user pool.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$change_password(
  PreviousPassword = "string",
  ProposedPassword = "string",
  AccessToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
