% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_change_password}
\alias{cognitoidentityprovider_change_password}
\title{Changes the password for a specified user in a user pool}
\usage{
cognitoidentityprovider_change_password(
  PreviousPassword,
  ProposedPassword,
  AccessToken
)
}
\arguments{
\item{PreviousPassword}{[required] The old password.}

\item{ProposedPassword}{[required] The new password.}

\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user whose
password you want to change.}
}
\description{
Changes the password for a specified user in a user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_change_password/} for full documentation.
}
\keyword{internal}
