% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_confirm_device}
\alias{cognitoidentityprovider_confirm_device}
\title{Confirms tracking of the device}
\usage{
cognitoidentityprovider_confirm_device(
  AccessToken,
  DeviceKey,
  DeviceSecretVerifierConfig = NULL,
  DeviceName = NULL
)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user whose device
you want to confirm.}

\item{DeviceKey}{[required] The device key.}

\item{DeviceSecretVerifierConfig}{The configuration of the device secret verifier.}

\item{DeviceName}{The device name.}
}
\description{
Confirms tracking of the device. This API call is the call that begins device tracking. For more information about device authentication, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/amazon-cognito-user-pools-device-tracking.html}{Working with user devices in your user pool}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_confirm_device/} for full documentation.
}
\keyword{internal}
