% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_create_user_import_job}
\alias{cognitoidentityprovider_create_user_import_job}
\title{Creates a user import job}
\usage{
cognitoidentityprovider_create_user_import_job(
  JobName,
  UserPoolId,
  CloudWatchLogsRoleArn
)
}
\arguments{
\item{JobName}{[required] The job name for the user import job.}

\item{UserPoolId}{[required] The user pool ID for the user pool that the users are being imported
into.}

\item{CloudWatchLogsRoleArn}{[required] The role ARN for the Amazon CloudWatch Logs Logging role for the user
import job.}
}
\description{
Creates a user import job.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_create_user_import_job/} for full documentation.
}
\keyword{internal}
