% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_identity_provider}
\alias{cognitoidentityprovider_delete_identity_provider}
\title{Deletes an IdP for a user pool}
\usage{
cognitoidentityprovider_delete_identity_provider(UserPoolId, ProviderName)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{ProviderName}{[required] The IdP name.}
}
\description{
Deletes an IdP for a user pool.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/delete_identity_provider.html} for full documentation.
}
\keyword{internal}
