% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_user}
\alias{cognitoidentityprovider_delete_user}
\title{Self-deletes a user profile}
\usage{
cognitoidentityprovider_delete_user(AccessToken)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user whose user
profile you want to delete.}
}
\description{
Self-deletes a user profile. A deleted user profile can no longer be used to sign in and can't be restored.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_delete_user/} for full documentation.
}
\keyword{internal}
