% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_user}
\alias{cognitoidentityprovider_delete_user}
\title{Allows a user to delete their own user profile}
\usage{
cognitoidentityprovider_delete_user(AccessToken)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user whose user
profile you want to delete.}
}
\description{
Allows a user to delete their own user profile.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_delete_user/} for full documentation.
}
\keyword{internal}
