% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_user_attributes}
\alias{cognitoidentityprovider_delete_user_attributes}
\title{Deletes the attributes for a user}
\usage{
cognitoidentityprovider_delete_user_attributes(UserAttributeNames, AccessToken)
}
\arguments{
\item{UserAttributeNames}{[required] An array of strings representing the user attribute names you want to
delete.

For custom attributes, you must prependattach the \verb{custom:} prefix to
the front of the attribute name.}

\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user whose
attributes you want to delete.}
}
\description{
Deletes the attributes for a user.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_delete_user_attributes/} for full documentation.
}
\keyword{internal}
