% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_user_pool}
\alias{cognitoidentityprovider_delete_user_pool}
\title{Deletes the specified Amazon Cognito user pool}
\usage{
cognitoidentityprovider_delete_user_pool(UserPoolId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool you want to delete.}
}
\value{
An empty list.
}
\description{
Deletes the specified Amazon Cognito user pool.
}
\section{Request syntax}{
\preformatted{svc$delete_user_pool(
  UserPoolId = "string"
)
}
}

\keyword{internal}
