% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_user_pool}
\alias{cognitoidentityprovider_delete_user_pool}
\title{Deletes the specified Amazon Cognito user pool}
\usage{
cognitoidentityprovider_delete_user_pool(UserPoolId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool you want to delete.}
}
\description{
Deletes the specified Amazon Cognito user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_delete_user_pool/} for full documentation.
}
\keyword{internal}
