% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_user_pool_domain}
\alias{cognitoidentityprovider_delete_user_pool_domain}
\title{Deletes a domain for a user pool}
\usage{
cognitoidentityprovider_delete_user_pool_domain(Domain, UserPoolId)
}
\arguments{
\item{Domain}{[required] The domain string. For custom domains, this is the fully-qualified
domain name, such as \code{auth.example.com}. For Amazon Cognito prefix
domains, this is the prefix alone, such as \code{auth}.}

\item{UserPoolId}{[required] The user pool ID.}
}
\description{
Deletes a domain for a user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_delete_user_pool_domain/} for full documentation.
}
\keyword{internal}
