% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_delete_web_authn_credential}
\alias{cognitoidentityprovider_delete_web_authn_credential}
\title{Deletes a registered passkey, or WebAuthn, authenticator for the
currently signed-in user}
\usage{
cognitoidentityprovider_delete_web_authn_credential(AccessToken, CredentialId)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the currently
signed-in user. Must include a scope claim for
\code{aws.cognito.signin.user.admin}.}

\item{CredentialId}{[required] The unique identifier of the passkey that you want to delete.}
}
\description{
Deletes a registered passkey, or WebAuthn, authenticator for the currently signed-in user.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_delete_web_authn_credential/} for full documentation.
}
\keyword{internal}
