% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_identity_provider}
\alias{cognitoidentityprovider_describe_identity_provider}
\title{Given a user pool ID and identity provider (IdP) name, returns details
about the IdP}
\usage{
cognitoidentityprovider_describe_identity_provider(UserPoolId, ProviderName)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that has the IdP that you want to describe..}

\item{ProviderName}{[required] The name of the IdP that you want to describe.}
}
\description{
Given a user pool ID and identity provider (IdP) name, returns details about the IdP.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_describe_identity_provider/} for full documentation.
}
\keyword{internal}
