% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_identity_provider}
\alias{cognitoidentityprovider_describe_identity_provider}
\title{Gets information about a specific IdP}
\usage{
cognitoidentityprovider_describe_identity_provider(UserPoolId, ProviderName)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{ProviderName}{[required] The IdP name.}
}
\description{
Gets information about a specific IdP.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_describe_identity_provider/} for full documentation.
}
\keyword{internal}
