% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_resource_server}
\alias{cognitoidentityprovider_describe_resource_server}
\title{Describes a resource server}
\usage{
cognitoidentityprovider_describe_resource_server(UserPoolId, Identifier)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool that hosts the resource server.}

\item{Identifier}{[required] The identifier for the resource server}
}
\description{
Describes a resource server.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_describe_resource_server/} for full documentation.
}
\keyword{internal}
