% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_risk_configuration}
\alias{cognitoidentityprovider_describe_risk_configuration}
\title{Describes the risk configuration}
\usage{
cognitoidentityprovider_describe_risk_configuration(
  UserPoolId,
  ClientId = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{ClientId}{The app client ID.}
}
\description{
Describes the risk configuration.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_describe_risk_configuration/} for full documentation.
}
\keyword{internal}
