% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_user_import_job}
\alias{cognitoidentityprovider_describe_user_import_job}
\title{Describes the user import job}
\usage{
cognitoidentityprovider_describe_user_import_job(UserPoolId, JobId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool that the users are being imported
into.}

\item{JobId}{[required] The job ID for the user import job.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UserImportJob = list(
    JobName = "string",
    JobId = "string",
    UserPoolId = "string",
    PreSignedUrl = "string",
    CreationDate = as.POSIXct(
      "2015-01-01"
    ),
    StartDate = as.POSIXct(
      "2015-01-01"
    ),
    CompletionDate = as.POSIXct(
      "2015-01-01"
    ),
    Status = "Created"|"Pending"|"InProgress"|"Stopping"|"Expired"|"Stopped"|"Failed"|"Succeeded",
    CloudWatchLogsRoleArn = "string",
    ImportedUsers = 123,
    SkippedUsers = 123,
    FailedUsers = 123,
    CompletionMessage = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Describes the user import job.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_user_import_job(
  UserPoolId = "string",
  JobId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
