% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_user_pool_client}
\alias{cognitoidentityprovider_describe_user_pool_client}
\title{Given an app client ID, returns configuration information}
\usage{
cognitoidentityprovider_describe_user_pool_client(UserPoolId, ClientId)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool that contains the app client you want to
describe.}

\item{ClientId}{[required] The ID of the app client that you want to describe.}
}
\description{
Given an app client ID, returns configuration information. This operation is useful when you want to inspect an existing app client and programmatically replicate the configuration to another app client. For more information about app clients, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/user-pool-settings-client-apps.html}{App clients}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_describe_user_pool_client/} for full documentation.
}
\keyword{internal}
