% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_user_pool_client}
\alias{cognitoidentityprovider_describe_user_pool_client}
\title{Client method for returning the configuration information and metadata
of the specified user pool app client}
\usage{
cognitoidentityprovider_describe_user_pool_client(UserPoolId, ClientId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool you want to describe.}

\item{ClientId}{[required] The app client ID of the app associated with the user pool.}
}
\description{
Client method for returning the configuration information and metadata of the specified user pool app client.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_describe_user_pool_client/} for full documentation.
}
\keyword{internal}
