% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_describe_user_pool_domain}
\alias{cognitoidentityprovider_describe_user_pool_domain}
\title{Gets information about a domain}
\usage{
cognitoidentityprovider_describe_user_pool_domain(Domain)
}
\arguments{
\item{Domain}{[required] The domain string.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DomainDescription = list(
    UserPoolId = "string",
    AWSAccountId = "string",
    Domain = "string",
    S3Bucket = "string",
    CloudFrontDistribution = "string",
    Version = "string",
    Status = "CREATING"|"DELETING"|"UPDATING"|"ACTIVE"|"FAILED",
    CustomDomainConfig = list(
      CertificateArn = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets information about a domain.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_user_pool_domain(
  Domain = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
