% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_forget_device}
\alias{cognitoidentityprovider_forget_device}
\title{Forgets the specified device}
\usage{
cognitoidentityprovider_forget_device(AccessToken, DeviceKey)
}
\arguments{
\item{AccessToken}{The access token for the forgotten device request.}

\item{DeviceKey}{[required] The device key.}
}
\value{
An empty list.
}
\description{
Forgets the specified device.
}
\section{Request syntax}{
\preformatted{svc$forget_device(
  AccessToken = "string",
  DeviceKey = "string"
)
}
}

\keyword{internal}
