% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_forget_device}
\alias{cognitoidentityprovider_forget_device}
\title{Forgets the specified device}
\usage{
cognitoidentityprovider_forget_device(AccessToken = NULL, DeviceKey)
}
\arguments{
\item{AccessToken}{A valid access token that Amazon Cognito issued to the user whose
registered device you want to forget.}

\item{DeviceKey}{[required] The device key.}
}
\description{
Forgets the specified device. For more information about device authentication, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/amazon-cognito-user-pools-device-tracking.html}{Working with user devices in your user pool}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_forget_device/} for full documentation.
}
\keyword{internal}
