% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_csv_header}
\alias{cognitoidentityprovider_get_csv_header}
\title{Gets the header information for the comma-separated value (CSV) file to
be used as input for the user import job}
\usage{
cognitoidentityprovider_get_csv_header(UserPoolId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool that the users are to be imported
into.}
}
\description{
Gets the header information for the comma-separated value (CSV) file to be used as input for the user import job.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_get_csv_header/} for full documentation.
}
\keyword{internal}
