% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_device}
\alias{cognitoidentityprovider_get_device}
\title{Gets the device}
\usage{
cognitoidentityprovider_get_device(DeviceKey, AccessToken)
}
\arguments{
\item{DeviceKey}{[required] The device key.}

\item{AccessToken}{The access token.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Device = list(
    DeviceKey = "string",
    DeviceAttributes = list(
      list(
        Name = "string",
        Value = "string"
      )
    ),
    DeviceCreateDate = as.POSIXct(
      "2015-01-01"
    ),
    DeviceLastModifiedDate = as.POSIXct(
      "2015-01-01"
    ),
    DeviceLastAuthenticatedDate = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the device.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_device(
  DeviceKey = "string",
  AccessToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
