% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_device}
\alias{cognitoidentityprovider_get_device}
\title{Gets the device}
\usage{
cognitoidentityprovider_get_device(DeviceKey, AccessToken = NULL)
}
\arguments{
\item{DeviceKey}{[required] The device key.}

\item{AccessToken}{A valid access token that Amazon Cognito issued to the user whose device
information you want to request.}
}
\description{
Gets the device. For more information about device authentication, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/amazon-cognito-user-pools-device-tracking.html}{Working with user devices in your user pool}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_get_device/} for full documentation.
}
\keyword{internal}
