% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_group}
\alias{cognitoidentityprovider_get_group}
\title{Gets a group}
\usage{
cognitoidentityprovider_get_group(GroupName, UserPoolId)
}
\arguments{
\item{GroupName}{[required] The name of the group.}

\item{UserPoolId}{[required] The user pool ID for the user pool.}
}
\description{
Gets a group.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_get_group/} for full documentation.
}
\keyword{internal}
