% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_identity_provider_by_identifier}
\alias{cognitoidentityprovider_get_identity_provider_by_identifier}
\title{Gets the specified IdP}
\usage{
cognitoidentityprovider_get_identity_provider_by_identifier(
  UserPoolId,
  IdpIdentifier
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{IdpIdentifier}{[required] The IdP identifier.}
}
\description{
Gets the specified IdP.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_get_identity_provider_by_identifier/} for full documentation.
}
\keyword{internal}
