% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_log_delivery_configuration}
\alias{cognitoidentityprovider_get_log_delivery_configuration}
\title{Gets the detailed activity logging configuration for a user pool}
\usage{
cognitoidentityprovider_get_log_delivery_configuration(UserPoolId)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to view detailed activity logging
configuration.}
}
\description{
Gets the detailed activity logging configuration for a user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_get_log_delivery_configuration/} for full documentation.
}
\keyword{internal}
