% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_signing_certificate}
\alias{cognitoidentityprovider_get_signing_certificate}
\title{Given a user pool ID, returns the signing certificate for SAML 2}
\usage{
cognitoidentityprovider_get_signing_certificate(UserPoolId)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to view the signing certificate.}
}
\description{
Given a user pool ID, returns the signing certificate for SAML 2.0 federation.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_get_signing_certificate/} for full documentation.
}
\keyword{internal}
