% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_signing_certificate}
\alias{cognitoidentityprovider_get_signing_certificate}
\title{This method takes a user pool ID, and returns the signing certificate}
\usage{
cognitoidentityprovider_get_signing_certificate(UserPoolId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}
}
\value{
A list with the following syntax:\preformatted{list(
  Certificate = "string"
)
}
}
\description{
This method takes a user pool ID, and returns the signing certificate.
}
\section{Request syntax}{
\preformatted{svc$get_signing_certificate(
  UserPoolId = "string"
)
}
}

\keyword{internal}
