% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_user}
\alias{cognitoidentityprovider_get_user}
\title{Gets the user attributes and metadata for a user}
\usage{
cognitoidentityprovider_get_user(AccessToken)
}
\arguments{
\item{AccessToken}{[required] The access token returned by the server response to get information
about the user.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Username = "string",
  UserAttributes = list(
    list(
      Name = "string",
      Value = "string"
    )
  ),
  MFAOptions = list(
    list(
      DeliveryMedium = "SMS"|"EMAIL",
      AttributeName = "string"
    )
  ),
  PreferredMfaSetting = "string",
  UserMFASettingList = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Gets the user attributes and metadata for a user.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_user(
  AccessToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
