% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_user}
\alias{cognitoidentityprovider_get_user}
\title{Gets the user attributes and metadata for a user}
\usage{
cognitoidentityprovider_get_user(AccessToken)
}
\arguments{
\item{AccessToken}{[required] A non-expired access token for the user whose information you want to
query.}
}
\description{
Gets the user attributes and metadata for a user.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_get_user/} for full documentation.
}
\keyword{internal}
