% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_user_attribute_verification_code}
\alias{cognitoidentityprovider_get_user_attribute_verification_code}
\title{Gets the user attribute verification code for the specified attribute
name}
\usage{
cognitoidentityprovider_get_user_attribute_verification_code(
  AccessToken, AttributeName)
}
\arguments{
\item{AccessToken}{[required] The access token returned by the server response to get the user
attribute verification code.}

\item{AttributeName}{[required] The attribute name returned by the server response to get the user
attribute verification code.}
}
\description{
Gets the user attribute verification code for the specified attribute
name.
}
\section{Request syntax}{
\preformatted{svc$get_user_attribute_verification_code(
  AccessToken = "string",
  AttributeName = "string"
)
}
}

\keyword{internal}
