% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_user_pool_mfa_config}
\alias{cognitoidentityprovider_get_user_pool_mfa_config}
\title{Given a user pool ID, returns configuration for sign-in with WebAuthn
authenticators and for multi-factor authentication (MFA)}
\usage{
cognitoidentityprovider_get_user_pool_mfa_config(UserPoolId)
}
\arguments{
\item{UserPoolId}{[required] The ID of the user pool where you want to query WebAuthn and MFA
configuration.}
}
\description{
Given a user pool ID, returns configuration for sign-in with WebAuthn authenticators and for multi-factor authentication (MFA). This operation describes the following:

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_get_user_pool_mfa_config/} for full documentation.
}
\keyword{internal}
