% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_get_user_pool_mfa_config}
\alias{cognitoidentityprovider_get_user_pool_mfa_config}
\title{Gets the user pool multi-factor authentication (MFA) configuration}
\usage{
cognitoidentityprovider_get_user_pool_mfa_config(UserPoolId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}
}
\description{
Gets the user pool multi-factor authentication (MFA) configuration.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_get_user_pool_mfa_config/} for full documentation.
}
\keyword{internal}
