% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_global_sign_out}
\alias{cognitoidentityprovider_global_sign_out}
\title{Invalidates the identity, access, and refresh tokens that Amazon Cognito
issued to a user}
\usage{
cognitoidentityprovider_global_sign_out(AccessToken)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user who you want
to sign out.}
}
\description{
Invalidates the identity, access, and refresh tokens that Amazon Cognito issued to a user. Call this operation when your user signs out of your app. This results in the following behavior.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_global_sign_out/} for full documentation.
}
\keyword{internal}
