% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_devices}
\alias{cognitoidentityprovider_list_devices}
\title{Lists the sign-in devices that Amazon Cognito has registered to the
current user}
\usage{
cognitoidentityprovider_list_devices(
  AccessToken,
  Limit = NULL,
  PaginationToken = NULL
)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user whose list
of devices you want to view.}

\item{Limit}{The limit of the device request.}

\item{PaginationToken}{The pagination token for the list request.}
}
\description{
Lists the sign-in devices that Amazon Cognito has registered to the current user.

See \url{https://paws-r.github.io/docs/cognitoidentityprovider/list_devices.html} for full documentation.
}
\keyword{internal}
