% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_groups}
\alias{cognitoidentityprovider_list_groups}
\title{Lists the groups associated with a user pool}
\usage{
cognitoidentityprovider_list_groups(UserPoolId, Limit = NULL, NextToken = NULL)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{Limit}{The limit of the request to list groups.}

\item{NextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Lists the groups associated with a user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_groups/} for full documentation.
}
\keyword{internal}
