% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_identity_providers}
\alias{cognitoidentityprovider_list_identity_providers}
\title{Lists information about all IdPs for a user pool}
\usage{
cognitoidentityprovider_list_identity_providers(
  UserPoolId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{MaxResults}{The maximum number of IdPs to return.}

\item{NextToken}{A pagination token.}
}
\description{
Lists information about all IdPs for a user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_identity_providers/} for full documentation.
}
\keyword{internal}
