% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_resource_servers}
\alias{cognitoidentityprovider_list_resource_servers}
\title{Lists the resource servers for a user pool}
\usage{
cognitoidentityprovider_list_resource_servers(
  UserPoolId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{MaxResults}{The maximum number of resource servers to return.}

\item{NextToken}{A pagination token.}
}
\description{
Lists the resource servers for a user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_resource_servers/} for full documentation.
}
\keyword{internal}
