% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_tags_for_resource}
\alias{cognitoidentityprovider_list_tags_for_resource}
\title{Lists the tags that are assigned to an Amazon Cognito user pool}
\usage{
cognitoidentityprovider_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the user pool that the tags are
assigned to.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists the tags that are assigned to an Amazon Cognito user pool.

A tag is a label that you can apply to user pools to categorize and
manage them in different ways, such as by purpose, owner, environment,
or other criteria.

You can use this action up to 10 times per second, per account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_tags_for_resource(
  ResourceArn = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
