% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_user_import_jobs}
\alias{cognitoidentityprovider_list_user_import_jobs}
\title{Lists user import jobs for a user pool}
\usage{
cognitoidentityprovider_list_user_import_jobs(
  UserPoolId,
  MaxResults,
  PaginationToken = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool that the users are being imported
into.}

\item{MaxResults}{[required] The maximum number of import jobs you want the request to return.}

\item{PaginationToken}{An identifier that was returned from the previous call to
\code{\link[=cognitoidentityprovider_list_user_import_jobs]{list_user_import_jobs}},
which can be used to return the next set of import jobs in the list.}
}
\description{
Lists user import jobs for a user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_user_import_jobs/} for full documentation.
}
\keyword{internal}
