% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_user_import_jobs}
\alias{cognitoidentityprovider_list_user_import_jobs}
\title{Lists user import jobs for a user pool}
\usage{
cognitoidentityprovider_list_user_import_jobs(
  UserPoolId,
  MaxResults,
  PaginationToken = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool that the users are being imported
into.}

\item{MaxResults}{[required] The maximum number of import jobs you want the request to return.}

\item{PaginationToken}{This API operation returns a limited number of results. The pagination
token is an identifier that you can present in an additional API request
with the same parameters. When you include the pagination token, Amazon
Cognito returns the next set of items after the current list. Subsequent
requests return a new pagination token. By use of this token, you can
paginate through the full list of items.}
}
\description{
Lists user import jobs for a user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_user_import_jobs/} for full documentation.
}
\keyword{internal}
