% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_user_pool_clients}
\alias{cognitoidentityprovider_list_user_pool_clients}
\title{Lists the clients that have been created for the specified user pool}
\usage{
cognitoidentityprovider_list_user_pool_clients(UserPoolId, MaxResults,
  NextToken)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool where you want to list user pool
clients.}

\item{MaxResults}{The maximum number of results you want the request to return when
listing the user pool clients.}

\item{NextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\value{
A list with the following syntax:\preformatted{list(
  UserPoolClients = list(
    list(
      ClientId = "string",
      UserPoolId = "string",
      ClientName = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Lists the clients that have been created for the specified user pool.
}
\section{Request syntax}{
\preformatted{svc$list_user_pool_clients(
  UserPoolId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
