% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_list_users_in_group}
\alias{cognitoidentityprovider_list_users_in_group}
\title{Lists the users in the specified group}
\usage{
cognitoidentityprovider_list_users_in_group(
  UserPoolId,
  GroupName,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool.}

\item{GroupName}{[required] The name of the group.}

\item{Limit}{The maximum number of users that you want to retrieve before pagination.}

\item{NextToken}{An identifier that was returned from the previous call to this
operation, which can be used to return the next set of items in the
list.}
}
\description{
Lists the users in the specified group.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_list_users_in_group/} for full documentation.
}
\keyword{internal}
