% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_revoke_token}
\alias{cognitoidentityprovider_revoke_token}
\title{Revokes all of the access tokens generated by, and at the same time as,
the specified refresh token}
\usage{
cognitoidentityprovider_revoke_token(Token, ClientId, ClientSecret = NULL)
}
\arguments{
\item{Token}{[required] The refresh token that you want to revoke.}

\item{ClientId}{[required] The client ID for the token that you want to revoke.}

\item{ClientSecret}{The secret for the client ID. This is required only if the client ID has
a secret.}
}
\description{
Revokes all of the access tokens generated by, and at the same time as, the specified refresh token. After a token is revoked, you can't use the revoked token to access Amazon Cognito user APIs, or to authorize access to your resource server.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_revoke_token/} for full documentation.
}
\keyword{internal}
