% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_set_risk_configuration}
\alias{cognitoidentityprovider_set_risk_configuration}
\title{Configures actions on detected risks}
\usage{
cognitoidentityprovider_set_risk_configuration(
  UserPoolId,
  ClientId = NULL,
  CompromisedCredentialsRiskConfiguration = NULL,
  AccountTakeoverRiskConfiguration = NULL,
  RiskExceptionConfiguration = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{ClientId}{The app client ID. If \code{ClientId} is null, then the risk configuration is
mapped to \code{userPoolId}. When the client ID is null, the same risk
configuration is applied to all the clients in the userPool.

Otherwise, \code{ClientId} is mapped to the client. When the client ID isn't
null, the user pool configuration is overridden and the risk
configuration for the client is used instead.}

\item{CompromisedCredentialsRiskConfiguration}{The compromised credentials risk configuration.}

\item{AccountTakeoverRiskConfiguration}{The account takeover risk configuration.}

\item{RiskExceptionConfiguration}{The configuration to override the risk decision.}
}
\description{
Configures actions on detected risks. To delete the risk configuration for \code{UserPoolId} or \code{ClientId}, pass null values for all four configuration types.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_set_risk_configuration/} for full documentation.
}
\keyword{internal}
