% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_set_user_pool_mfa_config}
\alias{cognitoidentityprovider_set_user_pool_mfa_config}
\title{Sets user pool multi-factor authentication (MFA) and passkey
configuration}
\usage{
cognitoidentityprovider_set_user_pool_mfa_config(
  UserPoolId,
  SmsMfaConfiguration = NULL,
  SoftwareTokenMfaConfiguration = NULL,
  EmailMfaConfiguration = NULL,
  MfaConfiguration = NULL,
  WebAuthnConfiguration = NULL
)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{SmsMfaConfiguration}{Configures user pool SMS messages for MFA. Sets the message template and
the SMS message sending configuration for Amazon SNS.}

\item{SoftwareTokenMfaConfiguration}{Configures a user pool for time-based one-time password (TOTP) MFA.
Enables or disables TOTP.}

\item{EmailMfaConfiguration}{Sets configuration for user pool email message MFA and sign-in with
one-time passwords (OTPs). Includes the subject and body of the email
message template for sign-in and MFA messages. To activate this setting,
your user pool must be in the \href{https://docs.aws.amazon.com/cognito/latest/developerguide/feature-plans-features-essentials.html}{Essentials tier}
or higher.}

\item{MfaConfiguration}{Sets multi-factor authentication (MFA) to be on, off, or optional. When
\code{ON}, all users must set up MFA before they can sign in. When
\code{OPTIONAL}, your application must make a client-side determination of
whether a user wants to register an MFA device. For user pools with
adaptive authentication with threat protection, choose \code{OPTIONAL}.

When \code{MfaConfiguration} is \code{OPTIONAL}, managed login doesn't
automatically prompt users to set up MFA. Amazon Cognito generates MFA
prompts in API responses and in managed login for users who have chosen
and configured a preferred MFA factor.}

\item{WebAuthnConfiguration}{The configuration of your user pool for passkey, or WebAuthn,
authentication and registration. You can set this configuration
independent of the MFA configuration options in this operation.}
}
\description{
Sets user pool multi-factor authentication (MFA) and passkey configuration. For more information about user pool MFA, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/user-pool-settings-mfa.html}{Adding MFA}. For more information about WebAuthn passkeys see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/amazon-cognito-user-pools-authentication-flow-methods.html#amazon-cognito-user-pools-authentication-flow-methods-passkey}{Authentication flows}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_set_user_pool_mfa_config/} for full documentation.
}
\keyword{internal}
