% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_set_user_pool_mfa_config}
\alias{cognitoidentityprovider_set_user_pool_mfa_config}
\title{Set the user pool multi-factor authentication (MFA) configuration}
\usage{
cognitoidentityprovider_set_user_pool_mfa_config(UserPoolId,
  SmsMfaConfiguration, SoftwareTokenMfaConfiguration, MfaConfiguration)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID.}

\item{SmsMfaConfiguration}{The SMS text message MFA configuration.}

\item{SoftwareTokenMfaConfiguration}{The software token MFA configuration.}

\item{MfaConfiguration}{The MFA configuration. Valid values include:
\itemize{
\item \code{OFF} MFA will not be used for any users.
\item \code{ON} MFA is required for all users to sign in.
\item \code{OPTIONAL} MFA will be required only for individual users who have
an MFA factor enabled.
}}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SmsMfaConfiguration = list(
    SmsAuthenticationMessage = "string",
    SmsConfiguration = list(
      SnsCallerArn = "string",
      ExternalId = "string"
    )
  ),
  SoftwareTokenMfaConfiguration = list(
    Enabled = TRUE|FALSE
  ),
  MfaConfiguration = "OFF"|"ON"|"OPTIONAL"
)
}\if{html}{\out{</div>}}
}
\description{
Set the user pool multi-factor authentication (MFA) configuration.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$set_user_pool_mfa_config(
  UserPoolId = "string",
  SmsMfaConfiguration = list(
    SmsAuthenticationMessage = "string",
    SmsConfiguration = list(
      SnsCallerArn = "string",
      ExternalId = "string"
    )
  ),
  SoftwareTokenMfaConfiguration = list(
    Enabled = TRUE|FALSE
  ),
  MfaConfiguration = "OFF"|"ON"|"OPTIONAL"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
