% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_stop_user_import_job}
\alias{cognitoidentityprovider_stop_user_import_job}
\title{Stops the user import job}
\usage{
cognitoidentityprovider_stop_user_import_job(UserPoolId, JobId)
}
\arguments{
\item{UserPoolId}{[required] The user pool ID for the user pool that the users are being imported
into.}

\item{JobId}{[required] The job ID for the user import job.}
}
\description{
Stops the user import job.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_stop_user_import_job/} for full documentation.
}
\keyword{internal}
