% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_untag_resource}
\alias{cognitoidentityprovider_untag_resource}
\title{Given tag IDs that you previously assigned to a user pool, removes them}
\usage{
cognitoidentityprovider_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the user pool that the tags are
assigned to.}

\item{TagKeys}{[required] An array of tag keys that you want to remove from the user pool.}
}
\description{
Given tag IDs that you previously assigned to a user pool, removes them.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_untag_resource/} for full documentation.
}
\keyword{internal}
