% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_untag_resource}
\alias{cognitoidentityprovider_untag_resource}
\title{Removes the specified tags from an Amazon Cognito user pool}
\usage{
cognitoidentityprovider_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the user pool that the tags are
assigned to.}

\item{TagKeys}{[required] The keys of the tags to remove from the user pool.}
}
\description{
Removes the specified tags from an Amazon Cognito user pool. You can use this action up to 5 times per second, per account.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_untag_resource/} for full documentation.
}
\keyword{internal}
