% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_update_device_status}
\alias{cognitoidentityprovider_update_device_status}
\title{Updates the device status}
\usage{
cognitoidentityprovider_update_device_status(
  AccessToken,
  DeviceKey,
  DeviceRememberedStatus = NULL
)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user whose device
status you want to update.}

\item{DeviceKey}{[required] The device key.}

\item{DeviceRememberedStatus}{The status of whether a device is remembered.}
}
\description{
Updates the device status. For more information about device authentication, see \href{https://docs.aws.amazon.com/cognito/latest/developerguide/amazon-cognito-user-pools-device-tracking.html}{Working with user devices in your user pool}.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_update_device_status/} for full documentation.
}
\keyword{internal}
