% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_update_user_attributes}
\alias{cognitoidentityprovider_update_user_attributes}
\title{Allows a user to update a specific attribute (one at a time)}
\usage{
cognitoidentityprovider_update_user_attributes(UserAttributes,
  AccessToken)
}
\arguments{
\item{UserAttributes}{[required] An array of name-value pairs representing user attributes.

For custom attributes, you must prepend the \code{custom:} prefix to the
attribute name.}

\item{AccessToken}{[required] The access token for the request to update user attributes.}
}
\description{
Allows a user to update a specific attribute (one at a time).
}
\section{Request syntax}{
\preformatted{svc$update_user_attributes(
  UserAttributes = list(
    list(
      Name = "string",
      Value = "string"
    )
  ),
  AccessToken = "string"
)
}
}

\keyword{internal}
