% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_verify_user_attribute}
\alias{cognitoidentityprovider_verify_user_attribute}
\title{Verifies the specified user attributes in the user pool}
\usage{
cognitoidentityprovider_verify_user_attribute(AccessToken,
  AttributeName, Code)
}
\arguments{
\item{AccessToken}{[required] Represents the access token of the request to verify user attributes.}

\item{AttributeName}{[required] The attribute name in the request to verify user attributes.}

\item{Code}{[required] The verification code in the request to verify user attributes.}
}
\description{
Verifies the specified user attributes in the user pool.
}
\section{Request syntax}{
\preformatted{svc$verify_user_attribute(
  AccessToken = "string",
  AttributeName = "string",
  Code = "string"
)
}
}

\keyword{internal}
