% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitoidentityprovider_operations.R
\name{cognitoidentityprovider_verify_user_attribute}
\alias{cognitoidentityprovider_verify_user_attribute}
\title{Verifies the specified user attributes in the user pool}
\usage{
cognitoidentityprovider_verify_user_attribute(AccessToken, AttributeName, Code)
}
\arguments{
\item{AccessToken}{[required] A valid access token that Amazon Cognito issued to the user whose user
attributes you want to verify.}

\item{AttributeName}{[required] The attribute name in the request to verify user attributes.}

\item{Code}{[required] The verification code in the request to verify user attributes.}
}
\description{
Verifies the specified user attributes in the user pool.

See \url{https://www.paws-r-sdk.com/docs/cognitoidentityprovider_verify_user_attribute/} for full documentation.
}
\keyword{internal}
