% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitosync_operations.R
\name{cognitosync_get_bulk_publish_details}
\alias{cognitosync_get_bulk_publish_details}
\title{Get the status of the last BulkPublish operation for an identity pool}
\usage{
cognitosync_get_bulk_publish_details(IdentityPoolId)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example,
us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon
Cognito. GUID generation is unique within a region.}
}
\description{
Get the status of the last BulkPublish operation for an identity pool.

See \url{https://www.paws-r-sdk.com/docs/cognitosync_get_bulk_publish_details/} for full documentation.
}
\keyword{internal}
