% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitosync_operations.R
\name{cognitosync_get_identity_pool_configuration}
\alias{cognitosync_get_identity_pool_configuration}
\title{Gets the configuration settings of an identity pool}
\usage{
cognitosync_get_identity_pool_configuration(IdentityPoolId)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example,
us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon
Cognito. This is the ID of the pool for which to return a configuration.}
}
\description{
Gets the configuration settings of an identity pool.

See \url{https://www.paws-r-sdk.com/docs/cognitosync_get_identity_pool_configuration/} for full documentation.
}
\keyword{internal}
