% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitosync_operations.R
\name{cognitosync_list_identity_pool_usage}
\alias{cognitosync_list_identity_pool_usage}
\title{Gets a list of identity pools registered with Cognito}
\usage{
cognitosync_list_identity_pool_usage(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A pagination token for obtaining the next page of results.}

\item{MaxResults}{The maximum number of results to be returned.}
}
\description{
Gets a list of identity pools registered with Cognito.

See \url{https://www.paws-r-sdk.com/docs/cognitosync_list_identity_pool_usage/} for full documentation.
}
\keyword{internal}
