% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitosync_operations.R
\name{cognitosync_register_device}
\alias{cognitosync_register_device}
\title{Registers a device to receive push sync notifications}
\usage{
cognitosync_register_device(IdentityPoolId, IdentityId, Platform, Token)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example,
us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon
Cognito. Here, the ID of the pool that the identity belongs to.}

\item{IdentityId}{[required] The unique ID for this identity.}

\item{Platform}{[required] The SNS platform type (e.g. GCM, SDM, APNS, APNS_SANDBOX).}

\item{Token}{[required] The push token.}
}
\description{
Registers a device to receive push sync notifications.

See \url{https://www.paws-r-sdk.com/docs/cognitosync_register_device/} for full documentation.
}
\keyword{internal}
