% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitosync_operations.R
\name{cognitosync_unsubscribe_from_dataset}
\alias{cognitosync_unsubscribe_from_dataset}
\title{Unsubscribes from receiving notifications when a dataset is modified by
another device}
\usage{
cognitosync_unsubscribe_from_dataset(
  IdentityPoolId,
  IdentityId,
  DatasetName,
  DeviceId
)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example,
us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon
Cognito. The ID of the pool to which this identity belongs.}

\item{IdentityId}{[required] Unique ID for this identity.}

\item{DatasetName}{[required] The name of the dataset from which to unsubcribe.}

\item{DeviceId}{[required] The unique ID generated for this device by Cognito.}
}
\description{
Unsubscribes from receiving notifications when a dataset is modified by another device.

See \url{https://www.paws-r-sdk.com/docs/cognitosync_unsubscribe_from_dataset/} for full documentation.
}
\keyword{internal}
