% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cognitosync_operations.R
\name{cognitosync_update_records}
\alias{cognitosync_update_records}
\title{Posts updates to records and adds and deletes records for a dataset and
user}
\usage{
cognitosync_update_records(
  IdentityPoolId,
  IdentityId,
  DatasetName,
  DeviceId = NULL,
  RecordPatches = NULL,
  SyncSessionToken,
  ClientContext = NULL
)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example,
us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon
Cognito. GUID generation is unique within a region.}

\item{IdentityId}{[required] A name-spaced GUID (for example,
us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon
Cognito. GUID generation is unique within a region.}

\item{DatasetName}{[required] A string of up to 128 characters. Allowed characters are a-z, A-Z, 0-9,
'_' (underscore), '-' (dash), and '.' (dot).}

\item{DeviceId}{The unique ID generated for this device by Cognito.}

\item{RecordPatches}{A list of patch operations.}

\item{SyncSessionToken}{[required] The SyncSessionToken returned by a previous call to ListRecords for this
dataset and identity.}

\item{ClientContext}{Intended to supply a device ID that will populate the lastModifiedBy
field referenced in other methods. The ClientContext field is not yet
implemented.}
}
\description{
Posts updates to records and adds and deletes records for a dataset and user.

See \url{https://www.paws-r-sdk.com/docs/cognitosync_update_records/} for full documentation.
}
\keyword{internal}
