% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_list_tags_for_resource}
\alias{detective_list_tags_for_resource}
\title{Returns the tag values that are assigned to a behavior graph}
\usage{
detective_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the behavior graph for which to retrieve the tag values.}
}
\description{
Returns the tag values that are assigned to a behavior graph.

See \url{https://www.paws-r-sdk.com/docs/detective_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
