% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_create_snapshot}
\alias{directoryservice_create_snapshot}
\title{Creates a snapshot of a Simple AD or Microsoft AD directory in the AWS
cloud}
\usage{
directoryservice_create_snapshot(DirectoryId, Name)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory of which to take a snapshot.}

\item{Name}{The descriptive name to apply to the snapshot.}
}
\description{
Creates a snapshot of a Simple AD or Microsoft AD directory in the AWS
cloud.

You cannot take snapshots of AD Connector directories.
}
\section{Request syntax}{
\preformatted{svc$create_snapshot(
  DirectoryId = "string",
  Name = "string"
)
}
}

\keyword{internal}
