% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_create_snapshot}
\alias{directoryservice_create_snapshot}
\title{Creates a snapshot of a Simple AD or Microsoft AD directory in the
Amazon Web Services cloud}
\usage{
directoryservice_create_snapshot(DirectoryId, Name = NULL)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory of which to take a snapshot.}

\item{Name}{The descriptive name to apply to the snapshot.}
}
\description{
Creates a snapshot of a Simple AD or Microsoft AD directory in the Amazon Web Services cloud.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_create_snapshot/} for full documentation.
}
\keyword{internal}
