% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_delete_directory}
\alias{directoryservice_delete_directory}
\title{Deletes an Directory Service directory}
\usage{
directoryservice_delete_directory(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory to delete.}
}
\description{
Deletes an Directory Service directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_delete_directory/} for full documentation.
}
\keyword{internal}
