% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_delete_trust}
\alias{directoryservice_delete_trust}
\title{Deletes an existing trust relationship between your Managed Microsoft AD
directory and an external domain}
\usage{
directoryservice_delete_trust(
  TrustId,
  DeleteAssociatedConditionalForwarder = NULL
)
}
\arguments{
\item{TrustId}{[required] The Trust ID of the trust relationship to be deleted.}

\item{DeleteAssociatedConditionalForwarder}{Delete a conditional forwarder as part of a DeleteTrustRequest.}
}
\description{
Deletes an existing trust relationship between your Managed Microsoft AD directory and an external domain.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_delete_trust/} for full documentation.
}
\keyword{internal}
