% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_deregister_certificate}
\alias{directoryservice_deregister_certificate}
\title{Deletes from the system the certificate that was registered for secure
LDAP or client certificate authentication}
\usage{
directoryservice_deregister_certificate(DirectoryId, CertificateId)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{CertificateId}{[required] The identifier of the certificate.}
}
\description{
Deletes from the system the certificate that was registered for secure LDAP or client certificate authentication.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_deregister_certificate/} for full documentation.
}
\keyword{internal}
