% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_directories}
\alias{directoryservice_describe_directories}
\title{Obtains information about the directories that belong to this account}
\usage{
directoryservice_describe_directories(
  DirectoryIds = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{DirectoryIds}{A list of identifiers of the directories for which to obtain the
information. If this member is null, all directories that belong to the
current account are returned.

An empty list results in an \code{InvalidParameterException} being thrown.}

\item{NextToken}{The \code{DescribeDirectoriesResult.NextToken} value from a previous call to
\code{\link[=directoryservice_describe_directories]{describe_directories}}. Pass
null if this is the first call.}

\item{Limit}{The maximum number of items to return. If this value is zero, the
maximum number of items is specified by the limitations of the
operation.}
}
\description{
Obtains information about the directories that belong to this account.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_describe_directories/} for full documentation.
}
\keyword{internal}
