% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_domain_controllers}
\alias{directoryservice_describe_domain_controllers}
\title{Provides information about any domain controllers in your directory}
\usage{
directoryservice_describe_domain_controllers(
  DirectoryId,
  DomainControllerIds = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{DirectoryId}{[required] Identifier of the directory for which to retrieve the domain controller
information.}

\item{DomainControllerIds}{A list of identifiers for the domain controllers whose information will
be provided.}

\item{NextToken}{The \emph{DescribeDomainControllers.NextToken} value from a previous call to
\code{\link[=directoryservice_describe_domain_controllers]{describe_domain_controllers}}.
Pass null if this is the first call.}

\item{Limit}{The maximum number of items to return.}
}
\description{
Provides information about any domain controllers in your directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_describe_domain_controllers/} for full documentation.
}
\keyword{internal}
