% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_event_topics}
\alias{directoryservice_describe_event_topics}
\title{Obtains information about which Amazon SNS topics receive status
messages from the specified directory}
\usage{
directoryservice_describe_event_topics(DirectoryId = NULL, TopicNames = NULL)
}
\arguments{
\item{DirectoryId}{The Directory ID for which to get the list of associated Amazon SNS
topics. If this member is null, associations for all Directory IDs are
returned.}

\item{TopicNames}{A list of Amazon SNS topic names for which to obtain the information. If
this member is null, all associations for the specified Directory ID are
returned.

An empty list results in an \code{InvalidParameterException} being thrown.}
}
\description{
Obtains information about which Amazon SNS topics receive status messages from the specified directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_describe_event_topics/} for full documentation.
}
\keyword{internal}
