% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_regions}
\alias{directoryservice_describe_regions}
\title{Provides information about the Regions that are configured for
multi-Region replication}
\usage{
directoryservice_describe_regions(
  DirectoryId,
  RegionName = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{RegionName}{The name of the Region. For example, \code{us-east-1}.}

\item{NextToken}{The \code{DescribeRegionsResult.NextToken} value from a previous call to
\code{\link[=directoryservice_describe_regions]{describe_regions}}. Pass null if
this is the first call.}
}
\description{
Provides information about the Regions that are configured for multi-Region replication.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_describe_regions/} for full documentation.
}
\keyword{internal}
