% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_shared_directories}
\alias{directoryservice_describe_shared_directories}
\title{Returns the shared directories in your account}
\usage{
directoryservice_describe_shared_directories(
  OwnerDirectoryId,
  SharedDirectoryIds = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{OwnerDirectoryId}{[required] Returns the identifier of the directory in the directory owner account.}

\item{SharedDirectoryIds}{A list of identifiers of all shared directories in your account.}

\item{NextToken}{The \code{DescribeSharedDirectoriesResult.NextToken} value from a previous
call to
\code{\link[=directoryservice_describe_shared_directories]{describe_shared_directories}}.
Pass null if this is the first call.}

\item{Limit}{The number of shared directories to return in the response object.}
}
\description{
Returns the shared directories in your account.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_describe_shared_directories/} for full documentation.
}
\keyword{internal}
