% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_snapshots}
\alias{directoryservice_describe_snapshots}
\title{Obtains information about the directory snapshots that belong to this
account}
\usage{
directoryservice_describe_snapshots(
  DirectoryId = NULL,
  SnapshotIds = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{DirectoryId}{The identifier of the directory for which to retrieve snapshot
information.}

\item{SnapshotIds}{A list of identifiers of the snapshots to obtain the information for. If
this member is null or empty, all snapshots are returned using the
\emph{Limit} and \emph{NextToken} members.}

\item{NextToken}{The \emph{DescribeSnapshotsResult.NextToken} value from a previous call to
\code{\link[=directoryservice_describe_snapshots]{describe_snapshots}}. Pass null
if this is the first call.}

\item{Limit}{The maximum number of objects to return.}
}
\description{
Obtains information about the directory snapshots that belong to this account.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_describe_snapshots/} for full documentation.
}
\keyword{internal}
