% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_disable_ldaps}
\alias{directoryservice_disable_ldaps}
\title{Deactivates LDAP secure calls for the specified directory}
\usage{
directoryservice_disable_ldaps(DirectoryId, Type)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{Type}{[required] The type of LDAP security to enable. Currently only the value \code{Client}
is supported.}
}
\description{
Deactivates LDAP secure calls for the specified directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_disable_ldaps/} for full documentation.
}
\keyword{internal}
