% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_disable_radius}
\alias{directoryservice_disable_radius}
\title{Disables multi-factor authentication (MFA) with the Remote
Authentication Dial In User Service (RADIUS) server for an AD Connector
or Microsoft AD directory}
\usage{
directoryservice_disable_radius(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory for which to disable MFA.}
}
\description{
Disables multi-factor authentication (MFA) with the Remote Authentication Dial In User Service (RADIUS) server for an AD Connector or Microsoft AD directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_disable_radius/} for full documentation.
}
\keyword{internal}
