% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_enable_client_authentication}
\alias{directoryservice_enable_client_authentication}
\title{Enables alternative client authentication methods for the specified
directory}
\usage{
directoryservice_enable_client_authentication(DirectoryId, Type)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the specified directory.}

\item{Type}{[required] The type of client authentication to enable. Currently only the value
\code{SmartCard} is supported. Smart card authentication in AD Connector
requires that you enable Kerberos Constrained Delegation for the Service
User to the LDAP service in your self-managed AD.}
}
\description{
Enables alternative client authentication methods for the specified directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_enable_client_authentication/} for full documentation.
}
\keyword{internal}
