% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_enable_ldaps}
\alias{directoryservice_enable_ldaps}
\title{Activates the switch for the specific directory to always use LDAP
secure calls}
\usage{
directoryservice_enable_ldaps(DirectoryId, Type)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{Type}{[required] The type of LDAP security to enable. Currently only the value \code{Client}
is supported.}
}
\description{
Activates the switch for the specific directory to always use LDAP secure calls.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_enable_ldaps/} for full documentation.
}
\keyword{internal}
