% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_get_directory_limits}
\alias{directoryservice_get_directory_limits}
\title{Obtains directory limit information for the current Region}
\usage{
directoryservice_get_directory_limits()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  DirectoryLimits = list(
    CloudOnlyDirectoriesLimit = 123,
    CloudOnlyDirectoriesCurrentCount = 123,
    CloudOnlyDirectoriesLimitReached = TRUE|FALSE,
    CloudOnlyMicrosoftADLimit = 123,
    CloudOnlyMicrosoftADCurrentCount = 123,
    CloudOnlyMicrosoftADLimitReached = TRUE|FALSE,
    ConnectedDirectoriesLimit = 123,
    ConnectedDirectoriesCurrentCount = 123,
    ConnectedDirectoriesLimitReached = TRUE|FALSE
  )
)
}\if{html}{\out{</div>}}
}
\description{
Obtains directory limit information for the current Region.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_directory_limits()
}\if{html}{\out{</div>}}
}

\keyword{internal}
