% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_get_directory_limits}
\alias{directoryservice_get_directory_limits}
\title{Obtains directory limit information for the current Region}
\usage{
directoryservice_get_directory_limits()
}
\description{
Obtains directory limit information for the current Region.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_get_directory_limits/} for full documentation.
}
\keyword{internal}
