% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_get_snapshot_limits}
\alias{directoryservice_get_snapshot_limits}
\title{Obtains the manual snapshot limits for a directory}
\usage{
directoryservice_get_snapshot_limits(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] Contains the identifier of the directory to obtain the limits for.}
}
\description{
Obtains the manual snapshot limits for a directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_get_snapshot_limits/} for full documentation.
}
\keyword{internal}
