% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_list_log_subscriptions}
\alias{directoryservice_list_log_subscriptions}
\title{Lists the active log subscriptions for the Amazon Web Services account}
\usage{
directoryservice_list_log_subscriptions(
  DirectoryId = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{DirectoryId}{If a \emph{DirectoryID} is provided, lists only the log subscription
associated with that directory. If no \emph{DirectoryId} is provided, lists
all log subscriptions associated with your Amazon Web Services account.
If there are no log subscriptions for the Amazon Web Services account or
the directory, an empty list will be returned.}

\item{NextToken}{The token for the next set of items to return.}

\item{Limit}{The maximum number of items returned.}
}
\description{
Lists the active log subscriptions for the Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_list_log_subscriptions/} for full documentation.
}
\keyword{internal}
