% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_list_schema_extensions}
\alias{directoryservice_list_schema_extensions}
\title{Lists all schema extensions applied to a Microsoft AD Directory}
\usage{
directoryservice_list_schema_extensions(
  DirectoryId,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory from which to retrieve the schema
extension information.}

\item{NextToken}{The \code{ListSchemaExtensions.NextToken} value from a previous call to
\code{\link[=directoryservice_list_schema_extensions]{list_schema_extensions}}.
Pass null if this is the first call.}

\item{Limit}{The maximum number of items to return.}
}
\description{
Lists all schema extensions applied to a Microsoft AD Directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_list_schema_extensions/} for full documentation.
}
\keyword{internal}
