% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_list_tags_for_resource}
\alias{directoryservice_list_tags_for_resource}
\title{Lists all tags on a directory}
\usage{
directoryservice_list_tags_for_resource(
  ResourceId,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{ResourceId}{[required] Identifier (ID) of the directory for which you want to retrieve tags.}

\item{NextToken}{Reserved for future use.}

\item{Limit}{Reserved for future use.}
}
\description{
Lists all tags on a directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
