% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_register_certificate}
\alias{directoryservice_register_certificate}
\title{Registers a certificate for a secure LDAP or client certificate
authentication}
\usage{
directoryservice_register_certificate(
  DirectoryId,
  CertificateData,
  Type = NULL,
  ClientCertAuthSettings = NULL
)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory.}

\item{CertificateData}{[required] The certificate PEM string that needs to be registered.}

\item{Type}{The function that the registered certificate performs. Valid values
include \code{ClientLDAPS} or \code{ClientCertAuth}. The default value is
\code{ClientLDAPS}.}

\item{ClientCertAuthSettings}{A \code{ClientCertAuthSettings} object that contains client certificate
authentication settings.}
}
\description{
Registers a certificate for a secure LDAP or client certificate authentication.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_register_certificate/} for full documentation.
}
\keyword{internal}
