% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_reject_shared_directory}
\alias{directoryservice_reject_shared_directory}
\title{Rejects a directory sharing request that was sent from the directory
owner account}
\usage{
directoryservice_reject_shared_directory(SharedDirectoryId)
}
\arguments{
\item{SharedDirectoryId}{[required] Identifier of the shared directory in the directory consumer account.
This identifier is different for each directory owner account.}
}
\description{
Rejects a directory sharing request that was sent from the directory owner account.

See \url{https://paws-r.github.io/docs/directoryservice/reject_shared_directory.html} for full documentation.
}
\keyword{internal}
