% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_remove_region}
\alias{directoryservice_remove_region}
\title{Stops all replication and removes the domain controllers from the
specified Region}
\usage{
directoryservice_remove_region(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory for which you want to remove Region
replication.}
}
\description{
Stops all replication and removes the domain controllers from the specified Region. You cannot remove the primary Region with this operation. Instead, use the \code{\link[=directoryservice_delete_directory]{delete_directory}} API.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_remove_region/} for full documentation.
}
\keyword{internal}
