% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_remove_tags_from_resource}
\alias{directoryservice_remove_tags_from_resource}
\title{Removes tags from a directory}
\usage{
directoryservice_remove_tags_from_resource(ResourceId, TagKeys)
}
\arguments{
\item{ResourceId}{[required] Identifier (ID) of the directory from which to remove the tag.}

\item{TagKeys}{[required] The tag key (name) of the tag to be removed.}
}
\description{
Removes tags from a directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_remove_tags_from_resource/} for full documentation.
}
\keyword{internal}
