% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_restore_from_snapshot}
\alias{directoryservice_restore_from_snapshot}
\title{Restores a directory using an existing directory snapshot}
\usage{
directoryservice_restore_from_snapshot(SnapshotId)
}
\arguments{
\item{SnapshotId}{[required] The identifier of the snapshot to restore from.}
}
\description{
Restores a directory using an existing directory snapshot.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_restore_from_snapshot/} for full documentation.
}
\keyword{internal}
