% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_update_conditional_forwarder}
\alias{directoryservice_update_conditional_forwarder}
\title{Updates a conditional forwarder that has been set up for your Amazon Web
Services directory}
\usage{
directoryservice_update_conditional_forwarder(
  DirectoryId,
  RemoteDomainName,
  DnsIpAddrs = NULL,
  DnsIpv6Addrs = NULL
)
}
\arguments{
\item{DirectoryId}{[required] The directory ID of the Amazon Web Services directory for which to
update the conditional forwarder.}

\item{RemoteDomainName}{[required] The fully qualified domain name (FQDN) of the remote domain with which
you will set up a trust relationship.}

\item{DnsIpAddrs}{The updated IP addresses of the remote DNS server associated with the
conditional forwarder.}

\item{DnsIpv6Addrs}{The updated IPv6 addresses of the remote DNS server associated with the
conditional forwarder.}
}
\description{
Updates a conditional forwarder that has been set up for your Amazon Web Services directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_update_conditional_forwarder/} for full documentation.
}
\keyword{internal}
