% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_update_directory_setup}
\alias{directoryservice_update_directory_setup}
\title{Updates directory configuration for the specified update type}
\usage{
directoryservice_update_directory_setup(
  DirectoryId,
  UpdateType,
  OSUpdateSettings = NULL,
  DirectorySizeUpdateSettings = NULL,
  NetworkUpdateSettings = NULL,
  CreateSnapshotBeforeUpdate = NULL
)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory to update.}

\item{UpdateType}{[required] The type of update to perform on the directory.}

\item{OSUpdateSettings}{Operating system configuration to apply during the directory update
operation.}

\item{DirectorySizeUpdateSettings}{Directory size configuration to apply during the update operation.}

\item{NetworkUpdateSettings}{Network configuration to apply during the directory update operation.}

\item{CreateSnapshotBeforeUpdate}{Specifies whether to create a directory snapshot before performing the
update.}
}
\description{
Updates directory configuration for the specified update type.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_update_directory_setup/} for full documentation.
}
\keyword{internal}
