% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_update_radius}
\alias{directoryservice_update_radius}
\title{Updates the Remote Authentication Dial In User Service (RADIUS) server
information for an AD Connector or Microsoft AD directory}
\usage{
directoryservice_update_radius(DirectoryId, RadiusSettings)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory for which to update the RADIUS server
information.}

\item{RadiusSettings}{[required] A RadiusSettings object that contains information about the RADIUS
server.}
}
\description{
Updates the Remote Authentication Dial In User Service (RADIUS) server information for an AD Connector or Microsoft AD directory.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_update_radius/} for full documentation.
}
\keyword{internal}
