% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_verify_trust}
\alias{directoryservice_verify_trust}
\title{Directory Service for Microsoft Active Directory allows you to configure
and verify trust relationships}
\usage{
directoryservice_verify_trust(TrustId)
}
\arguments{
\item{TrustId}{[required] The unique Trust ID of the trust relationship to verify.}
}
\description{
Directory Service for Microsoft Active Directory allows you to configure and verify trust relationships.

See \url{https://www.paws-r-sdk.com/docs/directoryservice_verify_trust/} for full documentation.
}
\keyword{internal}
