% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_delete_notification_channel}
\alias{fms_delete_notification_channel}
\title{Deletes an Firewall Manager association with the IAM role and the Amazon
Simple Notification Service (SNS) topic that is used to record Firewall
Manager SNS logs}
\usage{
fms_delete_notification_channel()
}
\description{
Deletes an Firewall Manager association with the IAM role and the Amazon Simple Notification Service (SNS) topic that is used to record Firewall Manager SNS logs.

See \url{https://www.paws-r-sdk.com/docs/fms_delete_notification_channel/} for full documentation.
}
\keyword{internal}
