% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_admin_account}
\alias{fms_get_admin_account}
\title{Returns the Organizations account that is associated with Firewall
Manager as the Firewall Manager administrator}
\usage{
fms_get_admin_account()
}
\description{
Returns the Organizations account that is associated with Firewall Manager as the Firewall Manager administrator.

See \url{https://paws-r.github.io/docs/fms/get_admin_account.html} for full documentation.
}
\keyword{internal}
