% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_admin_scope}
\alias{fms_get_admin_scope}
\title{Returns information about the specified account's administrative scope}
\usage{
fms_get_admin_scope(AdminAccount)
}
\arguments{
\item{AdminAccount}{[required] The administrator account that you want to get the details for.}
}
\description{
Returns information about the specified account's administrative scope. The administrative scope defines the resources that an Firewall Manager administrator can manage.

See \url{https://www.paws-r-sdk.com/docs/fms_get_admin_scope/} for full documentation.
}
\keyword{internal}
