% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_apps_list}
\alias{fms_get_apps_list}
\title{Returns information about the specified Firewall Manager applications
list}
\usage{
fms_get_apps_list(ListId, DefaultList = NULL)
}
\arguments{
\item{ListId}{[required] The ID of the Firewall Manager applications list that you want the
details for.}

\item{DefaultList}{Specifies whether the list to retrieve is a default list owned by
Firewall Manager.}
}
\description{
Returns information about the specified Firewall Manager applications list.

See \url{https://www.paws-r-sdk.com/docs/fms_get_apps_list/} for full documentation.
}
\keyword{internal}
