% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_notification_channel}
\alias{fms_get_notification_channel}
\title{Information about the Amazon Simple Notification Service (SNS) topic
that is used to record AWS Firewall Manager SNS logs}
\usage{
fms_get_notification_channel()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SnsTopicArn = "string",
  SnsRoleName = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Information about the Amazon Simple Notification Service (SNS) topic
that is used to record AWS Firewall Manager SNS logs.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_notification_channel()
}\if{html}{\out{</div>}}
}

\keyword{internal}
