% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_notification_channel}
\alias{fms_get_notification_channel}
\title{Information about the Amazon Simple Notification Service (SNS) topic
that is used to record Firewall Manager SNS logs}
\usage{
fms_get_notification_channel()
}
\description{
Information about the Amazon Simple Notification Service (SNS) topic that is used to record Firewall Manager SNS logs.

See \url{https://www.paws-r-sdk.com/docs/fms_get_notification_channel/} for full documentation.
}
\keyword{internal}
