% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_get_policy}
\alias{fms_get_policy}
\title{Returns information about the specified Firewall Manager policy}
\usage{
fms_get_policy(PolicyId)
}
\arguments{
\item{PolicyId}{[required] The ID of the Firewall Manager policy that you want the details for.}
}
\description{
Returns information about the specified Firewall Manager policy.

See \url{https://www.paws-r-sdk.com/docs/fms_get_policy/} for full documentation.
}
\keyword{internal}
