% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_list_tags_for_resource}
\alias{fms_list_tags_for_resource}
\title{Retrieves the list of tags for the specified Amazon Web Services
resource}
\usage{
fms_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to return tags for. The
Firewall Manager resources that support tagging are policies,
applications lists, and protocols lists.}
}
\description{
Retrieves the list of tags for the specified Amazon Web Services resource.

See \url{https://www.paws-r-sdk.com/docs/fms_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
