% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_put_apps_list}
\alias{fms_put_apps_list}
\title{Creates an Firewall Manager applications list}
\usage{
fms_put_apps_list(AppsList, TagList = NULL)
}
\arguments{
\item{AppsList}{[required] The details of the Firewall Manager applications list to be created.}

\item{TagList}{The tags associated with the resource.}
}
\description{
Creates an Firewall Manager applications list.

See \url{https://www.paws-r-sdk.com/docs/fms_put_apps_list/} for full documentation.
}
\keyword{internal}
