% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_put_notification_channel}
\alias{fms_put_notification_channel}
\title{Designates the IAM role and Amazon Simple Notification Service (SNS)
topic that Firewall Manager uses to record SNS logs}
\usage{
fms_put_notification_channel(SnsTopicArn, SnsRoleName)
}
\arguments{
\item{SnsTopicArn}{[required] The Amazon Resource Name (ARN) of the SNS topic that collects
notifications from Firewall Manager.}

\item{SnsRoleName}{[required] The Amazon Resource Name (ARN) of the IAM role that allows Amazon SNS to
record Firewall Manager activity.}
}
\description{
Designates the IAM role and Amazon Simple Notification Service (SNS) topic that Firewall Manager uses to record SNS logs.

See \url{https://www.paws-r-sdk.com/docs/fms_put_notification_channel/} for full documentation.
}
\keyword{internal}
