% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_put_policy}
\alias{fms_put_policy}
\title{Creates an Firewall Manager policy}
\usage{
fms_put_policy(Policy, TagList = NULL)
}
\arguments{
\item{Policy}{[required] The details of the Firewall Manager policy to be created.}

\item{TagList}{The tags to add to the Amazon Web Services resource.}
}
\description{
Creates an Firewall Manager policy.

See \url{https://www.paws-r-sdk.com/docs/fms_put_policy/} for full documentation.
}
\keyword{internal}
