% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_tag_resource}
\alias{fms_tag_resource}
\title{Adds one or more tags to an Amazon Web Services resource}
\usage{
fms_tag_resource(ResourceArn, TagList)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to return tags for. The
Firewall Manager resources that support tagging are policies,
applications lists, and protocols lists.}

\item{TagList}{[required] The tags to add to the resource.}
}
\description{
Adds one or more tags to an Amazon Web Services resource.

See \url{https://www.paws-r-sdk.com/docs/fms_tag_resource/} for full documentation.
}
\keyword{internal}
