% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fms_operations.R
\name{fms_untag_resource}
\alias{fms_untag_resource}
\title{Removes one or more tags from an Amazon Web Services resource}
\usage{
fms_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource to return tags for. The
Firewall Manager resources that support tagging are policies,
applications lists, and protocols lists.}

\item{TagKeys}{[required] The keys of the tags to remove from the resource.}
}
\description{
Removes one or more tags from an Amazon Web Services resource.

See \url{https://www.paws-r-sdk.com/docs/fms_untag_resource/} for full documentation.
}
\keyword{internal}
