% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_create_filter}
\alias{guardduty_create_filter}
\title{Creates a filter using the specified finding criteria}
\usage{
guardduty_create_filter(
  DetectorId,
  Name,
  Description = NULL,
  Action = NULL,
  Rank = NULL,
  FindingCriteria,
  ClientToken = NULL,
  Tags = NULL
)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector belonging to the GuardDuty account that you want
to create a filter for.}

\item{Name}{[required] The name of the filter. Valid characters include period (.), underscore
(_), dash (-), and alphanumeric characters. A whitespace is considered
to be an invalid character.}

\item{Description}{The description of the filter. Valid characters include alphanumeric
characters, and special characters such as hyphen, period, colon,
underscore, parentheses (\code{{ }}, \verb{[ ]}, and \verb{( )}), forward slash,
horizontal tab, vertical tab, newline, form feed, return, and
whitespace.

[ ]: R:\%20}

\item{Action}{Specifies the action that is to be applied to the findings that match
the filter.}

\item{Rank}{Specifies the position of the filter in the list of current filters.
Also specifies the order in which this filter is applied to the
findings.}

\item{FindingCriteria}{[required] Represents the criteria to be used in the filter for querying findings.

You can only use the following attributes to query findings:
\itemize{
\item accountId
\item region
\item id
\item resource.accessKeyDetails.accessKeyId
\item resource.accessKeyDetails.principalId
\item resource.accessKeyDetails.userName
\item resource.accessKeyDetails.userType
\item resource.instanceDetails.iamInstanceProfile.id
\item resource.instanceDetails.imageId
\item resource.instanceDetails.instanceId
\item resource.instanceDetails.outpostArn
\item resource.instanceDetails.networkInterfaces.ipv6Addresses
\item resource.instanceDetails.networkInterfaces.privateIpAddresses.privateIpAddress
\item resource.instanceDetails.networkInterfaces.publicDnsName
\item resource.instanceDetails.networkInterfaces.publicIp
\item resource.instanceDetails.networkInterfaces.securityGroups.groupId
\item resource.instanceDetails.networkInterfaces.securityGroups.groupName
\item resource.instanceDetails.networkInterfaces.subnetId
\item resource.instanceDetails.networkInterfaces.vpcId
\item resource.instanceDetails.tags.key
\item resource.instanceDetails.tags.value
\item resource.resourceType
\item service.action.actionType
\item service.action.awsApiCallAction.api
\item service.action.awsApiCallAction.callerType
\item service.action.awsApiCallAction.errorCode
\item service.action.awsApiCallAction.userAgent
\item service.action.awsApiCallAction.remoteIpDetails.city.cityName
\item service.action.awsApiCallAction.remoteIpDetails.country.countryName
\item service.action.awsApiCallAction.remoteIpDetails.ipAddressV4
\item service.action.awsApiCallAction.remoteIpDetails.organization.asn
\item service.action.awsApiCallAction.remoteIpDetails.organization.asnOrg
\item service.action.awsApiCallAction.serviceName
\item service.action.dnsRequestAction.domain
\item service.action.networkConnectionAction.blocked
\item service.action.networkConnectionAction.connectionDirection
\item service.action.networkConnectionAction.localPortDetails.port
\item service.action.networkConnectionAction.protocol
\item service.action.networkConnectionAction.localIpDetails.ipAddressV4
\item service.action.networkConnectionAction.remoteIpDetails.city.cityName
\item service.action.networkConnectionAction.remoteIpDetails.country.countryName
\item service.action.networkConnectionAction.remoteIpDetails.ipAddressV4
\item service.action.networkConnectionAction.remoteIpDetails.organization.asn
\item service.action.networkConnectionAction.remoteIpDetails.organization.asnOrg
\item service.action.networkConnectionAction.remotePortDetails.port
\item service.additionalInfo.threatListName
\item resource.s3BucketDetails.publicAccess.effectivePermissions
\item resource.s3BucketDetails.name
\item resource.s3BucketDetails.tags.key
\item resource.s3BucketDetails.tags.value
\item resource.s3BucketDetails.type
\item service.resourceRole
\item severity
\item type
\item updatedAt

Type: ISO 8601 string format: YYYY-MM-DDTHH:MM:SS.SSSZ or
YYYY-MM-DDTHH:MM:SSZ depending on whether the value contains
milliseconds.
}}

\item{ClientToken}{The idempotency token for the create request.}

\item{Tags}{The tags to be added to a new filter resource.}
}
\description{
Creates a filter using the specified finding criteria. The maximum number of saved filters per Amazon Web Services account per Region is 100. For more information, see \href{https://docs.aws.amazon.com/guardduty/latest/ug/guardduty_limits.html}{Quotas for GuardDuty}.

See \url{https://www.paws-r-sdk.com/docs/guardduty_create_filter/} for full documentation.
}
\keyword{internal}
