% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_create_sample_findings}
\alias{guardduty_create_sample_findings}
\title{Generates example findings of types specified by the list of finding
types}
\usage{
guardduty_create_sample_findings(DetectorId, FindingTypes)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector to create sample findings for.}

\item{FindingTypes}{The types of sample findings to generate.}
}
\description{
Generates example findings of types specified by the list of finding
types. If \'NULL\' is specified for \code{findingTypes}, the API generates
example findings of all supported finding types.
}
\section{Request syntax}{
\preformatted{svc$create_sample_findings(
  DetectorId = "string",
  FindingTypes = list(
    "string"
  )
)
}
}

\keyword{internal}
