% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_create_threat_intel_set}
\alias{guardduty_create_threat_intel_set}
\title{Creates a new ThreatIntelSet}
\usage{
guardduty_create_threat_intel_set(
  DetectorId,
  Name,
  Format,
  Location,
  Activate,
  ClientToken = NULL,
  Tags = NULL,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty account for which you
want to create a \code{threatIntelSet}.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{Name}{[required] A user-friendly ThreatIntelSet name displayed in all findings that are
generated by activity that involves IP addresses included in this
ThreatIntelSet.}

\item{Format}{[required] The format of the file that contains the ThreatIntelSet.}

\item{Location}{[required] The URI of the file that contains the ThreatIntelSet.}

\item{Activate}{[required] A Boolean value that indicates whether GuardDuty is to start using the
uploaded ThreatIntelSet.}

\item{ClientToken}{The idempotency token for the create request.}

\item{Tags}{The tags to be added to a new threat list resource.}

\item{ExpectedBucketOwner}{The Amazon Web Services account ID that owns the Amazon S3 bucket
specified in the \strong{location} parameter.}
}
\description{
Creates a new ThreatIntelSet. ThreatIntelSets consist of known malicious IP addresses. GuardDuty generates findings based on ThreatIntelSets. Only users of the administrator account can use this operation.

See \url{https://www.paws-r-sdk.com/docs/guardduty_create_threat_intel_set/} for full documentation.
}
\keyword{internal}
