% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_detector}
\alias{guardduty_delete_detector}
\title{Deletes an Amazon GuardDuty detector that is specified by the detector
ID}
\usage{
guardduty_delete_detector(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that you want to delete.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes an Amazon GuardDuty detector that is specified by the detector
ID.
}
\section{Request syntax}{
\preformatted{svc$delete_detector(
  DetectorId = "string"
)
}
}

\keyword{internal}
