% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_ip_set}
\alias{guardduty_delete_ip_set}
\title{Deletes the IPSet specified by the IPSet ID}
\usage{
guardduty_delete_ip_set(DetectorId, IpSetId)
}
\arguments{
\item{DetectorId}{[required] The detectorID that specifies the GuardDuty service whose IPSet you want to delete.}

\item{IpSetId}{[required] The unique ID that specifies the IPSet that you want to delete.}
}
\description{
Deletes the IPSet specified by the IPSet ID.
}
\section{Request syntax}{
\preformatted{svc$delete_ip_set(
  DetectorId = "string",
  IpSetId = "string"
)
}
}

\keyword{internal}
