% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_describe_organization_configuration}
\alias{guardduty_describe_organization_configuration}
\title{Returns information about the account selected as the delegated
administrator for GuardDuty}
\usage{
guardduty_describe_organization_configuration(
  DetectorId,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{DetectorId}{[required] The detector ID of the delegated administrator for which you need to
retrieve the information.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{MaxResults}{You can use this parameter to indicate the maximum number of items that
you want in the response.}

\item{NextToken}{You can use this parameter when paginating results. Set the value of
this parameter to null on your first call to the list action. For
subsequent calls to the action, fill \code{nextToken} in the request with the
value of \code{NextToken} from the previous response to continue listing
data.}
}
\description{
Returns information about the account selected as the delegated administrator for GuardDuty.

See \url{https://www.paws-r-sdk.com/docs/guardduty_describe_organization_configuration/} for full documentation.
}
\keyword{internal}
