% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_disable_organization_admin_account}
\alias{guardduty_disable_organization_admin_account}
\title{Disables an AWS account within the Organization as the GuardDuty
delegated administrator}
\usage{
guardduty_disable_organization_admin_account(AdminAccountId)
}
\arguments{
\item{AdminAccountId}{[required] The AWS Account ID for the organizations account to be disabled as a
GuardDuty delegated administrator.}
}
\value{
An empty list.
}
\description{
Disables an AWS account within the Organization as the GuardDuty
delegated administrator.
}
\section{Request syntax}{
\preformatted{svc$disable_organization_admin_account(
  AdminAccountId = "string"
)
}
}

\keyword{internal}
