% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_disable_organization_admin_account}
\alias{guardduty_disable_organization_admin_account}
\title{Removes the existing GuardDuty delegated administrator of the
organization}
\usage{
guardduty_disable_organization_admin_account(AdminAccountId)
}
\arguments{
\item{AdminAccountId}{[required] The Amazon Web Services Account ID for the organizations account to be
disabled as a GuardDuty delegated administrator.}
}
\description{
Removes the existing GuardDuty delegated administrator of the organization. Only the organization's management account can run this API operation.

See \url{https://www.paws-r-sdk.com/docs/guardduty_disable_organization_admin_account/} for full documentation.
}
\keyword{internal}
