% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_disassociate_from_master_account}
\alias{guardduty_disassociate_from_master_account}
\title{Disassociates the current GuardDuty member account from its
administrator account}
\usage{
guardduty_disassociate_from_master_account(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty member account.}
}
\value{
An empty list.
}
\description{
Disassociates the current GuardDuty member account from its
administrator account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disassociate_from_master_account(
  DetectorId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
