% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_disassociate_from_master_account}
\alias{guardduty_disassociate_from_master_account}
\title{Disassociates the current GuardDuty member account from its
administrator account}
\usage{
guardduty_disassociate_from_master_account(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty member account.}
}
\description{
Disassociates the current GuardDuty member account from its administrator account.

See \url{https://www.paws-r-sdk.com/docs/guardduty_disassociate_from_master_account/} for full documentation.
}
\keyword{internal}
