% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_disassociate_members}
\alias{guardduty_disassociate_members}
\title{Disassociates GuardDuty member accounts (from the current administrator
account) specified by the account IDs}
\usage{
guardduty_disassociate_members(DetectorId, AccountIds)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty account whose members you
want to disassociate from the administrator account.}

\item{AccountIds}{[required] A list of account IDs of the GuardDuty member accounts that you want to
disassociate from the administrator account.}
}
\description{
Disassociates GuardDuty member accounts (from the current administrator account) specified by the account IDs.

See \url{https://www.paws-r-sdk.com/docs/guardduty_disassociate_members/} for full documentation.
}
\keyword{internal}
