% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_enable_organization_admin_account}
\alias{guardduty_enable_organization_admin_account}
\title{Enables an AWS account within the organization as the GuardDuty
delegated administrator}
\usage{
guardduty_enable_organization_admin_account(AdminAccountId)
}
\arguments{
\item{AdminAccountId}{[required] The AWS Account ID for the organization account to be enabled as a
GuardDuty delegated administrator.}
}
\value{
An empty list.
}
\description{
Enables an AWS account within the organization as the GuardDuty
delegated administrator.
}
\section{Request syntax}{
\preformatted{svc$enable_organization_admin_account(
  AdminAccountId = "string"
)
}
}

\keyword{internal}
