% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_detector}
\alias{guardduty_get_detector}
\title{Retrieves an Amazon GuardDuty detector specified by the detectorId}
\usage{
guardduty_get_detector(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that you want to get.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  CreatedAt = "string",
  FindingPublishingFrequency = "FIFTEEN_MINUTES"|"ONE_HOUR"|"SIX_HOURS",
  ServiceRole = "string",
  Status = "ENABLED"|"DISABLED",
  UpdatedAt = "string",
  DataSources = list(
    CloudTrail = list(
      Status = "ENABLED"|"DISABLED"
    ),
    DNSLogs = list(
      Status = "ENABLED"|"DISABLED"
    ),
    FlowLogs = list(
      Status = "ENABLED"|"DISABLED"
    ),
    S3Logs = list(
      Status = "ENABLED"|"DISABLED"
    )
  ),
  Tags = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves an Amazon GuardDuty detector specified by the detectorId.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_detector(
  DetectorId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
