% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_detector}
\alias{guardduty_get_detector}
\title{Retrieves an Amazon GuardDuty detector specified by the detectorId}
\usage{
guardduty_get_detector(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that you want to retrieve.}
}
\description{
Retrieves an Amazon GuardDuty detector specified by the detectorId.
}
\section{Request syntax}{
\preformatted{svc$get_detector(
  DetectorId = "string"
)
}
}

\keyword{internal}
