% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_filter}
\alias{guardduty_get_filter}
\title{Returns the details of the filter specified by the filter name}
\usage{
guardduty_get_filter(DetectorId, FilterName)
}
\arguments{
\item{DetectorId}{[required] The detector ID that specifies the GuardDuty service where you want to list the details of the specified filter.}

\item{FilterName}{[required] The name of the filter whose details you want to get.}
}
\description{
Returns the details of the filter specified by the filter name.
}
\section{Request syntax}{
\preformatted{svc$get_filter(
  DetectorId = "string",
  FilterName = "string"
)
}
}

\keyword{internal}
