% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_filter}
\alias{guardduty_get_filter}
\title{Returns the details of the filter specified by the filter name}
\usage{
guardduty_get_filter(DetectorId, FilterName)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that the filter is associated with.}

\item{FilterName}{[required] The name of the filter you want to get.}
}
\description{
Returns the details of the filter specified by the filter name.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_filter/} for full documentation.
}
\keyword{internal}
