% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_findings}
\alias{guardduty_get_findings}
\title{Describes Amazon GuardDuty findings specified by finding IDs}
\usage{
guardduty_get_findings(DetectorId, FindingIds, SortCriteria = NULL)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector that specifies the GuardDuty service whose
findings you want to retrieve.}

\item{FindingIds}{[required] The IDs of the findings that you want to retrieve.}

\item{SortCriteria}{Represents the criteria used for sorting findings.}
}
\description{
Describes Amazon GuardDuty findings specified by finding IDs.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_findings/} for full documentation.
}
\keyword{internal}
