% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_findings_statistics}
\alias{guardduty_get_findings_statistics}
\title{Lists Amazon GuardDuty findings statistics for the specified detector ID}
\usage{
guardduty_get_findings_statistics(
  DetectorId,
  FindingStatisticTypes,
  FindingCriteria = NULL
)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector that specifies the GuardDuty service whose
findings' statistics you want to retrieve.}

\item{FindingStatisticTypes}{[required] The types of finding statistics to retrieve.}

\item{FindingCriteria}{Represents the criteria that is used for querying findings.}
}
\description{
Lists Amazon GuardDuty findings statistics for the specified detector ID.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_findings_statistics/} for full documentation.
}
\keyword{internal}
