% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_invitations_count}
\alias{guardduty_get_invitations_count}
\title{Returns the count of all GuardDuty membership invitations that were sent
to the current member account except the currently accepted invitation}
\usage{
guardduty_get_invitations_count()
}
\description{
Returns the count of all GuardDuty membership invitations that were sent to the current member account except the currently accepted invitation.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_invitations_count/} for full documentation.
}
\keyword{internal}
