% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_ip_set}
\alias{guardduty_get_ip_set}
\title{Retrieves the IPSet specified by the ipSetId}
\usage{
guardduty_get_ip_set(DetectorId, IpSetId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that the IPSet is associated with.}

\item{IpSetId}{[required] The unique ID of the IPSet to retrieve.}
}
\description{
Retrieves the IPSet specified by the \code{ipSetId}.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_ip_set/} for full documentation.
}
\keyword{internal}
