% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_malware_scan_settings}
\alias{guardduty_get_malware_scan_settings}
\title{Returns the details of the malware scan settings}
\usage{
guardduty_get_malware_scan_settings(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that the scan setting is associated with.}
}
\description{
Returns the details of the malware scan settings.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_malware_scan_settings/} for full documentation.
}
\keyword{internal}
