% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_master_account}
\alias{guardduty_get_master_account}
\title{Provides the details for the GuardDuty administrator account associated
with the current GuardDuty member account}
\usage{
guardduty_get_master_account(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty member account.}
}
\value{
A list with the following syntax:\preformatted{list(
  Master = list(
    AccountId = "string",
    InvitationId = "string",
    RelationshipStatus = "string",
    InvitedAt = "string"
  )
)
}
}
\description{
Provides the details for the GuardDuty administrator account associated
with the current GuardDuty member account.
}
\section{Request syntax}{
\preformatted{svc$get_master_account(
  DetectorId = "string"
)
}
}

\keyword{internal}
