% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_master_account}
\alias{guardduty_get_master_account}
\title{Provides the details for the GuardDuty administrator account associated
with the current GuardDuty member account}
\usage{
guardduty_get_master_account(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty member account.}
}
\description{
Provides the details for the GuardDuty administrator account associated with the current GuardDuty member account.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_master_account/} for full documentation.
}
\keyword{internal}
