% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_member_detectors}
\alias{guardduty_get_member_detectors}
\title{Describes which data sources are enabled for the member account's
detector}
\usage{
guardduty_get_member_detectors(DetectorId, AccountIds)
}
\arguments{
\item{DetectorId}{[required] The detector ID for the administrator account.}

\item{AccountIds}{[required] The account ID of the member account.}
}
\description{
Describes which data sources are enabled for the member account's detector.

See \url{https://paws-r.github.io/docs/guardduty/get_member_detectors.html} for full documentation.
}
\keyword{internal}
