% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_remaining_free_trial_days}
\alias{guardduty_get_remaining_free_trial_days}
\title{Provides the number of days left for each data source used in the free
trial period}
\usage{
guardduty_get_remaining_free_trial_days(DetectorId, AccountIds)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty member account.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{AccountIds}{[required] A list of account identifiers of the GuardDuty member account.}
}
\description{
Provides the number of days left for each data source used in the free trial period.

See \url{https://www.paws-r-sdk.com/docs/guardduty_get_remaining_free_trial_days/} for full documentation.
}
\keyword{internal}
