% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_tags_for_resource}
\alias{guardduty_list_tags_for_resource}
\title{Lists tags for a resource}
\usage{
guardduty_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the given GuardDuty resource.}
}
\description{
Lists tags for a resource. Tagging is currently supported for detectors, finding filters, IP sets, threat intel sets, and publishing destination, with a limit of 50 tags per resource. When invoked, this operation returns all assigned tags for a given resource.

See \url{https://www.paws-r-sdk.com/docs/guardduty_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
