% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_start_malware_scan}
\alias{guardduty_start_malware_scan}
\title{Initiates the malware scan}
\usage{
guardduty_start_malware_scan(
  ResourceArn,
  ClientToken = NULL,
  ScanConfiguration = NULL
)
}
\arguments{
\item{ResourceArn}{[required] Amazon Resource Name (ARN) of the resource for which you invoked the
API.}

\item{ClientToken}{The idempotency token for the create request.}

\item{ScanConfiguration}{Contains information about the configuration to be used for the malware
scan.}
}
\description{
Initiates the malware scan. Invoking this API will automatically create the \href{https://docs.aws.amazon.com/guardduty/latest/ug/slr-permissions-malware-protection.html}{Service-linked role} in the corresponding account if the resourceArn belongs to an EC2 instance.

See \url{https://www.paws-r-sdk.com/docs/guardduty_start_malware_scan/} for full documentation.
}
\keyword{internal}
