% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_stop_monitoring_members}
\alias{guardduty_stop_monitoring_members}
\title{Stops GuardDuty monitoring for the specified member accounts}
\usage{
guardduty_stop_monitoring_members(DetectorId, AccountIds)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector associated with the GuardDuty
administrator account that is monitoring member accounts.}

\item{AccountIds}{[required] A list of account IDs for the member accounts to stop monitoring.}
}
\description{
Stops GuardDuty monitoring for the specified member accounts. Use the \code{\link[=guardduty_start_monitoring_members]{start_monitoring_members}} operation to restart monitoring for those accounts.

See \url{https://www.paws-r-sdk.com/docs/guardduty_stop_monitoring_members/} for full documentation.
}
\keyword{internal}
