% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_unarchive_findings}
\alias{guardduty_unarchive_findings}
\title{Unarchives Amazon GuardDuty findings specified by the list of finding
IDs}
\usage{
guardduty_unarchive_findings(DetectorId, FindingIds)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector that specifies the GuardDuty service whose findings you want to unarchive.}

\item{FindingIds}{[required] IDs of the findings that you want to unarchive.}
}
\description{
Unarchives Amazon GuardDuty findings specified by the list of finding IDs.
}
\section{Request syntax}{
\preformatted{svc$unarchive_findings(
  DetectorId = "string",
  FindingIds = list(
    "string"
  )
)
}
}

\keyword{internal}
