% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_unarchive_findings}
\alias{guardduty_unarchive_findings}
\title{Unarchives GuardDuty findings specified by the findingIds}
\usage{
guardduty_unarchive_findings(DetectorId, FindingIds)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector associated with the findings to unarchive.}

\item{FindingIds}{[required] The IDs of the findings to unarchive.}
}
\value{
An empty list.
}
\description{
Unarchives GuardDuty findings specified by the \code{findingIds}.
}
\section{Request syntax}{
\preformatted{svc$unarchive_findings(
  DetectorId = "string",
  FindingIds = list(
    "string"
  )
)
}
}

\keyword{internal}
