% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_untag_resource}
\alias{guardduty_untag_resource}
\title{Removes tags from a resource}
\usage{
guardduty_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) for the resource to remove tags from.}

\item{TagKeys}{[required] The tag keys to remove from the resource.}
}
\description{
Removes tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/guardduty_untag_resource/} for full documentation.
}
\keyword{internal}
