% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_ip_set}
\alias{guardduty_update_ip_set}
\title{Updates the IPSet specified by the IPSet ID}
\usage{
guardduty_update_ip_set(
  DetectorId,
  IpSetId,
  Name = NULL,
  Location = NULL,
  Activate = NULL,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{DetectorId}{[required] The detectorID that specifies the GuardDuty service whose IPSet you want
to update.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{IpSetId}{[required] The unique ID that specifies the IPSet that you want to update.}

\item{Name}{The unique ID that specifies the IPSet that you want to update.}

\item{Location}{The updated URI of the file that contains the IPSet.}

\item{Activate}{The updated Boolean value that specifies whether the IPSet is active or
not.}

\item{ExpectedBucketOwner}{The Amazon Web Services account ID that owns the Amazon S3 bucket
specified in the \strong{location} parameter.}
}
\description{
Updates the IPSet specified by the IPSet ID.

See \url{https://www.paws-r-sdk.com/docs/guardduty_update_ip_set/} for full documentation.
}
\keyword{internal}
