% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_member_detectors}
\alias{guardduty_update_member_detectors}
\title{Contains information on member accounts to be updated}
\usage{
guardduty_update_member_detectors(DetectorId, AccountIds, DataSources)
}
\arguments{
\item{DetectorId}{[required] The detector ID of the administrator account.}

\item{AccountIds}{[required] A list of member account IDs to be updated.}

\item{DataSources}{Describes which data sources will be updated.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UnprocessedAccounts = list(
    list(
      AccountId = "string",
      Result = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Contains information on member accounts to be updated.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_member_detectors(
  DetectorId = "string",
  AccountIds = list(
    "string"
  ),
  DataSources = list(
    S3Logs = list(
      Enable = TRUE|FALSE
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
