% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_publishing_destination}
\alias{guardduty_update_publishing_destination}
\title{Updates information about the publishing destination specified by the
destinationId}
\usage{
guardduty_update_publishing_destination(
  DetectorId,
  DestinationId,
  DestinationProperties = NULL
)
}
\arguments{
\item{DetectorId}{[required] The ID of the detector associated with the publishing destinations to
update.}

\item{DestinationId}{[required] The ID of the publishing destination to update.}

\item{DestinationProperties}{A \code{DestinationProperties} object that includes the \code{DestinationArn} and
\code{KmsKeyArn} of the publishing destination.}
}
\description{
Updates information about the publishing destination specified by the \code{destinationId}.

See \url{https://www.paws-r-sdk.com/docs/guardduty_update_publishing_destination/} for full documentation.
}
\keyword{internal}
