% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_threat_intel_set}
\alias{guardduty_update_threat_intel_set}
\title{Updates the ThreatIntelSet specified by the ThreatIntelSet ID}
\usage{
guardduty_update_threat_intel_set(
  DetectorId,
  ThreatIntelSetId,
  Name = NULL,
  Location = NULL,
  Activate = NULL,
  ExpectedBucketOwner = NULL
)
}
\arguments{
\item{DetectorId}{[required] The detectorID that specifies the GuardDuty service whose ThreatIntelSet
you want to update.

To find the \code{detectorId} in the current Region, see the Settings page in
the GuardDuty console, or run the
\code{\link[=guardduty_list_detectors]{list_detectors}} API.}

\item{ThreatIntelSetId}{[required] The unique ID that specifies the ThreatIntelSet that you want to update.}

\item{Name}{The unique ID that specifies the ThreatIntelSet that you want to update.}

\item{Location}{The updated URI of the file that contains the ThreateIntelSet.}

\item{Activate}{The updated Boolean value that specifies whether the ThreateIntelSet is
active or not.}

\item{ExpectedBucketOwner}{The Amazon Web Services account ID that owns the Amazon S3 bucket
specified in the \strong{location} parameter.}
}
\description{
Updates the ThreatIntelSet specified by the ThreatIntelSet ID.

See \url{https://www.paws-r-sdk.com/docs/guardduty_update_threat_intel_set/} for full documentation.
}
\keyword{internal}
