% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_add_user_to_group}
\alias{iam_add_user_to_group}
\title{Adds the specified user to the specified group}
\usage{
iam_add_user_to_group(GroupName, UserName)
}
\arguments{
\item{GroupName}{[required] The name of the group to update.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}

\item{UserName}{[required] The name of the user to add.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Adds the specified user to the specified group.

See \url{https://www.paws-r-sdk.com/docs/iam_add_user_to_group/} for full documentation.
}
\keyword{internal}
