% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_create_access_key}
\alias{iam_create_access_key}
\title{Creates a new Amazon Web Services secret access key and corresponding
Amazon Web Services access key ID for the specified user}
\usage{
iam_create_access_key(UserName = NULL)
}
\arguments{
\item{UserName}{The name of the IAM user that the new key will belong to.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Creates a new Amazon Web Services secret access key and corresponding Amazon Web Services access key ID for the specified user. The default status for new keys is \code{Active}.

See \url{https://www.paws-r-sdk.com/docs/iam_create_access_key/} for full documentation.
}
\keyword{internal}
