% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_create_login_profile}
\alias{iam_create_login_profile}
\title{Creates a password for the specified IAM user}
\usage{
iam_create_login_profile(UserName, Password, PasswordResetRequired)
}
\arguments{
\item{UserName}{[required] The name of the IAM user to create a password for. The user must already
exist.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}

\item{Password}{[required] The new password for the user.

The \href{https://en.wikipedia.org/wiki/Regex}{regex pattern} that is used to
validate this parameter is a string of characters. That string can
include almost any printable ASCII character from the space (\verb{\\u0020})
through the end of the ASCII character range (\verb{\\u00FF}). You can also
include the tab (\verb{\\u0009}), line feed (\verb{\\u000A}), and carriage return
(\verb{\\u000D}) characters. Any of these characters are valid in a password.
However, many tools, such as the AWS Management Console, might restrict
the ability to type certain characters because they have special meaning
within that tool.}

\item{PasswordResetRequired}{Specifies whether the user is required to set a new password on next
sign-in.}
}
\value{
A list with the following syntax:\preformatted{list(
  LoginProfile = list(
    UserName = "string",
    CreateDate = as.POSIXct(
      "2015-01-01"
    ),
    PasswordResetRequired = TRUE|FALSE
  )
)
}
}
\description{
Creates a password for the specified IAM user. A password allows an IAM
user to access AWS services through the AWS Management Console.

You can use the AWS CLI, the AWS API, or the \strong{Users} page in the IAM
console to create a password for any IAM user. Use
\code{\link[=iam_change_password]{change_password}} to update your own existing
password in the \strong{My Security Credentials} page in the AWS Management
Console.

For more information about managing passwords, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_passwords_admin-change-user.html}{Managing passwords}
in the \emph{IAM User Guide}.
}
\section{Request syntax}{
\preformatted{svc$create_login_profile(
  UserName = "string",
  Password = "string",
  PasswordResetRequired = TRUE|FALSE
)
}
}

\examples{
\dontrun{
# The following command changes IAM user Bob's password and sets the flag
# that required Bob to change the password the next time he signs in.
svc$create_login_profile(
  Password = "h]6EszR\}vJ*m",
  PasswordResetRequired = TRUE,
  UserName = "Bob"
)
}

}
\keyword{internal}
