% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_create_service_specific_credential}
\alias{iam_create_service_specific_credential}
\title{Generates a set of credentials consisting of a user name and password
that can be used to access the service specified in the request}
\usage{
iam_create_service_specific_credential(UserName, ServiceName)
}
\arguments{
\item{UserName}{[required] The name of the IAM user that is to be associated with the credentials.
The new service-specific credentials have the same permissions as the
associated user except that they can be used only to access the
specified service.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}

\item{ServiceName}{[required] The name of the AWS service that is to be associated with the
credentials. The service you specify here is the only service that can
be accessed using these credentials.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  ServiceSpecificCredential = list(
    CreateDate = as.POSIXct(
      "2015-01-01"
    ),
    ServiceName = "string",
    ServiceUserName = "string",
    ServicePassword = "string",
    ServiceSpecificCredentialId = "string",
    UserName = "string",
    Status = "Active"|"Inactive"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Generates a set of credentials consisting of a user name and password
that can be used to access the service specified in the request. These
credentials are generated by IAM, and can be used only for the specified
service.

You can have a maximum of two sets of service-specific credentials for
each supported service per user.

The only supported service at this time is AWS CodeCommit.

You can reset the password to a new service-generated value by calling
\code{\link[=iam_reset_service_specific_credential]{reset_service_specific_credential}}.

For more information about service-specific credentials, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_ssh-keys.html}{Using IAM with AWS CodeCommit: Git Credentials, SSH Keys, and AWS Access Keys}
in the \emph{IAM User Guide}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_service_specific_credential(
  UserName = "string",
  ServiceName = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
