% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_group_policy}
\alias{iam_delete_group_policy}
\title{Deletes the specified inline policy that is embedded in the specified
IAM group}
\usage{
iam_delete_group_policy(GroupName, PolicyName)
}
\arguments{
\item{GroupName}{[required] The name (friendly name, not ARN) identifying the group that the policy
is embedded in.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}

\item{PolicyName}{[required] The name identifying the policy document to delete.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Deletes the specified inline policy that is embedded in the specified IAM group.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_group_policy/} for full documentation.
}
\keyword{internal}
