% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_instance_profile}
\alias{iam_delete_instance_profile}
\title{Deletes the specified instance profile}
\usage{
iam_delete_instance_profile(InstanceProfileName)
}
\arguments{
\item{InstanceProfileName}{[required] The name of the instance profile to delete.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Deletes the specified instance profile. The instance profile must not have an associated role.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_instance_profile/} for full documentation.
}
\keyword{internal}
