% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_open_id_connect_provider}
\alias{iam_delete_open_id_connect_provider}
\title{Deletes an OpenID Connect identity provider (IdP) resource object in IAM}
\usage{
iam_delete_open_id_connect_provider(OpenIDConnectProviderArn)
}
\arguments{
\item{OpenIDConnectProviderArn}{[required] The Amazon Resource Name (ARN) of the IAM OpenID Connect provider
resource object to delete. You can get a list of OpenID Connect provider
resource ARNs by using the
\code{\link[=iam_list_open_id_connect_providers]{list_open_id_connect_providers}}
operation.}
}
\description{
Deletes an OpenID Connect identity provider (IdP) resource object in IAM.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_open_id_connect_provider/} for full documentation.
}
\keyword{internal}
