% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_policy}
\alias{iam_delete_policy}
\title{Deletes the specified managed policy}
\usage{
iam_delete_policy(PolicyArn)
}
\arguments{
\item{PolicyArn}{[required] The Amazon Resource Name (ARN) of the IAM policy you want to delete.

For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{Amazon Web Services General Reference}.}
}
\description{
Deletes the specified managed policy.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_policy/} for full documentation.
}
\keyword{internal}
