% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_role}
\alias{iam_delete_role}
\title{Deletes the specified role}
\usage{
iam_delete_role(RoleName)
}
\arguments{
\item{RoleName}{[required] The name of the role to delete.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Deletes the specified role. Unlike the Amazon Web Services Management Console, when you delete a role programmatically, you must delete the items attached to the role manually, or the deletion fails. For more information, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_manage_delete.html#roles-managingrole-deleting-cli}{Deleting an IAM role}. Before attempting to delete a role, remove the following attached items:

See \url{https://www.paws-r-sdk.com/docs/iam_delete_role/} for full documentation.
}
\keyword{internal}
