% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_role_policy}
\alias{iam_delete_role_policy}
\title{Deletes the specified inline policy that is embedded in the specified
IAM role}
\usage{
iam_delete_role_policy(RoleName, PolicyName)
}
\arguments{
\item{RoleName}{[required] The name (friendly name, not ARN) identifying the role that the policy
is embedded in.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}

\item{PolicyName}{[required] The name of the inline policy to delete from the specified IAM role.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Deletes the specified inline policy that is embedded in the specified IAM role.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_role_policy/} for full documentation.
}
\keyword{internal}
