% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_server_certificate}
\alias{iam_delete_server_certificate}
\title{Deletes the specified server certificate}
\usage{
iam_delete_server_certificate(ServerCertificateName)
}
\arguments{
\item{ServerCertificateName}{[required] The name of the server certificate you want to delete.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Deletes the specified server certificate.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_server_certificate/} for full documentation.
}
\keyword{internal}
