% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_service_linked_role}
\alias{iam_delete_service_linked_role}
\title{Submits a service-linked role deletion request and returns a
DeletionTaskId, which you can use to check the status of the deletion}
\usage{
iam_delete_service_linked_role(RoleName)
}
\arguments{
\item{RoleName}{[required] The name of the service-linked role to be deleted.}
}
\description{
Submits a service-linked role deletion request and returns a \code{DeletionTaskId}, which you can use to check the status of the deletion. Before you call this operation, confirm that the role has no active sessions and that any resources used by the role in the linked service are deleted. If you call this operation more than once for the same service-linked role and an earlier deletion task is not complete, then the \code{DeletionTaskId} of the earlier request is returned.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_service_linked_role/} for full documentation.
}
\keyword{internal}
