% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_ssh_public_key}
\alias{iam_delete_ssh_public_key}
\title{Deletes the specified SSH public key}
\usage{
iam_delete_ssh_public_key(UserName, SSHPublicKeyId)
}
\arguments{
\item{UserName}{[required] The name of the IAM user associated with the SSH public key.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}

\item{SSHPublicKeyId}{[required] The unique identifier for the SSH public key.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
that can consist of any upper or lowercased letter or digit.}
}
\description{
Deletes the specified SSH public key.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_ssh_public_key/} for full documentation.
}
\keyword{internal}
