% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_delete_user_permissions_boundary}
\alias{iam_delete_user_permissions_boundary}
\title{Deletes the permissions boundary for the specified IAM user}
\usage{
iam_delete_user_permissions_boundary(UserName)
}
\arguments{
\item{UserName}{[required] The name (friendly name, not ARN) of the IAM user from which you want to
remove the permissions boundary.}
}
\description{
Deletes the permissions boundary for the specified IAM user.

See \url{https://www.paws-r-sdk.com/docs/iam_delete_user_permissions_boundary/} for full documentation.
}
\keyword{internal}
