% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_account_password_policy}
\alias{iam_get_account_password_policy}
\title{Retrieves the password policy for the Amazon Web Services account}
\usage{
iam_get_account_password_policy()
}
\description{
Retrieves the password policy for the Amazon Web Services account. This tells you the complexity requirements and mandatory rotation periods for the IAM user passwords in your account. For more information about using a password policy, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_passwords_account-policy.html}{Managing an IAM password policy}.

See \url{https://www.paws-r-sdk.com/docs/iam_get_account_password_policy/} for full documentation.
}
\keyword{internal}
