% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_account_summary}
\alias{iam_get_account_summary}
\title{Retrieves information about IAM entity usage and IAM quotas in the
Amazon Web Services account}
\usage{
iam_get_account_summary()
}
\description{
Retrieves information about IAM entity usage and IAM quotas in the Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/iam_get_account_summary/} for full documentation.
}
\keyword{internal}
