% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_credential_report}
\alias{iam_get_credential_report}
\title{Retrieves a credential report for the Amazon Web Services account}
\usage{
iam_get_credential_report()
}
\description{
Retrieves a credential report for the Amazon Web Services account. For more information about the credential report, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_getting-report.html}{Getting credential reports} in the \emph{IAM User Guide}.

See \url{https://www.paws-r-sdk.com/docs/iam_get_credential_report/} for full documentation.
}
\keyword{internal}
