% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_open_id_connect_provider}
\alias{iam_get_open_id_connect_provider}
\title{Returns information about the specified OpenID Connect (OIDC) provider
resource object in IAM}
\usage{
iam_get_open_id_connect_provider(OpenIDConnectProviderArn)
}
\arguments{
\item{OpenIDConnectProviderArn}{[required] The Amazon Resource Name (ARN) of the OIDC provider resource object in
IAM to get information for. You can get a list of OIDC provider resource
ARNs by using the
\code{\link[=iam_list_open_id_connect_providers]{list_open_id_connect_providers}}
operation.

For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{Amazon Web Services General Reference}.}
}
\description{
Returns information about the specified OpenID Connect (OIDC) provider resource object in IAM.

See \url{https://www.paws-r-sdk.com/docs/iam_get_open_id_connect_provider/} for full documentation.
}
\keyword{internal}
