% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_saml_provider}
\alias{iam_get_saml_provider}
\title{Returns the SAML provider metadocument that was uploaded when the IAM
SAML provider resource object was created or updated}
\usage{
iam_get_saml_provider(SAMLProviderArn)
}
\arguments{
\item{SAMLProviderArn}{[required] The Amazon Resource Name (ARN) of the SAML provider resource object in
IAM to get information about.

For more information about ARNs, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Names (ARNs)}
in the \emph{Amazon Web Services General Reference}.}
}
\description{
Returns the SAML provider metadocument that was uploaded when the IAM SAML provider resource object was created or updated.

See \url{https://www.paws-r-sdk.com/docs/iam_get_saml_provider/} for full documentation.
}
\keyword{internal}
