% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_get_user}
\alias{iam_get_user}
\title{Retrieves information about the specified IAM user, including the user's
creation date, path, unique ID, and ARN}
\usage{
iam_get_user(UserName = NULL)
}
\arguments{
\item{UserName}{The name of the user to get information about.

This parameter is optional. If it is not included, it defaults to the
user making the request. This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Retrieves information about the specified IAM user, including the user's creation date, path, unique ID, and ARN.

See \url{https://www.paws-r-sdk.com/docs/iam_get_user/} for full documentation.
}
\keyword{internal}
