% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_list_open_id_connect_providers}
\alias{iam_list_open_id_connect_providers}
\title{Lists information about the IAM OpenID Connect (OIDC) provider resource
objects defined in the Amazon Web Services account}
\usage{
iam_list_open_id_connect_providers()
}
\description{
Lists information about the IAM OpenID Connect (OIDC) provider resource objects defined in the Amazon Web Services account.

See \url{https://paws-r.github.io/docs/iam/list_open_id_connect_providers.html} for full documentation.
}
\keyword{internal}
