% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_list_saml_providers}
\alias{iam_list_saml_providers}
\title{Lists the SAML provider resource objects defined in IAM in the account}
\usage{
iam_list_saml_providers()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  SAMLProviderList = list(
    list(
      Arn = "string",
      ValidUntil = as.POSIXct(
        "2015-01-01"
      ),
      CreateDate = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Lists the SAML provider resource objects defined in IAM in the account.

This operation requires \href{https://docs.aws.amazon.com/general/latest/gr/signature-version-4.html}{Signature Version 4}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_saml_providers()
}\if{html}{\out{</div>}}
}

\keyword{internal}
