% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_remove_role_from_instance_profile}
\alias{iam_remove_role_from_instance_profile}
\title{Removes the specified IAM role from the specified Amazon EC2 instance
profile}
\usage{
iam_remove_role_from_instance_profile(InstanceProfileName, RoleName)
}
\arguments{
\item{InstanceProfileName}{[required] The name of the instance profile to update.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}

\item{RoleName}{[required] The name of the role to remove.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}
}
\description{
Removes the specified IAM role from the specified Amazon EC2 instance profile.

See \url{https://www.paws-r-sdk.com/docs/iam_remove_role_from_instance_profile/} for full documentation.
}
\keyword{internal}
