% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_set_security_token_service_preferences}
\alias{iam_set_security_token_service_preferences}
\title{Sets the specified version of the global endpoint token as the token
version used for the Amazon Web Services account}
\usage{
iam_set_security_token_service_preferences(GlobalEndpointTokenVersion)
}
\arguments{
\item{GlobalEndpointTokenVersion}{[required] The version of the global endpoint token. Version 1 tokens are valid
only in Amazon Web Services Regions that are available by default. These
tokens do not work in manually enabled Regions, such as Asia Pacific
(Hong Kong). Version 2 tokens are valid in all Regions. However, version
2 tokens are longer and might affect systems where you temporarily store
tokens.

For information, see \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/id_credentials_temp_enable-regions.html}{Activating and deactivating STS in an Amazon Web Services Region}
in the \emph{IAM User Guide}.}
}
\description{
Sets the specified version of the global endpoint token as the token version used for the Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/iam_set_security_token_service_preferences/} for full documentation.
}
\keyword{internal}
