% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_update_role_description}
\alias{iam_update_role_description}
\title{Use UpdateRole instead}
\usage{
iam_update_role_description(RoleName, Description)
}
\arguments{
\item{RoleName}{[required] The name of the role that you want to modify.}

\item{Description}{[required] The new description that you want to apply to the specified role.}
}
\description{
Use \code{\link[=iam_update_role]{update_role}} instead.

See \url{https://www.paws-r-sdk.com/docs/iam_update_role_description/} for full documentation.
}
\keyword{internal}
