% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iam_operations.R
\name{iam_update_ssh_public_key}
\alias{iam_update_ssh_public_key}
\title{Sets the status of an IAM user's SSH public key to active or inactive}
\usage{
iam_update_ssh_public_key(UserName, SSHPublicKeyId, Status)
}
\arguments{
\item{UserName}{[required] The name of the IAM user associated with the SSH public key.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
consisting of upper and lowercase alphanumeric characters with no
spaces. You can also include any of the following characters: _+=,.@-}

\item{SSHPublicKeyId}{[required] The unique identifier for the SSH public key.

This parameter allows (through its \href{https://en.wikipedia.org/wiki/Regex}{regex pattern}) a string of characters
that can consist of any upper or lowercased letter or digit.}

\item{Status}{[required] The status to assign to the SSH public key. \code{Active} means that the key
can be used for authentication with an CodeCommit repository. \code{Inactive}
means that the key cannot be used.}
}
\description{
Sets the status of an IAM user's SSH public key to active or inactive. SSH public keys that are inactive cannot be used for authentication. This operation can be used to disable a user's SSH public key as part of a key rotation work flow.

See \url{https://www.paws-r-sdk.com/docs/iam_update_ssh_public_key/} for full documentation.
}
\keyword{internal}
