% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_disable_profile}
\alias{iamrolesanywhere_disable_profile}
\title{Disables a profile}
\usage{
iamrolesanywhere_disable_profile(profileId)
}
\arguments{
\item{profileId}{[required] The unique identifier of the profile.}
}
\description{
Disables a profile. When disabled, temporary credential requests with this profile fail.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_disable_profile/} for full documentation.
}
\keyword{internal}
