% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_enable_profile}
\alias{iamrolesanywhere_enable_profile}
\title{Enables temporary credential requests for a profile}
\usage{
iamrolesanywhere_enable_profile(profileId)
}
\arguments{
\item{profileId}{[required] The unique identifier of the profile.}
}
\description{
Enables temporary credential requests for a profile.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_enable_profile/} for full documentation.
}
\keyword{internal}
