% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_get_crl}
\alias{iamrolesanywhere_get_crl}
\title{Gets a certificate revocation list (CRL)}
\usage{
iamrolesanywhere_get_crl(crlId)
}
\arguments{
\item{crlId}{[required] The unique identifier of the certificate revocation list (CRL).}
}
\description{
Gets a certificate revocation list (CRL).

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_get_crl/} for full documentation.
}
\keyword{internal}
