% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_get_subject}
\alias{iamrolesanywhere_get_subject}
\title{Gets a subject, which associates a certificate identity with
authentication attempts}
\usage{
iamrolesanywhere_get_subject(subjectId)
}
\arguments{
\item{subjectId}{[required] The unique identifier of the subject.}
}
\description{
Gets a \emph{subject}, which associates a certificate identity with authentication attempts. The subject stores auditing information such as the status of the last authentication attempt, the certificate data used in the attempt, and the last time the associated identity attempted authentication.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_get_subject/} for full documentation.
}
\keyword{internal}
