% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_list_tags_for_resource}
\alias{iamrolesanywhere_list_tags_for_resource}
\title{Lists the tags attached to the resource}
\usage{
iamrolesanywhere_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource.}
}
\description{
Lists the tags attached to the resource.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
