% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_reset_notification_settings}
\alias{iamrolesanywhere_reset_notification_settings}
\title{Resets the custom notification setting to IAM Roles Anywhere default
setting}
\usage{
iamrolesanywhere_reset_notification_settings(
  trustAnchorId,
  notificationSettingKeys
)
}
\arguments{
\item{trustAnchorId}{[required] The unique identifier of the trust anchor.}

\item{notificationSettingKeys}{[required] A list of notification setting keys to reset. A notification setting key
includes the event and the channel.}
}
\description{
Resets the \emph{custom notification setting} to IAM Roles Anywhere default setting.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_reset_notification_settings/} for full documentation.
}
\keyword{internal}
