% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_update_crl}
\alias{iamrolesanywhere_update_crl}
\title{Updates the certificate revocation list (CRL)}
\usage{
iamrolesanywhere_update_crl(crlId, name = NULL, crlData = NULL)
}
\arguments{
\item{crlId}{[required] The unique identifier of the certificate revocation list (CRL).}

\item{name}{The name of the Crl.}

\item{crlData}{The x509 v3 specified certificate revocation list (CRL).}
}
\description{
Updates the certificate revocation list (CRL). A CRL is a list of certificates that have been revoked by the issuing certificate authority (CA). IAM Roles Anywhere validates against the CRL before issuing credentials.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_update_crl/} for full documentation.
}
\keyword{internal}
