% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_update_profile}
\alias{iamrolesanywhere_update_profile}
\title{Updates a profile, a list of the roles that IAM Roles Anywhere service
is trusted to assume}
\usage{
iamrolesanywhere_update_profile(
  profileId,
  name = NULL,
  sessionPolicy = NULL,
  roleArns = NULL,
  managedPolicyArns = NULL,
  durationSeconds = NULL,
  acceptRoleSessionName = NULL
)
}
\arguments{
\item{profileId}{[required] The unique identifier of the profile.}

\item{name}{The name of the profile.}

\item{sessionPolicy}{A session policy that applies to the trust boundary of the vended
session credentials.}

\item{roleArns}{A list of IAM roles that this profile can assume in a temporary
credential request.}

\item{managedPolicyArns}{A list of managed policy ARNs that apply to the vended session
credentials.}

\item{durationSeconds}{Used to determine how long sessions vended using this profile are valid
for. See the \code{Expiration} section of the \href{https://docs.aws.amazon.com/rolesanywhere/latest/userguide/authentication-create-session.html#credentials-object}{CreateSession API documentation}
page for more details. In requests, if this value is not provided, the
default value will be 3600.}

\item{acceptRoleSessionName}{Used to determine if a custom role session name will be accepted in a
temporary credential request.}
}
\description{
Updates a \emph{profile}, a list of the roles that IAM Roles Anywhere service is trusted to assume. You use profiles to intersect permissions with IAM managed policies.

See \url{https://www.paws-r-sdk.com/docs/iamrolesanywhere_update_profile/} for full documentation.
}
\keyword{internal}
