% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_create_group}
\alias{identitystore_create_group}
\title{Creates a group within the specified identity store}
\usage{
identitystore_create_group(
  IdentityStoreId,
  DisplayName = NULL,
  Description = NULL
)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.}

\item{DisplayName}{A string containing the name of the group. This value is commonly
displayed when the group is referenced. \code{Administrator} and
\code{AWSAdministrators} are reserved names and can't be used for users or
groups.}

\item{Description}{A string containing the description of the group.}
}
\description{
Creates a group within the specified identity store.

See \url{https://www.paws-r-sdk.com/docs/identitystore_create_group/} for full documentation.
}
\keyword{internal}
