% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_create_group_membership}
\alias{identitystore_create_group_membership}
\title{Creates a relationship between a member and a group}
\usage{
identitystore_create_group_membership(IdentityStoreId, GroupId, MemberId)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.}

\item{GroupId}{[required] The identifier for a group in the identity store.}

\item{MemberId}{[required] An object that contains the identifier of a group member. Setting the
\code{UserID} field to the specific identifier for a user indicates that the
user is a member of the group.}
}
\description{
Creates a relationship between a member and a group. The following identifiers must be specified: \code{GroupId}, \code{IdentityStoreId}, and \code{MemberId}.

See \url{https://www.paws-r-sdk.com/docs/identitystore_create_group_membership/} for full documentation.
}
\keyword{internal}
