% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_delete_group}
\alias{identitystore_delete_group}
\title{Delete a group within an identity store given GroupId}
\usage{
identitystore_delete_group(IdentityStoreId, GroupId)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.}

\item{GroupId}{[required] The identifier for a group in the identity store.}
}
\description{
Delete a group within an identity store given \code{GroupId}.

See \url{https://www.paws-r-sdk.com/docs/identitystore_delete_group/} for full documentation.
}
\keyword{internal}
