% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_delete_group_membership}
\alias{identitystore_delete_group_membership}
\title{Delete a membership within a group given MembershipId}
\usage{
identitystore_delete_group_membership(IdentityStoreId, MembershipId)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.}

\item{MembershipId}{[required] The identifier for a \code{GroupMembership} in an identity store.}
}
\description{
Delete a membership within a group given \code{MembershipId}.

See \url{https://www.paws-r-sdk.com/docs/identitystore_delete_group_membership/} for full documentation.
}
\keyword{internal}
