% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_delete_user}
\alias{identitystore_delete_user}
\title{Deletes a user within an identity store given UserId}
\usage{
identitystore_delete_user(IdentityStoreId, UserId)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.}

\item{UserId}{[required] The identifier for a user in the identity store.}
}
\description{
Deletes a user within an identity store given \code{UserId}.

See \url{https://www.paws-r-sdk.com/docs/identitystore_delete_user/} for full documentation.
}
\keyword{internal}
