% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_describe_group_membership}
\alias{identitystore_describe_group_membership}
\title{Retrieves membership metadata and attributes from MembershipId in an
identity store}
\usage{
identitystore_describe_group_membership(IdentityStoreId, MembershipId)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.}

\item{MembershipId}{[required] The identifier for a \code{GroupMembership} in an identity store.}
}
\description{
Retrieves membership metadata and attributes from \code{MembershipId} in an identity store.

See \url{https://www.paws-r-sdk.com/docs/identitystore_describe_group_membership/} for full documentation.
}
\keyword{internal}
