% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_get_group_id}
\alias{identitystore_get_group_id}
\title{Retrieves GroupId in an identity store}
\usage{
identitystore_get_group_id(IdentityStoreId, AlternateIdentifier)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.}

\item{AlternateIdentifier}{[required] A unique identifier for a user or group that is not the primary
identifier. This value can be an identifier from an external identity
provider (IdP) that is associated with the user, the group, or a unique
attribute. For the unique attribute, the only valid path is
\code{ displayName}.}
}
\description{
Retrieves \code{GroupId} in an identity store.

See \url{https://www.paws-r-sdk.com/docs/identitystore_get_group_id/} for full documentation.
}
\keyword{internal}
