% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_get_group_membership_id}
\alias{identitystore_get_group_membership_id}
\title{Retrieves the MembershipId in an identity store}
\usage{
identitystore_get_group_membership_id(IdentityStoreId, GroupId, MemberId)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.}

\item{GroupId}{[required] The identifier for a group in the identity store.}

\item{MemberId}{[required] An object that contains the identifier of a group member. Setting the
\code{UserID} field to the specific identifier for a user indicates that the
user is a member of the group.}
}
\description{
Retrieves the \code{MembershipId} in an identity store.

See \url{https://www.paws-r-sdk.com/docs/identitystore_get_group_membership_id/} for full documentation.
}
\keyword{internal}
