% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_is_member_in_groups}
\alias{identitystore_is_member_in_groups}
\title{Checks the user's membership in all requested groups and returns if the
member exists in all queried groups}
\usage{
identitystore_is_member_in_groups(IdentityStoreId, MemberId, GroupIds)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.}

\item{MemberId}{[required] An object containing the identifier of a group member.}

\item{GroupIds}{[required] A list of identifiers for groups in the identity store.}
}
\description{
Checks the user's membership in all requested groups and returns if the member exists in all queried groups.

See \url{https://www.paws-r-sdk.com/docs/identitystore_is_member_in_groups/} for full documentation.
}
\keyword{internal}
