% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_list_users}
\alias{identitystore_list_users}
\title{Lists the attribute name and value of the user that you specified in the
search}
\usage{
identitystore_list_users(
  IdentityStoreId,
  MaxResults = NULL,
  NextToken = NULL,
  Filters = NULL
)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store, such as
\code{d-1234567890}. In this example, \verb{d-} is a fixed prefix, and
\code{1234567890} is a randomly generated string that contains number and
lower case letters. This value is generated at the time that a new
identity store is created.}

\item{MaxResults}{The maximum number of results to be returned per request. This parameter
is used in the \code{\link[=identitystore_list_users]{list_users}} and
\code{\link[=identitystore_list_groups]{list_groups}} request to specify how many
results to return in one page. The length limit is 50 characters.}

\item{NextToken}{The pagination token used for the
\code{\link[=identitystore_list_users]{list_users}} and
\code{\link[=identitystore_list_groups]{list_groups}} API operations. This value is
generated by the identity store service. It is returned in the API
response if the total results are more than the size of one page. This
token is also returned when it is used in the API request to search for
the next page.}

\item{Filters}{A list of \code{Filter} objects, which is used in the
\code{\link[=identitystore_list_users]{list_users}} and
\code{\link[=identitystore_list_groups]{list_groups}} request.}
}
\description{
Lists the attribute name and value of the user that you specified in the search. We only support \code{UserName} as a valid filter attribute path currently, and filter is required. This API returns minimum attributes, including \code{UserId} and \code{UserName} in the response.

See \url{https://paws-r.github.io/docs/identitystore/list_users.html} for full documentation.
}
\keyword{internal}
