% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_update_user}
\alias{identitystore_update_user}
\title{For the specified user in the specified identity store, updates the user
metadata and attributes}
\usage{
identitystore_update_user(IdentityStoreId, UserId, Operations)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.}

\item{UserId}{[required] The identifier for a user in the identity store.}

\item{Operations}{[required] A list of \code{AttributeOperation} objects to apply to the requested user.
These operations might add, replace, or remove an attribute.}
}
\description{
For the specified user in the specified identity store, updates the user metadata and attributes.

See \url{https://www.paws-r-sdk.com/docs/identitystore_update_user/} for full documentation.
}
\keyword{internal}
