% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identitystore_operations.R
\name{identitystore_update_user}
\alias{identitystore_update_user}
\title{Updates the specified user metadata and attributes in the specified
identity store}
\usage{
identitystore_update_user(IdentityStoreId, UserId, Operations)
}
\arguments{
\item{IdentityStoreId}{[required] The globally unique identifier for the identity store.}

\item{UserId}{[required] The identifier for a user in the identity store.}

\item{Operations}{[required] A list of \code{AttributeOperation} objects to apply to the requested user.
These operations might add, replace, or remove an attribute. For more
information on the attributes that can be added, replaced, or removed,
see
\href{https://docs.aws.amazon.com/singlesignon/latest/IdentityStoreAPIReference/API_User.html}{User}.}
}
\description{
Updates the specified user metadata and attributes in the specified identity store.

See \url{https://www.paws-r-sdk.com/docs/identitystore_update_user/} for full documentation.
}
\keyword{internal}
