% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_associate_member}
\alias{inspector2_associate_member}
\title{Associates an Amazon Web Services account with an Amazon Inspector
delegated administrator}
\usage{
inspector2_associate_member(accountId)
}
\arguments{
\item{accountId}{[required] The Amazon Web Services account ID of the member account to be
associated.}
}
\description{
Associates an Amazon Web Services account with an Amazon Inspector delegated administrator. An HTTP 200 response indicates the association was successfully started, but doesn’t indicate whether it was completed. You can check if the association completed by using \code{\link[=inspector2_list_members]{list_members}} for multiple accounts or \href{https://docs.aws.amazon.com/inspector/v2/APIReference/API_GetMember.html}{GetMembers} for a single account.

See \url{https://paws-r.github.io/docs/inspector2/associate_member.html} for full documentation.
}
\keyword{internal}
