% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_batch_get_member_ec_2_deep_inspection_status}
\alias{inspector2_batch_get_member_ec_2_deep_inspection_status}
\title{Retrieves Amazon Inspector deep inspection activation status of multiple
member accounts within your organization}
\usage{
inspector2_batch_get_member_ec_2_deep_inspection_status(accountIds = NULL)
}
\arguments{
\item{accountIds}{The unique identifiers for the Amazon Web Services accounts to retrieve
Amazon Inspector deep inspection activation status for.}
}
\description{
Retrieves Amazon Inspector deep inspection activation status of multiple member accounts within your organization. You must be the delegated administrator of an organization in Amazon Inspector to use this API.

See \url{https://www.paws-r-sdk.com/docs/inspector2_batch_get_member_ec_2_deep_inspection_status/} for full documentation.
}
\keyword{internal}
