% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_describe_organization_configuration}
\alias{inspector2_describe_organization_configuration}
\title{Describe Amazon Inspector configuration settings for an Amazon Web
Services organization}
\usage{
inspector2_describe_organization_configuration()
}
\description{
Describe Amazon Inspector configuration settings for an Amazon Web Services organization.

See \url{https://www.paws-r-sdk.com/docs/inspector2_describe_organization_configuration/} for full documentation.
}
\keyword{internal}
