% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_disassociate_member}
\alias{inspector2_disassociate_member}
\title{Disassociates a member account from an Amazon Inspector delegated
administrator}
\usage{
inspector2_disassociate_member(accountId)
}
\arguments{
\item{accountId}{[required] The Amazon Web Services account ID of the member account to
disassociate.}
}
\description{
Disassociates a member account from an Amazon Inspector delegated administrator.

See \url{https://www.paws-r-sdk.com/docs/inspector2_disassociate_member/} for full documentation.
}
\keyword{internal}
