% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_cis_scan_result_details}
\alias{inspector2_get_cis_scan_result_details}
\title{Retrieves CIS scan result details}
\usage{
inspector2_get_cis_scan_result_details(
  scanArn,
  targetResourceId,
  accountId,
  filterCriteria = NULL,
  sortBy = NULL,
  sortOrder = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{scanArn}{[required] The scan ARN.}

\item{targetResourceId}{[required] The target resource ID.}

\item{accountId}{[required] The account ID.}

\item{filterCriteria}{The filter criteria.}

\item{sortBy}{The sort by order.}

\item{sortOrder}{The sort order.}

\item{nextToken}{The pagination token from a previous request that's used to retrieve the
next page of results.}

\item{maxResults}{The maximum number of CIS scan result details to be returned in a single
page of results.}
}
\description{
Retrieves CIS scan result details.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_cis_scan_result_details/} for full documentation.
}
\keyword{internal}
