% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_configuration}
\alias{inspector2_get_configuration}
\title{Retrieves setting configurations for Inspector scans}
\usage{
inspector2_get_configuration()
}
\description{
Retrieves setting configurations for Inspector scans.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_configuration/} for full documentation.
}
\keyword{internal}
