% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_get_delegated_admin_account}
\alias{inspector2_get_delegated_admin_account}
\title{Retrieves information about the Amazon Inspector delegated administrator
for your organization}
\usage{
inspector2_get_delegated_admin_account()
}
\description{
Retrieves information about the Amazon Inspector delegated administrator for your organization.

See \url{https://www.paws-r-sdk.com/docs/inspector2_get_delegated_admin_account/} for full documentation.
}
\keyword{internal}
