% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_account_permissions}
\alias{inspector2_list_account_permissions}
\title{Lists the permissions an account has to configure Amazon Inspector}
\usage{
inspector2_list_account_permissions(
  service = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{service}{The service scan type to check permissions for.}

\item{maxResults}{The maximum number of results the response can return. If your request
would return more than the maximum the response will return a
\code{nextToken} value, use this value when you call the action again to get
the remaining results.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. If your response returns more than the \code{maxResults} maximum
value it will also return a \code{nextToken} value. For subsequent calls, use
the NextToken value returned from the previous request to continue
listing results after the first page.}
}
\description{
Lists the permissions an account has to configure Amazon Inspector. If the account is a member account or standalone account with resources managed by an Organizations policy, the operation returns fewer permissions.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_account_permissions/} for full documentation.
}
\keyword{internal}
