% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_filters}
\alias{inspector2_list_filters}
\title{Lists the filters associated with your account}
\usage{
inspector2_list_filters(
  arns = NULL,
  action = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{arns}{The Amazon resource number (ARN) of the filter.}

\item{action}{The action the filter applies to matched findings.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. If your response returns more than the \code{maxResults} maximum
value it will also return a \code{nextToken} value. For subsequent calls, use
the \code{nextToken} value returned from the previous request to continue
listing results after the first page.}

\item{maxResults}{The maximum number of results the response can return. If your request
would return more than the maximum the response will return a
\code{nextToken} value, use this value when you call the action again to get
the remaining results.}
}
\description{
Lists the filters associated with your account.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_filters/} for full documentation.
}
\keyword{internal}
