% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_usage_totals}
\alias{inspector2_list_usage_totals}
\title{Lists the Amazon Inspector usage totals over the last 30 days}
\usage{
inspector2_list_usage_totals(
  maxResults = NULL,
  nextToken = NULL,
  accountIds = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results the response can return. If your request
would return more than the maximum the response will return a
\code{nextToken} value, use this value when you call the action again to get
the remaining results.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. If your response returns more than the \code{maxResults} maximum
value it will also return a \code{nextToken} value. For subsequent calls, use
the \code{nextToken} value returned from the previous request to continue
listing results after the first page.}

\item{accountIds}{The Amazon Web Services account IDs to retrieve usage totals for.}
}
\description{
Lists the Amazon Inspector usage totals over the last 30 days.

See \url{https://www.paws-r-sdk.com/docs/inspector2_list_usage_totals/} for full documentation.
}
\keyword{internal}
