% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_tag_resource}
\alias{inspector2_tag_resource}
\title{Adds tags to a resource}
\usage{
inspector2_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource to apply a tag to.}

\item{tags}{[required] The tags to be added to a resource.}
}
\description{
Adds tags to a resource.

See \url{https://www.paws-r-sdk.com/docs/inspector2_tag_resource/} for full documentation.
}
\keyword{internal}
