% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_update_configuration}
\alias{inspector2_update_configuration}
\title{Updates setting configurations for your Amazon Inspector account}
\usage{
inspector2_update_configuration(
  ecrConfiguration = NULL,
  ec2Configuration = NULL
)
}
\arguments{
\item{ecrConfiguration}{Specifies how the ECR automated re-scan will be updated for your
environment.}

\item{ec2Configuration}{Specifies how the Amazon EC2 automated scan will be updated for your
environment.}
}
\description{
Updates setting configurations for your Amazon Inspector account. When you use this API as an Amazon Inspector delegated administrator this updates the setting for all accounts you manage. Member accounts in an organization cannot update this setting.

See \url{https://www.paws-r-sdk.com/docs/inspector2_update_configuration/} for full documentation.
}
\keyword{internal}
