% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_add_attributes_to_findings}
\alias{inspector_add_attributes_to_findings}
\title{Assigns attributes (key and value pairs) to the findings that are
specified by the ARNs of the findings}
\usage{
inspector_add_attributes_to_findings(findingArns, attributes)
}
\arguments{
\item{findingArns}{[required] The ARNs that specify the findings that you want to assign attributes
to.}

\item{attributes}{[required] The array of attributes that you want to assign to specified findings.}
}
\description{
Assigns attributes (key and value pairs) to the findings that are specified by the ARNs of the findings.

See \url{https://www.paws-r-sdk.com/docs/inspector_add_attributes_to_findings/} for full documentation.
}
\keyword{internal}
