% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_describe_assessment_runs}
\alias{inspector_describe_assessment_runs}
\title{Describes the assessment runs that are specified by the ARNs of the
assessment runs}
\usage{
inspector_describe_assessment_runs(assessmentRunArns)
}
\arguments{
\item{assessmentRunArns}{[required] The ARN that specifies the assessment run that you want to describe.}
}
\description{
Describes the assessment runs that are specified by the ARNs of the assessment runs.

See \url{https://www.paws-r-sdk.com/docs/inspector_describe_assessment_runs/} for full documentation.
}
\keyword{internal}
