% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_describe_assessment_targets}
\alias{inspector_describe_assessment_targets}
\title{Describes the assessment targets that are specified by the ARNs of the
assessment targets}
\usage{
inspector_describe_assessment_targets(assessmentTargetArns)
}
\arguments{
\item{assessmentTargetArns}{[required] The ARNs that specifies the assessment targets that you want to
describe.}
}
\description{
Describes the assessment targets that are specified by the ARNs of the assessment targets.

See \url{https://www.paws-r-sdk.com/docs/inspector_describe_assessment_targets/} for full documentation.
}
\keyword{internal}
