% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_describe_assessment_templates}
\alias{inspector_describe_assessment_templates}
\title{Describes the assessment templates that are specified by the ARNs of the
assessment templates}
\usage{
inspector_describe_assessment_templates(assessmentTemplateArns)
}
\arguments{
\item{assessmentTemplateArns}{[required]}
}
\description{
Describes the assessment templates that are specified by the ARNs of the assessment templates.

See \url{https://paws-r.github.io/docs/inspector/describe_assessment_templates.html} for full documentation.
}
\keyword{internal}
