% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_describe_exclusions}
\alias{inspector_describe_exclusions}
\title{Describes the exclusions that are specified by the exclusions' ARNs}
\usage{
inspector_describe_exclusions(exclusionArns, locale = NULL)
}
\arguments{
\item{exclusionArns}{[required] The list of ARNs that specify the exclusions that you want to describe.}

\item{locale}{The locale into which you want to translate the exclusion's title,
description, and recommendation.}
}
\description{
Describes the exclusions that are specified by the exclusions' ARNs.

See \url{https://www.paws-r-sdk.com/docs/inspector_describe_exclusions/} for full documentation.
}
\keyword{internal}
