% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_describe_findings}
\alias{inspector_describe_findings}
\title{Describes the findings that are specified by the ARNs of the findings}
\usage{
inspector_describe_findings(findingArns, locale = NULL)
}
\arguments{
\item{findingArns}{[required] The ARN that specifies the finding that you want to describe.}

\item{locale}{The locale into which you want to translate a finding description,
recommendation, and the short description that identifies the finding.}
}
\description{
Describes the findings that are specified by the ARNs of the findings.

See \url{https://www.paws-r-sdk.com/docs/inspector_describe_findings/} for full documentation.
}
\keyword{internal}
