% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_describe_resource_groups}
\alias{inspector_describe_resource_groups}
\title{Describes the resource groups that are specified by the ARNs of the
resource groups}
\usage{
inspector_describe_resource_groups(resourceGroupArns)
}
\arguments{
\item{resourceGroupArns}{[required] The ARN that specifies the resource group that you want to describe.}
}
\description{
Describes the resource groups that are specified by the ARNs of the resource groups.

See \url{https://www.paws-r-sdk.com/docs/inspector_describe_resource_groups/} for full documentation.
}
\keyword{internal}
