% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_describe_rules_packages}
\alias{inspector_describe_rules_packages}
\title{Describes the rules packages that are specified by the ARNs of the rules
packages}
\usage{
inspector_describe_rules_packages(rulesPackageArns, locale = NULL)
}
\arguments{
\item{rulesPackageArns}{[required] The ARN that specifies the rules package that you want to describe.}

\item{locale}{The locale that you want to translate a rules package description into.}
}
\description{
Describes the rules packages that are specified by the ARNs of the rules packages.

See \url{https://www.paws-r-sdk.com/docs/inspector_describe_rules_packages/} for full documentation.
}
\keyword{internal}
