% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_get_telemetry_metadata}
\alias{inspector_get_telemetry_metadata}
\title{Information about the data that is collected for the specified
assessment run}
\usage{
inspector_get_telemetry_metadata(assessmentRunArn)
}
\arguments{
\item{assessmentRunArn}{[required] The ARN that specifies the assessment run that has the telemetry data
that you want to obtain.}
}
\description{
Information about the data that is collected for the specified assessment run.

See \url{https://www.paws-r-sdk.com/docs/inspector_get_telemetry_metadata/} for full documentation.
}
\keyword{internal}
