% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_list_tags_for_resource}
\alias{inspector_list_tags_for_resource}
\title{Lists all tags associated with an assessment template}
\usage{
inspector_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN that specifies the assessment template whose tags you want to
list.}
}
\description{
Lists all tags associated with an assessment template.

See \url{https://www.paws-r-sdk.com/docs/inspector_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
