% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_remove_attributes_from_findings}
\alias{inspector_remove_attributes_from_findings}
\title{Removes entire attributes (key and value pairs) from the findings that
are specified by the ARNs of the findings where an attribute with the
specified key exists}
\usage{
inspector_remove_attributes_from_findings(findingArns, attributeKeys)
}
\arguments{
\item{findingArns}{[required] The ARNs that specify the findings that you want to remove attributes
from.}

\item{attributeKeys}{[required] The array of attribute keys that you want to remove from specified
findings.}
}
\description{
Removes entire attributes (key and value pairs) from the findings that are specified by the ARNs of the findings where an attribute with the specified key exists.

See \url{https://www.paws-r-sdk.com/docs/inspector_remove_attributes_from_findings/} for full documentation.
}
\keyword{internal}
