% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_set_tags_for_resource}
\alias{inspector_set_tags_for_resource}
\title{Sets tags (key and value pairs) to the assessment template that is
specified by the ARN of the assessment template}
\usage{
inspector_set_tags_for_resource(resourceArn, tags = NULL)
}
\arguments{
\item{resourceArn}{[required] The ARN of the assessment template that you want to set tags to.}

\item{tags}{A collection of key and value pairs that you want to set to the
assessment template.}
}
\description{
Sets tags (key and value pairs) to the assessment template that is specified by the ARN of the assessment template.

See \url{https://www.paws-r-sdk.com/docs/inspector_set_tags_for_resource/} for full documentation.
}
\keyword{internal}
