% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_unsubscribe_from_event}
\alias{inspector_unsubscribe_from_event}
\title{Disables the process of sending Amazon Simple Notification Service (SNS)
notifications about a specified event to a specified SNS topic}
\usage{
inspector_unsubscribe_from_event(resourceArn, event, topicArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the assessment template that is used during the event for
which you want to stop receiving SNS notifications.}

\item{event}{[required] The event for which you want to stop receiving SNS notifications.}

\item{topicArn}{[required] The ARN of the SNS topic to which SNS notifications are sent.}
}
\description{
Disables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic.

See \url{https://www.paws-r-sdk.com/docs/inspector_unsubscribe_from_event/} for full documentation.
}
\keyword{internal}
