% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms_operations.R
\name{kms_delete_alias}
\alias{kms_delete_alias}
\title{Deletes the specified alias}
\usage{
kms_delete_alias(AliasName)
}
\arguments{
\item{AliasName}{[required] The alias to be deleted. The alias name must begin with \verb{alias/}
followed by the alias name, such as \code{alias/ExampleAlias}.}
}
\description{
Deletes the specified alias.

See \url{https://www.paws-r-sdk.com/docs/kms_delete_alias/} for full documentation.
}
\keyword{internal}
