% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms_operations.R
\name{kms_disconnect_custom_key_store}
\alias{kms_disconnect_custom_key_store}
\title{Disconnects the custom key store from its backing key store}
\usage{
kms_disconnect_custom_key_store(CustomKeyStoreId)
}
\arguments{
\item{CustomKeyStoreId}{[required] Enter the ID of the custom key store you want to disconnect. To find the
ID of a custom key store, use the
\code{\link[=kms_describe_custom_key_stores]{describe_custom_key_stores}}
operation.}
}
\description{
Disconnects the \href{https://docs.aws.amazon.com/kms/latest/developerguide/key-store-overview.html}{custom key store} from its backing key store. This operation disconnects an CloudHSM key store from its associated CloudHSM cluster or disconnects an external key store from the external key store proxy that communicates with your external key manager.

See \url{https://www.paws-r-sdk.com/docs/kms_disconnect_custom_key_store/} for full documentation.
}
\keyword{internal}
