% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms_operations.R
\name{kms_import_key_material}
\alias{kms_import_key_material}
\title{Imports or reimports key material into an existing KMS key that was
created without key material}
\usage{
kms_import_key_material(
  KeyId,
  ImportToken,
  EncryptedKeyMaterial,
  ValidTo = NULL,
  ExpirationModel = NULL,
  ImportType = NULL,
  KeyMaterialDescription = NULL,
  KeyMaterialId = NULL
)
}
\arguments{
\item{KeyId}{[required] The identifier of the KMS key that will be associated with the imported
key material. This must be the same KMS key specified in the \code{KeyID}
parameter of the corresponding
\code{\link[=kms_get_parameters_for_import]{get_parameters_for_import}} request.
The \code{Origin} of the KMS key must be \code{EXTERNAL} and its \code{KeyState} must
be \code{PendingImport}.

The KMS key can be a symmetric encryption KMS key, HMAC KMS key,
asymmetric encryption KMS key, or asymmetric signing KMS key, including
a \href{https://docs.aws.amazon.com/kms/latest/developerguide/multi-region-keys-overview.html}{multi-Region key}
of any supported type. You cannot perform this operation on a KMS key in
a custom key store, or on a KMS key in a different Amazon Web Services
account.

Specify the key ID or key ARN of the KMS key.

For example:
\itemize{
\item Key ID: \verb{1234abcd-12ab-34cd-56ef-1234567890ab}
\item Key ARN:
\verb{arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab}
}

To get the key ID and key ARN for a KMS key, use
\code{\link[=kms_list_keys]{list_keys}} or \code{\link[=kms_describe_key]{describe_key}}.}

\item{ImportToken}{[required] The import token that you received in the response to a previous
\code{\link[=kms_get_parameters_for_import]{get_parameters_for_import}} request. It
must be from the same response that contained the public key that you
used to encrypt the key material.}

\item{EncryptedKeyMaterial}{[required] The encrypted key material to import. The key material must be encrypted
under the public wrapping key that
\code{\link[=kms_get_parameters_for_import]{get_parameters_for_import}} returned,
using the wrapping algorithm that you specified in the same
\code{\link[=kms_get_parameters_for_import]{get_parameters_for_import}} request.}

\item{ValidTo}{The date and time when the imported key material expires. This parameter
is required when the value of the \code{ExpirationModel} parameter is
\code{KEY_MATERIAL_EXPIRES}. Otherwise it is not valid.

The value of this parameter must be a future date and time. The maximum
value is 365 days from the request date.

When the key material expires, KMS deletes the key material from the KMS
key. Without its key material, the KMS key is unusable. To use the KMS
key in cryptographic operations, you must reimport the same key
material.

You cannot change the \code{ExpirationModel} or \code{ValidTo} values for the
current import after the request completes. To change either value, you
must delete
(\code{\link[=kms_delete_imported_key_material]{delete_imported_key_material}}) and
reimport the key material.}

\item{ExpirationModel}{Specifies whether the key material expires. The default is
\code{KEY_MATERIAL_EXPIRES}. For help with this choice, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/importing-keys-import-key-material.html#importing-keys-expiration}{Setting an expiration time}
in the \emph{Key Management Service Developer Guide}.

When the value of \code{ExpirationModel} is \code{KEY_MATERIAL_EXPIRES}, you must
specify a value for the \code{ValidTo} parameter. When value is
\code{KEY_MATERIAL_DOES_NOT_EXPIRE}, you must omit the \code{ValidTo} parameter.

You cannot change the \code{ExpirationModel} or \code{ValidTo} values for the
current import after the request completes. To change either value, you
must reimport the key material.}

\item{ImportType}{Indicates whether the key material being imported is previously
associated with this KMS key or not. This parameter is optional and only
usable with symmetric encryption keys. If no key material has ever been
imported into the KMS key, and this parameter is omitted, the parameter
defaults to \code{NEW_KEY_MATERIAL}. After the first key material is
imported, if this parameter is omitted then the parameter defaults to
\code{EXISTING_KEY_MATERIAL}.

For multi-Region keys, you must first import new key material into the
primary Region key. You should use the \code{NEW_KEY_MATERIAL} import type
when importing key material into the primary Region key. Then, you can
import the same key material into the replica Region key. The import
type for the replica Region key should be \code{EXISTING_KEY_MATERIAL}.}

\item{KeyMaterialDescription}{Description for the key material being imported. This parameter is
optional and only usable with symmetric encryption keys. If you do not
specify a key material description, KMS retains the value you specified
when you last imported the same key material into this KMS key.}

\item{KeyMaterialId}{Identifies the key material being imported. This parameter is optional
and only usable with symmetric encryption keys. You cannot specify a key
material ID with \code{ImportType} set to \code{NEW_KEY_MATERIAL}. Whenever you
import key material into a symmetric encryption key, KMS assigns a
unique identifier to the key material based on the KMS key ID and the
imported key material. When you re-import key material with a specified
key material ID, KMS:
\itemize{
\item Computes the identifier for the key material
\item Matches the computed identifier against the specified key material
ID
\item Verifies that the key material ID is already associated with the KMS
key
}

To get the list of key material IDs associated with a KMS key, use
\code{\link[=kms_list_key_rotations]{list_key_rotations}}.}
}
\description{
Imports or reimports key material into an existing KMS key that was created without key material. You can also use this operation to set or update the expiration model and expiration date of the imported key material.

See \url{https://www.paws-r-sdk.com/docs/kms_import_key_material/} for full documentation.
}
\keyword{internal}
