% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms_operations.R
\name{kms_list_key_rotations}
\alias{kms_list_key_rotations}
\title{Returns information about the key materials associated with the
specified KMS key}
\usage{
kms_list_key_rotations(
  KeyId,
  IncludeKeyMaterial = NULL,
  Limit = NULL,
  Marker = NULL
)
}
\arguments{
\item{KeyId}{[required] Gets the key rotations for the specified KMS key.

Specify the key ID or key ARN of the KMS key.

For example:
\itemize{
\item Key ID: \verb{1234abcd-12ab-34cd-56ef-1234567890ab}
\item Key ARN:
\verb{arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab}
}

To get the key ID and key ARN for a KMS key, use
\code{\link[=kms_list_keys]{list_keys}} or \code{\link[=kms_describe_key]{describe_key}}.}

\item{IncludeKeyMaterial}{Use this optional parameter to control which key materials associated
with this key are listed in the response. The default value of this
parameter is \code{ROTATIONS_ONLY}. If you omit this parameter, KMS returns
information on the key materials created by automatic or on-demand key
rotation. When you specify a value of \code{ALL_KEY_MATERIAL}, KMS adds the
first key material and any imported key material pending rotation to the
response. This parameter can only be used with KMS keys that support
automatic or on-demand key rotation.}

\item{Limit}{Use this parameter to specify the maximum number of items to return.
When this value is present, KMS does not return more than the specified
number of items, but it might return fewer.

This value is optional. If you include a value, it must be between 1 and
1000, inclusive. If you do not include a value, it defaults to 100.}

\item{Marker}{Use this parameter in a subsequent request after you receive a response
with truncated results. Set it to the value of \code{NextMarker} from the
truncated response you just received.}
}
\description{
Returns information about the key materials associated with the specified KMS key. You can use the optional \code{IncludeKeyMaterial} parameter to control which key materials are included in the response.

See \url{https://www.paws-r-sdk.com/docs/kms_list_key_rotations/} for full documentation.
}
\keyword{internal}
