% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kms_operations.R
\name{kms_verify}
\alias{kms_verify}
\title{Verifies a digital signature that was generated by the Sign operation}
\usage{
kms_verify(
  KeyId,
  Message,
  MessageType = NULL,
  Signature,
  SigningAlgorithm,
  GrantTokens = NULL,
  DryRun = NULL
)
}
\arguments{
\item{KeyId}{[required] Identifies the asymmetric KMS key that will be used to verify the
signature. This must be the same KMS key that was used to generate the
signature. If you specify a different KMS key, the signature
verification fails.

To specify a KMS key, use its key ID, key ARN, alias name, or alias ARN.
When using an alias name, prefix it with \code{"alias/"}. To specify a KMS
key in a different Amazon Web Services account, you must use the key ARN
or alias ARN.

For example:
\itemize{
\item Key ID: \verb{1234abcd-12ab-34cd-56ef-1234567890ab}
\item Key ARN:
\verb{arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab}
\item Alias name: \code{alias/ExampleAlias}
\item Alias ARN: \code{arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias}
}

To get the key ID and key ARN for a KMS key, use
\code{\link[=kms_list_keys]{list_keys}} or \code{\link[=kms_describe_key]{describe_key}}. To
get the alias name and alias ARN, use
\code{\link[=kms_list_aliases]{list_aliases}}.}

\item{Message}{[required] Specifies the message that was signed. You can submit a raw message of
up to 4096 bytes, or a hash digest of the message. If you submit a
digest, use the \code{MessageType} parameter with a value of \code{DIGEST}.

If the message specified here is different from the message that was
signed, the signature verification fails. A message and its hash digest
are considered to be the same message.}

\item{MessageType}{Tells KMS whether the value of the \code{Message} parameter should be hashed
as part of the signing algorithm. Use \code{RAW} for unhashed messages; use
\code{DIGEST} for message digests, which are already hashed; use
\code{EXTERNAL_MU} for 64-byte representative μ used in ML-DSA signing as
defined in NIST FIPS 204 Section 6.2.

When the value of \code{MessageType} is \code{RAW}, KMS uses the standard signing
algorithm, which begins with a hash function. When the value is
\code{DIGEST}, KMS skips the hashing step in the signing algorithm. When the
value is \code{EXTERNAL_MU} KMS skips the concatenated hashing of the public
key hash and the message done in the ML-DSA signing algorithm.

Use the \code{DIGEST} or \code{EXTERNAL_MU} value only when the value of the
\code{Message} parameter is a message digest. If you use the \code{DIGEST} value
with an unhashed message, the security of the signing operation can be
compromised.

When using ECC_NIST_EDWARDS25519 KMS keys:
\itemize{
\item ED25519_SHA_512 signing algorithm requires KMS \code{MessageType:RAW}
\item ED25519_PH_SHA_512 signing algorithm requires KMS
\code{MessageType:DIGEST}
}

When the value of \code{MessageType} is \code{DIGEST}, the length of the \code{Message}
value must match the length of hashed messages for the specified signing
algorithm.

When the value of \code{MessageType} is \code{EXTERNAL_MU} the length of the
\code{Message} value must be 64 bytes.

You can submit a message digest and omit the \code{MessageType} or specify
\code{RAW} so the digest is hashed again while signing. However, if the
signed message is hashed once while signing, but twice while verifying,
verification fails, even when the message hasn't changed.

The hashing algorithm that \code{\link[=kms_verify]{verify}} uses is based on the
\code{SigningAlgorithm} value.
\itemize{
\item Signing algorithms that end in SHA_256 use the SHA_256 hashing
algorithm.
\item Signing algorithms that end in SHA_384 use the SHA_384 hashing
algorithm.
\item Signing algorithms that end in SHA_512 use the SHA_512 hashing
algorithm.
\item Signing algorithms that end in SHAKE_256 use the SHAKE_256 hashing
algorithm.
\item SM2DSA uses the SM3 hashing algorithm. For details, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/offline-operations.html#key-spec-sm-offline-verification}{Offline verification with SM2 key pairs}.
}}

\item{Signature}{[required] The signature that the \code{\link[=kms_sign]{sign}} operation generated.}

\item{SigningAlgorithm}{[required] The signing algorithm that was used to sign the message. If you submit a
different algorithm, the signature verification fails.}

\item{GrantTokens}{A list of grant tokens.

Use a grant token when your permission to call this operation comes from
a new grant that has not yet achieved \emph{eventual consistency}. For more
information, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/grants.html#grant_token}{Grant token}
and \href{https://docs.aws.amazon.com/kms/latest/developerguide/using-grant-token.html}{Using a grant token}
in the \emph{Key Management Service Developer Guide}.}

\item{DryRun}{Checks if your request will succeed. \code{DryRun} is an optional parameter.

To learn more about how to use this parameter, see \href{https://docs.aws.amazon.com/kms/latest/developerguide/testing-permissions.html}{Testing your permissions}
in the \emph{Key Management Service Developer Guide}.}
}
\description{
Verifies a digital signature that was generated by the \code{\link[=kms_sign]{sign}} operation.

See \url{https://www.paws-r-sdk.com/docs/kms_verify/} for full documentation.
}
\keyword{internal}
