% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_delete_custom_data_identifier}
\alias{macie2_delete_custom_data_identifier}
\title{Soft deletes a custom data identifier}
\usage{
macie2_delete_custom_data_identifier(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}
}
\description{
Soft deletes a custom data identifier.

See \url{https://www.paws-r-sdk.com/docs/macie2_delete_custom_data_identifier/} for full documentation.
}
\keyword{internal}
