% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_findings_publication_configuration}
\alias{macie2_get_findings_publication_configuration}
\title{Retrieves the configuration settings for publishing findings to Security
Hub}
\usage{
macie2_get_findings_publication_configuration()
}
\description{
Retrieves the configuration settings for publishing findings to Security Hub.

See \url{https://paws-r.github.io/docs/macie2/get_findings_publication_configuration.html} for full documentation.
}
\keyword{internal}
