% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_macie_session}
\alias{macie2_get_macie_session}
\title{Retrieves the status and configuration settings for an Amazon Macie
account}
\usage{
macie2_get_macie_session()
}
\description{
Retrieves the status and configuration settings for an Amazon Macie account.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_macie_session/} for full documentation.
}
\keyword{internal}
