% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_member}
\alias{macie2_get_member}
\title{Retrieves information about an account that's associated with an Amazon
Macie administrator account}
\usage{
macie2_get_member(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource that the request
applies to.}
}
\description{
Retrieves information about an account that's associated with an Amazon Macie administrator account.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_member/} for full documentation.
}
\keyword{internal}
