% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_reveal_configuration}
\alias{macie2_get_reveal_configuration}
\title{Retrieves the status and configuration settings for retrieving
occurrences of sensitive data reported by findings}
\usage{
macie2_get_reveal_configuration()
}
\description{
Retrieves the status and configuration settings for retrieving occurrences of sensitive data reported by findings.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_reveal_configuration/} for full documentation.
}
\keyword{internal}
