% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_sensitive_data_occurrences_availability}
\alias{macie2_get_sensitive_data_occurrences_availability}
\title{Checks whether occurrences of sensitive data can be retrieved for a
finding}
\usage{
macie2_get_sensitive_data_occurrences_availability(findingId)
}
\arguments{
\item{findingId}{[required] The unique identifier for the finding.}
}
\description{
Checks whether occurrences of sensitive data can be retrieved for a finding.

See \url{https://www.paws-r-sdk.com/docs/macie2_get_sensitive_data_occurrences_availability/} for full documentation.
}
\keyword{internal}
