% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_tags_for_resource}
\alias{macie2_list_tags_for_resource}
\title{Retrieves the tags (keys and values) that are associated with an Amazon
Macie resource}
\usage{
macie2_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) of the resource.}
}
\description{
Retrieves the tags (keys and values) that are associated with an Amazon Macie resource.

See \url{https://www.paws-r-sdk.com/docs/macie2_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
