% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_automated_discovery_configuration}
\alias{macie2_update_automated_discovery_configuration}
\title{Enables or disables automated sensitive data discovery for an account}
\usage{
macie2_update_automated_discovery_configuration(status)
}
\arguments{
\item{status}{[required] The new status of automated sensitive data discovery for the account.
Valid values are: ENABLED, start or resume automated sensitive data
discovery activities for the account; and, DISABLED, stop performing
automated sensitive data discovery activities for the account.

When you enable automated sensitive data discovery for the first time,
Amazon Macie uses default configuration settings to determine which data
sources to analyze and which managed data identifiers to use. To change
these settings, use the UpdateClassificationScope and
UpdateSensitivityInspectionTemplate operations, respectively. If you
change the settings and subsequently disable the configuration, Amazon
Macie retains your changes.}
}
\description{
Enables or disables automated sensitive data discovery for an account.

See \url{https://www.paws-r-sdk.com/docs/macie2_update_automated_discovery_configuration/} for full documentation.
}
\keyword{internal}
