% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_update_reveal_configuration}
\alias{macie2_update_reveal_configuration}
\title{Updates the status and configuration settings for retrieving occurrences
of sensitive data reported by findings}
\usage{
macie2_update_reveal_configuration(
  configuration,
  retrievalConfiguration = NULL
)
}
\arguments{
\item{configuration}{[required] The KMS key to use to encrypt the sensitive data, and the status of the
configuration for the Amazon Macie account.}

\item{retrievalConfiguration}{The access method and settings to use when retrieving the sensitive
data.}
}
\description{
Updates the status and configuration settings for retrieving occurrences of sensitive data reported by findings.

See \url{https://www.paws-r-sdk.com/docs/macie2_update_reveal_configuration/} for full documentation.
}
\keyword{internal}
