% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie_operations.R
\name{macie_list_member_accounts}
\alias{macie_list_member_accounts}
\title{Lists all Amazon Macie Classic member accounts for the current Amazon
Macie Classic master account}
\usage{
macie_list_member_accounts(nextToken, maxResults)
}
\arguments{
\item{nextToken}{Use this parameter when paginating results. Set the value of this
parameter to null on your first call to the ListMemberAccounts action.
Subsequent calls to the action fill nextToken in the request with the
value of nextToken from the previous response to continue listing data.}

\item{maxResults}{Use this parameter to indicate the maximum number of items that you want
in the response. The default value is 250.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  memberAccounts = list(
    list(
      accountId = "string"
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Lists all Amazon Macie Classic member accounts for the current Amazon
Macie Classic master account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_member_accounts(
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
