% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie_operations.R
\name{macie_list_s3_resources}
\alias{macie_list_s3_resources}
\title{(Discontinued) Lists all the S3 resources associated with Amazon Macie
Classic}
\usage{
macie_list_s3_resources(
  memberAccountId = NULL,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{memberAccountId}{(Discontinued) The Amazon Macie Classic member account ID whose
associated S3 resources you want to list.}

\item{nextToken}{(Discontinued) Use this parameter when paginating results. Set its value
to null on your first call to the
\code{\link[=macie_list_s3_resources]{list_s3_resources}} action. Subsequent calls
to the action fill \code{nextToken} in the request with the value of
\code{nextToken} from the previous response to continue listing data.}

\item{maxResults}{(Discontinued) Use this parameter to indicate the maximum number of
items that you want in the response. The default value is 250.}
}
\description{
(Discontinued) Lists all the S3 resources associated with Amazon Macie Classic. If \code{memberAccountId} isn't specified, the action lists the S3 resources associated with Macie Classic for the current Macie Classic administrator account. If \code{memberAccountId} is specified, the action lists the S3 resources associated with Macie Classic for the specified member account.

See \url{https://www.paws-r-sdk.com/docs/macie_list_s3_resources/} for full documentation.
}
\keyword{internal}
