% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_disassociate_resource_share}
\alias{ram_disassociate_resource_share}
\title{Disassociates the specified principals or resources from the specified
resource share}
\usage{
ram_disassociate_resource_share(resourceShareArn, resourceArns,
  principals, clientToken)
}
\arguments{
\item{resourceShareArn}{[required] The Amazon Resource Name (ARN) of the resource share.}

\item{resourceArns}{The Amazon Resource Names (ARNs) of the resources.}

\item{principals}{The principals.}

\item{clientToken}{A unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  resourceShareAssociations = list(
    list(
      resourceShareArn = "string",
      resourceShareName = "string",
      associatedEntity = "string",
      associationType = "PRINCIPAL"|"RESOURCE",
      status = "ASSOCIATING"|"ASSOCIATED"|"FAILED"|"DISASSOCIATING"|"DISASSOCIATED",
      statusMessage = "string",
      creationTime = as.POSIXct(
        "2015-01-01"
      ),
      lastUpdatedTime = as.POSIXct(
        "2015-01-01"
      ),
      external = TRUE|FALSE
    )
  ),
  clientToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Disassociates the specified principals or resources from the specified
resource share.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$disassociate_resource_share(
  resourceShareArn = "string",
  resourceArns = list(
    "string"
  ),
  principals = list(
    "string"
  ),
  clientToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
