% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_get_resource_share_invitations}
\alias{ram_get_resource_share_invitations}
\title{Gets the invitations for resource sharing that you've received}
\usage{
ram_get_resource_share_invitations(resourceShareInvitationArns,
  resourceShareArns, nextToken, maxResults)
}
\arguments{
\item{resourceShareInvitationArns}{The Amazon Resource Names (ARN) of the invitations.}

\item{resourceShareArns}{The Amazon Resource Names (ARN) of the resource shares.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  resourceShareInvitations = list(
    list(
      resourceShareInvitationArn = "string",
      resourceShareName = "string",
      resourceShareArn = "string",
      senderAccountId = "string",
      receiverAccountId = "string",
      invitationTimestamp = as.POSIXct(
        "2015-01-01"
      ),
      status = "PENDING"|"ACCEPTED"|"REJECTED"|"EXPIRED",
      resourceShareAssociations = list(
        list(
          resourceShareArn = "string",
          resourceShareName = "string",
          associatedEntity = "string",
          associationType = "PRINCIPAL"|"RESOURCE",
          status = "ASSOCIATING"|"ASSOCIATED"|"FAILED"|"DISASSOCIATING"|"DISASSOCIATED",
          statusMessage = "string",
          creationTime = as.POSIXct(
            "2015-01-01"
          ),
          lastUpdatedTime = as.POSIXct(
            "2015-01-01"
          ),
          external = TRUE|FALSE
        )
      )
    )
  ),
  nextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Gets the invitations for resource sharing that you've received.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_resource_share_invitations(
  resourceShareInvitationArns = list(
    "string"
  ),
  resourceShareArns = list(
    "string"
  ),
  nextToken = "string",
  maxResults = 123
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
