% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_list_principals}
\alias{ram_list_principals}
\title{Lists the principals that you have shared resources with or that have
shared resources with you}
\usage{
ram_list_principals(resourceOwner, resourceArn, principals,
  resourceType, resourceShareArns, nextToken, maxResults)
}
\arguments{
\item{resourceOwner}{[required] The type of owner.}

\item{resourceArn}{The Amazon Resource Name (ARN) of the resource.}

\item{principals}{The principals.}

\item{resourceType}{The resource type.

Valid values: \code{acm-pca:CertificateAuthority} | \code{appmesh:Mesh} |
\code{codebuild:Project} | \code{codebuild:ReportGroup} |
\code{ec2:CapacityReservation} | \code{ec2:DedicatedHost} |
\code{ec2:LocalGatewayRouteTable} | \code{ec2:PrefixList} | \code{ec2:Subnet} |
\code{ec2:TrafficMirrorTarget} | \code{ec2:TransitGateway} |
\code{imagebuilder:Component} | \code{imagebuilder:Image} |
\code{imagebuilder:ImageRecipe} | \code{imagebuilder:ContainerRecipe} |
\code{glue:Catalog} | \code{glue:Database} | \code{glue:Table} |
\code{license-manager:LicenseConfiguration} I
\code{network-firewall:FirewallPolicy} | \code{network-firewall:StatefulRuleGroup}
| \code{network-firewall:StatelessRuleGroup} | \code{outposts:Outpost} |
\code{resource-groups:Group} | \code{rds:Cluster} |
\code{route53resolver:FirewallRuleGroup}
|\code{route53resolver:ResolverQueryLogConfig} |
\code{route53resolver:ResolverRule}}

\item{resourceShareArns}{The Amazon Resource Names (ARN) of the resource shares.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}
}
\value{
A list with the following syntax:\preformatted{list(
  principals = list(
    list(
      id = "string",
      resourceShareArn = "string",
      creationTime = as.POSIXct(
        "2015-01-01"
      ),
      lastUpdatedTime = as.POSIXct(
        "2015-01-01"
      ),
      external = TRUE|FALSE
    )
  ),
  nextToken = "string"
)
}
}
\description{
Lists the principals that you have shared resources with or that have
shared resources with you.
}
\section{Request syntax}{
\preformatted{svc$list_principals(
  resourceOwner = "SELF"|"OTHER-ACCOUNTS",
  resourceArn = "string",
  principals = list(
    "string"
  ),
  resourceType = "string",
  resourceShareArns = list(
    "string"
  ),
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
