% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_list_resource_share_permissions}
\alias{ram_list_resource_share_permissions}
\title{Lists the AWS RAM permissions that are associated with a resource share}
\usage{
ram_list_resource_share_permissions(resourceShareArn, nextToken,
  maxResults)
}
\arguments{
\item{resourceShareArn}{[required] The Amazon Resource Name (ARN) of the resource share.}

\item{nextToken}{The token for the next page of results.}

\item{maxResults}{The maximum number of results to return with a single call. To retrieve
the remaining results, make another call with the returned \code{nextToken}
value.}
}
\description{
Lists the AWS RAM permissions that are associated with a resource share.
}
\section{Request syntax}{
\preformatted{svc$list_resource_share_permissions(
  resourceShareArn = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
