% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_promote_resource_share_created_from_policy}
\alias{ram_promote_resource_share_created_from_policy}
\title{Resource shares that were created by attaching a policy to a resource
are visible only to the resource share owner, and the resource share
cannot be modified in AWS RAM}
\usage{
ram_promote_resource_share_created_from_policy(resourceShareArn)
}
\arguments{
\item{resourceShareArn}{[required] The ARN of the resource share to promote.}
}
\value{
A list with the following syntax:\preformatted{list(
  returnValue = TRUE|FALSE
)
}
}
\description{
Resource shares that were created by attaching a policy to a resource
are visible only to the resource share owner, and the resource share
cannot be modified in AWS RAM.

Use this API action to promote the resource share. When you promote the
resource share, it becomes:
\itemize{
\item Visible to all principals that it is shared with.
\item Modifiable in AWS RAM.
}
}
\section{Request syntax}{
\preformatted{svc$promote_resource_share_created_from_policy(
  resourceShareArn = "string"
)
}
}

\keyword{internal}
