% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_promote_resource_share_created_from_policy}
\alias{ram_promote_resource_share_created_from_policy}
\title{When you attach a resource-based policy to a resource, RAM automatically
creates a resource share of featureSet=CREATED_FROM_POLICY with a
managed permission that has the same IAM permissions as the original
resource-based policy}
\usage{
ram_promote_resource_share_created_from_policy(resourceShareArn)
}
\arguments{
\item{resourceShareArn}{[required] Specifies the \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/reference-arns.html}{Amazon Resource Name (ARN)}
of the resource share to promote.}
}
\description{
When you attach a resource-based policy to a resource, RAM automatically creates a resource share of \code{featureSet}=\code{CREATED_FROM_POLICY} with a managed permission that has the same IAM permissions as the original resource-based policy. However, this type of managed permission is visible to only the resource share owner, and the associated resource share can't be modified by using RAM.

See \url{https://www.paws-r-sdk.com/docs/ram_promote_resource_share_created_from_policy/} for full documentation.
}
\keyword{internal}
