% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_tag_resource}
\alias{ram_tag_resource}
\title{Adds the specified tags to the specified resource share that you own}
\usage{
ram_tag_resource(resourceShareArn, tags)
}
\arguments{
\item{resourceShareArn}{[required] The Amazon Resource Name (ARN) of the resource share.}

\item{tags}{[required] One or more tags.}
}
\description{
Adds the specified tags to the specified resource share that you own.
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  resourceShareArn = "string",
  tags = list(
    list(
      key = "string",
      value = "string"
    )
  )
)
}
}

\keyword{internal}
