% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_untag_resource}
\alias{ram_untag_resource}
\title{Removes the specified tags from the specified resource share that you
own}
\usage{
ram_untag_resource(resourceShareArn, tagKeys)
}
\arguments{
\item{resourceShareArn}{[required] The Amazon Resource Name (ARN) of the resource share.}

\item{tagKeys}{[required] The tag keys of the tags to remove.}
}
\value{
An empty list.
}
\description{
Removes the specified tags from the specified resource share that you
own.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$untag_resource(
  resourceShareArn = "string",
  tagKeys = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
