% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsmanager_operations.R
\name{secretsmanager_create_secret}
\alias{secretsmanager_create_secret}
\title{Creates a new secret}
\usage{
secretsmanager_create_secret(
  Name,
  ClientRequestToken = NULL,
  Description = NULL,
  KmsKeyId = NULL,
  SecretBinary = NULL,
  SecretString = NULL,
  Tags = NULL,
  AddReplicaRegions = NULL,
  ForceOverwriteReplicaSecret = NULL,
  Type = NULL
)
}
\arguments{
\item{Name}{[required] The name of the new secret.

The secret name can contain ASCII letters, numbers, and the following
characters: /_+=.@-

Do not end your secret name with a hyphen followed by six characters. If
you do so, you risk confusion and unexpected results when searching for
a secret by partial ARN. Secrets Manager automatically adds a hyphen and
six random characters after the secret name at the end of the ARN.}

\item{ClientRequestToken}{If you include \code{SecretString} or \code{SecretBinary}, then Secrets Manager
creates an initial version for the secret, and this parameter specifies
the unique identifier for the new version.

If you use the Amazon Web Services CLI or one of the Amazon Web Services
SDKs to call this operation, then you can leave this parameter empty.
The CLI or SDK generates a random UUID for you and includes it as the
value for this parameter in the request.

If you generate a raw HTTP request to the Secrets Manager service
endpoint, then you must generate a \code{ClientRequestToken} and include it
in the request.

This value helps ensure idempotency. Secrets Manager uses this value to
prevent the accidental creation of duplicate versions if there are
failures and retries during a rotation. We recommend that you generate a
\href{https://en.wikipedia.org/wiki/Universally_unique_identifier}{UUID-type}
value to ensure uniqueness of your versions within the specified secret.
\itemize{
\item If the \code{ClientRequestToken} value isn't already associated with a
version of the secret then a new version of the secret is created.
\item If a version with this value already exists and the version
\code{SecretString} and \code{SecretBinary} values are the same as those in
the request, then the request is ignored.
\item If a version with this value already exists and that version's
\code{SecretString} and \code{SecretBinary} values are different from those in
the request, then the request fails because you cannot modify an
existing version. Instead, use
\code{\link[=secretsmanager_put_secret_value]{put_secret_value}} to create a
new version.
}

This value becomes the \code{VersionId} of the new version.}

\item{Description}{The description of the secret.}

\item{KmsKeyId}{The ARN, key ID, or alias of the KMS key that Secrets Manager uses to
encrypt the secret value in the secret. An alias is always prefixed by
\verb{alias/}, for example \code{alias/aws/secretsmanager}. For more information,
see \href{https://docs.aws.amazon.com/kms/latest/developerguide/kms-alias.html#alias-about}{About aliases}.

To use a KMS key in a different account, use the key ARN or the alias
ARN.

If you don't specify this value, then Secrets Manager uses the key
\code{aws/secretsmanager}. If that key doesn't yet exist, then Secrets
Manager creates it for you automatically the first time it encrypts the
secret value.

If the secret is in a different Amazon Web Services account from the
credentials calling the API, then you can't use \code{aws/secretsmanager} to
encrypt the secret, and you must create and use a customer managed KMS
key.}

\item{SecretBinary}{The binary data to encrypt and store in the new version of the secret.
We recommend that you store your binary data in a file and then pass the
contents of the file as a parameter.

Either \code{SecretString} or \code{SecretBinary} must have a value, but not both.

This parameter is not available in the Secrets Manager console.

Sensitive: This field contains sensitive information, so the service
does not include it in CloudTrail log entries. If you create your own
log entries, you must also avoid logging the information in this field.}

\item{SecretString}{The text data to encrypt and store in this new version of the secret. We
recommend you use a JSON structure of key/value pairs for your secret
value.

Either \code{SecretString} or \code{SecretBinary} must have a value, but not both.

If you create a secret by using the Secrets Manager console then Secrets
Manager puts the protected secret text in only the \code{SecretString}
parameter. The Secrets Manager console stores the information as a JSON
structure of key/value pairs that a Lambda rotation function can parse.

Sensitive: This field contains sensitive information, so the service
does not include it in CloudTrail log entries. If you create your own
log entries, you must also avoid logging the information in this field.}

\item{Tags}{A list of tags to attach to the secret. Each tag is a key and value pair
of strings in a JSON text string, for example:

\verb{[\{"Key":"CostCenter","Value":"12345"\},\{"Key":"environment","Value":"production"\}]}

Secrets Manager tag key names are case sensitive. A tag with the key
"ABC" is a different tag from one with key "abc".

If you check tags in permissions policies as part of your security
strategy, then adding or removing a tag can change permissions. If the
completion of this operation would result in you losing your permissions
for this secret, then Secrets Manager blocks the operation and returns
an \verb{Access Denied} error. For more information, see \href{https://docs.aws.amazon.com/secretsmanager/latest/userguide/auth-and-access_iam-policies.html#tag-secrets-abac}{Control access to secrets using tags}
and \href{https://docs.aws.amazon.com/secretsmanager/latest/userguide/auth-and-access_iam-policies.html#auth-and-access_tags2}{Limit access to identities with tags that match secrets' tags}.

For information about how to format a JSON parameter for the various
command line tool environments, see \href{https://docs.aws.amazon.com/cli/latest/userguide/cli-usage-parameters.html#cli-using-param-json}{Using JSON for Parameters}.
If your command-line tool or SDK requires quotation marks around the
parameter, you should use single quotes to avoid confusion with the
double quotes required in the JSON text.

For tag quotas and naming restrictions, see \href{https://docs.aws.amazon.com/general/latest/gr/arg.html#taged-reference-quotas}{Service quotas for Tagging}
in the \emph{Amazon Web Services General Reference guide}.}

\item{AddReplicaRegions}{A list of Regions and KMS keys to replicate secrets.}

\item{ForceOverwriteReplicaSecret}{Specifies whether to overwrite a secret with the same name in the
destination Region. By default, secrets aren't overwritten.}

\item{Type}{The exact string that identifies the partner that holds the external
secret. For more information, see \href{https://docs.aws.amazon.com/secretsmanager/latest/userguide/managed-external-secrets.html}{Using Secrets Manager managed external secrets}.}
}
\description{
Creates a new secret. A \emph{secret} can be a password, a set of credentials such as a user name and password, an OAuth token, or other secret information that you store in an encrypted form in Secrets Manager. The secret also includes the connection information to access a database or other service, which Secrets Manager doesn't encrypt. A secret in Secrets Manager consists of both the protected secret data and the important information needed to manage the secret.

See \url{https://www.paws-r-sdk.com/docs/secretsmanager_create_secret/} for full documentation.
}
\keyword{internal}
