% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsmanager_operations.R
\name{secretsmanager_get_resource_policy}
\alias{secretsmanager_get_resource_policy}
\title{Retrieves the JSON text of the resource-based policy document attached
to the secret}
\usage{
secretsmanager_get_resource_policy(SecretId)
}
\arguments{
\item{SecretId}{[required] The ARN or name of the secret to retrieve the attached resource-based
policy for.

For an ARN, we recommend that you specify a complete ARN rather than a
partial ARN. See \href{https://docs.aws.amazon.com/secretsmanager/latest/userguide/troubleshoot.html#ARN_secretnamehyphen}{Finding a secret from a partial ARN}.}
}
\description{
Retrieves the JSON text of the resource-based policy document attached to the secret. For more information about permissions policies attached to a secret, see \href{https://docs.aws.amazon.com/secretsmanager/latest/userguide/auth-and-access_resource-policies.html}{Permissions policies attached to a secret}.

See \url{https://www.paws-r-sdk.com/docs/secretsmanager_get_resource_policy/} for full documentation.
}
\keyword{internal}
