% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsmanager_operations.R
\name{secretsmanager_remove_regions_from_replication}
\alias{secretsmanager_remove_regions_from_replication}
\title{For a secret that is replicated to other Regions, deletes the secret
replicas from the Regions you specify}
\usage{
secretsmanager_remove_regions_from_replication(SecretId, RemoveReplicaRegions)
}
\arguments{
\item{SecretId}{[required] The ARN or name of the secret.}

\item{RemoveReplicaRegions}{[required] The Regions of the replicas to remove.}
}
\description{
For a secret that is replicated to other Regions, deletes the secret replicas from the Regions you specify.

See \url{https://www.paws-r-sdk.com/docs/secretsmanager_remove_regions_from_replication/} for full documentation.
}
\keyword{internal}
