% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsmanager_operations.R
\name{secretsmanager_replicate_secret_to_regions}
\alias{secretsmanager_replicate_secret_to_regions}
\title{Replicates the secret to a new Regions}
\usage{
secretsmanager_replicate_secret_to_regions(
  SecretId,
  AddReplicaRegions,
  ForceOverwriteReplicaSecret = NULL
)
}
\arguments{
\item{SecretId}{[required] The ARN or name of the secret to replicate.}

\item{AddReplicaRegions}{[required] A list of Regions in which to replicate the secret.}

\item{ForceOverwriteReplicaSecret}{Specifies whether to overwrite a secret with the same name in the
destination Region. By default, secrets aren't overwritten.}
}
\description{
Replicates the secret to a new Regions. See \href{https://docs.aws.amazon.com/secretsmanager/latest/userguide/replicate-secrets.html}{Multi-Region secrets}.

See \url{https://www.paws-r-sdk.com/docs/secretsmanager_replicate_secret_to_regions/} for full documentation.
}
\keyword{internal}
