% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsmanager_operations.R
\name{secretsmanager_restore_secret}
\alias{secretsmanager_restore_secret}
\title{Cancels the scheduled deletion of a secret by removing the DeletedDate
time stamp}
\usage{
secretsmanager_restore_secret(SecretId)
}
\arguments{
\item{SecretId}{[required] The ARN or name of the secret to restore.

For an ARN, we recommend that you specify a complete ARN rather than a
partial ARN. See \href{https://docs.aws.amazon.com/secretsmanager/latest/userguide/troubleshoot.html#ARN_secretnamehyphen}{Finding a secret from a partial ARN}.}
}
\description{
Cancels the scheduled deletion of a secret by removing the \code{DeletedDate} time stamp. You can access a secret again after it has been restored.

See \url{https://www.paws-r-sdk.com/docs/secretsmanager_restore_secret/} for full documentation.
}
\keyword{internal}
