% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsmanager_operations.R
\name{secretsmanager_rotate_secret}
\alias{secretsmanager_rotate_secret}
\title{Configures and starts the asynchronous process of rotating the secret}
\usage{
secretsmanager_rotate_secret(
  SecretId,
  ClientRequestToken = NULL,
  RotationLambdaARN = NULL,
  RotationRules = NULL,
  RotateImmediately = NULL
)
}
\arguments{
\item{SecretId}{[required] The ARN or name of the secret to rotate.

For an ARN, we recommend that you specify a complete ARN rather than a
partial ARN. See \href{https://docs.aws.amazon.com/secretsmanager/latest/userguide/troubleshoot.html#ARN_secretnamehyphen}{Finding a secret from a partial ARN}.}

\item{ClientRequestToken}{A unique identifier for the new version of the secret that helps ensure
idempotency. Secrets Manager uses this value to prevent the accidental
creation of duplicate versions if there are failures and retries during
rotation. This value becomes the \code{VersionId} of the new version.

If you use the Amazon Web Services CLI or one of the Amazon Web Services
SDK to call this operation, then you can leave this parameter empty. The
CLI or SDK generates a random UUID for you and includes that in the
request for this parameter. If you don't use the SDK and instead
generate a raw HTTP request to the Secrets Manager service endpoint,
then you must generate a \code{ClientRequestToken} yourself for new versions
and include that value in the request.

You only need to specify this value if you implement your own retry
logic and you want to ensure that Secrets Manager doesn't attempt to
create a secret version twice. We recommend that you generate a
\href{https://en.wikipedia.org/wiki/Universally_unique_identifier}{UUID-type}
value to ensure uniqueness within the specified secret.}

\item{RotationLambdaARN}{The ARN of the Lambda rotation function that can rotate the secret.}

\item{RotationRules}{A structure that defines the rotation configuration for this secret.}

\item{RotateImmediately}{Specifies whether to rotate the secret immediately or wait until the
next scheduled rotation window. The rotation schedule is defined in
RotateSecretRequest$RotationRules.

If you don't immediately rotate the secret, Secrets Manager tests the
rotation configuration by running the \href{https://docs.aws.amazon.com/secretsmanager/latest/userguide/}{\code{testSecret} step} of
the Lambda rotation function. The test creates an \code{AWSPENDING} version
of the secret and then removes it.

If you don't specify this value, then by default, Secrets Manager
rotates the secret immediately.}
}
\description{
Configures and starts the asynchronous process of rotating the secret. For more information about rotation, see \href{https://docs.aws.amazon.com/secretsmanager/latest/userguide/rotating-secrets.html}{Rotate secrets}.

See \url{https://paws-r.github.io/docs/secretsmanager/rotate_secret.html} for full documentation.
}
\keyword{internal}
