% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secretsmanager_operations.R
\name{secretsmanager_stop_replication_to_replica}
\alias{secretsmanager_stop_replication_to_replica}
\title{Removes the link between the replica secret and the primary secret and
promotes the replica to a primary secret in the replica Region}
\usage{
secretsmanager_stop_replication_to_replica(SecretId)
}
\arguments{
\item{SecretId}{[required] The ARN of the primary secret.}
}
\description{
Removes the link between the replica secret and the primary secret and promotes the replica to a primary secret in the replica Region.

See \url{https://www.paws-r-sdk.com/docs/secretsmanager_stop_replication_to_replica/} for full documentation.
}
\keyword{internal}
