% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_accept_administrator_invitation}
\alias{securityhub_accept_administrator_invitation}
\title{We recommend using Organizations instead of Security Hub invitations to
manage your member accounts}
\usage{
securityhub_accept_administrator_invitation(AdministratorId, InvitationId)
}
\arguments{
\item{AdministratorId}{[required] The account ID of the Security Hub administrator account that sent the
invitation.}

\item{InvitationId}{[required] The identifier of the invitation sent from the Security Hub
administrator account.}
}
\description{
We recommend using Organizations instead of Security Hub invitations to manage your member accounts. For information, see \href{https://docs.aws.amazon.com/securityhub/latest/userguide/securityhub-accounts-orgs.html}{Managing Security Hub administrator and member accounts with Organizations} in the \emph{Security Hub User Guide}.

See \url{https://www.paws-r-sdk.com/docs/securityhub_accept_administrator_invitation/} for full documentation.
}
\keyword{internal}
