% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_accept_administrator_invitation}
\alias{securityhub_accept_administrator_invitation}
\title{Accepts the invitation to be a member account and be monitored by the
Security Hub administrator account that the invitation was sent from}
\usage{
securityhub_accept_administrator_invitation(AdministratorId, InvitationId)
}
\arguments{
\item{AdministratorId}{[required] The account ID of the Security Hub administrator account that sent the
invitation.}

\item{InvitationId}{[required] The identifier of the invitation sent from the Security Hub
administrator account.}
}
\description{
Accepts the invitation to be a member account and be monitored by the Security Hub administrator account that the invitation was sent from.

See \url{https://www.paws-r-sdk.com/docs/securityhub_accept_administrator_invitation/} for full documentation.
}
\keyword{internal}
