% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_accept_invitation}
\alias{securityhub_accept_invitation}
\title{Accepts the invitation to be monitored by a Security Hub master account}
\usage{
securityhub_accept_invitation(MasterId, InvitationId)
}
\arguments{
\item{MasterId}{The account ID of the Security Hub master account whose invitation
you're accepting.}

\item{InvitationId}{The ID of the invitation that the Security Hub master account sends to
the AWS account.}
}
\description{
Accepts the invitation to be monitored by a Security Hub master account.
}
\section{Request syntax}{
\preformatted{svc$accept_invitation(
  MasterId = "string",
  InvitationId = "string"
)
}
}

\keyword{internal}
