% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_accept_invitation}
\alias{securityhub_accept_invitation}
\title{This method is deprecated}
\usage{
securityhub_accept_invitation(MasterId, InvitationId)
}
\arguments{
\item{MasterId}{[required] The account ID of the Security Hub administrator account that sent the
invitation.}

\item{InvitationId}{[required] The identifier of the invitation sent from the Security Hub
administrator account.}
}
\description{
This method is deprecated. Instead, use \code{\link[=securityhub_accept_administrator_invitation]{accept_administrator_invitation}}.

See \url{https://www.paws-r-sdk.com/docs/securityhub_accept_invitation/} for full documentation.
}
\keyword{internal}
