% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_batch_enable_standards}
\alias{securityhub_batch_enable_standards}
\title{Enables the standards specified by the provided StandardsArn}
\usage{
securityhub_batch_enable_standards(StandardsSubscriptionRequests)
}
\arguments{
\item{StandardsSubscriptionRequests}{[required] The list of standards checks to enable.}
}
\description{
Enables the standards specified by the provided \code{StandardsArn}. To obtain the ARN for a standard, use the \code{\link[=securityhub_describe_standards]{describe_standards}} operation.

See \url{https://www.paws-r-sdk.com/docs/securityhub_batch_enable_standards/} for full documentation.
}
\keyword{internal}
