% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_batch_import_findings}
\alias{securityhub_batch_import_findings}
\title{Imports security findings generated by a finding provider into Security
Hub}
\usage{
securityhub_batch_import_findings(Findings)
}
\arguments{
\item{Findings}{[required] A list of findings to import. To successfully import a finding, it must
follow the \href{https://docs.aws.amazon.com/securityhub/latest/userguide/securityhub-findings-format.html}{Amazon Web Services Security Finding Format}.
Maximum of 100 findings per request.}
}
\description{
Imports security findings generated by a finding provider into Security Hub. This action is requested by the finding provider to import its findings into Security Hub.

See \url{https://www.paws-r-sdk.com/docs/securityhub_batch_import_findings/} for full documentation.
}
\keyword{internal}
