% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_batch_update_standards_control_associations}
\alias{securityhub_batch_update_standards_control_associations}
\title{For a batch of security controls and standards, this operation updates
the enablement status of a control in a standard}
\usage{
securityhub_batch_update_standards_control_associations(
  StandardsControlAssociationUpdates
)
}
\arguments{
\item{StandardsControlAssociationUpdates}{[required] Updates the enablement status of a security control in a specified
standard.}
}
\description{
For a batch of security controls and standards, this operation updates the enablement status of a control in a standard.

See \url{https://www.paws-r-sdk.com/docs/securityhub_batch_update_standards_control_associations/} for full documentation.
}
\keyword{internal}
