% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_decline_invitations}
\alias{securityhub_decline_invitations}
\title{Declines invitations to become a member account}
\usage{
securityhub_decline_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{[required] The list of account IDs for the accounts from which to decline the
invitations to Security Hub.}
}
\value{
A list with the following syntax:\preformatted{list(
  UnprocessedAccounts = list(
    list(
      AccountId = "string",
      ProcessingResult = "string"
    )
  )
)
}
}
\description{
Declines invitations to become a member account.

This operation is only used by accounts that are not part of an
organization. Organization accounts do not receive invitations.
}
\section{Request syntax}{
\preformatted{svc$decline_invitations(
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
