% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_decline_invitations}
\alias{securityhub_decline_invitations}
\title{Declines invitations to become a member account}
\usage{
securityhub_decline_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{[required] The list of prospective member account IDs for which to decline an
invitation.}
}
\description{
Declines invitations to become a member account.

See \url{https://www.paws-r-sdk.com/docs/securityhub_decline_invitations/} for full documentation.
}
\keyword{internal}
