% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_action_target}
\alias{securityhub_delete_action_target}
\title{Deletes a custom action target from Security Hub}
\usage{
securityhub_delete_action_target(ActionTargetArn)
}
\arguments{
\item{ActionTargetArn}{[required] The Amazon Resource Name (ARN) of the custom action target to delete.}
}
\description{
Deletes a custom action target from Security Hub.

See \url{https://www.paws-r-sdk.com/docs/securityhub_delete_action_target/} for full documentation.
}
\keyword{internal}
