% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_finding_aggregator}
\alias{securityhub_delete_finding_aggregator}
\title{Deletes a finding aggregator}
\usage{
securityhub_delete_finding_aggregator(FindingAggregatorArn)
}
\arguments{
\item{FindingAggregatorArn}{[required] The ARN of the finding aggregator to delete. To obtain the ARN, use
\code{\link[=securityhub_list_finding_aggregators]{list_finding_aggregators}}.}
}
\description{
Deletes a finding aggregator. When you delete the finding aggregator, you stop finding aggregation.

See \url{https://www.paws-r-sdk.com/docs/securityhub_delete_finding_aggregator/} for full documentation.
}
\keyword{internal}
