% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_insight}
\alias{securityhub_delete_insight}
\title{Deletes an insight that is specified by the insight ARN}
\usage{
securityhub_delete_insight(InsightArn)
}
\arguments{
\item{InsightArn}{[required] The ARN of the insight that you want to delete.}
}
\description{
Deletes an insight that is specified by the insight ARN.
}
\section{Request syntax}{
\preformatted{svc$delete_insight(
  InsightArn = "string"
)
}
}

\keyword{internal}
