% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_insight}
\alias{securityhub_delete_insight}
\title{Deletes the insight specified by the InsightArn}
\usage{
securityhub_delete_insight(InsightArn)
}
\arguments{
\item{InsightArn}{[required] The ARN of the insight to delete.}
}
\description{
Deletes the insight specified by the \code{InsightArn}.

See \url{https://www.paws-r-sdk.com/docs/securityhub_delete_insight/} for full documentation.
}
\keyword{internal}
