% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_invitations}
\alias{securityhub_delete_invitations}
\title{Deletes invitations received by the AWS account to become a member
account}
\usage{
securityhub_delete_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{[required] The list of the account IDs that sent the invitations to delete.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  UnprocessedAccounts = list(
    list(
      AccountId = "string",
      ProcessingResult = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Deletes invitations received by the AWS account to become a member
account.

This operation is only used by accounts that are not part of an
organization. Organization accounts do not receive invitations.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_invitations(
  AccountIds = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
