% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_delete_invitations}
\alias{securityhub_delete_invitations}
\title{Deletes invitations that are sent to this AWS account (invitee) by the
AWS accounts (inviters) that are specified by their account IDs}
\usage{
securityhub_delete_invitations(AccountIds)
}
\arguments{
\item{AccountIds}{A list of account IDs specifying accounts whose invitations to Security
Hub you want to delete.}
}
\description{
Deletes invitations that are sent to this AWS account (invitee) by the
AWS accounts (inviters) that are specified by their account IDs.
}
\section{Request syntax}{
\preformatted{svc$delete_invitations(
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
