% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_describe_organization_configuration}
\alias{securityhub_describe_organization_configuration}
\title{Returns information about the Organizations configuration for Security
Hub}
\usage{
securityhub_describe_organization_configuration()
}
\description{
Returns information about the Organizations configuration for Security Hub. Can only be called from a Security Hub administrator account.

See \url{https://www.paws-r-sdk.com/docs/securityhub_describe_organization_configuration/} for full documentation.
}
\keyword{internal}
