% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_describe_organization_configuration}
\alias{securityhub_describe_organization_configuration}
\title{Returns information about the Organizations configuration for Security
Hub}
\usage{
securityhub_describe_organization_configuration()
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  AutoEnable = TRUE|FALSE,
  MemberAccountLimitReached = TRUE|FALSE
)
}\if{html}{\out{</div>}}
}
\description{
Returns information about the Organizations configuration for Security
Hub. Can only be called from a Security Hub administrator account.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_organization_configuration()
}\if{html}{\out{</div>}}
}

\keyword{internal}
