% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_describe_organization_configuration}
\alias{securityhub_describe_organization_configuration}
\title{Returns information about the way your organization is configured in
Security Hub}
\usage{
securityhub_describe_organization_configuration()
}
\description{
Returns information about the way your organization is configured in Security Hub. Only the Security Hub administrator account can invoke this operation.

See \url{https://www.paws-r-sdk.com/docs/securityhub_describe_organization_configuration/} for full documentation.
}
\keyword{internal}
