% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_describe_standards}
\alias{securityhub_describe_standards}
\title{Returns a list of the available standards in Security Hub}
\usage{
securityhub_describe_standards(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The token that is required for pagination. On your first call to the
\code{\link[=securityhub_describe_standards]{describe_standards}} operation, set
the value of this parameter to \code{NULL}.

For subsequent calls to the operation, to continue listing data, set the
value of this parameter to the value returned from the previous
response.}

\item{MaxResults}{The maximum number of standards to return.}
}
\description{
Returns a list of the available standards in Security Hub.

See \url{https://www.paws-r-sdk.com/docs/securityhub_describe_standards/} for full documentation.
}
\keyword{internal}
