% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disable_import_findings_for_product}
\alias{securityhub_disable_import_findings_for_product}
\title{Disables the integration of the specified product with Security Hub}
\usage{
securityhub_disable_import_findings_for_product(ProductSubscriptionArn)
}
\arguments{
\item{ProductSubscriptionArn}{[required] The ARN of the integrated product to disable the integration for.}
}
\description{
Disables the integration of the specified product with Security Hub. After the integration is disabled, findings from that product are no longer sent to Security Hub.

See \url{https://www.paws-r-sdk.com/docs/securityhub_disable_import_findings_for_product/} for full documentation.
}
\keyword{internal}
