% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disable_organization_admin_account}
\alias{securityhub_disable_organization_admin_account}
\title{Disables a Security Hub administrator account}
\usage{
securityhub_disable_organization_admin_account(AdminAccountId, Feature = NULL)
}
\arguments{
\item{AdminAccountId}{[required] The Amazon Web Services account identifier of the Security Hub
administrator account.}

\item{Feature}{The feature for which the delegated admin account is disabled. Defaults
to Security Hub if not specified.}
}
\description{
Disables a Security Hub administrator account. Can only be called by the organization management account.

See \url{https://www.paws-r-sdk.com/docs/securityhub_disable_organization_admin_account/} for full documentation.
}
\keyword{internal}
