% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disable_security_hub_v2}
\alias{securityhub_disable_security_hub_v2}
\title{Disable the service for the current Amazon Web Services Region or
specified Amazon Web Services Region}
\usage{
securityhub_disable_security_hub_v2()
}
\description{
Disable the service for the current Amazon Web Services Region or specified Amazon Web Services Region.

See \url{https://www.paws-r-sdk.com/docs/securityhub_disable_security_hub_v2/} for full documentation.
}
\keyword{internal}
