% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disassociate_from_administrator_account}
\alias{securityhub_disassociate_from_administrator_account}
\title{Disassociates the current Security Hub member account from the
associated administrator account}
\usage{
securityhub_disassociate_from_administrator_account()
}
\value{
An empty list.
}
\description{
Disassociates the current Security Hub member account from the
associated administrator account.

This operation is only used by accounts that are not part of an
organization. For organization accounts, only the administrator account
can disassociate a member account.
}
\section{Request syntax}{
\preformatted{svc$disassociate_from_administrator_account()
}
}

\keyword{internal}
