% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disassociate_from_master_account}
\alias{securityhub_disassociate_from_master_account}
\title{Disassociates the current Security Hub member account from the
associated master account}
\usage{
securityhub_disassociate_from_master_account()
}
\value{
An empty list.
}
\description{
Disassociates the current Security Hub member account from the
associated master account.

This operation is only used by accounts that are not part of an
organization. For organization accounts, only the master account (the
designated Security Hub administrator) can disassociate a member
account.
}
\section{Request syntax}{
\preformatted{svc$disassociate_from_master_account()
}
}

\keyword{internal}
