% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_disassociate_from_master_account}
\alias{securityhub_disassociate_from_master_account}
\title{This method is deprecated}
\usage{
securityhub_disassociate_from_master_account()
}
\description{
This method is deprecated. Instead, use \code{\link[=securityhub_disassociate_from_administrator_account]{disassociate_from_administrator_account}}.

See \url{https://www.paws-r-sdk.com/docs/securityhub_disassociate_from_master_account/} for full documentation.
}
\keyword{internal}
