% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_enable_import_findings_for_product}
\alias{securityhub_enable_import_findings_for_product}
\title{Sets up the subscription that enables a findings-generating solution
(product) to import its findings into Security Hub}
\usage{
securityhub_enable_import_findings_for_product(ProductArn)
}
\arguments{
\item{ProductArn}{[required] The ARN of the product that generates findings that you want to import
into Security Hub.}
}
\description{
Sets up the subscription that enables a findings-generating solution
(product) to import its findings into Security Hub.
}
\section{Request syntax}{
\preformatted{svc$enable_import_findings_for_product(
  ProductArn = "string"
)
}
}

\keyword{internal}
