% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_enable_import_findings_for_product}
\alias{securityhub_enable_import_findings_for_product}
\title{Enables the integration of a partner product with Security Hub}
\usage{
securityhub_enable_import_findings_for_product(ProductArn)
}
\arguments{
\item{ProductArn}{[required] The ARN of the product to enable the integration for.}
}
\description{
Enables the integration of a partner product with Security Hub. Integrated products send findings to Security Hub.

See \url{https://www.paws-r-sdk.com/docs/securityhub_enable_import_findings_for_product/} for full documentation.
}
\keyword{internal}
