% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_finding_statistics_v2}
\alias{securityhub_get_finding_statistics_v2}
\title{Returns aggregated statistical data about findings}
\usage{
securityhub_get_finding_statistics_v2(
  GroupByRules,
  SortOrder = NULL,
  MaxStatisticResults = NULL
)
}
\arguments{
\item{GroupByRules}{[required] Specifies how security findings should be aggregated and organized in
the statistical analysis. It can accept up to 5 \code{groupBy} fields in a
single call.}

\item{SortOrder}{Orders the aggregation count in descending or ascending order.
Descending order is the default.}

\item{MaxStatisticResults}{The maximum number of results to be returned.}
}
\description{
Returns aggregated statistical data about findings. \code{\link[=securityhub_get_finding_statistics_v2]{get_finding_statistics_v2}} use \code{securityhub:GetAdhocInsightResults} in the \code{Action} element of an IAM policy statement. You must have permission to perform the \code{s} action.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_finding_statistics_v2/} for full documentation.
}
\keyword{internal}
