% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_findings}
\alias{securityhub_get_findings}
\title{Returns a list of findings that match the specified criteria}
\usage{
securityhub_get_findings(
  Filters = NULL,
  SortCriteria = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filters}{The finding attributes used to define a condition to filter the returned
findings.

You can filter by up to 10 finding attributes. For each attribute, you
can provide up to 20 filter values.

Note that in the available filter fields, \code{WorkflowState} is deprecated.
To search for a finding based on its workflow status, use
\code{WorkflowStatus}.}

\item{SortCriteria}{The finding attributes used to sort the list of returned findings.}

\item{NextToken}{The token that is required for pagination. On your first call to the
\code{\link[=securityhub_get_findings]{get_findings}} operation, set the value of
this parameter to \code{NULL}.

For subsequent calls to the operation, to continue listing data, set the
value of this parameter to the value returned from the previous
response.}

\item{MaxResults}{The maximum number of findings to return.}
}
\description{
Returns a list of findings that match the specified criteria.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_findings/} for full documentation.
}
\keyword{internal}
