% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_findings_v2}
\alias{securityhub_get_findings_v2}
\title{Return a list of findings that match the specified criteria}
\usage{
securityhub_get_findings_v2(
  Filters = NULL,
  SortCriteria = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filters}{The finding attributes used to define a condition to filter the returned
OCSF findings. You can filter up to 10 composite filters. For each
filter type inside of a composite filter, you can provide up to 20
filters.}

\item{SortCriteria}{The finding attributes used to sort the list of returned findings.}

\item{NextToken}{The token required for pagination. On your first call, set the value of
this parameter to \code{NULL}. For subsequent calls, to continue listing
data, set the value of this parameter to the value returned in the
previous response.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Return a list of findings that match the specified criteria. \code{\link[=securityhub_get_findings]{get_findings}} and \code{\link[=securityhub_get_findings_v2]{get_findings_v2}} both use \code{securityhub:GetFindings} in the \code{Action} element of an IAM policy statement. You must have permission to perform the \code{securityhub:GetFindings} action.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_findings_v2/} for full documentation.
}
\keyword{internal}
