% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_insight_results}
\alias{securityhub_get_insight_results}
\title{Lists the results of the Security Hub insight specified by the insight
ARN}
\usage{
securityhub_get_insight_results(InsightArn)
}
\arguments{
\item{InsightArn}{[required] The ARN of the insight for which to return results.}
}
\description{
Lists the results of the Security Hub insight specified by the insight ARN.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_insight_results/} for full documentation.
}
\keyword{internal}
