% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_insights}
\alias{securityhub_get_insights}
\title{Lists and describes insights for the specified insight ARNs}
\usage{
securityhub_get_insights(
  InsightArns = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{InsightArns}{The ARNs of the insights to describe. If you do not provide any insight
ARNs, then \code{\link[=securityhub_get_insights]{get_insights}} returns all of
your custom insights. It does not return any managed insights.}

\item{NextToken}{The token that is required for pagination. On your first call to the
\code{\link[=securityhub_get_insights]{get_insights}} operation, set the value of
this parameter to \code{NULL}.

For subsequent calls to the operation, to continue listing data, set the
value of this parameter to the value returned from the previous
response.}

\item{MaxResults}{The maximum number of items to return in the response.}
}
\description{
Lists and describes insights for the specified insight ARNs.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_insights/} for full documentation.
}
\keyword{internal}
