% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_invitations_count}
\alias{securityhub_get_invitations_count}
\title{Returns the count of all Security Hub membership invitations that were
sent to the current member account, not including the currently accepted
invitation}
\usage{
securityhub_get_invitations_count()
}
\description{
Returns the count of all Security Hub membership invitations that were sent to the current member account, not including the currently accepted invitation.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_invitations_count/} for full documentation.
}
\keyword{internal}
