% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_master_account}
\alias{securityhub_get_master_account}
\title{Provides the details for the Security Hub master account for the current
member account}
\usage{
securityhub_get_master_account()
}
\description{
Provides the details for the Security Hub master account for the current
member account.
}
\section{Request syntax}{
\preformatted{svc$get_master_account()
}
}

\keyword{internal}
