% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_resources_statistics_v2}
\alias{securityhub_get_resources_statistics_v2}
\title{Retrieves statistical information about Amazon Web Services resources
and their associated security findings}
\usage{
securityhub_get_resources_statistics_v2(
  GroupByRules,
  SortOrder = NULL,
  MaxStatisticResults = NULL
)
}
\arguments{
\item{GroupByRules}{[required] How resource statistics should be aggregated and organized in the
response.}

\item{SortOrder}{Sorts aggregated statistics.}

\item{MaxStatisticResults}{The maximum number of results to be returned.}
}
\description{
Retrieves statistical information about Amazon Web Services resources and their associated security findings.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_resources_statistics_v2/} for full documentation.
}
\keyword{internal}
