% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_get_resources_v2}
\alias{securityhub_get_resources_v2}
\title{Returns a list of resources}
\usage{
securityhub_get_resources_v2(
  Filters = NULL,
  SortCriteria = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{Filters}{Filters resources based on a set of criteria.}

\item{SortCriteria}{The finding attributes used to sort the list of returned findings.}

\item{NextToken}{The token required for pagination. On your first call, set the value of
this parameter to \code{NULL}. For subsequent calls, to continue listing
data, set the value of this parameter to the value returned in the
previous response.}

\item{MaxResults}{The maximum number of results to return.}
}
\description{
Returns a list of resources.

See \url{https://www.paws-r-sdk.com/docs/securityhub_get_resources_v2/} for full documentation.
}
\keyword{internal}
