% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_invite_members}
\alias{securityhub_invite_members}
\title{Invites other AWS accounts to enable Security Hub and become Security
Hub member accounts}
\usage{
securityhub_invite_members(AccountIds)
}
\arguments{
\item{AccountIds}{A list of IDs of the AWS accounts that you want to invite to Security
Hub as members.}
}
\description{
Invites other AWS accounts to enable Security Hub and become Security
Hub member accounts. When an account accepts the invitation and becomes
a member account, the master account can view Security Hub findings of
the member account.
}
\section{Request syntax}{
\preformatted{svc$invite_members(
  AccountIds = list(
    "string"
  )
)
}
}

\keyword{internal}
