% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_invite_members}
\alias{securityhub_invite_members}
\title{Invites other Amazon Web Services accounts to become member accounts for
the Security Hub administrator account that the invitation is sent from}
\usage{
securityhub_invite_members(AccountIds)
}
\arguments{
\item{AccountIds}{[required] The list of account IDs of the Amazon Web Services accounts to invite to
Security Hub as members.}
}
\description{
Invites other Amazon Web Services accounts to become member accounts for the Security Hub administrator account that the invitation is sent from.

See \url{https://www.paws-r-sdk.com/docs/securityhub_invite_members/} for full documentation.
}
\keyword{internal}
