% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_invitations}
\alias{securityhub_list_invitations}
\title{Lists all Security Hub membership invitations that were sent to the
current AWS account}
\usage{
securityhub_list_invitations(MaxResults, NextToken)
}
\arguments{
\item{MaxResults}{The maximum number of items that you want in the response.}

\item{NextToken}{Paginates results. On your first call to the \code{ListInvitations}
operation, set the value of this parameter to \code{NULL}. For subsequent
calls to the operation, fill \code{nextToken} in the request with the value
of \code{NextToken} from the previous response to continue listing data.}
}
\description{
Lists all Security Hub membership invitations that were sent to the
current AWS account.
}
\section{Request syntax}{
\preformatted{svc$list_invitations(
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
