% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_invitations}
\alias{securityhub_list_invitations}
\title{Lists all Security Hub membership invitations that were sent to the
current Amazon Web Services account}
\usage{
securityhub_list_invitations(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of items to return in the response.}

\item{NextToken}{The token that is required for pagination. On your first call to the
\code{\link[=securityhub_list_invitations]{list_invitations}} operation, set the
value of this parameter to \code{NULL}.

For subsequent calls to the operation, to continue listing data, set the
value of this parameter to the value returned from the previous
response.}
}
\description{
Lists all Security Hub membership invitations that were sent to the current Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/securityhub_list_invitations/} for full documentation.
}
\keyword{internal}
