% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_list_tags_for_resource}
\alias{securityhub_list_tags_for_resource}
\title{Returns a list of tags associated with a resource}
\usage{
securityhub_list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource to retrieve tags for.}
}
\description{
Returns a list of tags associated with a resource.

See \url{https://www.paws-r-sdk.com/docs/securityhub_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
