% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_untag_resource}
\alias{securityhub_untag_resource}
\title{Removes one or more tags from a resource}
\usage{
securityhub_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The ARN of the resource to remove the tags from.}

\item{TagKeys}{[required] The tag keys associated with the tags to remove from the resource. You
can remove up to 50 tags at a time.}
}
\description{
Removes one or more tags from a resource.

See \url{https://www.paws-r-sdk.com/docs/securityhub_untag_resource/} for full documentation.
}
\keyword{internal}
