% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_action_target}
\alias{securityhub_update_action_target}
\title{Updates the name and description of a custom action target in Security
Hub}
\usage{
securityhub_update_action_target(
  ActionTargetArn,
  Name = NULL,
  Description = NULL
)
}
\arguments{
\item{ActionTargetArn}{[required] The ARN of the custom action target to update.}

\item{Name}{The updated name of the custom action target.}

\item{Description}{The updated description for the custom action target.}
}
\description{
Updates the name and description of a custom action target in Security Hub.

See \url{https://www.paws-r-sdk.com/docs/securityhub_update_action_target/} for full documentation.
}
\keyword{internal}
