% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_findings}
\alias{securityhub_update_findings}
\title{UpdateFindings is deprecated}
\usage{
securityhub_update_findings(Filters, Note = NULL, RecordState = NULL)
}
\arguments{
\item{Filters}{[required] A collection of attributes that specify which findings you want to
update.}

\item{Note}{The updated note for the finding.}

\item{RecordState}{The updated record state for the finding.}
}
\description{
\code{\link[=securityhub_update_findings]{update_findings}} is deprecated. Instead of \code{\link[=securityhub_update_findings]{update_findings}}, use \code{\link[=securityhub_batch_update_findings]{batch_update_findings}}.

See \url{https://www.paws-r-sdk.com/docs/securityhub_update_findings/} for full documentation.
}
\keyword{internal}
