% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_security_hub_configuration}
\alias{securityhub_update_security_hub_configuration}
\title{Updates configuration options for Security Hub}
\usage{
securityhub_update_security_hub_configuration(
  AutoEnableControls = NULL,
  ControlFindingGenerator = NULL
)
}
\arguments{
\item{AutoEnableControls}{Whether to automatically enable new controls when they are added to
standards that are enabled.

By default, this is set to \code{true}, and new controls are enabled
automatically. To not automatically enable new controls, set this to
\code{false}.}

\item{ControlFindingGenerator}{Updates whether the calling account has consolidated control findings
turned on. If the value for this field is set to \code{SECURITY_CONTROL},
Security Hub generates a single finding for a control check even when
the check applies to multiple enabled standards.

If the value for this field is set to \code{STANDARD_CONTROL}, Security Hub
generates separate findings for a control check when the check applies
to multiple enabled standards.

For accounts that are part of an organization, this value can only be
updated in the administrator account.}
}
\description{
Updates configuration options for Security Hub.

See \url{https://www.paws-r-sdk.com/docs/securityhub_update_security_hub_configuration/} for full documentation.
}
\keyword{internal}
