% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityhub_operations.R
\name{securityhub_update_standards_control}
\alias{securityhub_update_standards_control}
\title{Used to control whether an individual security standard control is
enabled or disabled}
\usage{
securityhub_update_standards_control(
  StandardsControlArn,
  ControlStatus = NULL,
  DisabledReason = NULL
)
}
\arguments{
\item{StandardsControlArn}{[required] The ARN of the security standard control to enable or disable.}

\item{ControlStatus}{The updated status of the security standard control.}

\item{DisabledReason}{A description of the reason why you are disabling a security standard
control. If you are disabling a control, then this is required.}
}
\description{
Used to control whether an individual security standard control is enabled or disabled.

See \url{https://www.paws-r-sdk.com/docs/securityhub_update_standards_control/} for full documentation.
}
\keyword{internal}
