% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_get_data_lake_exception_subscription}
\alias{securitylake_get_data_lake_exception_subscription}
\title{Retrieves the details of exception notifications for the account in
Amazon Security Lake}
\usage{
securitylake_get_data_lake_exception_subscription()
}
\description{
Retrieves the details of exception notifications for the account in Amazon Security Lake.

See \url{https://www.paws-r-sdk.com/docs/securitylake_get_data_lake_exception_subscription/} for full documentation.
}
\keyword{internal}
