% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securitylake_operations.R
\name{securitylake_update_data_lake}
\alias{securitylake_update_data_lake}
\title{Specifies where to store your security data and for how long}
\usage{
securitylake_update_data_lake(configurations, metaStoreManagerRoleArn = NULL)
}
\arguments{
\item{configurations}{[required] Specify the Region or Regions that will contribute data to the rollup
region.}

\item{metaStoreManagerRoleArn}{The Amazon Resource Name (ARN) used to create and update the Glue table.
This table contains partitions generated by the ingestion and
normalization of Amazon Web Services log sources and custom sources.}
}
\description{
Specifies where to store your security data and for how long. You can add a rollup Region to consolidate data from multiple Amazon Web Services Regions.

See \url{https://www.paws-r-sdk.com/docs/securitylake_update_data_lake/} for full documentation.
}
\keyword{internal}
