% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_associate_drt_log_bucket}
\alias{shield_associate_drt_log_bucket}
\title{Authorizes the DDoS Response Team (DRT) to access the specified Amazon
S3 bucket containing your AWS WAF logs}
\usage{
shield_associate_drt_log_bucket(LogBucket)
}
\arguments{
\item{LogBucket}{[required] The Amazon S3 bucket that contains your AWS WAF logs.}
}
\value{
An empty list.
}
\description{
Authorizes the DDoS Response Team (DRT) to access the specified Amazon
S3 bucket containing your AWS WAF logs. You can associate up to 10
Amazon S3 buckets with your subscription.

To use the services of the DRT and make an
\code{\link[=shield_associate_drt_log_bucket]{associate_drt_log_bucket}} request,
you must be subscribed to the \href{https://aws.amazon.com/premiumsupport/plans/business/}{Business Support plan} or the
\href{https://aws.amazon.com/premiumsupport/plans/enterprise/}{Enterprise Support plan}.
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$associate_drt_log_bucket(
  LogBucket = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
