% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_associate_drt_log_bucket}
\alias{shield_associate_drt_log_bucket}
\title{Authorizes the Shield Response Team (SRT) to access the specified Amazon
S3 bucket containing log data such as Application Load Balancer access
logs, CloudFront logs, or logs from third party sources}
\usage{
shield_associate_drt_log_bucket(LogBucket)
}
\arguments{
\item{LogBucket}{[required] The Amazon S3 bucket that contains the logs that you want to share.}
}
\description{
Authorizes the Shield Response Team (SRT) to access the specified Amazon S3 bucket containing log data such as Application Load Balancer access logs, CloudFront logs, or logs from third party sources. You can associate up to 10 Amazon S3 buckets with your subscription.

See \url{https://www.paws-r-sdk.com/docs/shield_associate_drt_log_bucket/} for full documentation.
}
\keyword{internal}
