% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_delete_protection}
\alias{shield_delete_protection}
\title{Deletes an Shield Advanced Protection}
\usage{
shield_delete_protection(ProtectionId)
}
\arguments{
\item{ProtectionId}{[required] The unique identifier (ID) for the Protection object to be deleted.}
}
\description{
Deletes an Shield Advanced Protection.

See \url{https://www.paws-r-sdk.com/docs/shield_delete_protection/} for full documentation.
}
\keyword{internal}
