% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_delete_protection}
\alias{shield_delete_protection}
\title{Deletes an AWS Shield Advanced Protection}
\usage{
shield_delete_protection(ProtectionId)
}
\arguments{
\item{ProtectionId}{[required] The unique identifier (ID) for the <span>Protection</span> object to be
deleted.}
}
\value{
An empty list.
}
\description{
Deletes an AWS Shield Advanced <span>Protection</span>.
}
\section{Request syntax}{
\preformatted{svc$delete_protection(
  ProtectionId = "string"
)
}
}

\keyword{internal}
