% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_attack}
\alias{shield_describe_attack}
\title{Describes the details of a DDoS attack}
\usage{
shield_describe_attack(AttackId)
}
\arguments{
\item{AttackId}{[required] The unique identifier (ID) for the attack.}
}
\description{
Describes the details of a DDoS attack.

See \url{https://www.paws-r-sdk.com/docs/shield_describe_attack/} for full documentation.
}
\keyword{internal}
