% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_drt_access}
\alias{shield_describe_drt_access}
\title{Returns the current role and list of Amazon S3 log buckets used by the
Shield Response Team (SRT) to access your Amazon Web Services account
while assisting with attack mitigation}
\usage{
shield_describe_drt_access()
}
\description{
Returns the current role and list of Amazon S3 log buckets used by the Shield Response Team (SRT) to access your Amazon Web Services account while assisting with attack mitigation.

See \url{https://www.paws-r-sdk.com/docs/shield_describe_drt_access/} for full documentation.
}
\keyword{internal}
