% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_protection}
\alias{shield_describe_protection}
\title{Lists the details of a Protection object}
\usage{
shield_describe_protection(ProtectionId = NULL, ResourceArn = NULL)
}
\arguments{
\item{ProtectionId}{The unique identifier (ID) for the Protection object to describe. You
must provide either the \code{ResourceArn} of the protected resource or the
\code{ProtectionID} of the protection, but not both.}

\item{ResourceArn}{The ARN (Amazon Resource Name) of the protected Amazon Web Services
resource. You must provide either the \code{ResourceArn} of the protected
resource or the \code{ProtectionID} of the protection, but not both.}
}
\description{
Lists the details of a Protection object.

See \url{https://www.paws-r-sdk.com/docs/shield_describe_protection/} for full documentation.
}
\keyword{internal}
