% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_protection_group}
\alias{shield_describe_protection_group}
\title{Returns the specification for the specified protection group}
\usage{
shield_describe_protection_group(ProtectionGroupId)
}
\arguments{
\item{ProtectionGroupId}{[required] The name of the protection group. You use this to identify the
protection group in lists and to manage the protection group, for
example to update, delete, or describe it.}
}
\description{
Returns the specification for the specified protection group.

See \url{https://www.paws-r-sdk.com/docs/shield_describe_protection_group/} for full documentation.
}
\keyword{internal}
