% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_subscription}
\alias{shield_describe_subscription}
\title{Provides details about the Shield Advanced subscription for an account}
\usage{
shield_describe_subscription()
}
\description{
Provides details about the Shield Advanced subscription for an account.

See \url{https://www.paws-r-sdk.com/docs/shield_describe_subscription/} for full documentation.
}
\keyword{internal}
