% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_disable_proactive_engagement}
\alias{shield_disable_proactive_engagement}
\title{Removes authorization from the Shield Response Team (SRT) to notify
contacts about escalations to the SRT and to initiate proactive customer
support}
\usage{
shield_disable_proactive_engagement()
}
\description{
Removes authorization from the Shield Response Team (SRT) to notify contacts about escalations to the SRT and to initiate proactive customer support.

See \url{https://www.paws-r-sdk.com/docs/shield_disable_proactive_engagement/} for full documentation.
}
\keyword{internal}
