% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_disassociate_drt_role}
\alias{shield_disassociate_drt_role}
\title{Removes the Shield Response Team's (SRT) access to your Amazon Web
Services account}
\usage{
shield_disassociate_drt_role()
}
\description{
Removes the Shield Response Team's (SRT) access to your Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/shield_disassociate_drt_role/} for full documentation.
}
\keyword{internal}
