% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_disassociate_health_check}
\alias{shield_disassociate_health_check}
\title{Removes health-based detection from the Shield Advanced protection for a
resource}
\usage{
shield_disassociate_health_check(ProtectionId, HealthCheckArn)
}
\arguments{
\item{ProtectionId}{[required] The unique identifier (ID) for the Protection object to remove the
health check association from.}

\item{HealthCheckArn}{[required] The Amazon Resource Name (ARN) of the health check that is associated
with the protection.}
}
\description{
Removes health-based detection from the Shield Advanced protection for a resource. Shield Advanced health-based detection uses the health of your Amazon Web Services resource to improve responsiveness and accuracy in attack detection and response.

See \url{https://www.paws-r-sdk.com/docs/shield_disassociate_health_check/} for full documentation.
}
\keyword{internal}
