% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_get_subscription_state}
\alias{shield_get_subscription_state}
\title{Returns the SubscriptionState, either Active or Inactive}
\usage{
shield_get_subscription_state()
}
\description{
Returns the \code{SubscriptionState}, either \code{Active} or \code{Inactive}.

See \url{https://www.paws-r-sdk.com/docs/shield_get_subscription_state/} for full documentation.
}
\keyword{internal}
