% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_list_tags_for_resource}
\alias{shield_list_tags_for_resource}
\title{Gets information about AWS tags for a specified Amazon Resource Name
(ARN) in AWS Shield}
\usage{
shield_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource to get tags for.}
}
\value{
A list with the following syntax:\preformatted{list(
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}
\description{
Gets information about AWS tags for a specified Amazon Resource Name
(ARN) in AWS Shield.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceARN = "string"
)
}
}

\keyword{internal}
