% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_list_tags_for_resource}
\alias{shield_list_tags_for_resource}
\title{Gets information about Amazon Web Services tags for a specified Amazon
Resource Name (ARN) in Shield}
\usage{
shield_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource to get tags for.}
}
\description{
Gets information about Amazon Web Services tags for a specified Amazon Resource Name (ARN) in Shield.

See \url{https://www.paws-r-sdk.com/docs/shield_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
