% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_untag_resource}
\alias{shield_untag_resource}
\title{Removes tags from a resource in Shield}
\usage{
shield_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name (ARN) of the resource that you want to remove
tags from.}

\item{TagKeys}{[required] The tag key for each tag that you want to remove from the resource.}
}
\description{
Removes tags from a resource in Shield.

See \url{https://www.paws-r-sdk.com/docs/shield_untag_resource/} for full documentation.
}
\keyword{internal}
