% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_update_emergency_contact_settings}
\alias{shield_update_emergency_contact_settings}
\title{Updates the details of the list of email addresses and phone numbers
that the Shield Response Team (SRT) can use to contact you if you have
proactive engagement enabled, for escalations to the SRT and to initiate
proactive customer support}
\usage{
shield_update_emergency_contact_settings(EmergencyContactList = NULL)
}
\arguments{
\item{EmergencyContactList}{A list of email addresses and phone numbers that the Shield Response
Team (SRT) can use to contact you if you have proactive engagement
enabled, for escalations to the SRT and to initiate proactive customer
support.

If you have proactive engagement enabled, the contact list must include
at least one phone number.}
}
\description{
Updates the details of the list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

See \url{https://www.paws-r-sdk.com/docs/shield_update_emergency_contact_settings/} for full documentation.
}
\keyword{internal}
