% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_describe_account_assignment_creation_status}
\alias{ssoadmin_describe_account_assignment_creation_status}
\title{Describes the status of the assignment creation request}
\usage{
ssoadmin_describe_account_assignment_creation_status(
  InstanceArn,
  AccountAssignmentCreationRequestId
)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and AWS Service Namespaces in the \emph{AWS General Reference}.}

\item{AccountAssignmentCreationRequestId}{[required] The identifier that is used to track the request operation progress.}
}
\description{
Describes the status of the assignment creation request.

See \url{https://paws-r.github.io/docs/ssoadmin/describe_account_assignment_creation_status.html} for full documentation.
}
\keyword{internal}
