% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_describe_account_assignment_creation_status}
\alias{ssoadmin_describe_account_assignment_creation_status}
\title{Describes the status of the assignment creation request}
\usage{
ssoadmin_describe_account_assignment_creation_status(
  AccountAssignmentCreationRequestId,
  InstanceArn
)
}
\arguments{
\item{AccountAssignmentCreationRequestId}{[required] The identifier that is used to track the request operation progress.}

\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and Amazon Web Services Service Namespaces in the \emph{Amazon
Web Services General Reference}.}
}
\description{
Describes the status of the assignment creation request.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_describe_account_assignment_creation_status/} for full documentation.
}
\keyword{internal}
