% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_application_assignments_for_principal}
\alias{ssoadmin_list_application_assignments_for_principal}
\title{Lists the applications to which a specified principal is assigned}
\usage{
ssoadmin_list_application_assignments_for_principal(
  Filter = NULL,
  InstanceArn,
  MaxResults = NULL,
  NextToken = NULL,
  PrincipalId,
  PrincipalType
)
}
\arguments{
\item{Filter}{Filters the output to include only assignments associated with the
application that has the specified ARN.}

\item{InstanceArn}{[required] Specifies the instance of IAM Identity Center that contains principal
and applications.}

\item{MaxResults}{Specifies the total number of results that you want included in each
response. If additional items exist beyond the number you specify, the
\code{NextToken} response element is returned with a value (not null).
Include the specified value as the \code{NextToken} request parameter in the
next call to the operation to get the next set of results. Note that the
service might return fewer results than the maximum even when there are
more results available. You should check \code{NextToken} after every
operation to ensure that you receive all of the results.}

\item{NextToken}{Specifies that you want to receive the next page of results. Valid only
if you received a \code{NextToken} response in the previous request. If you
did, it indicates that more output is available. Set this parameter to
the value provided by the previous call's \code{NextToken} response to
request the next page of results.}

\item{PrincipalId}{[required] Specifies the unique identifier of the principal for which you want to
retrieve its assignments.}

\item{PrincipalType}{[required] Specifies the type of the principal for which you want to retrieve its
assignments.}
}
\description{
Lists the applications to which a specified principal is assigned.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_list_application_assignments_for_principal/} for full documentation.
}
\keyword{internal}
