% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_managed_policies_in_permission_set}
\alias{ssoadmin_list_managed_policies_in_permission_set}
\title{Lists the Amazon Web Services managed policy that is attached to a
specified permission set}
\usage{
ssoadmin_list_managed_policies_in_permission_set(
  InstanceArn,
  MaxResults = NULL,
  NextToken = NULL,
  PermissionSetArn
)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and Amazon Web Services Service Namespaces in the \emph{Amazon
Web Services General Reference}.}

\item{MaxResults}{The maximum number of results to display for the PermissionSet.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}

\item{PermissionSetArn}{[required] The ARN of the PermissionSet whose managed policies will be listed.}
}
\description{
Lists the Amazon Web Services managed policy that is attached to a specified permission set.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_list_managed_policies_in_permission_set/} for full documentation.
}
\keyword{internal}
