% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_list_permission_sets_provisioned_to_account}
\alias{ssoadmin_list_permission_sets_provisioned_to_account}
\title{Lists all the permission sets that are provisioned to a specified Amazon
Web Services account}
\usage{
ssoadmin_list_permission_sets_provisioned_to_account(
  AccountId,
  InstanceArn,
  MaxResults = NULL,
  NextToken = NULL,
  ProvisioningStatus = NULL
)
}
\arguments{
\item{AccountId}{[required] The identifier of the Amazon Web Services account from which to list the
assignments.}

\item{InstanceArn}{[required] The ARN of the IAM Identity Center instance under which the operation
will be executed. For more information about ARNs, see Amazon Resource
Names (ARNs) and Amazon Web Services Service Namespaces in the \emph{Amazon
Web Services General Reference}.}

\item{MaxResults}{The maximum number of results to display for the assignment.}

\item{NextToken}{The pagination token for the list API. Initially the value is null. Use
the output of previous API calls to make subsequent calls.}

\item{ProvisioningStatus}{The status object for the permission set provisioning operation.}
}
\description{
Lists all the permission sets that are provisioned to a specified Amazon Web Services account.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_list_permission_sets_provisioned_to_account/} for full documentation.
}
\keyword{internal}
