% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_put_application_grant}
\alias{ssoadmin_put_application_grant}
\title{Adds a grant to an application}
\usage{
ssoadmin_put_application_grant(ApplicationArn, Grant, GrantType)
}
\arguments{
\item{ApplicationArn}{[required] Specifies the ARN of the application to update.}

\item{Grant}{[required] Specifies a structure that describes the grant to update.}

\item{GrantType}{[required] Specifies the type of grant to update.}
}
\description{
Adds a grant to an application.

See \url{https://www.paws-r-sdk.com/docs/ssoadmin_put_application_grant/} for full documentation.
}
\keyword{internal}
