% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts_operations.R
\name{sts_decode_authorization_message}
\alias{sts_decode_authorization_message}
\title{Decodes additional information about the authorization status of a
request from an encoded message returned in response to an Amazon Web
Services request}
\usage{
sts_decode_authorization_message(EncodedMessage)
}
\arguments{
\item{EncodedMessage}{[required] The encoded message that was returned with the response.}
}
\description{
Decodes additional information about the authorization status of a request from an encoded message returned in response to an Amazon Web Services request.

See \url{https://www.paws-r-sdk.com/docs/sts_decode_authorization_message/} for full documentation.
}
\keyword{internal}
