% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sts_operations.R
\name{sts_get_access_key_info}
\alias{sts_get_access_key_info}
\title{Returns the account identifier for the specified access key ID}
\usage{
sts_get_access_key_info(AccessKeyId)
}
\arguments{
\item{AccessKeyId}{[required] The identifier of an access key.

This parameter allows (through its regex pattern) a string of characters
that can consist of any upper- or lowercase letter or digit.}
}
\description{
Returns the account identifier for the specified access key ID.

See \url{https://www.paws-r-sdk.com/docs/sts_get_access_key_info/} for full documentation.
}
\keyword{internal}
