% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_delete_identity_source}
\alias{verifiedpermissions_delete_identity_source}
\title{Deletes an identity source that references an identity provider (IdP)
such as Amazon Cognito}
\usage{
verifiedpermissions_delete_identity_source(policyStoreId, identitySourceId)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that contains the identity source
that you want to delete.}

\item{identitySourceId}{[required] Specifies the ID of the identity source that you want to delete.}
}
\description{
Deletes an identity source that references an identity provider (IdP) such as Amazon Cognito. After you delete the identity source, you can no longer use tokens for identities from that identity source to represent principals in authorization queries made using \code{\link[=verifiedpermissions_is_authorized_with_token]{is_authorized_with_token}}. operations.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_delete_identity_source/} for full documentation.
}
\keyword{internal}
