% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_delete_policy}
\alias{verifiedpermissions_delete_policy}
\title{Deletes the specified policy from the policy store}
\usage{
verifiedpermissions_delete_policy(policyStoreId, policyId)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that contains the policy that you
want to delete.}

\item{policyId}{[required] Specifies the ID of the policy that you want to delete.}
}
\description{
Deletes the specified policy from the policy store.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_delete_policy/} for full documentation.
}
\keyword{internal}
