% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_delete_policy_template}
\alias{verifiedpermissions_delete_policy_template}
\title{Deletes the specified policy template from the policy store}
\usage{
verifiedpermissions_delete_policy_template(policyStoreId, policyTemplateId)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that contains the policy template
that you want to delete.}

\item{policyTemplateId}{[required] Specifies the ID of the policy template that you want to delete.}
}
\description{
Deletes the specified policy template from the policy store.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_delete_policy_template/} for full documentation.
}
\keyword{internal}
