% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_get_identity_source}
\alias{verifiedpermissions_get_identity_source}
\title{Retrieves the details about the specified identity source}
\usage{
verifiedpermissions_get_identity_source(policyStoreId, identitySourceId)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that contains the identity source
you want information about.}

\item{identitySourceId}{[required] Specifies the ID of the identity source you want information about.}
}
\description{
Retrieves the details about the specified identity source.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_get_identity_source/} for full documentation.
}
\keyword{internal}
