% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_get_policy}
\alias{verifiedpermissions_get_policy}
\title{Retrieves information about the specified policy}
\usage{
verifiedpermissions_get_policy(policyStoreId, policyId)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that contains the policy that you
want information about.}

\item{policyId}{[required] Specifies the ID of the policy you want information about.}
}
\description{
Retrieves information about the specified policy.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_get_policy/} for full documentation.
}
\keyword{internal}
