% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_get_policy_store}
\alias{verifiedpermissions_get_policy_store}
\title{Retrieves details about a policy store}
\usage{
verifiedpermissions_get_policy_store(policyStoreId)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that you want information about.}
}
\description{
Retrieves details about a policy store.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_get_policy_store/} for full documentation.
}
\keyword{internal}
