% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_get_policy_template}
\alias{verifiedpermissions_get_policy_template}
\title{Retrieve the details for the specified policy template in the specified
policy store}
\usage{
verifiedpermissions_get_policy_template(policyStoreId, policyTemplateId)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that contains the policy template
that you want information about.}

\item{policyTemplateId}{[required] Specifies the ID of the policy template that you want information about.}
}
\description{
Retrieve the details for the specified policy template in the specified policy store.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_get_policy_template/} for full documentation.
}
\keyword{internal}
