% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verifiedpermissions_operations.R
\name{verifiedpermissions_update_policy_store}
\alias{verifiedpermissions_update_policy_store}
\title{Modifies the validation setting for a policy store}
\usage{
verifiedpermissions_update_policy_store(
  policyStoreId,
  validationSettings,
  description = NULL
)
}
\arguments{
\item{policyStoreId}{[required] Specifies the ID of the policy store that you want to update}

\item{validationSettings}{[required] A structure that defines the validation settings that want to enable for
the policy store.}

\item{description}{Descriptive text that you can provide to help with identification of the
current policy store.}
}
\description{
Modifies the validation setting for a policy store.

See \url{https://www.paws-r-sdk.com/docs/verifiedpermissions_update_policy_store/} for full documentation.
}
\keyword{internal}
