% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_create_size_constraint_set}
\alias{waf_create_size_constraint_set}
\title{This is AWS WAF Classic documentation}
\usage{
waf_create_size_constraint_set(Name, ChangeToken)
}
\arguments{
\item{Name}{[required] A friendly name or description of the SizeConstraintSet. You can't
change \code{Name} after you create a \code{SizeConstraintSet}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=waf_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_create_size_constraint_set/} for full documentation.
}
\keyword{internal}
