% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_delete_logging_configuration}
\alias{waf_delete_logging_configuration}
\title{This is AWS WAF Classic documentation}
\usage{
waf_delete_logging_configuration(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the web ACL from which you want to
delete the LoggingConfiguration.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_delete_logging_configuration/} for full documentation.
}
\keyword{internal}
