% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_get_regex_match_set}
\alias{waf_get_regex_match_set}
\title{This is AWS WAF Classic documentation}
\usage{
waf_get_regex_match_set(RegexMatchSetId)
}
\arguments{
\item{RegexMatchSetId}{[required] The \code{RegexMatchSetId} of the RegexMatchSet that you want to get.
\code{RegexMatchSetId} is returned by
\code{\link[=waf_create_regex_match_set]{create_regex_match_set}} and by
\code{\link[=waf_list_regex_match_sets]{list_regex_match_sets}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_get_regex_match_set/} for full documentation.
}
\keyword{internal}
