% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_get_regex_pattern_set}
\alias{waf_get_regex_pattern_set}
\title{This is AWS WAF Classic documentation}
\usage{
waf_get_regex_pattern_set(RegexPatternSetId)
}
\arguments{
\item{RegexPatternSetId}{[required] The \code{RegexPatternSetId} of the RegexPatternSet that you want to get.
\code{RegexPatternSetId} is returned by
\code{\link[=waf_create_regex_pattern_set]{create_regex_pattern_set}} and by
\code{\link[=waf_list_regex_pattern_sets]{list_regex_pattern_sets}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_get_regex_pattern_set/} for full documentation.
}
\keyword{internal}
