% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_get_web_acl}
\alias{waf_get_web_acl}
\title{Returns the WebACL that is specified by WebACLId}
\usage{
waf_get_web_acl(WebACLId)
}
\arguments{
\item{WebACLId}{[required] The \code{WebACLId} of the WebACL that you want to get. \code{WebACLId} is
returned by CreateWebACL and by ListWebACLs.}
}
\description{
Returns the WebACL that is specified by \code{WebACLId}.
}
\section{Request syntax}{
\preformatted{svc$get_web_acl(
  WebACLId = "string"
)
}
}

\examples{
# The following example returns the details of a web ACL with the ID
# createwebacl-1472061481310.
\dontrun{svc$get_web_acl(
  WebACLId = "createwebacl-1472061481310"
)}

}
\keyword{internal}
