% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waf_operations.R
\name{waf_list_tags_for_resource}
\alias{waf_list_tags_for_resource}
\title{This is AWS WAF Classic documentation}
\usage{
waf_list_tags_for_resource(NextMarker = NULL, Limit = NULL, ResourceARN)
}
\arguments{
\item{NextMarker}{}

\item{Limit}{}

\item{ResourceARN}{[required]}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/waf_list_tags_for_resource/} for full documentation.
}
\keyword{internal}
