% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_delete_ip_set}
\alias{wafregional_delete_ip_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_delete_ip_set(IPSetId, ChangeToken)
}
\arguments{
\item{IPSetId}{[required] The \code{IPSetId} of the IPSet that you want to delete. \code{IPSetId} is
returned by \code{\link[=wafregional_create_ip_set]{create_ip_set}} and by
\code{\link[=wafregional_list_ip_sets]{list_ip_sets}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=wafregional_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_delete_ip_set/} for full documentation.
}
\keyword{internal}
