% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_delete_rate_based_rule}
\alias{wafregional_delete_rate_based_rule}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_delete_rate_based_rule(RuleId, ChangeToken)
}
\arguments{
\item{RuleId}{[required] The \code{RuleId} of the RateBasedRule that you want to delete. \code{RuleId} is
returned by
\code{\link[=wafregional_create_rate_based_rule]{create_rate_based_rule}} and by
\code{\link[=wafregional_list_rate_based_rules]{list_rate_based_rules}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=wafregional_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_delete_rate_based_rule/} for full documentation.
}
\keyword{internal}
