% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_delete_regex_pattern_set}
\alias{wafregional_delete_regex_pattern_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_delete_regex_pattern_set(RegexPatternSetId, ChangeToken)
}
\arguments{
\item{RegexPatternSetId}{[required] The \code{RegexPatternSetId} of the RegexPatternSet that you want to delete.
\code{RegexPatternSetId} is returned by
\code{\link[=wafregional_create_regex_pattern_set]{create_regex_pattern_set}} and
by \code{\link[=wafregional_list_regex_pattern_sets]{list_regex_pattern_sets}}.}

\item{ChangeToken}{[required] The value returned by the most recent call to
\code{\link[=wafregional_get_change_token]{get_change_token}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_delete_regex_pattern_set/} for full documentation.
}
\keyword{internal}
