% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_ip_set}
\alias{wafregional_get_ip_set}
\title{Returns the IPSet that is specified by IPSetId}
\usage{
wafregional_get_ip_set(IPSetId)
}
\arguments{
\item{IPSetId}{[required] The \code{IPSetId} of the IPSet that you want to get. \code{IPSetId} is returned
by CreateIPSet and by ListIPSets.}
}
\description{
Returns the IPSet that is specified by \code{IPSetId}.
}
\section{Request syntax}{
\preformatted{svc$get_ip_set(
  IPSetId = "string"
)
}
}

\section{Examples}{
\preformatted{# The following example returns the details of an IP match set with the ID
# example1ds3t-46da-4fdb-b8d5-abc321j569j5.
svc$get_ip_set(
  IPSetId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)
}
}

\keyword{internal}
