% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_rule}
\alias{wafregional_get_rule}
\title{Returns the Rule that is specified by the RuleId that you included in
the GetRule request}
\usage{
wafregional_get_rule(RuleId)
}
\arguments{
\item{RuleId}{[required] The \code{RuleId} of the Rule that you want to get. \code{RuleId} is returned by
CreateRule and by ListRules.}
}
\description{
Returns the Rule that is specified by the \code{RuleId} that you included in
the \code{GetRule} request.
}
\section{Request syntax}{
\preformatted{svc$get_rule(
  RuleId = "string"
)
}
}

\section{Examples}{
\preformatted{# The following example returns the details of a rule with the ID
# example1ds3t-46da-4fdb-b8d5-abc321j569j5.
svc$get_rule(
  RuleId = "example1ds3t-46da-4fdb-b8d5-abc321j569j5"
)
}
}

\keyword{internal}
