% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_rule}
\alias{wafregional_get_rule}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_get_rule(RuleId)
}
\arguments{
\item{RuleId}{[required] The \code{RuleId} of the Rule that you want to get. \code{RuleId} is returned by
\code{\link[=wafregional_create_rule]{create_rule}} and by
\code{\link[=wafregional_list_rules]{list_rules}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_get_rule/} for full documentation.
}
\keyword{internal}
