% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_get_size_constraint_set}
\alias{wafregional_get_size_constraint_set}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_get_size_constraint_set(SizeConstraintSetId)
}
\arguments{
\item{SizeConstraintSetId}{[required] The \code{SizeConstraintSetId} of the SizeConstraintSet that you want to get.
\code{SizeConstraintSetId} is returned by
\code{\link[=wafregional_create_size_constraint_set]{create_size_constraint_set}}
and by
\code{\link[=wafregional_list_size_constraint_sets]{list_size_constraint_sets}}.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_get_size_constraint_set/} for full documentation.
}
\keyword{internal}
