% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_list_tags_for_resource}
\alias{wafregional_list_tags_for_resource}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_list_tags_for_resource(NextMarker, Limit, ResourceARN)
}
\arguments{
\item{NextMarker}{}

\item{Limit}{}

\item{ResourceARN}{[required]}
}
\value{
A list with the following syntax:\preformatted{list(
  NextMarker = "string",
  TagInfoForResource = list(
    ResourceARN = "string",
    TagList = list(
      list(
        Key = "string",
        Value = "string"
      )
    )
  )
)
}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see
\href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic}
in the developer guide.

\strong{For the latest version of AWS WAF}, use the AWS WAFV2 API and see the
\href{https://docs.aws.amazon.com/waf/latest/developerguide/waf-chapter.html}{AWS WAF Developer Guide}.
With the latest version, AWS WAF has a single set of endpoints for
regional and global use.

Retrieves the tags associated with the specified AWS resource. Tags are
key:value pairs that you can use to categorize and manage your
resources, for purposes like billing. For example, you might set the tag
key to "customer" and the value to the customer name or ID. You can
specify one or more tags to add to each AWS resource, up to 50 tags for
a resource.

Tagging is only available through the API, SDKs, and CLI. You can't
manage or view tags through the AWS WAF Classic console. You can tag the
AWS resources that you manage through AWS WAF Classic: web ACLs, rule
groups, and rules.
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  NextMarker = "string",
  Limit = 123,
  ResourceARN = "string"
)
}
}

\keyword{internal}
