% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafregional_operations.R
\name{wafregional_list_web_ac_ls}
\alias{wafregional_list_web_ac_ls}
\title{This is AWS WAF Classic documentation}
\usage{
wafregional_list_web_ac_ls(NextMarker = NULL, Limit = NULL)
}
\arguments{
\item{NextMarker}{If you specify a value for \code{Limit} and you have more \code{WebACL} objects
than the number that you specify for \code{Limit}, AWS WAF returns a
\code{NextMarker} value in the response that allows you to list another group
of \code{WebACL} objects. For the second and subsequent
\code{\link[=wafregional_list_web_ac_ls]{list_web_ac_ls}} requests, specify the
value of \code{NextMarker} from the previous response to get information
about another batch of \code{WebACL} objects.}

\item{Limit}{Specifies the number of \code{WebACL} objects that you want AWS WAF to return
for this request. If you have more \code{WebACL} objects than the number that
you specify for \code{Limit}, the response includes a \code{NextMarker} value that
you can use to get another batch of \code{WebACL} objects.}
}
\description{
This is \strong{AWS WAF Classic} documentation. For more information, see \href{https://docs.aws.amazon.com/waf/latest/developerguide/classic-waf-chapter.html}{AWS WAF Classic} in the developer guide.

See \url{https://www.paws-r-sdk.com/docs/wafregional_list_web_ac_ls/} for full documentation.
}
\keyword{internal}
