% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_delete_logging_configuration}
\alias{wafv2_delete_logging_configuration}
\title{Deletes the LoggingConfiguration from the specified web ACL}
\usage{
wafv2_delete_logging_configuration(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the web ACL from which you want to
delete the LoggingConfiguration.}
}
\description{
Deletes the LoggingConfiguration from the specified web ACL.

See \url{https://www.paws-r-sdk.com/docs/wafv2_delete_logging_configuration/} for full documentation.
}
\keyword{internal}
