% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafv2_operations.R
\name{wafv2_describe_all_managed_products}
\alias{wafv2_describe_all_managed_products}
\title{Provides high-level information for the Amazon Web Services Managed
Rules rule groups and Amazon Web Services Marketplace managed rule
groups}
\usage{
wafv2_describe_all_managed_products(Scope)
}
\arguments{
\item{Scope}{[required] Specifies whether this is for an Amazon CloudFront distribution or for a
regional application. A regional application can be an Application Load
Balancer (ALB), an Amazon API Gateway REST API, an AppSync GraphQL API,
an Amazon Cognito user pool, an App Runner service, or an Amazon Web
Services Verified Access instance.

To work with CloudFront, you must also specify the Region US East (N.
Virginia) as follows:
\itemize{
\item CLI - Specify the Region when you use the CloudFront scope:
\code{--scope=CLOUDFRONT --region=us-east-1}.
\item API and SDKs - For all calls, use the Region endpoint us-east-1.
}}
}
\description{
Provides high-level information for the Amazon Web Services Managed Rules rule groups and Amazon Web Services Marketplace managed rule groups.

See \url{https://www.paws-r-sdk.com/docs/wafv2_describe_all_managed_products/} for full documentation.
}
\keyword{internal}
